% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_create_api_key}
\alias{apigateway_create_api_key}
\title{Create an ApiKey resource}
\usage{
apigateway_create_api_key(
  name = NULL,
  description = NULL,
  enabled = NULL,
  generateDistinctId = NULL,
  value = NULL,
  stageKeys = NULL,
  customerId = NULL,
  tags = NULL
)
}
\arguments{
\item{name}{The name of the ApiKey.}

\item{description}{The description of the ApiKey.}

\item{enabled}{Specifies whether the ApiKey can be used by callers.}

\item{generateDistinctId}{Specifies whether (\code{true}) or not (\code{false}) the key identifier is
distinct from the created API key value. This parameter is deprecated
and should not be used.}

\item{value}{Specifies a value of the API key.}

\item{stageKeys}{DEPRECATED FOR USAGE PLANS - Specifies stages associated with the API
key.}

\item{customerId}{An Amazon Web Services Marketplace customer identifier, when integrating
with the Amazon Web Services SaaS Marketplace.}

\item{tags}{The key-value map of strings. The valid character set is
[a-zA-Z+-=._:/]. The tag key can be up to 128 characters and must not
start with \verb{aws:}. The tag value can be up to 256 characters.}
}
\description{
Create an ApiKey resource.

See \url{https://www.paws-r-sdk.com/docs/apigateway_create_api_key/} for full documentation.
}
\keyword{internal}
