% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_create_documentation_version}
\alias{apigateway_create_documentation_version}
\title{Create documentation version}
\usage{
apigateway_create_documentation_version(restApiId, documentationVersion,
  stageName, description)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated RestApi.}

\item{documentationVersion}{[required] [Required] The version identifier of the new snapshot.}

\item{stageName}{The stage name to be associated with the new documentation snapshot.}

\item{description}{A description about the new documentation snapshot.}
}
\value{
A list with the following syntax:\preformatted{list(
  version = "string",
  createdDate = as.POSIXct(
    "2015-01-01"
  ),
  description = "string"
)
}
}
\description{
Create documentation version
}
\section{Request syntax}{
\preformatted{svc$create_documentation_version(
  restApiId = "string",
  documentationVersion = "string",
  stageName = "string",
  description = "string"
)
}
}

\keyword{internal}
