% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_delete_base_path_mapping}
\alias{apigateway_delete_base_path_mapping}
\title{Deletes the BasePathMapping resource}
\usage{
apigateway_delete_base_path_mapping(domainName, basePath)
}
\arguments{
\item{domainName}{[required] The domain name of the BasePathMapping resource to delete.}

\item{basePath}{[required] The base path name of the BasePathMapping resource to delete.

To specify an empty base path, set this parameter to \code{'(none)'}.}
}
\description{
Deletes the BasePathMapping resource.

See \url{https://paws-r.github.io/docs/apigateway/delete_base_path_mapping.html} for full documentation.
}
\keyword{internal}
