% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_delete_client_certificate}
\alias{apigateway_delete_client_certificate}
\title{Deletes the ClientCertificate resource}
\usage{
apigateway_delete_client_certificate(clientCertificateId)
}
\arguments{
\item{clientCertificateId}{[required] [Required] The identifier of the ClientCertificate resource to be
deleted.}
}
\value{
An empty list.
}
\description{
Deletes the ClientCertificate resource.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_client_certificate(
  clientCertificateId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
