% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_delete_deployment}
\alias{apigateway_delete_deployment}
\title{Deletes a Deployment resource}
\usage{
apigateway_delete_deployment(restApiId, deploymentId)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated
<span>RestApi</span>.}

\item{deploymentId}{[required] [Required] The identifier of the <span>Deployment</span> resource to
delete.}
}
\value{
An empty list.
}
\description{
Deletes a <span>Deployment</span> resource. Deleting a deployment will
only succeed if there are no <span>Stage</span> resources associated
with it.
}
\section{Request syntax}{
\preformatted{svc$delete_deployment(
  restApiId = "string",
  deploymentId = "string"
)
}
}

\keyword{internal}
