% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_delete_documentation_part}
\alias{apigateway_delete_documentation_part}
\title{Deletes a documentation part}
\usage{
apigateway_delete_documentation_part(restApiId, documentationPartId)
}
\arguments{
\item{restApiId}{[required] The string identifier of the associated RestApi.}

\item{documentationPartId}{[required] The identifier of the to-be-deleted documentation part.}
}
\description{
Deletes a documentation part

See \url{https://www.paws-r-sdk.com/docs/apigateway_delete_documentation_part/} for full documentation.
}
\keyword{internal}
