% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_delete_documentation_version}
\alias{apigateway_delete_documentation_version}
\title{Deletes a documentation version}
\usage{
apigateway_delete_documentation_version(restApiId, documentationVersion)
}
\arguments{
\item{restApiId}{[required] The string identifier of the associated RestApi.}

\item{documentationVersion}{[required] The version identifier of a to-be-deleted documentation snapshot.}
}
\description{
Deletes a documentation version.

See \url{https://www.paws-r-sdk.com/docs/apigateway_delete_documentation_version/} for full documentation.
}
\keyword{internal}
