% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_delete_domain_name}
\alias{apigateway_delete_domain_name}
\title{Deletes the DomainName resource}
\usage{
apigateway_delete_domain_name(domainName, domainNameId = NULL)
}
\arguments{
\item{domainName}{[required] The name of the DomainName resource to be deleted.}

\item{domainNameId}{The identifier for the domain name resource. Supported only for private
custom domain names.}
}
\description{
Deletes the DomainName resource.

See \url{https://www.paws-r-sdk.com/docs/apigateway_delete_domain_name/} for full documentation.
}
\keyword{internal}
