% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_delete_integration}
\alias{apigateway_delete_integration}
\title{Represents a delete integration}
\usage{
apigateway_delete_integration(restApiId, resourceId, httpMethod)
}
\arguments{
\item{restApiId}{[required] The string identifier of the associated RestApi.}

\item{resourceId}{[required] Specifies a delete integration request's resource identifier.}

\item{httpMethod}{[required] Specifies a delete integration request's HTTP method.}
}
\description{
Represents a delete integration.

See \url{https://www.paws-r-sdk.com/docs/apigateway_delete_integration/} for full documentation.
}
\keyword{internal}
