% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_delete_method}
\alias{apigateway_delete_method}
\title{Deletes an existing Method resource}
\usage{
apigateway_delete_method(restApiId, resourceId, httpMethod)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated
<span>RestApi</span>.}

\item{resourceId}{[required] [Required] The <span>Resource</span> identifier for the
<span>Method</span> resource.}

\item{httpMethod}{[required] [Required] The HTTP verb of the <span>Method</span> resource.}
}
\value{
An empty list.
}
\description{
Deletes an existing <span>Method</span> resource.
}
\section{Request syntax}{
\preformatted{svc$delete_method(
  restApiId = "string",
  resourceId = "string",
  httpMethod = "string"
)
}
}

\keyword{internal}
