% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_delete_model}
\alias{apigateway_delete_model}
\title{Deletes a model}
\usage{
apigateway_delete_model(restApiId, modelName)
}
\arguments{
\item{restApiId}{[required] The string identifier of the associated RestApi.}

\item{modelName}{[required] The name of the model to delete.}
}
\description{
Deletes a model.

See \url{https://paws-r.github.io/docs/apigateway/delete_model.html} for full documentation.
}
\keyword{internal}
