% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_delete_request_validator}
\alias{apigateway_delete_request_validator}
\title{Deletes a RequestValidator of a given RestApi}
\usage{
apigateway_delete_request_validator(restApiId, requestValidatorId)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated
<span>RestApi</span>.}

\item{requestValidatorId}{[required] [Required] The identifier of the <span>RequestValidator</span> to be
deleted.}
}
\value{
An empty list.
}
\description{
Deletes a <span>RequestValidator</span> of a given <span>RestApi</span>.
}
\section{Request syntax}{
\preformatted{svc$delete_request_validator(
  restApiId = "string",
  requestValidatorId = "string"
)
}
}

\keyword{internal}
