% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_delete_resource}
\alias{apigateway_delete_resource}
\title{Deletes a Resource resource}
\usage{
apigateway_delete_resource(restApiId, resourceId)
}
\arguments{
\item{restApiId}{[required] The string identifier of the associated RestApi.}

\item{resourceId}{[required] The identifier of the Resource resource.}
}
\description{
Deletes a Resource resource.

See \url{https://www.paws-r-sdk.com/docs/apigateway_delete_resource/} for full documentation.
}
\keyword{internal}
