% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_delete_stage}
\alias{apigateway_delete_stage}
\title{Deletes a Stage resource}
\usage{
apigateway_delete_stage(restApiId, stageName)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated
<span>RestApi</span>.}

\item{stageName}{[required] [Required] The name of the <span>Stage</span> resource to delete.}
}
\value{
An empty list.
}
\description{
Deletes a <span>Stage</span> resource.
}
\section{Request syntax}{
\preformatted{svc$delete_stage(
  restApiId = "string",
  stageName = "string"
)
}
}

\keyword{internal}
