% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_delete_usage_plan}
\alias{apigateway_delete_usage_plan}
\title{Deletes a usage plan of a given plan Id}
\usage{
apigateway_delete_usage_plan(usagePlanId)
}
\arguments{
\item{usagePlanId}{[required] [Required] The Id of the to-be-deleted usage plan.}
}
\value{
An empty list.
}
\description{
Deletes a usage plan of a given plan Id.
}
\section{Request syntax}{
\preformatted{svc$delete_usage_plan(
  usagePlanId = "string"
)
}
}

\keyword{internal}
