% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_flush_stage_cache}
\alias{apigateway_flush_stage_cache}
\title{Flushes a stage's cache}
\usage{
apigateway_flush_stage_cache(restApiId, stageName)
}
\arguments{
\item{restApiId}{[required] The string identifier of the associated RestApi.}

\item{stageName}{[required] The name of the stage to flush its cache.}
}
\description{
Flushes a stage's cache.

See \url{https://paws-r.github.io/docs/apigateway/flush_stage_cache.html} for full documentation.
}
\keyword{internal}
