% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_generate_client_certificate}
\alias{apigateway_generate_client_certificate}
\title{Generates a ClientCertificate resource}
\usage{
apigateway_generate_client_certificate(description = NULL, tags = NULL)
}
\arguments{
\item{description}{The description of the ClientCertificate.}

\item{tags}{The key-value map of strings. The valid character set is
[a-zA-Z+-=._:/]. The tag key can be up to 128 characters and must not
start with \verb{aws:}. The tag value can be up to 256 characters.}
}
\description{
Generates a ClientCertificate resource.

See \url{https://www.paws-r-sdk.com/docs/apigateway_generate_client_certificate/} for full documentation.
}
\keyword{internal}
