% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_api_keys}
\alias{apigateway_get_api_keys}
\title{Gets information about the current ApiKeys resource}
\usage{
apigateway_get_api_keys(position, limit, nameQuery, customerId,
  includeValues)
}
\arguments{
\item{position}{The current pagination position in the paged result set.}

\item{limit}{The maximum number of returned results per page. The default value is 25
and the maximum value is 500.}

\item{nameQuery}{The name of queried API keys.}

\item{customerId}{The identifier of a customer in AWS Marketplace or an external system,
such as a developer portal.}

\item{includeValues}{A boolean flag to specify whether (\code{true}) or not (\code{false}) the result
contains key values.}
}
\value{
A list with the following syntax:\preformatted{list(
  warnings = list(
    "string"
  ),
  position = "string",
  items = list(
    list(
      id = "string",
      value = "string",
      name = "string",
      customerId = "string",
      description = "string",
      enabled = TRUE|FALSE,
      createdDate = as.POSIXct(
        "2015-01-01"
      ),
      lastUpdatedDate = as.POSIXct(
        "2015-01-01"
      ),
      stageKeys = list(
        "string"
      ),
      tags = list(
        "string"
      )
    )
  )
)
}
}
\description{
Gets information about the current ApiKeys resource.
}
\section{Request syntax}{
\preformatted{svc$get_api_keys(
  position = "string",
  limit = 123,
  nameQuery = "string",
  customerId = "string",
  includeValues = TRUE|FALSE
)
}
}

\keyword{internal}
