% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_base_path_mapping}
\alias{apigateway_get_base_path_mapping}
\title{Describe a BasePathMapping resource}
\usage{
apigateway_get_base_path_mapping(domainName, basePath)
}
\arguments{
\item{domainName}{[required] The domain name of the BasePathMapping resource to be described.}

\item{basePath}{[required] The base path name that callers of the API must provide as part of the
URL after the domain name. This value must be unique for all of the
mappings across a single API. Specify '(none)' if you do not want
callers to specify any base path name after the domain name.}
}
\description{
Describe a BasePathMapping resource.

See \url{https://www.paws-r-sdk.com/docs/apigateway_get_base_path_mapping/} for full documentation.
}
\keyword{internal}
