% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_client_certificates}
\alias{apigateway_get_client_certificates}
\title{Gets a collection of ClientCertificate resources}
\usage{
apigateway_get_client_certificates(position, limit)
}
\arguments{
\item{position}{The current pagination position in the paged result set.}

\item{limit}{The maximum number of returned results per page. The default value is 25
and the maximum value is 500.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  position = "string",
  items = list(
    list(
      clientCertificateId = "string",
      description = "string",
      pemEncodedCertificate = "string",
      createdDate = as.POSIXct(
        "2015-01-01"
      ),
      expirationDate = as.POSIXct(
        "2015-01-01"
      ),
      tags = list(
        "string"
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Gets a collection of ClientCertificate resources.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_client_certificates(
  position = "string",
  limit = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
