% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_documentation_version}
\alias{apigateway_get_documentation_version}
\title{Gets a documentation version}
\usage{
apigateway_get_documentation_version(restApiId, documentationVersion)
}
\arguments{
\item{restApiId}{[required] The string identifier of the associated RestApi.}

\item{documentationVersion}{[required] The version identifier of the to-be-retrieved documentation snapshot.}
}
\description{
Gets a documentation version.

See \url{https://www.paws-r-sdk.com/docs/apigateway_get_documentation_version/} for full documentation.
}
\keyword{internal}
