% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_documentation_versions}
\alias{apigateway_get_documentation_versions}
\title{Gets documentation versions}
\usage{
apigateway_get_documentation_versions(restApiId, position = NULL, limit = NULL)
}
\arguments{
\item{restApiId}{[required] The string identifier of the associated RestApi.}

\item{position}{The current pagination position in the paged result set.}

\item{limit}{The maximum number of returned results per page. The default value is 25
and the maximum value is 500.}
}
\description{
Gets documentation versions.

See \url{https://paws-r.github.io/docs/apigateway/get_documentation_versions.html} for full documentation.
}
\keyword{internal}
