% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_gateway_response}
\alias{apigateway_get_gateway_response}
\title{Gets a GatewayResponse of a specified response type on the given RestApi}
\usage{
apigateway_get_gateway_response(restApiId, responseType)
}
\arguments{
\item{restApiId}{[required] The string identifier of the associated RestApi.}

\item{responseType}{[required] The response type of the associated GatewayResponse.}
}
\description{
Gets a GatewayResponse of a specified response type on the given RestApi.

See \url{https://paws-r.github.io/docs/apigateway/get_gateway_response.html} for full documentation.
}
\keyword{internal}
