% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_gateway_responses}
\alias{apigateway_get_gateway_responses}
\title{Gets the GatewayResponses collection on the given RestApi}
\usage{
apigateway_get_gateway_responses(restApiId, position, limit)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated RestApi.}

\item{position}{The current pagination position in the paged result set. The
GatewayResponse collection does not support pagination and the position
does not apply here.}

\item{limit}{The maximum number of returned results per page. The default value is 25
and the maximum value is 500. The GatewayResponses collection does not
support pagination and the limit does not apply here.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  position = "string",
  items = list(
    list(
      responseType = "DEFAULT_4XX"|"DEFAULT_5XX"|"RESOURCE_NOT_FOUND"|"UNAUTHORIZED"|"INVALID_API_KEY"|"ACCESS_DENIED"|"AUTHORIZER_FAILURE"|"AUTHORIZER_CONFIGURATION_ERROR"|"INVALID_SIGNATURE"|"EXPIRED_TOKEN"|"MISSING_AUTHENTICATION_TOKEN"|"INTEGRATION_FAILURE"|"INTEGRATION_TIMEOUT"|"API_CONFIGURATION_ERROR"|"UNSUPPORTED_MEDIA_TYPE"|"BAD_REQUEST_PARAMETERS"|"BAD_REQUEST_BODY"|"REQUEST_TOO_LARGE"|"THROTTLED"|"QUOTA_EXCEEDED",
      statusCode = "string",
      responseParameters = list(
        "string"
      ),
      responseTemplates = list(
        "string"
      ),
      defaultResponse = TRUE|FALSE
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Gets the GatewayResponses collection on the given RestApi. If an API
developer has not added any definitions for gateway responses, the
result will be the API Gateway-generated default GatewayResponses
collection for the supported response types.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_gateway_responses(
  restApiId = "string",
  position = "string",
  limit = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
