% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_method}
\alias{apigateway_get_method}
\title{Describe an existing Method resource}
\usage{
apigateway_get_method(restApiId, resourceId, httpMethod)
}
\arguments{
\item{restApiId}{[required] The string identifier of the associated RestApi.}

\item{resourceId}{[required] The Resource identifier for the Method resource.}

\item{httpMethod}{[required] Specifies the method request's HTTP method type.}
}
\description{
Describe an existing Method resource.

See \url{https://paws-r.github.io/docs/apigateway/get_method.html} for full documentation.
}
\keyword{internal}
