% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_method_response}
\alias{apigateway_get_method_response}
\title{Describes a MethodResponse resource}
\usage{
apigateway_get_method_response(restApiId, resourceId, httpMethod,
  statusCode)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated
<span>RestApi</span>.}

\item{resourceId}{[required] [Required] The <span>Resource</span> identifier for the
<span>MethodResponse</span> resource.}

\item{httpMethod}{[required] [Required] The HTTP verb of the <span>Method</span> resource.}

\item{statusCode}{[required] [Required] The status code for the <span>MethodResponse</span>
resource.}
}
\value{
A list with the following syntax:\preformatted{list(
  statusCode = "string",
  responseParameters = list(
    TRUE|FALSE
  ),
  responseModels = list(
    "string"
  )
)
}
}
\description{
Describes a <span>MethodResponse</span> resource.
}
\section{Request syntax}{
\preformatted{svc$get_method_response(
  restApiId = "string",
  resourceId = "string",
  httpMethod = "string",
  statusCode = "string"
)
}
}

\keyword{internal}
