% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_request_validator}
\alias{apigateway_get_request_validator}
\title{Gets a RequestValidator of a given RestApi}
\usage{
apigateway_get_request_validator(restApiId, requestValidatorId)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated RestApi.}

\item{requestValidatorId}{[required] [Required] The identifier of the RequestValidator to be retrieved.}
}
\value{
A list with the following syntax:\preformatted{list(
  id = "string",
  name = "string",
  validateRequestBody = TRUE|FALSE,
  validateRequestParameters = TRUE|FALSE
)
}
}
\description{
Gets a RequestValidator of a given RestApi.
}
\section{Request syntax}{
\preformatted{svc$get_request_validator(
  restApiId = "string",
  requestValidatorId = "string"
)
}
}

\keyword{internal}
