% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_request_validators}
\alias{apigateway_get_request_validators}
\title{Gets the RequestValidators collection of a given RestApi}
\usage{
apigateway_get_request_validators(restApiId, position, limit)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated RestApi.}

\item{position}{The current pagination position in the paged result set.}

\item{limit}{The maximum number of returned results per page. The default value is 25
and the maximum value is 500.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  position = "string",
  items = list(
    list(
      id = "string",
      name = "string",
      validateRequestBody = TRUE|FALSE,
      validateRequestParameters = TRUE|FALSE
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Gets the RequestValidators collection of a given RestApi.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_request_validators(
  restApiId = "string",
  position = "string",
  limit = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
