% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_resource}
\alias{apigateway_get_resource}
\title{Lists information about a resource}
\usage{
apigateway_get_resource(restApiId, resourceId, embed)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated RestApi.}

\item{resourceId}{[required] [Required] The identifier for the Resource resource.}

\item{embed}{A query parameter to retrieve the specified resources embedded in the
returned Resource representation in the response. This \code{embed} parameter
value is a list of comma-separated strings. Currently, the request
supports only retrieval of the embedded Method resources this way. The
query parameter value must be a single-valued list and contain the
\code{"methods"} string. For example,
\verb{GET /restapis/\\\{restapi_id\\\}/resources/\\\{resource_id\\\}?embed=methods}.}
}
\description{
Lists information about a resource.
}
\section{Request syntax}{
\preformatted{svc$get_resource(
  restApiId = "string",
  resourceId = "string",
  embed = list(
    "string"
  )
)
}
}

\keyword{internal}
