% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_rest_apis}
\alias{apigateway_get_rest_apis}
\title{Lists the RestApis resources for your collection}
\usage{
apigateway_get_rest_apis(position, limit)
}
\arguments{
\item{position}{The current pagination position in the paged result set.}

\item{limit}{The maximum number of returned results per page. The default value is 25
and the maximum value is 500.}
}
\description{
Lists the RestApis resources for your collection.
}
\section{Request syntax}{
\preformatted{svc$get_rest_apis(
  position = "string",
  limit = 123
)
}
}

\keyword{internal}
