% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_stages}
\alias{apigateway_get_stages}
\title{Gets information about one or more Stage resources}
\usage{
apigateway_get_stages(restApiId, deploymentId)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated RestApi.}

\item{deploymentId}{The stages' deployment identifiers.}
}
\description{
Gets information about one or more Stage resources.
}
\section{Request syntax}{
\preformatted{svc$get_stages(
  restApiId = "string",
  deploymentId = "string"
)
}
}

\keyword{internal}
