% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_usage_plan_key}
\alias{apigateway_get_usage_plan_key}
\title{Gets a usage plan key of a given key identifier}
\usage{
apigateway_get_usage_plan_key(usagePlanId, keyId)
}
\arguments{
\item{usagePlanId}{[required] [Required] The Id of the UsagePlan resource representing the usage
plan containing the to-be-retrieved UsagePlanKey resource representing a
plan customer.}

\item{keyId}{[required] [Required] The key Id of the to-be-retrieved UsagePlanKey resource
representing a plan customer.}
}
\description{
Gets a usage plan key of a given key identifier.
}
\section{Request syntax}{
\preformatted{svc$get_usage_plan_key(
  usagePlanId = "string",
  keyId = "string"
)
}
}

\keyword{internal}
