% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_vpc_link}
\alias{apigateway_get_vpc_link}
\title{Gets a specified VPC link under the caller's account in a region}
\usage{
apigateway_get_vpc_link(vpcLinkId)
}
\arguments{
\item{vpcLinkId}{[required] [Required] The identifier of the <span>VpcLink</span>. It is used in
an <span>Integration</span> to reference this <span>VpcLink</span>.}
}
\value{
A list with the following syntax:\preformatted{list(
  id = "string",
  name = "string",
  description = "string",
  targetArns = list(
    "string"
  ),
  status = "AVAILABLE"|"PENDING"|"DELETING"|"FAILED",
  statusMessage = "string",
  tags = list(
    "string"
  )
)
}
}
\description{
Gets a specified VPC link under the caller's account in a region.
}
\section{Request syntax}{
\preformatted{svc$get_vpc_link(
  vpcLinkId = "string"
)
}
}

\keyword{internal}
