% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_vpc_links}
\alias{apigateway_get_vpc_links}
\title{Gets the VpcLinks collection under the caller's account in a selected
region}
\usage{
apigateway_get_vpc_links(position, limit)
}
\arguments{
\item{position}{The current pagination position in the paged result set.}

\item{limit}{The maximum number of returned results per page. The default value is 25
and the maximum value is 500.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  position = "string",
  items = list(
    list(
      id = "string",
      name = "string",
      description = "string",
      targetArns = list(
        "string"
      ),
      status = "AVAILABLE"|"PENDING"|"DELETING"|"FAILED",
      statusMessage = "string",
      tags = list(
        "string"
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Gets the VpcLinks collection under the caller's account in a selected
region.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_vpc_links(
  position = "string",
  limit = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
