% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_import_api_keys}
\alias{apigateway_import_api_keys}
\title{Import API keys from an external source, such as a CSV-formatted file}
\usage{
apigateway_import_api_keys(body, format, failOnWarnings = NULL)
}
\arguments{
\item{body}{[required] The payload of the POST request to import API keys. For the payload
format, see API Key File Format.}

\item{format}{[required] A query parameter to specify the input format to imported API keys.
Currently, only the \code{csv} format is supported.}

\item{failOnWarnings}{A query parameter to indicate whether to rollback ApiKey importation
(\code{true}) or not (\code{false}) when error is encountered.}
}
\description{
Import API keys from an external source, such as a CSV-formatted file.

See \url{https://paws-r.github.io/docs/apigateway/import_api_keys.html} for full documentation.
}
\keyword{internal}
