% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_import_documentation_parts}
\alias{apigateway_import_documentation_parts}
\title{Imports documentation parts}
\usage{
apigateway_import_documentation_parts(
  restApiId,
  mode = NULL,
  failOnWarnings = NULL,
  body
)
}
\arguments{
\item{restApiId}{[required] The string identifier of the associated RestApi.}

\item{mode}{A query parameter to indicate whether to overwrite (\code{OVERWRITE}) any
existing DocumentationParts definition or to merge (\code{MERGE}) the new
definition into the existing one. The default value is \code{MERGE}.}

\item{failOnWarnings}{A query parameter to specify whether to rollback the documentation
importation (\code{true}) or not (\code{false}) when a warning is encountered. The
default value is \code{false}.}

\item{body}{[required] Raw byte array representing the to-be-imported documentation parts. To
import from an OpenAPI file, this is a JSON object.}
}
\description{
Imports documentation parts

See \url{https://paws-r.github.io/docs/apigateway/import_documentation_parts.html} for full documentation.
}
\keyword{internal}
