% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_import_rest_api}
\alias{apigateway_import_rest_api}
\title{A feature of the API Gateway control service for creating a new API from
an external API definition file}
\usage{
apigateway_import_rest_api(failOnWarnings, parameters, body)
}
\arguments{
\item{failOnWarnings}{A query parameter to indicate whether to rollback the API creation
(\code{true}) or not (\code{false}) when a warning is encountered. The default
value is \code{false}.}

\item{parameters}{A key-value map of context-specific query string parameters specifying
the behavior of different API importing operations. The following shows
operation-specific parameters and their supported values.

To exclude DocumentationParts from the import, set \code{parameters} as
\code{ignore=documentation}.

To configure the endpoint type, set \code{parameters} as
\code{endpointConfigurationTypes=EDGE},
\code{endpointConfigurationTypes=REGIONAL}, or
\code{endpointConfigurationTypes=PRIVATE}. The default endpoint type is
\code{EDGE}.

To handle imported \code{basepath}, set \code{parameters} as \code{basepath=ignore},
\code{basepath=prepend} or \code{basepath=split}.

For example, the AWS CLI command to exclude documentation from the
imported API is:

\if{html}{\out{<div class="sourceCode">}}\preformatted{aws apigateway import-rest-api --parameters ignore=documentation --body 'file:///path/to/imported-api-body.json'
}\if{html}{\out{</div>}}

The AWS CLI command to set the regional endpoint on the imported API is:

\if{html}{\out{<div class="sourceCode">}}\preformatted{aws apigateway import-rest-api --parameters endpointConfigurationTypes=REGIONAL --body 'file:///path/to/imported-api-body.json'
}\if{html}{\out{</div>}}}

\item{body}{[required] [Required] The POST request body containing external API definitions.
Currently, only OpenAPI definition JSON/YAML files are supported. The
maximum size of the API definition file is 6MB.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  id = "string",
  name = "string",
  description = "string",
  createdDate = as.POSIXct(
    "2015-01-01"
  ),
  version = "string",
  warnings = list(
    "string"
  ),
  binaryMediaTypes = list(
    "string"
  ),
  minimumCompressionSize = 123,
  apiKeySource = "HEADER"|"AUTHORIZER",
  endpointConfiguration = list(
    types = list(
      "REGIONAL"|"EDGE"|"PRIVATE"
    ),
    vpcEndpointIds = list(
      "string"
    )
  ),
  policy = "string",
  tags = list(
    "string"
  ),
  disableExecuteApiEndpoint = TRUE|FALSE
)
}\if{html}{\out{</div>}}
}
\description{
A feature of the API Gateway control service for creating a new API from
an external API definition file.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$import_rest_api(
  failOnWarnings = TRUE|FALSE,
  parameters = list(
    "string"
  ),
  body = raw
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
