% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_put_gateway_response}
\alias{apigateway_put_gateway_response}
\title{Creates a customization of a GatewayResponse of a specified response
type and status code on the given RestApi}
\usage{
apigateway_put_gateway_response(
  restApiId,
  responseType,
  statusCode = NULL,
  responseParameters = NULL,
  responseTemplates = NULL
)
}
\arguments{
\item{restApiId}{[required] The string identifier of the associated RestApi.}

\item{responseType}{[required] The response type of the associated GatewayResponse}

\item{statusCode}{The HTTP status code of the GatewayResponse.}

\item{responseParameters}{Response parameters (paths, query strings and headers) of the
GatewayResponse as a string-to-string map of key-value pairs.}

\item{responseTemplates}{Response templates of the GatewayResponse as a string-to-string map of
key-value pairs.}
}
\description{
Creates a customization of a GatewayResponse of a specified response type and status code on the given RestApi.

See \url{https://www.paws-r-sdk.com/docs/apigateway_put_gateway_response/} for full documentation.
}
\keyword{internal}
