% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_reject_domain_name_access_association}
\alias{apigateway_reject_domain_name_access_association}
\title{Rejects a domain name access association with a private custom domain
name}
\usage{
apigateway_reject_domain_name_access_association(
  domainNameAccessAssociationArn,
  domainNameArn
)
}
\arguments{
\item{domainNameAccessAssociationArn}{[required] The ARN of the domain name access association resource.}

\item{domainNameArn}{[required] The ARN of the domain name.}
}
\description{
Rejects a domain name access association with a private custom domain name.

See \url{https://www.paws-r-sdk.com/docs/apigateway_reject_domain_name_access_association/} for full documentation.
}
\keyword{internal}
