% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_update_account}
\alias{apigateway_update_account}
\title{Changes information about the current Account resource}
\usage{
apigateway_update_account(patchOperations)
}
\arguments{
\item{patchOperations}{A list of update operations to be applied to the specified resource and
in the order specified in this list.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  cloudwatchRoleArn = "string",
  throttleSettings = list(
    burstLimit = 123,
    rateLimit = 123.0
  ),
  features = list(
    "string"
  ),
  apiKeyVersion = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Changes information about the current Account resource.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_account(
  patchOperations = list(
    list(
      op = "add"|"remove"|"replace"|"move"|"copy"|"test",
      path = "string",
      value = "string",
      from = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
