% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_update_api_key}
\alias{apigateway_update_api_key}
\title{Changes information about an ApiKey resource}
\usage{
apigateway_update_api_key(apiKey, patchOperations)
}
\arguments{
\item{apiKey}{[required] [Required] The identifier of the ApiKey resource to be updated.}

\item{patchOperations}{A list of update operations to be applied to the specified resource and
in the order specified in this list.}
}
\value{
A list with the following syntax:\preformatted{list(
  id = "string",
  value = "string",
  name = "string",
  customerId = "string",
  description = "string",
  enabled = TRUE|FALSE,
  createdDate = as.POSIXct(
    "2015-01-01"
  ),
  lastUpdatedDate = as.POSIXct(
    "2015-01-01"
  ),
  stageKeys = list(
    "string"
  ),
  tags = list(
    "string"
  )
)
}
}
\description{
Changes information about an ApiKey resource.
}
\section{Request syntax}{
\preformatted{svc$update_api_key(
  apiKey = "string",
  patchOperations = list(
    list(
      op = "add"|"remove"|"replace"|"move"|"copy"|"test",
      path = "string",
      value = "string",
      from = "string"
    )
  )
)
}
}

\keyword{internal}
