% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_update_base_path_mapping}
\alias{apigateway_update_base_path_mapping}
\title{Changes information about the BasePathMapping resource}
\usage{
apigateway_update_base_path_mapping(
  domainName,
  basePath,
  patchOperations = NULL
)
}
\arguments{
\item{domainName}{[required] The domain name of the BasePathMapping resource to change.}

\item{basePath}{[required] The base path of the BasePathMapping resource to change.

To specify an empty base path, set this parameter to \code{'(none)'}.}

\item{patchOperations}{For more information about supported patch operations, see \href{https://docs.aws.amazon.com/apigateway/latest/api/patch-operations.html}{Patch Operations}.}
}
\description{
Changes information about the BasePathMapping resource.

See \url{https://paws-r.github.io/docs/apigateway/update_base_path_mapping.html} for full documentation.
}
\keyword{internal}
