% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_update_deployment}
\alias{apigateway_update_deployment}
\title{Changes information about a Deployment resource}
\usage{
apigateway_update_deployment(restApiId, deploymentId, patchOperations)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated RestApi.}

\item{deploymentId}{[required] The replacement identifier for the Deployment resource to change
information about.}

\item{patchOperations}{A list of update operations to be applied to the specified resource and
in the order specified in this list.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  id = "string",
  description = "string",
  createdDate = as.POSIXct(
    "2015-01-01"
  ),
  apiSummary = list(
    list(
      list(
        authorizationType = "string",
        apiKeyRequired = TRUE|FALSE
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Changes information about a Deployment resource.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_deployment(
  restApiId = "string",
  deploymentId = "string",
  patchOperations = list(
    list(
      op = "add"|"remove"|"replace"|"move"|"copy"|"test",
      path = "string",
      value = "string",
      from = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
