% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_update_domain_name}
\alias{apigateway_update_domain_name}
\title{Changes information about the DomainName resource}
\usage{
apigateway_update_domain_name(domainName, patchOperations = NULL)
}
\arguments{
\item{domainName}{[required] The name of the DomainName resource to be changed.}

\item{patchOperations}{For more information about supported patch operations, see \href{https://docs.aws.amazon.com/apigateway/latest/api/patch-operations.html}{Patch Operations}.}
}
\description{
Changes information about the DomainName resource.

See \url{https://www.paws-r-sdk.com/docs/apigateway_update_domain_name/} for full documentation.
}
\keyword{internal}
