% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_update_integration_response}
\alias{apigateway_update_integration_response}
\title{Represents an update integration response}
\usage{
apigateway_update_integration_response(
  restApiId,
  resourceId,
  httpMethod,
  statusCode,
  patchOperations = NULL
)
}
\arguments{
\item{restApiId}{[required] The string identifier of the associated RestApi.}

\item{resourceId}{[required] Specifies an update integration response request's resource identifier.}

\item{httpMethod}{[required] Specifies an update integration response request's HTTP method.}

\item{statusCode}{[required] Specifies an update integration response request's status code.}

\item{patchOperations}{For more information about supported patch operations, see \href{https://docs.aws.amazon.com/apigateway/latest/api/patch-operations.html}{Patch Operations}.}
}
\description{
Represents an update integration response.

See \url{https://www.paws-r-sdk.com/docs/apigateway_update_integration_response/} for full documentation.
}
\keyword{internal}
