% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_update_method_response}
\alias{apigateway_update_method_response}
\title{Updates an existing MethodResponse resource}
\usage{
apigateway_update_method_response(
  restApiId,
  resourceId,
  httpMethod,
  statusCode,
  patchOperations = NULL
)
}
\arguments{
\item{restApiId}{[required] The string identifier of the associated RestApi.}

\item{resourceId}{[required] The Resource identifier for the MethodResponse resource.}

\item{httpMethod}{[required] The HTTP verb of the Method resource.}

\item{statusCode}{[required] The status code for the MethodResponse resource.}

\item{patchOperations}{For more information about supported patch operations, see \href{https://docs.aws.amazon.com/apigateway/latest/api/patch-operations.html}{Patch Operations}.}
}
\description{
Updates an existing MethodResponse resource.

See \url{https://paws-r.github.io/docs/apigateway/update_method_response.html} for full documentation.
}
\keyword{internal}
