% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_update_model}
\alias{apigateway_update_model}
\title{Changes information about a model}
\usage{
apigateway_update_model(restApiId, modelName, patchOperations)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated RestApi.}

\item{modelName}{[required] [Required] The name of the model to update.}

\item{patchOperations}{A list of update operations to be applied to the specified resource and
in the order specified in this list.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  id = "string",
  name = "string",
  description = "string",
  schema = "string",
  contentType = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Changes information about a model.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_model(
  restApiId = "string",
  modelName = "string",
  patchOperations = list(
    list(
      op = "add"|"remove"|"replace"|"move"|"copy"|"test",
      path = "string",
      value = "string",
      from = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
