% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_update_resource}
\alias{apigateway_update_resource}
\title{Changes information about a Resource resource}
\usage{
apigateway_update_resource(restApiId, resourceId, patchOperations = NULL)
}
\arguments{
\item{restApiId}{[required] The string identifier of the associated RestApi.}

\item{resourceId}{[required] The identifier of the Resource resource.}

\item{patchOperations}{For more information about supported patch operations, see \href{https://docs.aws.amazon.com/apigateway/latest/api/patch-operations.html}{Patch Operations}.}
}
\description{
Changes information about a Resource resource.

See \url{https://paws-r.github.io/docs/apigateway/update_resource.html} for full documentation.
}
\keyword{internal}
