% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_update_stage}
\alias{apigateway_update_stage}
\title{Changes information about a Stage resource}
\usage{
apigateway_update_stage(restApiId, stageName, patchOperations)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated RestApi.}

\item{stageName}{[required] [Required] The name of the Stage resource to change information about.}

\item{patchOperations}{A list of update operations to be applied to the specified resource and
in the order specified in this list.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  deploymentId = "string",
  clientCertificateId = "string",
  stageName = "string",
  description = "string",
  cacheClusterEnabled = TRUE|FALSE,
  cacheClusterSize = "0.5"|"1.6"|"6.1"|"13.5"|"28.4"|"58.2"|"118"|"237",
  cacheClusterStatus = "CREATE_IN_PROGRESS"|"AVAILABLE"|"DELETE_IN_PROGRESS"|"NOT_AVAILABLE"|"FLUSH_IN_PROGRESS",
  methodSettings = list(
    list(
      metricsEnabled = TRUE|FALSE,
      loggingLevel = "string",
      dataTraceEnabled = TRUE|FALSE,
      throttlingBurstLimit = 123,
      throttlingRateLimit = 123.0,
      cachingEnabled = TRUE|FALSE,
      cacheTtlInSeconds = 123,
      cacheDataEncrypted = TRUE|FALSE,
      requireAuthorizationForCacheControl = TRUE|FALSE,
      unauthorizedCacheControlHeaderStrategy = "FAIL_WITH_403"|"SUCCEED_WITH_RESPONSE_HEADER"|"SUCCEED_WITHOUT_RESPONSE_HEADER"
    )
  ),
  variables = list(
    "string"
  ),
  documentationVersion = "string",
  accessLogSettings = list(
    format = "string",
    destinationArn = "string"
  ),
  canarySettings = list(
    percentTraffic = 123.0,
    deploymentId = "string",
    stageVariableOverrides = list(
      "string"
    ),
    useStageCache = TRUE|FALSE
  ),
  tracingEnabled = TRUE|FALSE,
  webAclArn = "string",
  tags = list(
    "string"
  ),
  createdDate = as.POSIXct(
    "2015-01-01"
  ),
  lastUpdatedDate = as.POSIXct(
    "2015-01-01"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Changes information about a Stage resource.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_stage(
  restApiId = "string",
  stageName = "string",
  patchOperations = list(
    list(
      op = "add"|"remove"|"replace"|"move"|"copy"|"test",
      path = "string",
      value = "string",
      from = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
