% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_update_usage}
\alias{apigateway_update_usage}
\title{Grants a temporary extension to the remaining quota of a usage plan
associated with a specified API key}
\usage{
apigateway_update_usage(usagePlanId, keyId, patchOperations = NULL)
}
\arguments{
\item{usagePlanId}{[required] The Id of the usage plan associated with the usage data.}

\item{keyId}{[required] The identifier of the API key associated with the usage plan in which a
temporary extension is granted to the remaining quota.}

\item{patchOperations}{For more information about supported patch operations, see \href{https://docs.aws.amazon.com/apigateway/latest/api/patch-operations.html}{Patch Operations}.}
}
\description{
Grants a temporary extension to the remaining quota of a usage plan associated with a specified API key.

See \url{https://www.paws-r-sdk.com/docs/apigateway_update_usage/} for full documentation.
}
\keyword{internal}
