% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_update_usage_plan}
\alias{apigateway_update_usage_plan}
\title{Updates a usage plan of a given plan Id}
\usage{
apigateway_update_usage_plan(usagePlanId, patchOperations = NULL)
}
\arguments{
\item{usagePlanId}{[required] The Id of the to-be-updated usage plan.}

\item{patchOperations}{For more information about supported patch operations, see \href{https://docs.aws.amazon.com/apigateway/latest/api/patch-operations.html}{Patch Operations}.}
}
\description{
Updates a usage plan of a given plan Id.

See \url{https://paws-r.github.io/docs/apigateway/update_usage_plan.html} for full documentation.
}
\keyword{internal}
