% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_update_vpc_link}
\alias{apigateway_update_vpc_link}
\title{Updates an existing VpcLink of a specified identifier}
\usage{
apigateway_update_vpc_link(vpcLinkId, patchOperations = NULL)
}
\arguments{
\item{vpcLinkId}{[required] The identifier of the VpcLink. It is used in an Integration to reference
this VpcLink.}

\item{patchOperations}{For more information about supported patch operations, see \href{https://docs.aws.amazon.com/apigateway/latest/api/patch-operations.html}{Patch Operations}.}
}
\description{
Updates an existing VpcLink of a specified identifier.

See \url{https://www.paws-r-sdk.com/docs/apigateway_update_vpc_link/} for full documentation.
}
\keyword{internal}
