% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewaymanagementapi_operations.R
\name{apigatewaymanagementapi_get_connection}
\alias{apigatewaymanagementapi_get_connection}
\title{Get information about the connection with the provided id}
\usage{
apigatewaymanagementapi_get_connection(ConnectionId)
}
\arguments{
\item{ConnectionId}{[required]}
}
\description{
Get information about the connection with the provided id.

See \url{https://paws-r.github.io/docs/apigatewaymanagementapi/get_connection.html} for full documentation.
}
\keyword{internal}
