% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_create_api_mapping}
\alias{apigatewayv2_create_api_mapping}
\title{Creates an API mapping}
\usage{
apigatewayv2_create_api_mapping(ApiId, ApiMappingKey = NULL, DomainName, Stage)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{ApiMappingKey}{The API mapping key.}

\item{DomainName}{[required] The domain name.}

\item{Stage}{[required] The API stage.}
}
\description{
Creates an API mapping.

See \url{https://paws-r.github.io/docs/apigatewayv2/create_api_mapping.html} for full documentation.
}
\keyword{internal}
