% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_create_portal}
\alias{apigatewayv2_create_portal}
\title{Creates a portal}
\usage{
apigatewayv2_create_portal(
  Authorization,
  EndpointConfiguration,
  IncludedPortalProductArns = NULL,
  LogoUri = NULL,
  PortalContent,
  RumAppMonitorName = NULL,
  Tags = NULL
)
}
\arguments{
\item{Authorization}{[required] The authentication configuration for the portal.}

\item{EndpointConfiguration}{[required] The domain configuration for the portal. Use a default domain provided
by API Gateway or provide a fully-qualified domain name that you own.}

\item{IncludedPortalProductArns}{The ARNs of the portal products included in the portal.}

\item{LogoUri}{The URI for the portal logo image that is displayed in the portal
header.}

\item{PortalContent}{[required] The content of the portal.}

\item{RumAppMonitorName}{The name of the Amazon CloudWatch RUM app monitor for the portal.}

\item{Tags}{The collection of tags. Each tag element is associated with a given
resource.}
}
\description{
Creates a portal.

See \url{https://www.paws-r-sdk.com/docs/apigatewayv2_create_portal/} for full documentation.
}
\keyword{internal}
