% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_create_portal_product}
\alias{apigatewayv2_create_portal_product}
\title{Creates a new portal product}
\usage{
apigatewayv2_create_portal_product(
  Description = NULL,
  DisplayName,
  Tags = NULL
)
}
\arguments{
\item{Description}{A description of the portal product.}

\item{DisplayName}{[required] The name of the portal product as it appears in a published portal.}

\item{Tags}{The collection of tags. Each tag element is associated with a given
resource.}
}
\description{
Creates a new portal product.

See \url{https://www.paws-r-sdk.com/docs/apigatewayv2_create_portal_product/} for full documentation.
}
\keyword{internal}
