% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_create_product_page}
\alias{apigatewayv2_create_product_page}
\title{Creates a new product page for a portal product}
\usage{
apigatewayv2_create_product_page(DisplayContent, PortalProductId)
}
\arguments{
\item{DisplayContent}{[required] The content of the product page.}

\item{PortalProductId}{[required] The portal product identifier.}
}
\description{
Creates a new product page for a portal product.

See \url{https://www.paws-r-sdk.com/docs/apigatewayv2_create_product_page/} for full documentation.
}
\keyword{internal}
