% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_create_product_rest_endpoint_page}
\alias{apigatewayv2_create_product_rest_endpoint_page}
\title{Creates a product REST endpoint page for a portal product}
\usage{
apigatewayv2_create_product_rest_endpoint_page(
  DisplayContent = NULL,
  PortalProductId,
  RestEndpointIdentifier,
  TryItState = NULL
)
}
\arguments{
\item{DisplayContent}{The content of the product REST endpoint page.}

\item{PortalProductId}{[required] The portal product identifier.}

\item{RestEndpointIdentifier}{[required] The REST endpoint identifier.}

\item{TryItState}{The try it state of the product REST endpoint page.}
}
\description{
Creates a product REST endpoint page for a portal product.

See \url{https://www.paws-r-sdk.com/docs/apigatewayv2_create_product_rest_endpoint_page/} for full documentation.
}
\keyword{internal}
