% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_create_route_response}
\alias{apigatewayv2_create_route_response}
\title{Creates a RouteResponse for a Route}
\usage{
apigatewayv2_create_route_response(
  ApiId,
  ModelSelectionExpression = NULL,
  ResponseModels = NULL,
  ResponseParameters = NULL,
  RouteId,
  RouteResponseKey
)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{ModelSelectionExpression}{The model selection expression for the route response. Supported only
for WebSocket APIs.}

\item{ResponseModels}{The response models for the route response.}

\item{ResponseParameters}{The route response parameters.}

\item{RouteId}{[required] The route ID.}

\item{RouteResponseKey}{[required] The route response key.}
}
\description{
Creates a RouteResponse for a Route.

See \url{https://paws-r.github.io/docs/apigatewayv2/create_route_response.html} for full documentation.
}
\keyword{internal}
