% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_create_routing_rule}
\alias{apigatewayv2_create_routing_rule}
\title{Creates a RoutingRule}
\usage{
apigatewayv2_create_routing_rule(
  Actions,
  Conditions,
  DomainName,
  DomainNameId = NULL,
  Priority
)
}
\arguments{
\item{Actions}{[required] Represents a routing rule action. The only supported action is
invokeApi.}

\item{Conditions}{[required] Represents a condition. Conditions can contain up to two matchHeaders
conditions and one matchBasePaths conditions. API Gateway evaluates
header conditions and base path conditions together. You can only use
AND between header and base path conditions.}

\item{DomainName}{[required] The domain name.}

\item{DomainNameId}{The domain name ID.}

\item{Priority}{[required] Represents the priority of the routing rule.}
}
\description{
Creates a RoutingRule.

See \url{https://www.paws-r-sdk.com/docs/apigatewayv2_create_routing_rule/} for full documentation.
}
\keyword{internal}
