% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_create_vpc_link}
\alias{apigatewayv2_create_vpc_link}
\title{Creates a VPC link}
\usage{
apigatewayv2_create_vpc_link(
  Name,
  SecurityGroupIds = NULL,
  SubnetIds,
  Tags = NULL
)
}
\arguments{
\item{Name}{[required] The name of the VPC link.}

\item{SecurityGroupIds}{A list of security group IDs for the VPC link.}

\item{SubnetIds}{[required] A list of subnet IDs to include in the VPC link.}

\item{Tags}{A list of tags.}
}
\description{
Creates a VPC link.

See \url{https://www.paws-r-sdk.com/docs/apigatewayv2_create_vpc_link/} for full documentation.
}
\keyword{internal}
