% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_delete_api}
\alias{apigatewayv2_delete_api}
\title{Deletes an Api resource}
\usage{
apigatewayv2_delete_api(ApiId)
}
\arguments{
\item{ApiId}{[required] The API identifier.}
}
\value{
An empty list.
}
\description{
Deletes an Api resource.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_api(
  ApiId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
