% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_delete_api_mapping}
\alias{apigatewayv2_delete_api_mapping}
\title{Deletes an API mapping}
\usage{
apigatewayv2_delete_api_mapping(ApiMappingId, DomainName)
}
\arguments{
\item{ApiMappingId}{[required] The API mapping identifier.}

\item{DomainName}{[required] The domain name.}
}
\description{
Deletes an API mapping.

See \url{https://www.paws-r-sdk.com/docs/apigatewayv2_delete_api_mapping/} for full documentation.
}
\keyword{internal}
