% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_delete_cors_configuration}
\alias{apigatewayv2_delete_cors_configuration}
\title{Deletes a CORS configuration}
\usage{
apigatewayv2_delete_cors_configuration(ApiId)
}
\arguments{
\item{ApiId}{[required] The API identifier.}
}
\description{
Deletes a CORS configuration.

See \url{https://www.paws-r-sdk.com/docs/apigatewayv2_delete_cors_configuration/} for full documentation.
}
\keyword{internal}
