% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_delete_domain_name}
\alias{apigatewayv2_delete_domain_name}
\title{Deletes a domain name}
\usage{
apigatewayv2_delete_domain_name(DomainName)
}
\arguments{
\item{DomainName}{[required] The domain name.}
}
\value{
An empty list.
}
\description{
Deletes a domain name.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_domain_name(
  DomainName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
