% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_delete_model}
\alias{apigatewayv2_delete_model}
\title{Deletes a Model}
\usage{
apigatewayv2_delete_model(ApiId, ModelId)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{ModelId}{[required] The model ID.}
}
\description{
Deletes a Model.

See \url{https://www.paws-r-sdk.com/docs/apigatewayv2_delete_model/} for full documentation.
}
\keyword{internal}
