% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_delete_portal}
\alias{apigatewayv2_delete_portal}
\title{Deletes a portal}
\usage{
apigatewayv2_delete_portal(PortalId)
}
\arguments{
\item{PortalId}{[required] The portal identifier.}
}
\description{
Deletes a portal.

See \url{https://www.paws-r-sdk.com/docs/apigatewayv2_delete_portal/} for full documentation.
}
\keyword{internal}
