% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_delete_portal_product_sharing_policy}
\alias{apigatewayv2_delete_portal_product_sharing_policy}
\title{Deletes the sharing policy for a portal product}
\usage{
apigatewayv2_delete_portal_product_sharing_policy(PortalProductId)
}
\arguments{
\item{PortalProductId}{[required] The portal product identifier.}
}
\description{
Deletes the sharing policy for a portal product.

See \url{https://www.paws-r-sdk.com/docs/apigatewayv2_delete_portal_product_sharing_policy/} for full documentation.
}
\keyword{internal}
