% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_delete_route_request_parameter}
\alias{apigatewayv2_delete_route_request_parameter}
\title{Deletes a route request parameter}
\usage{
apigatewayv2_delete_route_request_parameter(
  ApiId,
  RequestParameterKey,
  RouteId
)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{RequestParameterKey}{[required] The route request parameter key.}

\item{RouteId}{[required] The route ID.}
}
\description{
Deletes a route request parameter. Supported only for WebSocket APIs.

See \url{https://www.paws-r-sdk.com/docs/apigatewayv2_delete_route_request_parameter/} for full documentation.
}
\keyword{internal}
