% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_delete_route_response}
\alias{apigatewayv2_delete_route_response}
\title{Deletes a RouteResponse}
\usage{
apigatewayv2_delete_route_response(ApiId, RouteId, RouteResponseId)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{RouteId}{[required] The route ID.}

\item{RouteResponseId}{[required] The route response ID.}
}
\value{
An empty list.
}
\description{
Deletes a RouteResponse.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_route_response(
  ApiId = "string",
  RouteId = "string",
  RouteResponseId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
