% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_delete_routing_rule}
\alias{apigatewayv2_delete_routing_rule}
\title{Deletes a routing rule}
\usage{
apigatewayv2_delete_routing_rule(
  DomainName,
  DomainNameId = NULL,
  RoutingRuleId
)
}
\arguments{
\item{DomainName}{[required] The domain name.}

\item{DomainNameId}{The domain name ID.}

\item{RoutingRuleId}{[required] The routing rule ID.}
}
\description{
Deletes a routing rule.

See \url{https://www.paws-r-sdk.com/docs/apigatewayv2_delete_routing_rule/} for full documentation.
}
\keyword{internal}
