% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_delete_stage}
\alias{apigatewayv2_delete_stage}
\title{Deletes a Stage}
\usage{
apigatewayv2_delete_stage(ApiId, StageName)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{StageName}{[required] The stage name. Stage names can only contain alphanumeric characters,
hyphens, and underscores. Maximum length is 128 characters.}
}
\description{
Deletes a Stage.

See \url{https://www.paws-r-sdk.com/docs/apigatewayv2_delete_stage/} for full documentation.
}
\keyword{internal}
