% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_disable_portal}
\alias{apigatewayv2_disable_portal}
\title{Deletes the publication of a portal portal}
\usage{
apigatewayv2_disable_portal(PortalId)
}
\arguments{
\item{PortalId}{[required] The portal identifier.}
}
\description{
Deletes the publication of a portal portal.

See \url{https://www.paws-r-sdk.com/docs/apigatewayv2_disable_portal/} for full documentation.
}
\keyword{internal}
