% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_get_api}
\alias{apigatewayv2_get_api}
\title{Gets an Api resource}
\usage{
apigatewayv2_get_api(ApiId)
}
\arguments{
\item{ApiId}{[required] The API identifier.}
}
\description{
Gets an Api resource.

See \url{https://www.paws-r-sdk.com/docs/apigatewayv2_get_api/} for full documentation.
}
\keyword{internal}
