% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_get_api_mappings}
\alias{apigatewayv2_get_api_mappings}
\title{Gets API mappings}
\usage{
apigatewayv2_get_api_mappings(DomainName, MaxResults, NextToken)
}
\arguments{
\item{DomainName}{[required] The domain name.}

\item{MaxResults}{The maximum number of elements to be returned for this resource.}

\item{NextToken}{The next page of elements from this collection. Not valid for the last
element of the collection.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Items = list(
    list(
      ApiId = "string",
      ApiMappingId = "string",
      ApiMappingKey = "string",
      Stage = "string"
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Gets API mappings.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_api_mappings(
  DomainName = "string",
  MaxResults = "string",
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
