% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_get_domain_name}
\alias{apigatewayv2_get_domain_name}
\title{Gets a domain name}
\usage{
apigatewayv2_get_domain_name(DomainName)
}
\arguments{
\item{DomainName}{[required] The domain name.}
}
\description{
Gets a domain name.

See \url{https://paws-r.github.io/docs/apigatewayv2/get_domain_name.html} for full documentation.
}
\keyword{internal}
