% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_get_domain_names}
\alias{apigatewayv2_get_domain_names}
\title{Gets the domain names for an AWS account}
\usage{
apigatewayv2_get_domain_names(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{The maximum number of elements to be returned for this resource.}

\item{NextToken}{The next page of elements from this collection. Not valid for the last
element of the collection.}
}
\description{
Gets the domain names for an AWS account.

See \url{https://paws-r.github.io/docs/apigatewayv2/get_domain_names.html} for full documentation.
}
\keyword{internal}
