% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_get_integration_response}
\alias{apigatewayv2_get_integration_response}
\title{Gets an IntegrationResponses}
\usage{
apigatewayv2_get_integration_response(ApiId, IntegrationId,
  IntegrationResponseId)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{IntegrationId}{[required] The integration ID.}

\item{IntegrationResponseId}{[required] The integration response ID.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ContentHandlingStrategy = "CONVERT_TO_BINARY"|"CONVERT_TO_TEXT",
  IntegrationResponseId = "string",
  IntegrationResponseKey = "string",
  ResponseParameters = list(
    "string"
  ),
  ResponseTemplates = list(
    "string"
  ),
  TemplateSelectionExpression = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Gets an IntegrationResponses.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_integration_response(
  ApiId = "string",
  IntegrationId = "string",
  IntegrationResponseId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
