% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_get_model}
\alias{apigatewayv2_get_model}
\title{Gets a Model}
\usage{
apigatewayv2_get_model(ApiId, ModelId)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{ModelId}{[required] The model ID.}
}
\description{
Gets a Model.

See \url{https://paws-r.github.io/docs/apigatewayv2/get_model.html} for full documentation.
}
\keyword{internal}
