% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_get_model_template}
\alias{apigatewayv2_get_model_template}
\title{Gets a model template}
\usage{
apigatewayv2_get_model_template(ApiId, ModelId)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{ModelId}{[required] The model ID.}
}
\description{
Gets a model template.

See \url{https://paws-r.github.io/docs/apigatewayv2/get_model_template.html} for full documentation.
}
\keyword{internal}
