% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_get_models}
\alias{apigatewayv2_get_models}
\title{Gets the Models for an API}
\usage{
apigatewayv2_get_models(ApiId, MaxResults, NextToken)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{MaxResults}{The maximum number of elements to be returned for this resource.}

\item{NextToken}{The next page of elements from this collection. Not valid for the last
element of the collection.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Items = list(
    list(
      ContentType = "string",
      Description = "string",
      ModelId = "string",
      Name = "string",
      Schema = "string"
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Gets the Models for an API.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_models(
  ApiId = "string",
  MaxResults = "string",
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
