% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_get_portal}
\alias{apigatewayv2_get_portal}
\title{Gets a portal}
\usage{
apigatewayv2_get_portal(PortalId)
}
\arguments{
\item{PortalId}{[required] The portal identifier.}
}
\description{
Gets a portal.

See \url{https://www.paws-r-sdk.com/docs/apigatewayv2_get_portal/} for full documentation.
}
\keyword{internal}
