% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_get_portal_product}
\alias{apigatewayv2_get_portal_product}
\title{Gets a portal product}
\usage{
apigatewayv2_get_portal_product(PortalProductId, ResourceOwnerAccountId = NULL)
}
\arguments{
\item{PortalProductId}{[required] The portal product identifier.}

\item{ResourceOwnerAccountId}{The account ID of the resource owner of the portal product.}
}
\description{
Gets a portal product.

See \url{https://www.paws-r-sdk.com/docs/apigatewayv2_get_portal_product/} for full documentation.
}
\keyword{internal}
