% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_get_route_response}
\alias{apigatewayv2_get_route_response}
\title{Gets a RouteResponse}
\usage{
apigatewayv2_get_route_response(ApiId, RouteId, RouteResponseId)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{RouteId}{[required] The route ID.}

\item{RouteResponseId}{[required] The route response ID.}
}
\description{
Gets a RouteResponse.

See \url{https://paws-r.github.io/docs/apigatewayv2/get_route_response.html} for full documentation.
}
\keyword{internal}
