% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_get_route_responses}
\alias{apigatewayv2_get_route_responses}
\title{Gets the RouteResponses for a Route}
\usage{
apigatewayv2_get_route_responses(
  ApiId,
  MaxResults = NULL,
  NextToken = NULL,
  RouteId
)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{MaxResults}{The maximum number of elements to be returned for this resource.}

\item{NextToken}{The next page of elements from this collection. Not valid for the last
element of the collection.}

\item{RouteId}{[required] The route ID.}
}
\description{
Gets the RouteResponses for a Route.

See \url{https://www.paws-r-sdk.com/docs/apigatewayv2_get_route_responses/} for full documentation.
}
\keyword{internal}
