% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_get_stage}
\alias{apigatewayv2_get_stage}
\title{Gets a Stage}
\usage{
apigatewayv2_get_stage(ApiId, StageName)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{StageName}{[required] The stage name. Stage names can only contain alphanumeric characters,
hyphens, and underscores. Maximum length is 128 characters.}
}
\description{
Gets a Stage.

See \url{https://paws-r.github.io/docs/apigatewayv2/get_stage.html} for full documentation.
}
\keyword{internal}
