% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_get_tags}
\alias{apigatewayv2_get_tags}
\title{Gets a collection of Tag resources}
\usage{
apigatewayv2_get_tags(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The resource ARN for the tag.}
}
\description{
Gets a collection of Tag resources.

See \url{https://paws-r-sdk.com/docs/get_tags/} for full documentation.
}
\keyword{internal}
