% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_get_vpc_links}
\alias{apigatewayv2_get_vpc_links}
\title{Gets a collection of VPC links}
\usage{
apigatewayv2_get_vpc_links(MaxResults, NextToken)
}
\arguments{
\item{MaxResults}{The maximum number of elements to be returned for this resource.}

\item{NextToken}{The next page of elements from this collection. Not valid for the last
element of the collection.}
}
\value{
A list with the following syntax:\preformatted{list(
  Items = list(
    list(
      CreatedDate = as.POSIXct(
        "2015-01-01"
      ),
      Name = "string",
      SecurityGroupIds = list(
        "string"
      ),
      SubnetIds = list(
        "string"
      ),
      Tags = list(
        "string"
      ),
      VpcLinkId = "string",
      VpcLinkStatus = "PENDING"|"AVAILABLE"|"DELETING"|"FAILED"|"INACTIVE",
      VpcLinkStatusMessage = "string",
      VpcLinkVersion = "V2"
    )
  ),
  NextToken = "string"
)
}
}
\description{
Gets a collection of VPC links.
}
\section{Request syntax}{
\preformatted{svc$get_vpc_links(
  MaxResults = "string",
  NextToken = "string"
)
}
}

\keyword{internal}
