% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_import_api}
\alias{apigatewayv2_import_api}
\title{Imports an API}
\usage{
apigatewayv2_import_api(Basepath, Body, FailOnWarnings)
}
\arguments{
\item{Basepath}{Specifies how to interpret the base path of the API during import. Valid
values are ignore, prepend, and split. The default value is ignore. To
learn more, see \href{https://docs.aws.amazon.com/apigateway/latest/developerguide/api-gateway-import-api-basePath.html}{Set the OpenAPI basePath Property}.
Supported only for HTTP APIs.}

\item{Body}{[required] The OpenAPI definition. Supported only for HTTP APIs.}

\item{FailOnWarnings}{Specifies whether to rollback the API creation when a warning is
encountered. By default, API creation continues if a warning is
encountered.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ApiEndpoint = "string",
  ApiGatewayManaged = TRUE|FALSE,
  ApiId = "string",
  ApiKeySelectionExpression = "string",
  CorsConfiguration = list(
    AllowCredentials = TRUE|FALSE,
    AllowHeaders = list(
      "string"
    ),
    AllowMethods = list(
      "string"
    ),
    AllowOrigins = list(
      "string"
    ),
    ExposeHeaders = list(
      "string"
    ),
    MaxAge = 123
  ),
  CreatedDate = as.POSIXct(
    "2015-01-01"
  ),
  Description = "string",
  DisableSchemaValidation = TRUE|FALSE,
  DisableExecuteApiEndpoint = TRUE|FALSE,
  ImportInfo = list(
    "string"
  ),
  Name = "string",
  ProtocolType = "WEBSOCKET"|"HTTP",
  RouteSelectionExpression = "string",
  Tags = list(
    "string"
  ),
  Version = "string",
  Warnings = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Imports an API.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$import_api(
  Basepath = "string",
  Body = "string",
  FailOnWarnings = TRUE|FALSE
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
