% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_list_portals}
\alias{apigatewayv2_list_portals}
\title{Lists portals}
\usage{
apigatewayv2_list_portals(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{The maximum number of elements to be returned for this resource.}

\item{NextToken}{The next page of elements from this collection. Not valid for the last
element of the collection.}
}
\description{
Lists portals.

See \url{https://www.paws-r-sdk.com/docs/apigatewayv2_list_portals/} for full documentation.
}
\keyword{internal}
