% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_list_product_rest_endpoint_pages}
\alias{apigatewayv2_list_product_rest_endpoint_pages}
\title{Lists the product REST endpoint pages of a portal product}
\usage{
apigatewayv2_list_product_rest_endpoint_pages(
  MaxResults = NULL,
  NextToken = NULL,
  PortalProductId,
  ResourceOwnerAccountId = NULL
)
}
\arguments{
\item{MaxResults}{The maximum number of elements to be returned for this resource.}

\item{NextToken}{The next page of elements from this collection. Not valid for the last
element of the collection.}

\item{PortalProductId}{[required] The portal product identifier.}

\item{ResourceOwnerAccountId}{The account ID of the resource owner of the portal product.}
}
\description{
Lists the product REST endpoint pages of a portal product.

See \url{https://www.paws-r-sdk.com/docs/apigatewayv2_list_product_rest_endpoint_pages/} for full documentation.
}
\keyword{internal}
