% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_list_routing_rules}
\alias{apigatewayv2_list_routing_rules}
\title{Lists routing rules}
\usage{
apigatewayv2_list_routing_rules(
  DomainName,
  DomainNameId = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{DomainName}{[required] The domain name.}

\item{DomainNameId}{The domain name ID.}

\item{MaxResults}{The maximum number of elements to be returned for this resource.}

\item{NextToken}{The next page of elements from this collection. Not valid for the last
element of the collection.}
}
\description{
Lists routing rules.

See \url{https://www.paws-r-sdk.com/docs/apigatewayv2_list_routing_rules/} for full documentation.
}
\keyword{internal}
