% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_preview_portal}
\alias{apigatewayv2_preview_portal}
\title{Creates a portal preview}
\usage{
apigatewayv2_preview_portal(PortalId)
}
\arguments{
\item{PortalId}{[required] The portal identifier.}
}
\description{
Creates a portal preview.

See \url{https://www.paws-r-sdk.com/docs/apigatewayv2_preview_portal/} for full documentation.
}
\keyword{internal}
