% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_publish_portal}
\alias{apigatewayv2_publish_portal}
\title{Publishes a portal}
\usage{
apigatewayv2_publish_portal(Description = NULL, PortalId)
}
\arguments{
\item{Description}{The description of the portal. When the portal is published, this
description becomes the last published description.}

\item{PortalId}{[required] The portal identifier.}
}
\description{
Publishes a portal.

See \url{https://www.paws-r-sdk.com/docs/apigatewayv2_publish_portal/} for full documentation.
}
\keyword{internal}
