% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_put_routing_rule}
\alias{apigatewayv2_put_routing_rule}
\title{Updates a routing rule}
\usage{
apigatewayv2_put_routing_rule(
  Actions,
  Conditions,
  DomainName,
  DomainNameId = NULL,
  Priority,
  RoutingRuleId
)
}
\arguments{
\item{Actions}{[required] The routing rule action.}

\item{Conditions}{[required] The routing rule condition.}

\item{DomainName}{[required] The domain name.}

\item{DomainNameId}{The domain name ID.}

\item{Priority}{[required] The routing rule priority.}

\item{RoutingRuleId}{[required] The routing rule ID.}
}
\description{
Updates a routing rule.

See \url{https://www.paws-r-sdk.com/docs/apigatewayv2_put_routing_rule/} for full documentation.
}
\keyword{internal}
