% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_reimport_api}
\alias{apigatewayv2_reimport_api}
\title{Puts an Api resource}
\usage{
apigatewayv2_reimport_api(ApiId, Basepath, Body, FailOnWarnings)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{Basepath}{Represents the base path of the imported API. Supported only for HTTP
APIs.}

\item{Body}{[required] The OpenAPI definition. Supported only for HTTP APIs.}

\item{FailOnWarnings}{Specifies whether to rollback the API creation (true) or not (false)
when a warning is encountered. The default value is false.}
}
\description{
Puts an Api resource.
}
\section{Request syntax}{
\preformatted{svc$reimport_api(
  ApiId = "string",
  Basepath = "string",
  Body = "string",
  FailOnWarnings = TRUE|FALSE
)
}
}

\keyword{internal}
