% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_reset_authorizers_cache}
\alias{apigatewayv2_reset_authorizers_cache}
\title{Resets all authorizer cache entries on a stage}
\usage{
apigatewayv2_reset_authorizers_cache(ApiId, StageName)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{StageName}{[required] The stage name. Stage names can contain only alphanumeric characters,
hyphens, and underscores, or be $default. Maximum length is 128
characters.}
}
\value{
An empty list.
}
\description{
Resets all authorizer cache entries on a stage. Supported only for HTTP
APIs.
}
\section{Request syntax}{
\preformatted{svc$reset_authorizers_cache(
  ApiId = "string",
  StageName = "string"
)
}
}

\keyword{internal}
