% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_untag_resource}
\alias{apigatewayv2_untag_resource}
\title{Deletes a Tag}
\usage{
apigatewayv2_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The resource ARN for the tag.}

\item{TagKeys}{[required] The Tag keys to delete}
}
\description{
Deletes a Tag.

See \url{https://www.paws-r-sdk.com/docs/apigatewayv2_untag_resource/} for full documentation.
}
\keyword{internal}
