% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_update_deployment}
\alias{apigatewayv2_update_deployment}
\title{Updates a Deployment}
\usage{
apigatewayv2_update_deployment(ApiId, DeploymentId, Description)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{DeploymentId}{[required] The deployment ID.}

\item{Description}{The description for the deployment resource.}
}
\value{
A list with the following syntax:\preformatted{list(
  AutoDeployed = TRUE|FALSE,
  CreatedDate = as.POSIXct(
    "2015-01-01"
  ),
  DeploymentId = "string",
  DeploymentStatus = "PENDING"|"FAILED"|"DEPLOYED",
  DeploymentStatusMessage = "string",
  Description = "string"
)
}
}
\description{
Updates a Deployment.
}
\section{Request syntax}{
\preformatted{svc$update_deployment(
  ApiId = "string",
  DeploymentId = "string",
  Description = "string"
)
}
}

\keyword{internal}
