% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_update_model}
\alias{apigatewayv2_update_model}
\title{Updates a Model}
\usage{
apigatewayv2_update_model(
  ApiId,
  ContentType = NULL,
  Description = NULL,
  ModelId,
  Name = NULL,
  Schema = NULL
)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{ContentType}{The content-type for the model, for example, "application/json".}

\item{Description}{The description of the model.}

\item{ModelId}{[required] The model ID.}

\item{Name}{The name of the model.}

\item{Schema}{The schema for the model. For application/json models, this should be
JSON schema draft 4 model.}
}
\description{
Updates a Model.

See \url{https://paws-r.github.io/docs/apigatewayv2/update_model.html} for full documentation.
}
\keyword{internal}
