% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_update_portal}
\alias{apigatewayv2_update_portal}
\title{Updates a portal}
\usage{
apigatewayv2_update_portal(
  Authorization = NULL,
  EndpointConfiguration = NULL,
  IncludedPortalProductArns = NULL,
  LogoUri = NULL,
  PortalContent = NULL,
  PortalId,
  RumAppMonitorName = NULL
)
}
\arguments{
\item{Authorization}{The authorization of the portal.}

\item{EndpointConfiguration}{Represents an endpoint configuration.}

\item{IncludedPortalProductArns}{The ARNs of the portal products included in the portal.}

\item{LogoUri}{The logo URI.}

\item{PortalContent}{Contains the content that is visible to portal consumers including the
themes, display names, and description.}

\item{PortalId}{[required] The portal identifier.}

\item{RumAppMonitorName}{The CloudWatch RUM app monitor name.}
}
\description{
Updates a portal.

See \url{https://www.paws-r-sdk.com/docs/apigatewayv2_update_portal/} for full documentation.
}
\keyword{internal}
