% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_update_portal_product}
\alias{apigatewayv2_update_portal_product}
\title{Updates the portal product}
\usage{
apigatewayv2_update_portal_product(
  Description = NULL,
  DisplayName = NULL,
  DisplayOrder = NULL,
  PortalProductId
)
}
\arguments{
\item{Description}{The description.}

\item{DisplayName}{The displayName.}

\item{DisplayOrder}{The display order.}

\item{PortalProductId}{[required] The portal product identifier.}
}
\description{
Updates the portal product.

See \url{https://www.paws-r-sdk.com/docs/apigatewayv2_update_portal_product/} for full documentation.
}
\keyword{internal}
