% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_update_product_page}
\alias{apigatewayv2_update_product_page}
\title{Updates a product page of a portal product}
\usage{
apigatewayv2_update_product_page(
  DisplayContent = NULL,
  PortalProductId,
  ProductPageId
)
}
\arguments{
\item{DisplayContent}{The content of the product page.}

\item{PortalProductId}{[required] The portal product identifier.}

\item{ProductPageId}{[required] The portal product identifier.}
}
\description{
Updates a product page of a portal product.

See \url{https://www.paws-r-sdk.com/docs/apigatewayv2_update_product_page/} for full documentation.
}
\keyword{internal}
