% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_update_product_rest_endpoint_page}
\alias{apigatewayv2_update_product_rest_endpoint_page}
\title{Updates a product REST endpoint page}
\usage{
apigatewayv2_update_product_rest_endpoint_page(
  DisplayContent = NULL,
  PortalProductId,
  ProductRestEndpointPageId,
  TryItState = NULL
)
}
\arguments{
\item{DisplayContent}{The display content.}

\item{PortalProductId}{[required] The portal product identifier.}

\item{ProductRestEndpointPageId}{[required] The product REST endpoint identifier.}

\item{TryItState}{The try it state of a product REST endpoint page.}
}
\description{
Updates a product REST endpoint page.

See \url{https://www.paws-r-sdk.com/docs/apigatewayv2_update_product_rest_endpoint_page/} for full documentation.
}
\keyword{internal}
