% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_update_route_response}
\alias{apigatewayv2_update_route_response}
\title{Updates a RouteResponse}
\usage{
apigatewayv2_update_route_response(
  ApiId,
  ModelSelectionExpression = NULL,
  ResponseModels = NULL,
  ResponseParameters = NULL,
  RouteId,
  RouteResponseId,
  RouteResponseKey = NULL
)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{ModelSelectionExpression}{The model selection expression for the route response. Supported only
for WebSocket APIs.}

\item{ResponseModels}{The response models for the route response.}

\item{ResponseParameters}{The route response parameters.}

\item{RouteId}{[required] The route ID.}

\item{RouteResponseId}{[required] The route response ID.}

\item{RouteResponseKey}{The route response key.}
}
\description{
Updates a RouteResponse.

See \url{https://www.paws-r-sdk.com/docs/apigatewayv2_update_route_response/} for full documentation.
}
\keyword{internal}
