% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_update_vpc_link}
\alias{apigatewayv2_update_vpc_link}
\title{Updates a VPC link}
\usage{
apigatewayv2_update_vpc_link(Name = NULL, VpcLinkId)
}
\arguments{
\item{Name}{The name of the VPC link.}

\item{VpcLinkId}{[required] The ID of the VPC link.}
}
\description{
Updates a VPC link.

See \url{https://paws-r.github.io/docs/apigatewayv2/update_vpc_link.html} for full documentation.
}
\keyword{internal}
