% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appfabric_operations.R
\name{appfabric_batch_get_user_access_tasks}
\alias{appfabric_batch_get_user_access_tasks}
\title{Gets user access details in a batch request}
\usage{
appfabric_batch_get_user_access_tasks(appBundleIdentifier, taskIdList)
}
\arguments{
\item{appBundleIdentifier}{[required] The Amazon Resource Name (ARN) or Universal Unique Identifier (UUID) of
the app bundle to use for the request.}

\item{taskIdList}{[required] The tasks IDs to use for the request.}
}
\description{
Gets user access details in a batch request.

See \url{https://www.paws-r-sdk.com/docs/appfabric_batch_get_user_access_tasks/} for full documentation.
}
\keyword{internal}
