% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appfabric_operations.R
\name{appfabric_delete_app_authorization}
\alias{appfabric_delete_app_authorization}
\title{Deletes an app authorization}
\usage{
appfabric_delete_app_authorization(
  appBundleIdentifier,
  appAuthorizationIdentifier
)
}
\arguments{
\item{appBundleIdentifier}{[required] The Amazon Resource Name (ARN) or Universal Unique Identifier (UUID) of
the app bundle to use for the request.}

\item{appAuthorizationIdentifier}{[required] The Amazon Resource Name (ARN) or Universal Unique Identifier (UUID) of
the app authorization to use for the request.}
}
\description{
Deletes an app authorization. You must delete the associated ingestion before you can delete an app authorization.

See \url{https://www.paws-r-sdk.com/docs/appfabric_delete_app_authorization/} for full documentation.
}
\keyword{internal}
