% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appfabric_operations.R
\name{appfabric_delete_app_bundle}
\alias{appfabric_delete_app_bundle}
\title{Deletes an app bundle}
\usage{
appfabric_delete_app_bundle(appBundleIdentifier)
}
\arguments{
\item{appBundleIdentifier}{[required] The ID or Amazon Resource Name (ARN) of the app bundle that needs to be
deleted.}
}
\description{
Deletes an app bundle. You must delete all associated app authorizations before you can delete an app bundle.

See \url{https://www.paws-r-sdk.com/docs/appfabric_delete_app_bundle/} for full documentation.
}
\keyword{internal}
