% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appfabric_operations.R
\name{appfabric_delete_ingestion}
\alias{appfabric_delete_ingestion}
\title{Deletes an ingestion}
\usage{
appfabric_delete_ingestion(appBundleIdentifier, ingestionIdentifier)
}
\arguments{
\item{appBundleIdentifier}{[required] The Amazon Resource Name (ARN) or Universal Unique Identifier (UUID) of
the app bundle to use for the request.}

\item{ingestionIdentifier}{[required] The Amazon Resource Name (ARN) or Universal Unique Identifier (UUID) of
the ingestion to use for the request.}
}
\description{
Deletes an ingestion. You must stop (disable) the ingestion and you must delete all associated ingestion destinations before you can delete an app ingestion.

See \url{https://www.paws-r-sdk.com/docs/appfabric_delete_ingestion/} for full documentation.
}
\keyword{internal}
