% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appfabric_operations.R
\name{appfabric_get_app_bundle}
\alias{appfabric_get_app_bundle}
\title{Returns information about an app bundle}
\usage{
appfabric_get_app_bundle(appBundleIdentifier)
}
\arguments{
\item{appBundleIdentifier}{[required] The Amazon Resource Name (ARN) or Universal Unique Identifier (UUID) of
the app bundle to use for the request.}
}
\description{
Returns information about an app bundle.

See \url{https://www.paws-r-sdk.com/docs/appfabric_get_app_bundle/} for full documentation.
}
\keyword{internal}
