% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appfabric_operations.R
\name{appfabric_get_ingestion_destination}
\alias{appfabric_get_ingestion_destination}
\title{Returns information about an ingestion destination}
\usage{
appfabric_get_ingestion_destination(
  appBundleIdentifier,
  ingestionIdentifier,
  ingestionDestinationIdentifier
)
}
\arguments{
\item{appBundleIdentifier}{[required] The Amazon Resource Name (ARN) or Universal Unique Identifier (UUID) of
the app bundle to use for the request.}

\item{ingestionIdentifier}{[required] The Amazon Resource Name (ARN) or Universal Unique Identifier (UUID) of
the ingestion to use for the request.}

\item{ingestionDestinationIdentifier}{[required] The Amazon Resource Name (ARN) or Universal Unique Identifier (UUID) of
the ingestion destination to use for the request.}
}
\description{
Returns information about an ingestion destination.

See \url{https://www.paws-r-sdk.com/docs/appfabric_get_ingestion_destination/} for full documentation.
}
\keyword{internal}
