% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appfabric_operations.R
\name{appfabric_list_tags_for_resource}
\alias{appfabric_list_tags_for_resource}
\title{Returns a list of tags for a resource}
\usage{
appfabric_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource for which you want to
retrieve tags.}
}
\description{
Returns a list of tags for a resource.

See \url{https://www.paws-r-sdk.com/docs/appfabric_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
