% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appfabric_operations.R
\name{appfabric_start_ingestion}
\alias{appfabric_start_ingestion}
\title{Starts (enables) an ingestion, which collects data from an application}
\usage{
appfabric_start_ingestion(ingestionIdentifier, appBundleIdentifier)
}
\arguments{
\item{ingestionIdentifier}{[required] The Amazon Resource Name (ARN) or Universal Unique Identifier (UUID) of
the ingestion to use for the request.}

\item{appBundleIdentifier}{[required] The Amazon Resource Name (ARN) or Universal Unique Identifier (UUID) of
the app bundle to use for the request.}
}
\description{
Starts (enables) an ingestion, which collects data from an application.

See \url{https://www.paws-r-sdk.com/docs/appfabric_start_ingestion/} for full documentation.
}
\keyword{internal}
