% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appfabric_operations.R
\name{appfabric_start_user_access_tasks}
\alias{appfabric_start_user_access_tasks}
\title{Starts the tasks to search user access status for a specific email
address}
\usage{
appfabric_start_user_access_tasks(appBundleIdentifier, email)
}
\arguments{
\item{appBundleIdentifier}{[required] The Amazon Resource Name (ARN) or Universal Unique Identifier (UUID) of
the app bundle to use for the request.}

\item{email}{[required] The email address of the target user.}
}
\description{
Starts the tasks to search user access status for a specific email address.

See \url{https://www.paws-r-sdk.com/docs/appfabric_start_user_access_tasks/} for full documentation.
}
\keyword{internal}
