% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appfabric_operations.R
\name{appfabric_stop_ingestion}
\alias{appfabric_stop_ingestion}
\title{Stops (disables) an ingestion}
\usage{
appfabric_stop_ingestion(ingestionIdentifier, appBundleIdentifier)
}
\arguments{
\item{ingestionIdentifier}{[required] The Amazon Resource Name (ARN) or Universal Unique Identifier (UUID) of
the ingestion to use for the request.}

\item{appBundleIdentifier}{[required] The Amazon Resource Name (ARN) or Universal Unique Identifier (UUID) of
the app bundle to use for the request.}
}
\description{
Stops (disables) an ingestion.

See \url{https://www.paws-r-sdk.com/docs/appfabric_stop_ingestion/} for full documentation.
}
\keyword{internal}
