% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appfabric_operations.R
\name{appfabric_tag_resource}
\alias{appfabric_tag_resource}
\title{Assigns one or more tags (key-value pairs) to the specified resource}
\usage{
appfabric_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource that you want to tag.}

\item{tags}{[required] A map of the key-value pairs of the tag or tags to assign to the
resource.}
}
\description{
Assigns one or more tags (key-value pairs) to the specified resource.

See \url{https://www.paws-r-sdk.com/docs/appfabric_tag_resource/} for full documentation.
}
\keyword{internal}
