% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appfabric_operations.R
\name{appfabric_untag_resource}
\alias{appfabric_untag_resource}
\title{Removes a tag or tags from a resource}
\usage{
appfabric_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource that you want to untag.}

\item{tagKeys}{[required] The keys of the key-value pairs for the tag or tags you want to remove
from the specified resource.}
}
\description{
Removes a tag or tags from a resource.

See \url{https://www.paws-r-sdk.com/docs/appfabric_untag_resource/} for full documentation.
}
\keyword{internal}
