% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appfabric_operations.R
\name{appfabric_update_ingestion_destination}
\alias{appfabric_update_ingestion_destination}
\title{Updates an ingestion destination, which specifies how an application's
ingested data is processed by Amazon Web Services AppFabric and where
it's delivered}
\usage{
appfabric_update_ingestion_destination(
  appBundleIdentifier,
  ingestionIdentifier,
  ingestionDestinationIdentifier,
  destinationConfiguration
)
}
\arguments{
\item{appBundleIdentifier}{[required] The Amazon Resource Name (ARN) or Universal Unique Identifier (UUID) of
the app bundle to use for the request.}

\item{ingestionIdentifier}{[required] The Amazon Resource Name (ARN) or Universal Unique Identifier (UUID) of
the ingestion to use for the request.}

\item{ingestionDestinationIdentifier}{[required] The Amazon Resource Name (ARN) or Universal Unique Identifier (UUID) of
the ingestion destination to use for the request.}

\item{destinationConfiguration}{[required] Contains information about the destination of ingested data.}
}
\description{
Updates an ingestion destination, which specifies how an application's ingested data is processed by Amazon Web Services AppFabric and where it's delivered.

See \url{https://www.paws-r-sdk.com/docs/appfabric_update_ingestion_destination/} for full documentation.
}
\keyword{internal}
