% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appmesh_operations.R
\name{appmesh_create_mesh}
\alias{appmesh_create_mesh}
\title{Creates a service mesh}
\usage{
appmesh_create_mesh(clientToken = NULL, meshName, spec = NULL, tags = NULL)
}
\arguments{
\item{clientToken}{Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. Up to 36 letters, numbers, hyphens, and
underscores are allowed.}

\item{meshName}{[required] The name to use for the service mesh.}

\item{spec}{The service mesh specification to apply.}

\item{tags}{Optional metadata that you can apply to the service mesh to assist with
categorization and organization. Each tag consists of a key and an
optional value, both of which you define. Tag keys can have a maximum
character length of 128 characters, and tag values can have a maximum
length of 256 characters.}
}
\description{
Creates a service mesh.

See \url{https://www.paws-r-sdk.com/docs/appmesh_create_mesh/} for full documentation.
}
\keyword{internal}
