% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appmesh_operations.R
\name{appmesh_delete_virtual_node}
\alias{appmesh_delete_virtual_node}
\title{Deletes an existing virtual node}
\usage{
appmesh_delete_virtual_node(meshName, meshOwner = NULL, virtualNodeName)
}
\arguments{
\item{meshName}{[required] The name of the service mesh to delete the virtual node in.}

\item{meshOwner}{The Amazon Web Services IAM account ID of the service mesh owner. If the
account ID is not your own, then it's the ID of the account that shared
the mesh with your account. For more information about mesh sharing, see
\href{https://docs.aws.amazon.com/app-mesh/latest/userguide/sharing.html}{Working with shared meshes}.}

\item{virtualNodeName}{[required] The name of the virtual node to delete.}
}
\description{
Deletes an existing virtual node.

See \url{https://paws-r.github.io/docs/appmesh/delete_virtual_node.html} for full documentation.
}
\keyword{internal}
