% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appmesh_operations.R
\name{appmesh_delete_virtual_router}
\alias{appmesh_delete_virtual_router}
\title{Deletes an existing virtual router}
\usage{
appmesh_delete_virtual_router(meshName, meshOwner, virtualRouterName)
}
\arguments{
\item{meshName}{[required] The name of the service mesh to delete the virtual router in.}

\item{meshOwner}{The AWS IAM account ID of the service mesh owner. If the account ID is
not your own, then it's the ID of the account that shared the mesh with
your account. For more information about mesh sharing, see \href{https://docs.aws.amazon.com/app-mesh/latest/userguide/sharing.html}{Working with shared meshes}.}

\item{virtualRouterName}{[required] The name of the virtual router to delete.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  virtualRouter = list(
    meshName = "string",
    metadata = list(
      arn = "string",
      createdAt = as.POSIXct(
        "2015-01-01"
      ),
      lastUpdatedAt = as.POSIXct(
        "2015-01-01"
      ),
      meshOwner = "string",
      resourceOwner = "string",
      uid = "string",
      version = 123
    ),
    spec = list(
      listeners = list(
        list(
          portMapping = list(
            port = 123,
            protocol = "http"|"tcp"|"http2"|"grpc"
          )
        )
      )
    ),
    status = list(
      status = "ACTIVE"|"INACTIVE"|"DELETED"
    ),
    virtualRouterName = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Deletes an existing virtual router.

You must delete any routes associated with the virtual router before you
can delete the router itself.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_virtual_router(
  meshName = "string",
  meshOwner = "string",
  virtualRouterName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
