% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appmesh_operations.R
\name{appmesh_describe_virtual_service}
\alias{appmesh_describe_virtual_service}
\title{Describes an existing virtual service}
\usage{
appmesh_describe_virtual_service(
  meshName,
  meshOwner = NULL,
  virtualServiceName
)
}
\arguments{
\item{meshName}{[required] The name of the service mesh that the virtual service resides in.}

\item{meshOwner}{The Amazon Web Services IAM account ID of the service mesh owner. If the
account ID is not your own, then it's the ID of the account that shared
the mesh with your account. For more information about mesh sharing, see
\href{https://docs.aws.amazon.com/app-mesh/latest/userguide/sharing.html}{Working with shared meshes}.}

\item{virtualServiceName}{[required] The name of the virtual service to describe.}
}
\description{
Describes an existing virtual service.

See \url{https://paws-r.github.io/docs/appmesh/describe_virtual_service.html} for full documentation.
}
\keyword{internal}
