% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appmesh_operations.R
\name{appmesh_list_meshes}
\alias{appmesh_list_meshes}
\title{Returns a list of existing service meshes}
\usage{
appmesh_list_meshes(limit, nextToken)
}
\arguments{
\item{limit}{The maximum number of results returned by
\code{\link[=appmesh_list_meshes]{list_meshes}} in paginated output. When you use
this parameter, \code{\link[=appmesh_list_meshes]{list_meshes}} returns only
\code{limit} results in a single page along with a \code{nextToken} response
element. You can see the remaining results of the initial request by
sending another \code{\link[=appmesh_list_meshes]{list_meshes}} request with the
returned \code{nextToken} value. This value can be between 1 and 100. If you
don't use this parameter, \code{\link[=appmesh_list_meshes]{list_meshes}} returns
up to 100 results and a \code{nextToken} value if applicable.}

\item{nextToken}{The \code{nextToken} value returned from a previous paginated
\code{\link[=appmesh_list_meshes]{list_meshes}} request where \code{limit} was used and
the results exceeded the value of that parameter. Pagination continues
from the end of the previous results that returned the \code{nextToken}
value.

This token should be treated as an opaque identifier that is used only
to retrieve the next items in a list and not for other programmatic
purposes.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  meshes = list(
    list(
      arn = "string",
      createdAt = as.POSIXct(
        "2015-01-01"
      ),
      lastUpdatedAt = as.POSIXct(
        "2015-01-01"
      ),
      meshName = "string",
      meshOwner = "string",
      resourceOwner = "string",
      version = 123
    )
  ),
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Returns a list of existing service meshes.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_meshes(
  limit = 123,
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
