% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appmesh_operations.R
\name{appmesh_list_virtual_gateways}
\alias{appmesh_list_virtual_gateways}
\title{Returns a list of existing virtual gateways in a service mesh}
\usage{
appmesh_list_virtual_gateways(limit, meshName, meshOwner, nextToken)
}
\arguments{
\item{limit}{The maximum number of results returned by
\code{\link[=appmesh_list_virtual_gateways]{list_virtual_gateways}} in paginated
output. When you use this parameter,
\code{\link[=appmesh_list_virtual_gateways]{list_virtual_gateways}} returns only
\code{limit} results in a single page along with a \code{nextToken} response
element. You can see the remaining results of the initial request by
sending another \code{\link[=appmesh_list_virtual_gateways]{list_virtual_gateways}}
request with the returned \code{nextToken} value. This value can be between 1
and 100. If you don't use this parameter,
\code{\link[=appmesh_list_virtual_gateways]{list_virtual_gateways}} returns up to
100 results and a \code{nextToken} value if applicable.}

\item{meshName}{[required] The name of the service mesh to list virtual gateways in.}

\item{meshOwner}{The AWS IAM account ID of the service mesh owner. If the account ID is
not your own, then it's the ID of the account that shared the mesh with
your account. For more information about mesh sharing, see \href{https://docs.aws.amazon.com/app-mesh/latest/userguide/sharing.html}{Working with shared meshes}.}

\item{nextToken}{The \code{nextToken} value returned from a previous paginated
\code{\link[=appmesh_list_virtual_gateways]{list_virtual_gateways}} request where
\code{limit} was used and the results exceeded the value of that parameter.
Pagination continues from the end of the previous results that returned
the \code{nextToken} value.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  nextToken = "string",
  virtualGateways = list(
    list(
      arn = "string",
      createdAt = as.POSIXct(
        "2015-01-01"
      ),
      lastUpdatedAt = as.POSIXct(
        "2015-01-01"
      ),
      meshName = "string",
      meshOwner = "string",
      resourceOwner = "string",
      version = 123,
      virtualGatewayName = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Returns a list of existing virtual gateways in a service mesh.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_virtual_gateways(
  limit = 123,
  meshName = "string",
  meshOwner = "string",
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
