% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appmesh_operations.R
\name{appmesh_list_virtual_routers}
\alias{appmesh_list_virtual_routers}
\title{Returns a list of existing virtual routers in a service mesh}
\usage{
appmesh_list_virtual_routers(
  limit = NULL,
  meshName,
  meshOwner = NULL,
  nextToken = NULL
)
}
\arguments{
\item{limit}{The maximum number of results returned by
\code{\link[=appmesh_list_virtual_routers]{list_virtual_routers}} in paginated
output. When you use this parameter,
\code{\link[=appmesh_list_virtual_routers]{list_virtual_routers}} returns only
\code{limit} results in a single page along with a \code{nextToken} response
element. You can see the remaining results of the initial request by
sending another \code{\link[=appmesh_list_virtual_routers]{list_virtual_routers}}
request with the returned \code{nextToken} value. This value can be between 1
and 100. If you don't use this parameter,
\code{\link[=appmesh_list_virtual_routers]{list_virtual_routers}} returns up to 100
results and a \code{nextToken} value if applicable.}

\item{meshName}{[required] The name of the service mesh to list virtual routers in.}

\item{meshOwner}{The Amazon Web Services IAM account ID of the service mesh owner. If the
account ID is not your own, then it's the ID of the account that shared
the mesh with your account. For more information about mesh sharing, see
\href{https://docs.aws.amazon.com/app-mesh/latest/userguide/sharing.html}{Working with shared meshes}.}

\item{nextToken}{The \code{nextToken} value returned from a previous paginated
\code{\link[=appmesh_list_virtual_routers]{list_virtual_routers}} request where
\code{limit} was used and the results exceeded the value of that parameter.
Pagination continues from the end of the previous results that returned
the \code{nextToken} value.}
}
\description{
Returns a list of existing virtual routers in a service mesh.

See \url{https://www.paws-r-sdk.com/docs/appmesh_list_virtual_routers/} for full documentation.
}
\keyword{internal}
