% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appmesh_operations.R
\name{appmesh_tag_resource}
\alias{appmesh_tag_resource}
\title{Associates the specified tags to a resource with the specified
resourceArn}
\usage{
appmesh_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource to add tags to.}

\item{tags}{[required] The tags to add to the resource. A tag is an array of key-value pairs.
Tag keys can have a maximum character length of 128 characters, and tag
values can have a maximum length of 256 characters.}
}
\description{
Associates the specified tags to a resource with the specified
\code{resourceArn}. If existing tags on a resource aren\'t specified in the
request parameters, they aren\'t changed. When a resource is deleted,
the tags associated with that resource are also deleted.
}
\section{Request syntax}{
\preformatted{svc$tag_resource(
  resourceArn = "string",
  tags = list(
    list(
      key = "string",
      value = "string"
    )
  )
)
}
}

\keyword{internal}
