% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arczonalshift_operations.R
\name{arczonalshift_cancel_zonal_shift}
\alias{arczonalshift_cancel_zonal_shift}
\title{Cancel a zonal shift in Amazon Route 53 Application Recovery Controller
that you've started for a resource in your AWS account in an AWS Region}
\usage{
arczonalshift_cancel_zonal_shift(zonalShiftId)
}
\arguments{
\item{zonalShiftId}{[required] The internally-generated identifier of a zonal shift.}
}
\description{
Cancel a zonal shift in Amazon Route 53 Application Recovery Controller that you've started for a resource in your AWS account in an AWS Region.

See \url{https://www.paws-r-sdk.com/docs/arczonalshift_cancel_zonal_shift/} for full documentation.
}
\keyword{internal}
