% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arczonalshift_operations.R
\name{arczonalshift_delete_practice_run_configuration}
\alias{arczonalshift_delete_practice_run_configuration}
\title{Deletes the practice run configuration for a resource}
\usage{
arczonalshift_delete_practice_run_configuration(resourceIdentifier)
}
\arguments{
\item{resourceIdentifier}{[required] The identifier for the resource that you want to delete the practice run
configuration for. The identifier is the Amazon Resource Name (ARN) for
the resource.}
}
\description{
Deletes the practice run configuration for a resource. Before you can delete a practice run configuration for a resource., you must disable zonal autoshift for the resource. Practice runs must be configured for zonal autoshift to be enabled.

See \url{https://www.paws-r-sdk.com/docs/arczonalshift_delete_practice_run_configuration/} for full documentation.
}
\keyword{internal}
