% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arczonalshift_operations.R
\name{arczonalshift_list_autoshifts}
\alias{arczonalshift_list_autoshifts}
\title{Returns the active autoshifts for a specified resource}
\usage{
arczonalshift_list_autoshifts(
  maxResults = NULL,
  nextToken = NULL,
  status = NULL
)
}
\arguments{
\item{maxResults}{The number of objects that you want to return with this call.}

\item{nextToken}{Specifies that you want to receive the next page of results. Valid only
if you received a \code{NextToken} response in the previous request. If you
did, it indicates that more output is available. Set this parameter to
the value provided by the previous call's \code{NextToken} response to
request the next page of results.}

\item{status}{The status of the autoshift.}
}
\description{
Returns the active autoshifts for a specified resource.

See \url{https://www.paws-r-sdk.com/docs/arczonalshift_list_autoshifts/} for full documentation.
}
\keyword{internal}
