% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arczonalshift_operations.R
\name{arczonalshift_update_autoshift_observer_notification_status}
\alias{arczonalshift_update_autoshift_observer_notification_status}
\title{Update the status of autoshift observer notification}
\usage{
arczonalshift_update_autoshift_observer_notification_status(status)
}
\arguments{
\item{status}{[required] The status to set for autoshift observer notification. If the status is
\code{ENABLED}, Route 53 ARC includes all autoshift events when you use the
Amazon EventBridge pattern \verb{Autoshift In Progress}. When the status is
\code{DISABLED}, Route 53 ARC includes only autoshift events for autoshifts
when one or more of your resources is included in the autoshift.}
}
\description{
Update the status of autoshift observer notification. Autoshift observer notification enables you to be notified, through Amazon EventBridge, when there is an autoshift event for zonal autoshift.

See \url{https://www.paws-r-sdk.com/docs/arczonalshift_update_autoshift_observer_notification_status/} for full documentation.
}
\keyword{internal}
