% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backupgateway_operations.R
\name{backupgateway_associate_gateway_to_server}
\alias{backupgateway_associate_gateway_to_server}
\title{Associates a backup gateway with your server}
\usage{
backupgateway_associate_gateway_to_server(GatewayArn, ServerArn)
}
\arguments{
\item{GatewayArn}{[required] The Amazon Resource Name (ARN) of the gateway. Use the
\code{\link[=backupgateway_list_gateways]{list_gateways}} operation to return a
list of gateways for your account and Amazon Web Services Region.}

\item{ServerArn}{[required] The Amazon Resource Name (ARN) of the server that hosts your virtual
machines.}
}
\description{
Associates a backup gateway with your server. After you complete the association process, you can back up and restore your VMs through the gateway.

See \url{https://www.paws-r-sdk.com/docs/backupgateway_associate_gateway_to_server/} for full documentation.
}
\keyword{internal}
