% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backupgateway_operations.R
\name{backupgateway_create_gateway}
\alias{backupgateway_create_gateway}
\title{Creates a backup gateway}
\usage{
backupgateway_create_gateway(
  ActivationKey,
  GatewayDisplayName,
  GatewayType,
  Tags = NULL
)
}
\arguments{
\item{ActivationKey}{[required] The activation key of the created gateway.}

\item{GatewayDisplayName}{[required] The display name of the created gateway.}

\item{GatewayType}{[required] The type of created gateway.}

\item{Tags}{A list of up to 50 tags to assign to the gateway. Each tag is a
key-value pair.}
}
\description{
Creates a backup gateway. After you create a gateway, you can associate it with a server using the \code{\link[=backupgateway_associate_gateway_to_server]{associate_gateway_to_server}} operation.

See \url{https://paws-r.github.io/docs/backupgateway/create_gateway.html} for full documentation.
}
\keyword{internal}
