% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backupgateway_operations.R
\name{backupgateway_delete_gateway}
\alias{backupgateway_delete_gateway}
\title{Deletes a backup gateway}
\usage{
backupgateway_delete_gateway(GatewayArn)
}
\arguments{
\item{GatewayArn}{[required] The Amazon Resource Name (ARN) of the gateway to delete.}
}
\description{
Deletes a backup gateway.

See \url{https://paws-r.github.io/docs/backupgateway/delete_gateway.html} for full documentation.
}
\keyword{internal}
