% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backupgateway_operations.R
\name{backupgateway_delete_hypervisor}
\alias{backupgateway_delete_hypervisor}
\title{Deletes a hypervisor}
\usage{
backupgateway_delete_hypervisor(HypervisorArn)
}
\arguments{
\item{HypervisorArn}{[required] The Amazon Resource Name (ARN) of the hypervisor to delete.}
}
\description{
Deletes a hypervisor.

See \url{https://www.paws-r-sdk.com/docs/backupgateway_delete_hypervisor/} for full documentation.
}
\keyword{internal}
