% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backupgateway_operations.R
\name{backupgateway_get_gateway}
\alias{backupgateway_get_gateway}
\title{By providing the ARN (Amazon Resource Name), this API returns the
gateway}
\usage{
backupgateway_get_gateway(GatewayArn)
}
\arguments{
\item{GatewayArn}{[required] The Amazon Resource Name (ARN) of the gateway.}
}
\description{
By providing the ARN (Amazon Resource Name), this API returns the gateway.

See \url{https://www.paws-r-sdk.com/docs/backupgateway_get_gateway/} for full documentation.
}
\keyword{internal}
