% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backupgateway_operations.R
\name{backupgateway_get_hypervisor}
\alias{backupgateway_get_hypervisor}
\title{This action requests information about the specified hypervisor to which
the gateway will connect}
\usage{
backupgateway_get_hypervisor(HypervisorArn)
}
\arguments{
\item{HypervisorArn}{[required] The Amazon Resource Name (ARN) of the hypervisor.}
}
\description{
This action requests information about the specified hypervisor to which the gateway will connect. A hypervisor is hardware, software, or firmware that creates and manages virtual machines, and allocates resources to them.

See \url{https://paws-r.github.io/docs/backupgateway/get_hypervisor.html} for full documentation.
}
\keyword{internal}
