% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backupgateway_operations.R
\name{backupgateway_get_hypervisor_property_mappings}
\alias{backupgateway_get_hypervisor_property_mappings}
\title{This action retrieves the property mappings for the specified hypervisor}
\usage{
backupgateway_get_hypervisor_property_mappings(HypervisorArn)
}
\arguments{
\item{HypervisorArn}{[required] The Amazon Resource Name (ARN) of the hypervisor.}
}
\description{
This action retrieves the property mappings for the specified hypervisor. A hypervisor property mapping displays the relationship of entity properties available from the on-premises hypervisor to the properties available in Amazon Web Services.

See \url{https://paws-r.github.io/docs/backupgateway/get_hypervisor_property_mappings.html} for full documentation.
}
\keyword{internal}
