% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backupgateway_operations.R
\name{backupgateway_get_virtual_machine}
\alias{backupgateway_get_virtual_machine}
\title{By providing the ARN (Amazon Resource Name), this API returns the
virtual machine}
\usage{
backupgateway_get_virtual_machine(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the virtual machine.}
}
\description{
By providing the ARN (Amazon Resource Name), this API returns the virtual machine.

See \url{https://www.paws-r-sdk.com/docs/backupgateway_get_virtual_machine/} for full documentation.
}
\keyword{internal}
