% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backupgateway_operations.R
\name{backupgateway_import_hypervisor_configuration}
\alias{backupgateway_import_hypervisor_configuration}
\title{Connect to a hypervisor by importing its configuration}
\usage{
backupgateway_import_hypervisor_configuration(
  Name,
  Host,
  Username = NULL,
  Password = NULL,
  KmsKeyArn = NULL,
  Tags = NULL
)
}
\arguments{
\item{Name}{[required] The name of the hypervisor.}

\item{Host}{[required] The server host of the hypervisor. This can be either an IP address or a
fully-qualified domain name (FQDN).}

\item{Username}{The username for the hypervisor.}

\item{Password}{The password for the hypervisor.}

\item{KmsKeyArn}{The Key Management Service for the hypervisor.}

\item{Tags}{The tags of the hypervisor configuration to import.}
}
\description{
Connect to a hypervisor by importing its configuration.

See \url{https://www.paws-r-sdk.com/docs/backupgateway_import_hypervisor_configuration/} for full documentation.
}
\keyword{internal}
