% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backupgateway_operations.R
\name{backupgateway_list_gateways}
\alias{backupgateway_list_gateways}
\title{Lists backup gateways owned by an Amazon Web Services account in an
Amazon Web Services Region}
\usage{
backupgateway_list_gateways(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{The maximum number of gateways to list.}

\item{NextToken}{The next item following a partial list of returned resources. For
example, if a request is made to return \code{MaxResults} number of
resources, \code{NextToken} allows you to return more items in your list
starting at the location pointed to by the next token.}
}
\description{
Lists backup gateways owned by an Amazon Web Services account in an Amazon Web Services Region. The returned list is ordered by gateway Amazon Resource Name (ARN).

See \url{https://www.paws-r-sdk.com/docs/backupgateway_list_gateways/} for full documentation.
}
\keyword{internal}
