% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backupgateway_operations.R
\name{backupgateway_list_hypervisors}
\alias{backupgateway_list_hypervisors}
\title{Lists your hypervisors}
\usage{
backupgateway_list_hypervisors(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{The maximum number of hypervisors to list.}

\item{NextToken}{The next item following a partial list of returned resources. For
example, if a request is made to return \code{maxResults} number of
resources, \code{NextToken} allows you to return more items in your list
starting at the location pointed to by the next token.}
}
\description{
Lists your hypervisors.

See \url{https://paws-r.github.io/docs/backupgateway/list_hypervisors.html} for full documentation.
}
\keyword{internal}
