% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backupgateway_operations.R
\name{backupgateway_list_tags_for_resource}
\alias{backupgateway_list_tags_for_resource}
\title{Lists the tags applied to the resource identified by its Amazon Resource
Name (ARN)}
\usage{
backupgateway_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource's tags to list.}
}
\description{
Lists the tags applied to the resource identified by its Amazon Resource Name (ARN).

See \url{https://paws-r.github.io/docs/backupgateway/list_tags_for_resource.html} for full documentation.
}
\keyword{internal}
