% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backupgateway_operations.R
\name{backupgateway_start_virtual_machines_metadata_sync}
\alias{backupgateway_start_virtual_machines_metadata_sync}
\title{This action sends a request to sync metadata across the specified
virtual machines}
\usage{
backupgateway_start_virtual_machines_metadata_sync(HypervisorArn)
}
\arguments{
\item{HypervisorArn}{[required] The Amazon Resource Name (ARN) of the hypervisor.}
}
\description{
This action sends a request to sync metadata across the specified virtual machines.

See \url{https://paws-r.github.io/docs/backupgateway/start_virtual_machines_metadata_sync.html} for full documentation.
}
\keyword{internal}
