% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backupgateway_operations.R
\name{backupgateway_tag_resource}
\alias{backupgateway_tag_resource}
\title{Tag the resource}
\usage{
backupgateway_tag_resource(ResourceARN, Tags)
}
\arguments{
\item{ResourceARN}{[required] The Amazon Resource Name (ARN) of the resource to tag.}

\item{Tags}{[required] A list of tags to assign to the resource.}
}
\description{
Tag the resource.

See \url{https://www.paws-r-sdk.com/docs/backupgateway_tag_resource/} for full documentation.
}
\keyword{internal}
