% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backupgateway_operations.R
\name{backupgateway_test_hypervisor_configuration}
\alias{backupgateway_test_hypervisor_configuration}
\title{Tests your hypervisor configuration to validate that backup gateway can
connect with the hypervisor and its resources}
\usage{
backupgateway_test_hypervisor_configuration(
  GatewayArn,
  Host,
  Password = NULL,
  Username = NULL
)
}
\arguments{
\item{GatewayArn}{[required] The Amazon Resource Name (ARN) of the gateway to the hypervisor to test.}

\item{Host}{[required] The server host of the hypervisor. This can be either an IP address or a
fully-qualified domain name (FQDN).}

\item{Password}{The password for the hypervisor.}

\item{Username}{The username for the hypervisor.}
}
\description{
Tests your hypervisor configuration to validate that backup gateway can connect with the hypervisor and its resources.

See \url{https://www.paws-r-sdk.com/docs/backupgateway_test_hypervisor_configuration/} for full documentation.
}
\keyword{internal}
