% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backupgateway_operations.R
\name{backupgateway_untag_resource}
\alias{backupgateway_untag_resource}
\title{Removes tags from the resource}
\usage{
backupgateway_untag_resource(ResourceARN, TagKeys)
}
\arguments{
\item{ResourceARN}{[required] The Amazon Resource Name (ARN) of the resource from which to remove
tags.}

\item{TagKeys}{[required] The list of tag keys specifying which tags to remove.}
}
\description{
Removes tags from the resource.

See \url{https://paws-r.github.io/docs/backupgateway/untag_resource.html} for full documentation.
}
\keyword{internal}
