% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backupgateway_operations.R
\name{backupgateway_update_gateway_information}
\alias{backupgateway_update_gateway_information}
\title{Updates a gateway's name}
\usage{
backupgateway_update_gateway_information(GatewayArn, GatewayDisplayName = NULL)
}
\arguments{
\item{GatewayArn}{[required] The Amazon Resource Name (ARN) of the gateway to update.}

\item{GatewayDisplayName}{The updated display name of the gateway.}
}
\description{
Updates a gateway's name. Specify which gateway to update using the Amazon Resource Name (ARN) of the gateway in your request.

See \url{https://www.paws-r-sdk.com/docs/backupgateway_update_gateway_information/} for full documentation.
}
\keyword{internal}
