% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backupgateway_operations.R
\name{backupgateway_update_gateway_software_now}
\alias{backupgateway_update_gateway_software_now}
\title{Updates the gateway virtual machine (VM) software}
\usage{
backupgateway_update_gateway_software_now(GatewayArn)
}
\arguments{
\item{GatewayArn}{[required] The Amazon Resource Name (ARN) of the gateway to be updated.}
}
\description{
Updates the gateway virtual machine (VM) software. The request immediately triggers the software update.

See \url{https://paws-r.github.io/docs/backupgateway/update_gateway_software_now.html} for full documentation.
}
\keyword{internal}
