% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_service.R
\name{cloudfront}
\alias{cloudfront}
\title{Amazon CloudFront}
\usage{
cloudfront(
  config = list(),
  credentials = list(),
  endpoint = NULL,
  region = NULL
)
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.
\itemize{
\item{\strong{credentials}: \itemize{
\item{\strong{creds}: \itemize{
\item{\strong{access_key_id}: AWS access key ID}
\item{\strong{secret_access_key}: AWS secret access key}
\item{\strong{session_token}: AWS temporary session token}
}}
\item{\strong{profile}: The name of a profile to use. If not given, then the default profile is used.}
\item{\strong{anonymous}: Set anonymous credentials.}
}}
\item{\strong{endpoint}: The complete URL to use for the constructed client.}
\item{\strong{region}: The AWS Region used in instantiating the client.}
\item{\strong{close_connection}: Immediately close all HTTP connections.}
\item{\strong{timeout}: The time in seconds till a timeout exception is thrown when attempting to make a connection. The default is 60 seconds.}
\item{\strong{s3_force_path_style}: Set this to \code{true} to force the request to use path-style addressing, i.e. \verb{http://s3.amazonaws.com/BUCKET/KEY}.}
\item{\strong{sts_regional_endpoint}: Set sts regional endpoint resolver to regional or legacy \url{https://docs.aws.amazon.com/sdkref/latest/guide/feature-sts-regionalized-endpoints.html}}
\item{\strong{use_dual_stack}: Set this to \code{true} to use the dualstack (IPv4 and IPv6) endpoint for a service, where available, falling back to the regular endpoint when it isn't. Defaults to the \code{AWS_USE_DUALSTACK_ENDPOINT} environment variable when unset.}
}}

\item{credentials}{Optional credentials shorthand for the config parameter
\itemize{
\item{\strong{creds}: \itemize{
\item{\strong{access_key_id}: AWS access key ID}
\item{\strong{secret_access_key}: AWS secret access key}
\item{\strong{session_token}: AWS temporary session token}
}}
\item{\strong{profile}: The name of a profile to use. If not given, then the default profile is used.}
\item{\strong{anonymous}: Set anonymous credentials.}
}}

\item{endpoint}{Optional shorthand for complete URL to use for the constructed client.}

\item{region}{Optional shorthand for AWS Region used in instantiating the client.}
}
\value{
A client for the service. You can call the service's operations using
syntax like \code{svc$operation(...)}, where \code{svc} is the name you've assigned
to the client. The available operations are listed in the
Operations section.
}
\description{
This is the \emph{Amazon CloudFront API Reference}. This guide is for developers who need detailed information about CloudFront API actions, data types, and errors. For detailed information about CloudFront features, see the \href{https://docs.aws.amazon.com/AmazonCloudFront/latest/DeveloperGuide/Introduction.html}{Amazon CloudFront Developer Guide}.
}
\section{Service syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc <- cloudfront(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string",
      anonymous = "logical"
    ),
    endpoint = "string",
    region = "string",
    close_connection = "logical",
    timeout = "numeric",
    s3_force_path_style = "logical",
    sts_regional_endpoint = "string",
    use_dual_stack = "logical"
  ),
  credentials = list(
    creds = list(
      access_key_id = "string",
      secret_access_key = "string",
      session_token = "string"
    ),
    profile = "string",
    anonymous = "logical"
  ),
  endpoint = "string",
  region = "string"
)
}\if{html}{\out{</div>}}
}

\section{Operations}{

\tabular{ll}{
\link[=cloudfront_associate_alias]{associate_alias} \tab The AssociateAlias API operation only supports standard distributions\cr
\link[=cloudfront_associate_distribution_tenant_web_acl]{associate_distribution_tenant_web_acl} \tab Associates the WAF web ACL with a distribution tenant\cr
\link[=cloudfront_associate_distribution_web_acl]{associate_distribution_web_acl} \tab Associates the WAF web ACL with a distribution\cr
\link[=cloudfront_copy_distribution]{copy_distribution} \tab Creates a staging distribution using the configuration of the provided primary distribution\cr
\link[=cloudfront_create_anycast_ip_list]{create_anycast_ip_list} \tab Creates an Anycast static IP list\cr
\link[=cloudfront_create_cache_policy]{create_cache_policy} \tab Creates a cache policy\cr
\link[=cloudfront_create_cloud_front_origin_access_identity]{create_cloud_front_origin_access_identity} \tab Creates a new origin access identity\cr
\link[=cloudfront_create_connection_function]{create_connection_function} \tab Creates a connection function\cr
\link[=cloudfront_create_connection_group]{create_connection_group} \tab Creates a connection group\cr
\link[=cloudfront_create_continuous_deployment_policy]{create_continuous_deployment_policy} \tab Creates a continuous deployment policy that distributes traffic for a custom domain name to two different CloudFront distributions\cr
\link[=cloudfront_create_distribution]{create_distribution} \tab Creates a CloudFront distribution\cr
\link[=cloudfront_create_distribution_tenant]{create_distribution_tenant} \tab Creates a distribution tenant\cr
\link[=cloudfront_create_distribution_with_tags]{create_distribution_with_tags} \tab Create a new distribution with tags\cr
\link[=cloudfront_create_field_level_encryption_config]{create_field_level_encryption_config} \tab Create a new field-level encryption configuration\cr
\link[=cloudfront_create_field_level_encryption_profile]{create_field_level_encryption_profile} \tab Create a field-level encryption profile\cr
\link[=cloudfront_create_function]{create_function} \tab Creates a CloudFront function\cr
\link[=cloudfront_create_invalidation]{create_invalidation} \tab Create a new invalidation\cr
\link[=cloudfront_create_invalidation_for_distribution_tenant]{create_invalidation_for_distribution_tenant} \tab Creates an invalidation for a distribution tenant\cr
\link[=cloudfront_create_key_group]{create_key_group} \tab Creates a key group that you can use with CloudFront signed URLs and signed cookies\cr
\link[=cloudfront_create_key_value_store]{create_key_value_store} \tab Specifies the key value store resource to add to your account\cr
\link[=cloudfront_create_monitoring_subscription]{create_monitoring_subscription} \tab Enables or disables additional Amazon CloudWatch metrics for the specified CloudFront distribution\cr
\link[=cloudfront_create_origin_access_control]{create_origin_access_control} \tab Creates a new origin access control in CloudFront\cr
\link[=cloudfront_create_origin_request_policy]{create_origin_request_policy} \tab Creates an origin request policy\cr
\link[=cloudfront_create_public_key]{create_public_key} \tab Uploads a public key to CloudFront that you can use with signed URLs and signed cookies, or with field-level encryption\cr
\link[=cloudfront_create_realtime_log_config]{create_realtime_log_config} \tab Creates a real-time log configuration\cr
\link[=cloudfront_create_response_headers_policy]{create_response_headers_policy} \tab Creates a response headers policy\cr
\link[=cloudfront_create_streaming_distribution]{create_streaming_distribution} \tab This API is deprecated\cr
\link[=cloudfront_create_streaming_distribution_with_tags]{create_streaming_distribution_with_tags} \tab This API is deprecated\cr
\link[=cloudfront_create_trust_store]{create_trust_store} \tab Creates a trust store\cr
\link[=cloudfront_create_vpc_origin]{create_vpc_origin} \tab Create an Amazon CloudFront VPC origin\cr
\link[=cloudfront_delete_anycast_ip_list]{delete_anycast_ip_list} \tab Deletes an Anycast static IP list\cr
\link[=cloudfront_delete_cache_policy]{delete_cache_policy} \tab Deletes a cache policy\cr
\link[=cloudfront_delete_cloud_front_origin_access_identity]{delete_cloud_front_origin_access_identity} \tab Delete an origin access identity\cr
\link[=cloudfront_delete_connection_function]{delete_connection_function} \tab Deletes a connection function\cr
\link[=cloudfront_delete_connection_group]{delete_connection_group} \tab Deletes a connection group\cr
\link[=cloudfront_delete_continuous_deployment_policy]{delete_continuous_deployment_policy} \tab Deletes a continuous deployment policy\cr
\link[=cloudfront_delete_distribution]{delete_distribution} \tab Delete a distribution\cr
\link[=cloudfront_delete_distribution_tenant]{delete_distribution_tenant} \tab Deletes a distribution tenant\cr
\link[=cloudfront_delete_field_level_encryption_config]{delete_field_level_encryption_config} \tab Remove a field-level encryption configuration\cr
\link[=cloudfront_delete_field_level_encryption_profile]{delete_field_level_encryption_profile} \tab Remove a field-level encryption profile\cr
\link[=cloudfront_delete_function]{delete_function} \tab Deletes a CloudFront function\cr
\link[=cloudfront_delete_key_group]{delete_key_group} \tab Deletes a key group\cr
\link[=cloudfront_delete_key_value_store]{delete_key_value_store} \tab Specifies the key value store to delete\cr
\link[=cloudfront_delete_monitoring_subscription]{delete_monitoring_subscription} \tab Disables additional CloudWatch metrics for the specified CloudFront distribution\cr
\link[=cloudfront_delete_origin_access_control]{delete_origin_access_control} \tab Deletes a CloudFront origin access control\cr
\link[=cloudfront_delete_origin_request_policy]{delete_origin_request_policy} \tab Deletes an origin request policy\cr
\link[=cloudfront_delete_public_key]{delete_public_key} \tab Remove a public key you previously added to CloudFront\cr
\link[=cloudfront_delete_realtime_log_config]{delete_realtime_log_config} \tab Deletes a real-time log configuration\cr
\link[=cloudfront_delete_resource_policy]{delete_resource_policy} \tab Deletes the resource policy attached to the CloudFront resource\cr
\link[=cloudfront_delete_response_headers_policy]{delete_response_headers_policy} \tab Deletes a response headers policy\cr
\link[=cloudfront_delete_streaming_distribution]{delete_streaming_distribution} \tab Delete a streaming distribution\cr
\link[=cloudfront_delete_trust_store]{delete_trust_store} \tab Deletes a trust store\cr
\link[=cloudfront_delete_vpc_origin]{delete_vpc_origin} \tab Delete an Amazon CloudFront VPC origin\cr
\link[=cloudfront_describe_connection_function]{describe_connection_function} \tab Describes a connection function\cr
\link[=cloudfront_describe_function]{describe_function} \tab Gets configuration information and metadata about a CloudFront function, but not the function's code\cr
\link[=cloudfront_describe_key_value_store]{describe_key_value_store} \tab Specifies the key value store and its configuration\cr
\link[=cloudfront_disassociate_distribution_tenant_web_acl]{disassociate_distribution_tenant_web_acl} \tab Disassociates a distribution tenant from the WAF web ACL\cr
\link[=cloudfront_disassociate_distribution_web_acl]{disassociate_distribution_web_acl} \tab Disassociates a distribution from the WAF web ACL\cr
\link[=cloudfront_get_anycast_ip_list]{get_anycast_ip_list} \tab Gets an Anycast static IP list\cr
\link[=cloudfront_get_cache_policy]{get_cache_policy} \tab Gets a cache policy, including the following metadata:\cr
\link[=cloudfront_get_cache_policy_config]{get_cache_policy_config} \tab Gets a cache policy configuration\cr
\link[=cloudfront_get_cloud_front_origin_access_identity]{get_cloud_front_origin_access_identity} \tab Get the information about an origin access identity\cr
\link[=cloudfront_get_cloud_front_origin_access_identity_config]{get_cloud_front_origin_access_identity_config} \tab Get the configuration information about an origin access identity\cr
\link[=cloudfront_get_connection_function]{get_connection_function} \tab Gets a connection function\cr
\link[=cloudfront_get_connection_group]{get_connection_group} \tab Gets information about a connection group\cr
\link[=cloudfront_get_connection_group_by_routing_endpoint]{get_connection_group_by_routing_endpoint} \tab Gets information about a connection group by using the endpoint that you specify\cr
\link[=cloudfront_get_continuous_deployment_policy]{get_continuous_deployment_policy} \tab Gets a continuous deployment policy, including metadata (the policy's identifier and the date and time when the policy was last modified)\cr
\link[=cloudfront_get_continuous_deployment_policy_config]{get_continuous_deployment_policy_config} \tab Gets configuration information about a continuous deployment policy\cr
\link[=cloudfront_get_distribution]{get_distribution} \tab Get the information about a distribution\cr
\link[=cloudfront_get_distribution_config]{get_distribution_config} \tab Get the configuration information about a distribution\cr
\link[=cloudfront_get_distribution_tenant]{get_distribution_tenant} \tab Gets information about a distribution tenant\cr
\link[=cloudfront_get_distribution_tenant_by_domain]{get_distribution_tenant_by_domain} \tab Gets information about a distribution tenant by the associated domain\cr
\link[=cloudfront_get_field_level_encryption]{get_field_level_encryption} \tab Get the field-level encryption configuration information\cr
\link[=cloudfront_get_field_level_encryption_config]{get_field_level_encryption_config} \tab Get the field-level encryption configuration information\cr
\link[=cloudfront_get_field_level_encryption_profile]{get_field_level_encryption_profile} \tab Get the field-level encryption profile information\cr
\link[=cloudfront_get_field_level_encryption_profile_config]{get_field_level_encryption_profile_config} \tab Get the field-level encryption profile configuration information\cr
\link[=cloudfront_get_function]{get_function} \tab Gets the code of a CloudFront function\cr
\link[=cloudfront_get_invalidation]{get_invalidation} \tab Get the information about an invalidation\cr
\link[=cloudfront_get_invalidation_for_distribution_tenant]{get_invalidation_for_distribution_tenant} \tab Gets information about a specific invalidation for a distribution tenant\cr
\link[=cloudfront_get_key_group]{get_key_group} \tab Gets a key group, including the date and time when the key group was last modified\cr
\link[=cloudfront_get_key_group_config]{get_key_group_config} \tab Gets a key group configuration\cr
\link[=cloudfront_get_managed_certificate_details]{get_managed_certificate_details} \tab Gets details about the CloudFront managed ACM certificate\cr
\link[=cloudfront_get_monitoring_subscription]{get_monitoring_subscription} \tab Gets information about whether additional CloudWatch metrics are enabled for the specified CloudFront distribution\cr
\link[=cloudfront_get_origin_access_control]{get_origin_access_control} \tab Gets a CloudFront origin access control, including its unique identifier\cr
\link[=cloudfront_get_origin_access_control_config]{get_origin_access_control_config} \tab Gets a CloudFront origin access control configuration\cr
\link[=cloudfront_get_origin_request_policy]{get_origin_request_policy} \tab Gets an origin request policy, including the following metadata:\cr
\link[=cloudfront_get_origin_request_policy_config]{get_origin_request_policy_config} \tab Gets an origin request policy configuration\cr
\link[=cloudfront_get_public_key]{get_public_key} \tab Gets a public key\cr
\link[=cloudfront_get_public_key_config]{get_public_key_config} \tab Gets a public key configuration\cr
\link[=cloudfront_get_realtime_log_config]{get_realtime_log_config} \tab Gets a real-time log configuration\cr
\link[=cloudfront_get_resource_policy]{get_resource_policy} \tab Retrieves the resource policy for the specified CloudFront resource that you own and have shared\cr
\link[=cloudfront_get_response_headers_policy]{get_response_headers_policy} \tab Gets a response headers policy, including metadata (the policy's identifier and the date and time when the policy was last modified)\cr
\link[=cloudfront_get_response_headers_policy_config]{get_response_headers_policy_config} \tab Gets a response headers policy configuration\cr
\link[=cloudfront_get_streaming_distribution]{get_streaming_distribution} \tab Gets information about a specified RTMP distribution, including the distribution configuration\cr
\link[=cloudfront_get_streaming_distribution_config]{get_streaming_distribution_config} \tab Get the configuration information about a streaming distribution\cr
\link[=cloudfront_get_trust_store]{get_trust_store} \tab Gets a trust store\cr
\link[=cloudfront_get_vpc_origin]{get_vpc_origin} \tab Get the details of an Amazon CloudFront VPC origin\cr
\link[=cloudfront_list_anycast_ip_lists]{list_anycast_ip_lists} \tab Lists your Anycast static IP lists\cr
\link[=cloudfront_list_cache_policies]{list_cache_policies} \tab Gets a list of cache policies\cr
\link[=cloudfront_list_cloud_front_origin_access_identities]{list_cloud_front_origin_access_identities} \tab Lists origin access identities\cr
\link[=cloudfront_list_conflicting_aliases]{list_conflicting_aliases} \tab The ListConflictingAliases API operation only supports standard distributions\cr
\link[=cloudfront_list_connection_functions]{list_connection_functions} \tab Lists connection functions\cr
\link[=cloudfront_list_connection_groups]{list_connection_groups} \tab Lists the connection groups in your Amazon Web Services account\cr
\link[=cloudfront_list_continuous_deployment_policies]{list_continuous_deployment_policies} \tab Gets a list of the continuous deployment policies in your Amazon Web Services account\cr
\link[=cloudfront_list_distributions]{list_distributions} \tab List CloudFront distributions\cr
\link[=cloudfront_list_distributions_by_anycast_ip_list_id]{list_distributions_by_anycast_ip_list_id} \tab Lists the distributions in your account that are associated with the specified AnycastIpListId\cr
\link[=cloudfront_list_distributions_by_cache_policy_id]{list_distributions_by_cache_policy_id} \tab Gets a list of distribution IDs for distributions that have a cache behavior that's associated with the specified cache policy\cr
\link[=cloudfront_list_distributions_by_connection_function]{list_distributions_by_connection_function} \tab Lists distributions by connection function\cr
\link[=cloudfront_list_distributions_by_connection_mode]{list_distributions_by_connection_mode} \tab Lists the distributions by the connection mode that you specify\cr
\link[=cloudfront_list_distributions_by_key_group]{list_distributions_by_key_group} \tab Gets a list of distribution IDs for distributions that have a cache behavior that references the specified key group\cr
\link[=cloudfront_list_distributions_by_origin_request_policy_id]{list_distributions_by_origin_request_policy_id} \tab Gets a list of distribution IDs for distributions that have a cache behavior that's associated with the specified origin request policy\cr
\link[=cloudfront_list_distributions_by_owned_resource]{list_distributions_by_owned_resource} \tab Lists the CloudFront distributions that are associated with the specified resource that you own\cr
\link[=cloudfront_list_distributions_by_realtime_log_config]{list_distributions_by_realtime_log_config} \tab Gets a list of distributions that have a cache behavior that's associated with the specified real-time log configuration\cr
\link[=cloudfront_list_distributions_by_response_headers_policy_id]{list_distributions_by_response_headers_policy_id} \tab Gets a list of distribution IDs for distributions that have a cache behavior that's associated with the specified response headers policy\cr
\link[=cloudfront_list_distributions_by_trust_store]{list_distributions_by_trust_store} \tab Lists distributions by trust store\cr
\link[=cloudfront_list_distributions_by_vpc_origin_id]{list_distributions_by_vpc_origin_id} \tab List CloudFront distributions by their VPC origin ID\cr
\link[=cloudfront_list_distributions_by_web_acl_id]{list_distributions_by_web_acl_id} \tab List the distributions that are associated with a specified WAF web ACL\cr
\link[=cloudfront_list_distribution_tenants]{list_distribution_tenants} \tab Lists the distribution tenants in your Amazon Web Services account\cr
\link[=cloudfront_list_distribution_tenants_by_customization]{list_distribution_tenants_by_customization} \tab Lists distribution tenants by the customization that you specify\cr
\link[=cloudfront_list_domain_conflicts]{list_domain_conflicts} \tab We recommend that you use the ListDomainConflicts API operation to check for domain conflicts, as it supports both standard distributions and distribution tenants\cr
\link[=cloudfront_list_field_level_encryption_configs]{list_field_level_encryption_configs} \tab List all field-level encryption configurations that have been created in CloudFront for this account\cr
\link[=cloudfront_list_field_level_encryption_profiles]{list_field_level_encryption_profiles} \tab Request a list of field-level encryption profiles that have been created in CloudFront for this account\cr
\link[=cloudfront_list_functions]{list_functions} \tab Gets a list of all CloudFront functions in your Amazon Web Services account\cr
\link[=cloudfront_list_invalidations]{list_invalidations} \tab Lists invalidation batches\cr
\link[=cloudfront_list_invalidations_for_distribution_tenant]{list_invalidations_for_distribution_tenant} \tab Lists the invalidations for a distribution tenant\cr
\link[=cloudfront_list_key_groups]{list_key_groups} \tab Gets a list of key groups\cr
\link[=cloudfront_list_key_value_stores]{list_key_value_stores} \tab Specifies the key value stores to list\cr
\link[=cloudfront_list_origin_access_controls]{list_origin_access_controls} \tab Gets the list of CloudFront origin access controls (OACs) in this Amazon Web Services account\cr
\link[=cloudfront_list_origin_request_policies]{list_origin_request_policies} \tab Gets a list of origin request policies\cr
\link[=cloudfront_list_public_keys]{list_public_keys} \tab List all public keys that have been added to CloudFront for this account\cr
\link[=cloudfront_list_realtime_log_configs]{list_realtime_log_configs} \tab Gets a list of real-time log configurations\cr
\link[=cloudfront_list_response_headers_policies]{list_response_headers_policies} \tab Gets a list of response headers policies\cr
\link[=cloudfront_list_streaming_distributions]{list_streaming_distributions} \tab List streaming distributions\cr
\link[=cloudfront_list_tags_for_resource]{list_tags_for_resource} \tab List tags for a CloudFront resource\cr
\link[=cloudfront_list_trust_stores]{list_trust_stores} \tab Lists trust stores\cr
\link[=cloudfront_list_vpc_origins]{list_vpc_origins} \tab List the CloudFront VPC origins in your account\cr
\link[=cloudfront_publish_connection_function]{publish_connection_function} \tab Publishes a connection function\cr
\link[=cloudfront_publish_function]{publish_function} \tab Publishes a CloudFront function by copying the function code from the DEVELOPMENT stage to LIVE\cr
\link[=cloudfront_put_resource_policy]{put_resource_policy} \tab Creates a resource control policy for a given CloudFront resource\cr
\link[=cloudfront_tag_resource]{tag_resource} \tab Add tags to a CloudFront resource\cr
\link[=cloudfront_test_connection_function]{test_connection_function} \tab Tests a connection function\cr
\link[=cloudfront_test_function]{test_function} \tab Tests a CloudFront function\cr
\link[=cloudfront_untag_resource]{untag_resource} \tab Remove tags from a CloudFront resource\cr
\link[=cloudfront_update_anycast_ip_list]{update_anycast_ip_list} \tab Updates an Anycast static IP list\cr
\link[=cloudfront_update_cache_policy]{update_cache_policy} \tab Updates a cache policy configuration\cr
\link[=cloudfront_update_cloud_front_origin_access_identity]{update_cloud_front_origin_access_identity} \tab Update an origin access identity\cr
\link[=cloudfront_update_connection_function]{update_connection_function} \tab Updates a connection function\cr
\link[=cloudfront_update_connection_group]{update_connection_group} \tab Updates a connection group\cr
\link[=cloudfront_update_continuous_deployment_policy]{update_continuous_deployment_policy} \tab Updates a continuous deployment policy\cr
\link[=cloudfront_update_distribution]{update_distribution} \tab Updates the configuration for a CloudFront distribution\cr
\link[=cloudfront_update_distribution_tenant]{update_distribution_tenant} \tab Updates a distribution tenant\cr
\link[=cloudfront_update_distribution_with_staging_config]{update_distribution_with_staging_config} \tab Copies the staging distribution's configuration to its corresponding primary distribution\cr
\link[=cloudfront_update_domain_association]{update_domain_association} \tab We recommend that you use the UpdateDomainAssociation API operation to move a domain association, as it supports both standard distributions and distribution tenants\cr
\link[=cloudfront_update_field_level_encryption_config]{update_field_level_encryption_config} \tab Update a field-level encryption configuration\cr
\link[=cloudfront_update_field_level_encryption_profile]{update_field_level_encryption_profile} \tab Update a field-level encryption profile\cr
\link[=cloudfront_update_function]{update_function} \tab Updates a CloudFront function\cr
\link[=cloudfront_update_key_group]{update_key_group} \tab Updates a key group\cr
\link[=cloudfront_update_key_value_store]{update_key_value_store} \tab Specifies the key value store to update\cr
\link[=cloudfront_update_origin_access_control]{update_origin_access_control} \tab Updates a CloudFront origin access control\cr
\link[=cloudfront_update_origin_request_policy]{update_origin_request_policy} \tab Updates an origin request policy configuration\cr
\link[=cloudfront_update_public_key]{update_public_key} \tab Update public key information\cr
\link[=cloudfront_update_realtime_log_config]{update_realtime_log_config} \tab Updates a real-time log configuration\cr
\link[=cloudfront_update_response_headers_policy]{update_response_headers_policy} \tab Updates a response headers policy\cr
\link[=cloudfront_update_streaming_distribution]{update_streaming_distribution} \tab Update a streaming distribution\cr
\link[=cloudfront_update_trust_store]{update_trust_store} \tab Updates a trust store\cr
\link[=cloudfront_update_vpc_origin]{update_vpc_origin} \tab Update an Amazon CloudFront VPC origin in your account\cr
\link[=cloudfront_verify_dns_configuration]{verify_dns_configuration} \tab Verify the DNS configuration for your domain names
}
}

\examples{
\dontrun{
svc <- cloudfront()
svc$associate_alias(
  Foo = 123
)
}

}
