% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_associate_alias}
\alias{cloudfront_associate_alias}
\title{Associates an alias (also known as a CNAME or an alternate domain name)
with a CloudFront distribution}
\usage{
cloudfront_associate_alias(TargetDistributionId, Alias)
}
\arguments{
\item{TargetDistributionId}{[required] The ID of the distribution that you're associating the alias with.}

\item{Alias}{[required] The alias (also known as a CNAME) to add to the target distribution.}
}
\description{
Associates an alias (also known as a CNAME or an alternate domain name) with a CloudFront distribution.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_associate_alias/} for full documentation.
}
\keyword{internal}
