% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_associate_distribution_tenant_web_acl}
\alias{cloudfront_associate_distribution_tenant_web_acl}
\title{Associates the WAF web ACL with a distribution tenant}
\usage{
cloudfront_associate_distribution_tenant_web_acl(Id, WebACLArn, IfMatch = NULL)
}
\arguments{
\item{Id}{[required] The ID of the distribution tenant.}

\item{WebACLArn}{[required] The Amazon Resource Name (ARN) of the WAF web ACL to associate.}

\item{IfMatch}{The current \code{ETag} of the distribution tenant. This value is returned in
the response of the
\code{\link[=cloudfront_get_distribution_tenant]{get_distribution_tenant}} API
operation.}
}
\description{
Associates the WAF web ACL with a distribution tenant.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_associate_distribution_tenant_web_acl/} for full documentation.
}
\keyword{internal}
