% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_associate_distribution_web_acl}
\alias{cloudfront_associate_distribution_web_acl}
\title{Associates the WAF web ACL with a distribution}
\usage{
cloudfront_associate_distribution_web_acl(Id, WebACLArn, IfMatch = NULL)
}
\arguments{
\item{Id}{[required] The ID of the distribution.}

\item{WebACLArn}{[required] The Amazon Resource Name (ARN) of the WAF web ACL to associate.}

\item{IfMatch}{The value of the \code{ETag} header that you received when retrieving the
distribution that you're associating with the WAF web ACL.}
}
\description{
Associates the WAF web ACL with a distribution.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_associate_distribution_web_acl/} for full documentation.
}
\keyword{internal}
