% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_create_anycast_ip_list}
\alias{cloudfront_create_anycast_ip_list}
\title{Creates an Anycast static IP list}
\usage{
cloudfront_create_anycast_ip_list(Name, IpCount, Tags = NULL)
}
\arguments{
\item{Name}{[required] Name of the Anycast static IP list.}

\item{IpCount}{[required] The number of static IP addresses that are allocated to the Anycast
static IP list.}

\item{Tags}{}
}
\description{
Creates an Anycast static IP list.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_create_anycast_ip_list/} for full documentation.
}
\keyword{internal}
