% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_create_connection_function}
\alias{cloudfront_create_connection_function}
\title{Creates a connection function}
\usage{
cloudfront_create_connection_function(
  Name,
  ConnectionFunctionConfig,
  ConnectionFunctionCode,
  Tags = NULL
)
}
\arguments{
\item{Name}{[required] A name for the connection function.}

\item{ConnectionFunctionConfig}{[required]}

\item{ConnectionFunctionCode}{[required] The code for the connection function.}

\item{Tags}{}
}
\description{
Creates a connection function.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_create_connection_function/} for full documentation.
}
\keyword{internal}
