% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_create_connection_group}
\alias{cloudfront_create_connection_group}
\title{Creates a connection group}
\usage{
cloudfront_create_connection_group(
  Name,
  Ipv6Enabled = NULL,
  Tags = NULL,
  AnycastIpListId = NULL,
  Enabled = NULL
)
}
\arguments{
\item{Name}{[required] The name of the connection group. Enter a friendly identifier that is
unique within your Amazon Web Services account. This name can't be
updated after you create the connection group.}

\item{Ipv6Enabled}{Enable IPv6 for the connection group. The default is \code{true}. For more
information, see \href{https://docs.aws.amazon.com/AmazonCloudFront/latest/DeveloperGuide/distribution-web-values-specify.html#DownloadDistValuesEnableIPv6}{Enable IPv6}
in the \emph{Amazon CloudFront Developer Guide}.}

\item{Tags}{}

\item{AnycastIpListId}{The ID of the Anycast static IP list.}

\item{Enabled}{Enable the connection group.}
}
\description{
Creates a connection group.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_create_connection_group/} for full documentation.
}
\keyword{internal}
