% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_create_continuous_deployment_policy}
\alias{cloudfront_create_continuous_deployment_policy}
\title{Creates a continuous deployment policy that distributes traffic for a
custom domain name to two different CloudFront distributions}
\usage{
cloudfront_create_continuous_deployment_policy(
  ContinuousDeploymentPolicyConfig
)
}
\arguments{
\item{ContinuousDeploymentPolicyConfig}{[required] Contains the configuration for a continuous deployment policy.}
}
\description{
Creates a continuous deployment policy that distributes traffic for a custom domain name to two different CloudFront distributions.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_create_continuous_deployment_policy/} for full documentation.
}
\keyword{internal}
