% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_create_distribution}
\alias{cloudfront_create_distribution}
\title{Creates a CloudFront distribution}
\usage{
cloudfront_create_distribution(DistributionConfig)
}
\arguments{
\item{DistributionConfig}{[required] The distribution's configuration information.}
}
\description{
Creates a CloudFront distribution.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_create_distribution/} for full documentation.
}
\keyword{internal}
