% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_create_field_level_encryption_config}
\alias{cloudfront_create_field_level_encryption_config}
\title{Create a new field-level encryption configuration}
\usage{
cloudfront_create_field_level_encryption_config(FieldLevelEncryptionConfig)
}
\arguments{
\item{FieldLevelEncryptionConfig}{[required] The request to create a new field-level encryption configuration.}
}
\description{
Create a new field-level encryption configuration.

See \url{https://paws-r.github.io/docs/cloudfront/create_field_level_encryption_config.html} for full documentation.
}
\keyword{internal}
