% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_create_field_level_encryption_profile}
\alias{cloudfront_create_field_level_encryption_profile}
\title{Create a field-level encryption profile}
\usage{
cloudfront_create_field_level_encryption_profile(
  FieldLevelEncryptionProfileConfig
)
}
\arguments{
\item{FieldLevelEncryptionProfileConfig}{[required] The request to create a field-level encryption profile.}
}
\description{
Create a field-level encryption profile.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_create_field_level_encryption_profile/} for full documentation.
}
\keyword{internal}
