% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_create_function}
\alias{cloudfront_create_function}
\title{Creates a CloudFront function}
\usage{
cloudfront_create_function(Name, FunctionConfig, FunctionCode)
}
\arguments{
\item{Name}{[required] A name to identify the function.}

\item{FunctionConfig}{[required] Configuration information about the function, including an optional
comment and the function’s runtime.}

\item{FunctionCode}{[required] The function code. For more information about writing a CloudFront
function, see \href{https://docs.aws.amazon.com/AmazonCloudFront/latest/DeveloperGuide/writing-function-code.html}{Writing function code for CloudFront Functions}
in the \emph{Amazon CloudFront Developer Guide}.}
}
\value{
A list with the following syntax:\preformatted{list(
  FunctionSummary = list(
    Name = "string",
    Status = "string",
    FunctionConfig = list(
      Comment = "string",
      Runtime = "cloudfront-js-1.0"
    ),
    FunctionMetadata = list(
      FunctionARN = "string",
      Stage = "DEVELOPMENT"|"LIVE",
      CreatedTime = as.POSIXct(
        "2015-01-01"
      ),
      LastModifiedTime = as.POSIXct(
        "2015-01-01"
      )
    )
  ),
  Location = "string",
  ETag = "string"
)
}
}
\description{
Creates a CloudFront function.

To create a function, you provide the function code and some
configuration information about the function. The response contains an
Amazon Resource Name (ARN) that uniquely identifies the function.

When you create a function, it’s in the \code{DEVELOPMENT} stage. In this
stage, you can test the function with
\code{\link[=cloudfront_test_function]{test_function}}, and update it with
\code{\link[=cloudfront_update_function]{update_function}}.

When you’re ready to use your function with a CloudFront distribution,
use \code{\link[=cloudfront_publish_function]{publish_function}} to copy the
function from the \code{DEVELOPMENT} stage to \code{LIVE}. When it’s live, you can
attach the function to a distribution’s cache behavior, using the
function’s ARN.
}
\section{Request syntax}{
\preformatted{svc$create_function(
  Name = "string",
  FunctionConfig = list(
    Comment = "string",
    Runtime = "cloudfront-js-1.0"
  ),
  FunctionCode = raw
)
}
}

\keyword{internal}
