% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_create_invalidation_for_distribution_tenant}
\alias{cloudfront_create_invalidation_for_distribution_tenant}
\title{Creates an invalidation for a distribution tenant}
\usage{
cloudfront_create_invalidation_for_distribution_tenant(Id, InvalidationBatch)
}
\arguments{
\item{Id}{[required] The ID of the distribution tenant.}

\item{InvalidationBatch}{[required]}
}
\description{
Creates an invalidation for a distribution tenant. For more information, see \href{https://docs.aws.amazon.com/AmazonCloudFront/latest/DeveloperGuide/Invalidation.html}{Invalidating files} in the \emph{Amazon CloudFront Developer Guide}.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_create_invalidation_for_distribution_tenant/} for full documentation.
}
\keyword{internal}
