% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_create_monitoring_subscription}
\alias{cloudfront_create_monitoring_subscription}
\title{Enables or disables additional Amazon CloudWatch metrics for the
specified CloudFront distribution}
\usage{
cloudfront_create_monitoring_subscription(
  DistributionId,
  MonitoringSubscription
)
}
\arguments{
\item{DistributionId}{[required] The ID of the distribution that you are enabling metrics for.}

\item{MonitoringSubscription}{[required] A monitoring subscription. This structure contains information about
whether additional CloudWatch metrics are enabled for a given CloudFront
distribution.}
}
\description{
Enables or disables additional Amazon CloudWatch metrics for the specified CloudFront distribution. The additional metrics incur an additional cost.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_create_monitoring_subscription/} for full documentation.
}
\keyword{internal}
