% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_create_origin_access_control}
\alias{cloudfront_create_origin_access_control}
\title{Creates a new origin access control in CloudFront}
\usage{
cloudfront_create_origin_access_control(OriginAccessControlConfig)
}
\arguments{
\item{OriginAccessControlConfig}{[required] Contains the origin access control.}
}
\description{
Creates a new origin access control in CloudFront. After you create an origin access control, you can add it to an origin in a CloudFront distribution so that CloudFront sends authenticated (signed) requests to the origin.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_create_origin_access_control/} for full documentation.
}
\keyword{internal}
