% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_create_origin_request_policy}
\alias{cloudfront_create_origin_request_policy}
\title{Creates an origin request policy}
\usage{
cloudfront_create_origin_request_policy(OriginRequestPolicyConfig)
}
\arguments{
\item{OriginRequestPolicyConfig}{[required] An origin request policy configuration.}
}
\description{
Creates an origin request policy.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_create_origin_request_policy/} for full documentation.
}
\keyword{internal}
