% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_create_response_headers_policy}
\alias{cloudfront_create_response_headers_policy}
\title{Creates a response headers policy}
\usage{
cloudfront_create_response_headers_policy(ResponseHeadersPolicyConfig)
}
\arguments{
\item{ResponseHeadersPolicyConfig}{[required] Contains metadata about the response headers policy, and a set of
configurations that specify the response headers.}
}
\description{
Creates a response headers policy.

See \url{https://paws-r.github.io/docs/cloudfront/create_response_headers_policy.html} for full documentation.
}
\keyword{internal}
