% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_create_streaming_distribution_with_tags}
\alias{cloudfront_create_streaming_distribution_with_tags}
\title{This API is deprecated}
\usage{
cloudfront_create_streaming_distribution_with_tags(
  StreamingDistributionConfigWithTags
)
}
\arguments{
\item{StreamingDistributionConfigWithTags}{[required] The streaming distribution's configuration information.}
}
\description{
This API is deprecated. Amazon CloudFront is deprecating real-time messaging protocol (RTMP) distributions on December 31, 2020. For more information, read the announcement on the Amazon CloudFront discussion forum.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_create_streaming_distribution_with_tags/} for full documentation.
}
\keyword{internal}
