% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_delete_anycast_ip_list}
\alias{cloudfront_delete_anycast_ip_list}
\title{Deletes an Anycast static IP list}
\usage{
cloudfront_delete_anycast_ip_list(Id, IfMatch)
}
\arguments{
\item{Id}{[required] The ID of the Anycast static IP list.}

\item{IfMatch}{[required] The current version (\code{ETag} value) of the Anycast static IP list that
you are deleting.}
}
\description{
Deletes an Anycast static IP list.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_delete_anycast_ip_list/} for full documentation.
}
\keyword{internal}
