% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_delete_connection_group}
\alias{cloudfront_delete_connection_group}
\title{Deletes a connection group}
\usage{
cloudfront_delete_connection_group(Id, IfMatch)
}
\arguments{
\item{Id}{[required] The ID of the connection group to delete.}

\item{IfMatch}{[required] The value of the \code{ETag} header that you received when retrieving the
connection group to delete.}
}
\description{
Deletes a connection group.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_delete_connection_group/} for full documentation.
}
\keyword{internal}
