% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_delete_continuous_deployment_policy}
\alias{cloudfront_delete_continuous_deployment_policy}
\title{Deletes a continuous deployment policy}
\usage{
cloudfront_delete_continuous_deployment_policy(Id, IfMatch = NULL)
}
\arguments{
\item{Id}{[required] The identifier of the continuous deployment policy that you are
deleting.}

\item{IfMatch}{The current version (\code{ETag} value) of the continuous deployment policy
that you are deleting.}
}
\description{
Deletes a continuous deployment policy.

See \url{https://paws-r.github.io/docs/cloudfront/delete_continuous_deployment_policy.html} for full documentation.
}
\keyword{internal}
