% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_delete_distribution}
\alias{cloudfront_delete_distribution}
\title{Delete a distribution}
\usage{
cloudfront_delete_distribution(Id, IfMatch = NULL)
}
\arguments{
\item{Id}{[required] The distribution ID.}

\item{IfMatch}{The value of the \code{ETag} header that you received when you disabled the
distribution. For example: \code{E2QWRUHAPOMQZL}.}
}
\description{
Delete a distribution.

See \url{https://paws-r.github.io/docs/cloudfront/delete_distribution.html} for full documentation.
}
\keyword{internal}
