% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_delete_field_level_encryption_config}
\alias{cloudfront_delete_field_level_encryption_config}
\title{Remove a field-level encryption configuration}
\usage{
cloudfront_delete_field_level_encryption_config(Id, IfMatch)
}
\arguments{
\item{Id}{[required] The ID of the configuration you want to delete from CloudFront.}

\item{IfMatch}{The value of the \code{ETag} header that you received when retrieving the
configuration identity to delete. For example: \code{E2QWRUHAPOMQZL}.}
}
\description{
Remove a field-level encryption configuration.
}
\section{Request syntax}{
\preformatted{svc$delete_field_level_encryption_config(
  Id = "string",
  IfMatch = "string"
)
}
}

\keyword{internal}
