% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_delete_function}
\alias{cloudfront_delete_function}
\title{Deletes a CloudFront function}
\usage{
cloudfront_delete_function(Name, IfMatch)
}
\arguments{
\item{Name}{[required] The name of the function that you are deleting.}

\item{IfMatch}{[required] The current version (\code{ETag} value) of the function that you are
deleting, which you can get using
\code{\link[=cloudfront_describe_function]{describe_function}}.}
}
\value{
An empty list.
}
\description{
Deletes a CloudFront function.

You cannot delete a function if it’s associated with a cache behavior.
First, update your distributions to remove the function association from
all cache behaviors, then delete the function.

To delete a function, you must provide the function’s name and version
(\code{ETag} value). To get these values, you can use
\code{\link[=cloudfront_list_functions]{list_functions}} and
\code{\link[=cloudfront_describe_function]{describe_function}}.
}
\section{Request syntax}{
\preformatted{svc$delete_function(
  Name = "string",
  IfMatch = "string"
)
}
}

\keyword{internal}
