% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_delete_origin_access_control}
\alias{cloudfront_delete_origin_access_control}
\title{Deletes a CloudFront origin access control}
\usage{
cloudfront_delete_origin_access_control(Id, IfMatch = NULL)
}
\arguments{
\item{Id}{[required] The unique identifier of the origin access control that you are
deleting.}

\item{IfMatch}{The current version (\code{ETag} value) of the origin access control that you
are deleting.}
}
\description{
Deletes a CloudFront origin access control.

See \url{https://paws-r-sdk.com/docs/delete_origin_access_control/} for full documentation.
}
\keyword{internal}
