% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_delete_public_key}
\alias{cloudfront_delete_public_key}
\title{Remove a public key you previously added to CloudFront}
\usage{
cloudfront_delete_public_key(Id, IfMatch = NULL)
}
\arguments{
\item{Id}{[required] The ID of the public key you want to remove from CloudFront.}

\item{IfMatch}{The value of the \code{ETag} header that you received when retrieving the
public key identity to delete. For example: \code{E2QWRUHAPOMQZL}.}
}
\description{
Remove a public key you previously added to CloudFront.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_delete_public_key/} for full documentation.
}
\keyword{internal}
