% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_delete_realtime_log_config}
\alias{cloudfront_delete_realtime_log_config}
\title{Deletes a real-time log configuration}
\usage{
cloudfront_delete_realtime_log_config(Name = NULL, ARN = NULL)
}
\arguments{
\item{Name}{The name of the real-time log configuration to delete.}

\item{ARN}{The Amazon Resource Name (ARN) of the real-time log configuration to
delete.}
}
\description{
Deletes a real-time log configuration.

See \url{https://paws-r.github.io/docs/cloudfront/delete_realtime_log_config.html} for full documentation.
}
\keyword{internal}
