% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_delete_resource_policy}
\alias{cloudfront_delete_resource_policy}
\title{Deletes the resource policy attached to the CloudFront resource}
\usage{
cloudfront_delete_resource_policy(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the CloudFront resource for which the
resource policy should be deleted.}
}
\description{
Deletes the resource policy attached to the CloudFront resource.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_delete_resource_policy/} for full documentation.
}
\keyword{internal}
