% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_delete_streaming_distribution}
\alias{cloudfront_delete_streaming_distribution}
\title{Delete a streaming distribution}
\usage{
cloudfront_delete_streaming_distribution(Id, IfMatch = NULL)
}
\arguments{
\item{Id}{[required] The distribution ID.}

\item{IfMatch}{The value of the \code{ETag} header that you received when you disabled the
streaming distribution. For example: \code{E2QWRUHAPOMQZL}.}
}
\description{
Delete a streaming distribution. To delete an RTMP distribution using the CloudFront API, perform the following steps.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_delete_streaming_distribution/} for full documentation.
}
\keyword{internal}
