% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_delete_trust_store}
\alias{cloudfront_delete_trust_store}
\title{Deletes a trust store}
\usage{
cloudfront_delete_trust_store(Id, IfMatch)
}
\arguments{
\item{Id}{[required] The trust store's ID.}

\item{IfMatch}{[required] The current version (\code{ETag} value) of the trust store you are deleting.}
}
\description{
Deletes a trust store.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_delete_trust_store/} for full documentation.
}
\keyword{internal}
