% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_describe_connection_function}
\alias{cloudfront_describe_connection_function}
\title{Describes a connection function}
\usage{
cloudfront_describe_connection_function(Identifier, Stage = NULL)
}
\arguments{
\item{Identifier}{[required] The connection function's identifier.}

\item{Stage}{The connection function's stage.}
}
\description{
Describes a connection function.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_describe_connection_function/} for full documentation.
}
\keyword{internal}
