% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_describe_function}
\alias{cloudfront_describe_function}
\title{Gets configuration information and metadata about a CloudFront function,
but not the function’s code}
\usage{
cloudfront_describe_function(Name, Stage)
}
\arguments{
\item{Name}{[required] The name of the function that you are getting information about.}

\item{Stage}{The function’s stage, either \code{DEVELOPMENT} or \code{LIVE}.}
}
\value{
A list with the following syntax:\preformatted{list(
  FunctionSummary = list(
    Name = "string",
    Status = "string",
    FunctionConfig = list(
      Comment = "string",
      Runtime = "cloudfront-js-1.0"
    ),
    FunctionMetadata = list(
      FunctionARN = "string",
      Stage = "DEVELOPMENT"|"LIVE",
      CreatedTime = as.POSIXct(
        "2015-01-01"
      ),
      LastModifiedTime = as.POSIXct(
        "2015-01-01"
      )
    )
  ),
  ETag = "string"
)
}
}
\description{
Gets configuration information and metadata about a CloudFront function,
but not the function’s code. To get a function’s code, use
\code{\link[=cloudfront_get_function]{get_function}}.

To get configuration information and metadata about a function, you must
provide the function’s name and stage. To get these values, you can use
\code{\link[=cloudfront_list_functions]{list_functions}}.
}
\section{Request syntax}{
\preformatted{svc$describe_function(
  Name = "string",
  Stage = "DEVELOPMENT"|"LIVE"
)
}
}

\keyword{internal}
