% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_get_anycast_ip_list}
\alias{cloudfront_get_anycast_ip_list}
\title{Gets an Anycast static IP list}
\usage{
cloudfront_get_anycast_ip_list(Id)
}
\arguments{
\item{Id}{[required] The ID of the Anycast static IP list.}
}
\description{
Gets an Anycast static IP list.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_get_anycast_ip_list/} for full documentation.
}
\keyword{internal}
