% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_get_cloud_front_origin_access_identity}
\alias{cloudfront_get_cloud_front_origin_access_identity}
\title{Get the information about an origin access identity}
\usage{
cloudfront_get_cloud_front_origin_access_identity(Id)
}
\arguments{
\item{Id}{[required] The identity's ID.}
}
\description{
Get the information about an origin access identity.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_get_cloud_front_origin_access_identity/} for full documentation.
}
\keyword{internal}
