% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_get_connection_function}
\alias{cloudfront_get_connection_function}
\title{Gets a connection function}
\usage{
cloudfront_get_connection_function(Identifier, Stage = NULL)
}
\arguments{
\item{Identifier}{[required] The connection function's identifier.}

\item{Stage}{The connection function's stage.}
}
\description{
Gets a connection function.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_get_connection_function/} for full documentation.
}
\keyword{internal}
