% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_get_connection_group}
\alias{cloudfront_get_connection_group}
\title{Gets information about a connection group}
\usage{
cloudfront_get_connection_group(Identifier)
}
\arguments{
\item{Identifier}{[required] The ID, name, or Amazon Resource Name (ARN) of the connection group.}
}
\description{
Gets information about a connection group.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_get_connection_group/} for full documentation.
}
\keyword{internal}
