% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_get_connection_group_by_routing_endpoint}
\alias{cloudfront_get_connection_group_by_routing_endpoint}
\title{Gets information about a connection group by using the endpoint that you
specify}
\usage{
cloudfront_get_connection_group_by_routing_endpoint(RoutingEndpoint)
}
\arguments{
\item{RoutingEndpoint}{[required] The routing endpoint for the target connection group, such as
d111111abcdef8.cloudfront.net.}
}
\description{
Gets information about a connection group by using the endpoint that you specify.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_get_connection_group_by_routing_endpoint/} for full documentation.
}
\keyword{internal}
