% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_get_continuous_deployment_policy}
\alias{cloudfront_get_continuous_deployment_policy}
\title{Gets a continuous deployment policy, including metadata (the policy's
identifier and the date and time when the policy was last modified)}
\usage{
cloudfront_get_continuous_deployment_policy(Id)
}
\arguments{
\item{Id}{[required] The identifier of the continuous deployment policy that you are getting.}
}
\description{
Gets a continuous deployment policy, including metadata (the policy's identifier and the date and time when the policy was last modified).

See \url{https://www.paws-r-sdk.com/docs/cloudfront_get_continuous_deployment_policy/} for full documentation.
}
\keyword{internal}
