% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_get_continuous_deployment_policy_config}
\alias{cloudfront_get_continuous_deployment_policy_config}
\title{Gets configuration information about a continuous deployment policy}
\usage{
cloudfront_get_continuous_deployment_policy_config(Id)
}
\arguments{
\item{Id}{[required] The identifier of the continuous deployment policy whose configuration
you are getting.}
}
\description{
Gets configuration information about a continuous deployment policy.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_get_continuous_deployment_policy_config/} for full documentation.
}
\keyword{internal}
