% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_get_distribution}
\alias{cloudfront_get_distribution}
\title{Get the information about a distribution}
\usage{
cloudfront_get_distribution(Id)
}
\arguments{
\item{Id}{[required] The distribution's ID. If the ID is empty, an empty distribution
configuration is returned.}
}
\description{
Get the information about a distribution.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_get_distribution/} for full documentation.
}
\keyword{internal}
