% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_get_distribution_config}
\alias{cloudfront_get_distribution_config}
\title{Get the configuration information about a distribution}
\usage{
cloudfront_get_distribution_config(Id)
}
\arguments{
\item{Id}{[required] The distribution's ID. If the ID is empty, an empty distribution
configuration is returned.}
}
\description{
Get the configuration information about a distribution.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_get_distribution_config/} for full documentation.
}
\keyword{internal}
