% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_get_distribution_tenant}
\alias{cloudfront_get_distribution_tenant}
\title{Gets information about a distribution tenant}
\usage{
cloudfront_get_distribution_tenant(Identifier)
}
\arguments{
\item{Identifier}{[required] The identifier of the distribution tenant. You can specify the ARN, ID,
or name of the distribution tenant.}
}
\description{
Gets information about a distribution tenant.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_get_distribution_tenant/} for full documentation.
}
\keyword{internal}
