% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_get_distribution_tenant_by_domain}
\alias{cloudfront_get_distribution_tenant_by_domain}
\title{Gets information about a distribution tenant by the associated domain}
\usage{
cloudfront_get_distribution_tenant_by_domain(Domain)
}
\arguments{
\item{Domain}{[required] A domain name associated with the target distribution tenant.}
}
\description{
Gets information about a distribution tenant by the associated domain.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_get_distribution_tenant_by_domain/} for full documentation.
}
\keyword{internal}
