% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_get_field_level_encryption_profile}
\alias{cloudfront_get_field_level_encryption_profile}
\title{Get the field-level encryption profile information}
\usage{
cloudfront_get_field_level_encryption_profile(Id)
}
\arguments{
\item{Id}{[required] Get the ID for the field-level encryption profile information.}
}
\value{
A list with the following syntax:\preformatted{list(
  FieldLevelEncryptionProfile = list(
    Id = "string",
    LastModifiedTime = as.POSIXct(
      "2015-01-01"
    ),
    FieldLevelEncryptionProfileConfig = list(
      Name = "string",
      CallerReference = "string",
      Comment = "string",
      EncryptionEntities = list(
        Quantity = 123,
        Items = list(
          list(
            PublicKeyId = "string",
            ProviderId = "string",
            FieldPatterns = list(
              Quantity = 123,
              Items = list(
                "string"
              )
            )
          )
        )
      )
    )
  ),
  ETag = "string"
)
}
}
\description{
Get the field-level encryption profile information.
}
\section{Request syntax}{
\preformatted{svc$get_field_level_encryption_profile(
  Id = "string"
)
}
}

\keyword{internal}
