% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_get_function}
\alias{cloudfront_get_function}
\title{Gets the code of a CloudFront function}
\usage{
cloudfront_get_function(Name, Stage)
}
\arguments{
\item{Name}{[required] The name of the function whose code you are getting.}

\item{Stage}{The function’s stage, either \code{DEVELOPMENT} or \code{LIVE}.}
}
\value{
A list with the following syntax:\preformatted{list(
  FunctionCode = raw,
  ETag = "string",
  ContentType = "string"
)
}
}
\description{
Gets the code of a CloudFront function. To get configuration information
and metadata about a function, use
\code{\link[=cloudfront_describe_function]{describe_function}}.

To get a function’s code, you must provide the function’s name and
stage. To get these values, you can use
\code{\link[=cloudfront_list_functions]{list_functions}}.
}
\section{Request syntax}{
\preformatted{svc$get_function(
  Name = "string",
  Stage = "DEVELOPMENT"|"LIVE"
)
}
}

\keyword{internal}
