% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_get_invalidation_for_distribution_tenant}
\alias{cloudfront_get_invalidation_for_distribution_tenant}
\title{Gets information about a specific invalidation for a distribution tenant}
\usage{
cloudfront_get_invalidation_for_distribution_tenant(DistributionTenantId, Id)
}
\arguments{
\item{DistributionTenantId}{[required] The ID of the distribution tenant.}

\item{Id}{[required] The ID of the invalidation to retrieve.}
}
\description{
Gets information about a specific invalidation for a distribution tenant.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_get_invalidation_for_distribution_tenant/} for full documentation.
}
\keyword{internal}
