% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_get_key_group}
\alias{cloudfront_get_key_group}
\title{Gets a key group, including the date and time when the key group was
last modified}
\usage{
cloudfront_get_key_group(Id)
}
\arguments{
\item{Id}{[required] The identifier of the key group that you are getting. To get the
identifier, use \code{ListKeyGroups}.}
}
\description{
Gets a key group, including the date and time when the key group was
last modified.
}
\details{
To get a key group, you must provide the key group’s identifier. If the
key group is referenced in a distribution’s cache behavior, you can get
the key group’s identifier using \code{ListDistributions} or
\code{GetDistribution}. If the key group is not referenced in a cache
behavior, you can get the identifier using \code{ListKeyGroups}.
}
\section{Request syntax}{
\preformatted{svc$get_key_group(
  Id = "string"
)
}
}

\keyword{internal}
