% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_get_key_group_config}
\alias{cloudfront_get_key_group_config}
\title{Gets a key group configuration}
\usage{
cloudfront_get_key_group_config(Id)
}
\arguments{
\item{Id}{[required] The identifier of the key group whose configuration you are getting. To
get the identifier, use \code{\link[=cloudfront_list_key_groups]{list_key_groups}}.}
}
\description{
Gets a key group configuration.

See \url{https://paws-r.github.io/docs/cloudfront/get_key_group_config.html} for full documentation.
}
\keyword{internal}
