% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_get_managed_certificate_details}
\alias{cloudfront_get_managed_certificate_details}
\title{Gets details about the CloudFront managed ACM certificate}
\usage{
cloudfront_get_managed_certificate_details(Identifier)
}
\arguments{
\item{Identifier}{[required] The identifier of the multi-tenant distribution.}
}
\description{
Gets details about the CloudFront managed ACM certificate.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_get_managed_certificate_details/} for full documentation.
}
\keyword{internal}
