% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_get_monitoring_subscription}
\alias{cloudfront_get_monitoring_subscription}
\title{Gets information about whether additional CloudWatch metrics are enabled
for the specified CloudFront distribution}
\usage{
cloudfront_get_monitoring_subscription(DistributionId)
}
\arguments{
\item{DistributionId}{[required] The ID of the distribution that you are getting metrics information for.}
}
\description{
Gets information about whether additional CloudWatch metrics are enabled
for the specified CloudFront distribution.
}
\section{Request syntax}{
\preformatted{svc$get_monitoring_subscription(
  DistributionId = "string"
)
}
}

\keyword{internal}
