% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_get_origin_access_control}
\alias{cloudfront_get_origin_access_control}
\title{Gets a CloudFront origin access control, including its unique identifier}
\usage{
cloudfront_get_origin_access_control(Id)
}
\arguments{
\item{Id}{[required] The unique identifier of the origin access control.}
}
\description{
Gets a CloudFront origin access control, including its unique identifier.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_get_origin_access_control/} for full documentation.
}
\keyword{internal}
