% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_get_public_key_config}
\alias{cloudfront_get_public_key_config}
\title{Gets a public key configuration}
\usage{
cloudfront_get_public_key_config(Id)
}
\arguments{
\item{Id}{[required] The identifier of the public key whose configuration you are getting.}
}
\value{
A list with the following syntax:\preformatted{list(
  PublicKeyConfig = list(
    CallerReference = "string",
    Name = "string",
    EncodedKey = "string",
    Comment = "string"
  ),
  ETag = "string"
)
}
}
\description{
Gets a public key configuration.
}
\section{Request syntax}{
\preformatted{svc$get_public_key_config(
  Id = "string"
)
}
}

\keyword{internal}
