% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_get_realtime_log_config}
\alias{cloudfront_get_realtime_log_config}
\title{Gets a real-time log configuration}
\usage{
cloudfront_get_realtime_log_config(Name = NULL, ARN = NULL)
}
\arguments{
\item{Name}{The name of the real-time log configuration to get.}

\item{ARN}{The Amazon Resource Name (ARN) of the real-time log configuration to
get.}
}
\description{
Gets a real-time log configuration.

See \url{https://paws-r.github.io/docs/cloudfront/get_realtime_log_config.html} for full documentation.
}
\keyword{internal}
