% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_get_resource_policy}
\alias{cloudfront_get_resource_policy}
\title{Retrieves the resource policy for the specified CloudFront resource that
you own and have shared}
\usage{
cloudfront_get_resource_policy(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the CloudFront resource that is
associated with the resource policy.}
}
\description{
Retrieves the resource policy for the specified CloudFront resource that you own and have shared.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_get_resource_policy/} for full documentation.
}
\keyword{internal}
