% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_get_trust_store}
\alias{cloudfront_get_trust_store}
\title{Gets a trust store}
\usage{
cloudfront_get_trust_store(Identifier)
}
\arguments{
\item{Identifier}{[required] The trust store's identifier.}
}
\description{
Gets a trust store.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_get_trust_store/} for full documentation.
}
\keyword{internal}
