% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_get_vpc_origin}
\alias{cloudfront_get_vpc_origin}
\title{Get the details of an Amazon CloudFront VPC origin}
\usage{
cloudfront_get_vpc_origin(Id)
}
\arguments{
\item{Id}{[required] The VPC origin ID.}
}
\description{
Get the details of an Amazon CloudFront VPC origin.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_get_vpc_origin/} for full documentation.
}
\keyword{internal}
