% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_list_connection_functions}
\alias{cloudfront_list_connection_functions}
\title{Lists connection functions}
\usage{
cloudfront_list_connection_functions(
  Marker = NULL,
  MaxItems = NULL,
  Stage = NULL
)
}
\arguments{
\item{Marker}{Use this field when paginating results to indicate where to begin in
your list. The response includes items in the list that occur after the
marker. To get the next page of the list, set this field's value to the
value of \code{NextMarker} from the current page's response.}

\item{MaxItems}{The maximum number of connection functions that you want returned in the
response.}

\item{Stage}{The connection function's stage.}
}
\description{
Lists connection functions.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_list_connection_functions/} for full documentation.
}
\keyword{internal}
