% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_list_connection_groups}
\alias{cloudfront_list_connection_groups}
\title{Lists the connection groups in your Amazon Web Services account}
\usage{
cloudfront_list_connection_groups(
  AssociationFilter = NULL,
  Marker = NULL,
  MaxItems = NULL
)
}
\arguments{
\item{AssociationFilter}{Filter by associated Anycast IP list ID.}

\item{Marker}{The marker for the next set of connection groups to retrieve.}

\item{MaxItems}{The maximum number of connection groups to return.}
}
\description{
Lists the connection groups in your Amazon Web Services account.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_list_connection_groups/} for full documentation.
}
\keyword{internal}
