% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_list_continuous_deployment_policies}
\alias{cloudfront_list_continuous_deployment_policies}
\title{Gets a list of the continuous deployment policies in your Amazon Web
Services account}
\usage{
cloudfront_list_continuous_deployment_policies(Marker = NULL, MaxItems = NULL)
}
\arguments{
\item{Marker}{Use this field when paginating results to indicate where to begin in
your list of continuous deployment policies. The response includes
policies in the list that occur after the marker. To get the next page
of the list, set this field's value to the value of \code{NextMarker} from
the current page's response.}

\item{MaxItems}{The maximum number of continuous deployment policies that you want
returned in the response.}
}
\description{
Gets a list of the continuous deployment policies in your Amazon Web Services account.

See \url{https://paws-r.github.io/docs/cloudfront/list_continuous_deployment_policies.html} for full documentation.
}
\keyword{internal}
