% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_list_distribution_tenants}
\alias{cloudfront_list_distribution_tenants}
\title{Lists the distribution tenants in your Amazon Web Services account}
\usage{
cloudfront_list_distribution_tenants(
  AssociationFilter = NULL,
  Marker = NULL,
  MaxItems = NULL
)
}
\arguments{
\item{AssociationFilter}{}

\item{Marker}{The marker for the next set of results.}

\item{MaxItems}{The maximum number of distribution tenants to return.}
}
\description{
Lists the distribution tenants in your Amazon Web Services account.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_list_distribution_tenants/} for full documentation.
}
\keyword{internal}
