% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_list_distributions_by_anycast_ip_list_id}
\alias{cloudfront_list_distributions_by_anycast_ip_list_id}
\title{Lists the distributions in your account that are associated with the
specified AnycastIpListId}
\usage{
cloudfront_list_distributions_by_anycast_ip_list_id(
  Marker = NULL,
  MaxItems = NULL,
  AnycastIpListId
)
}
\arguments{
\item{Marker}{Use this field when paginating results to indicate where to begin in
your list. The response includes items in the list that occur after the
marker. To get the next page of the list, set this field's value to the
value of \code{NextMarker} from the current page's response.}

\item{MaxItems}{The maximum number of distributions that you want returned in the
response.}

\item{AnycastIpListId}{[required] The ID of the Anycast static IP list.}
}
\description{
Lists the distributions in your account that are associated with the specified \code{AnycastIpListId}.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_list_distributions_by_anycast_ip_list_id/} for full documentation.
}
\keyword{internal}
