% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_list_distributions_by_connection_mode}
\alias{cloudfront_list_distributions_by_connection_mode}
\title{Lists the distributions by the connection mode that you specify}
\usage{
cloudfront_list_distributions_by_connection_mode(
  Marker = NULL,
  MaxItems = NULL,
  ConnectionMode
)
}
\arguments{
\item{Marker}{The marker for the next set of distributions to retrieve.}

\item{MaxItems}{The maximum number of distributions to return.}

\item{ConnectionMode}{[required] This field specifies whether the connection mode is through a standard
distribution (direct) or a multi-tenant distribution with distribution
tenants (tenant-only).}
}
\description{
Lists the distributions by the connection mode that you specify.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_list_distributions_by_connection_mode/} for full documentation.
}
\keyword{internal}
