% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_list_distributions_by_key_group}
\alias{cloudfront_list_distributions_by_key_group}
\title{Gets a list of distribution IDs for distributions that have a cache
behavior that references the specified key group}
\usage{
cloudfront_list_distributions_by_key_group(
  Marker = NULL,
  MaxItems = NULL,
  KeyGroupId
)
}
\arguments{
\item{Marker}{Use this field when paginating results to indicate where to begin in
your list of distribution IDs. The response includes distribution IDs in
the list that occur after the marker. To get the next page of the list,
set this field's value to the value of \code{NextMarker} from the current
page's response.}

\item{MaxItems}{The maximum number of distribution IDs that you want in the response.}

\item{KeyGroupId}{[required] The ID of the key group whose associated distribution IDs you are
listing.}
}
\description{
Gets a list of distribution IDs for distributions that have a cache behavior that references the specified key group.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_list_distributions_by_key_group/} for full documentation.
}
\keyword{internal}
