% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_list_distributions_by_owned_resource}
\alias{cloudfront_list_distributions_by_owned_resource}
\title{Lists the CloudFront distributions that are associated with the
specified resource that you own}
\usage{
cloudfront_list_distributions_by_owned_resource(
  ResourceArn,
  Marker = NULL,
  MaxItems = NULL
)
}
\arguments{
\item{ResourceArn}{[required] The ARN of the CloudFront resource that you've shared with other Amazon
Web Services accounts.}

\item{Marker}{Use this field when paginating results to indicate where to begin in
your list of distributions. The response includes distributions in the
list that occur after the marker. To get the next page of the list, set
this field's value to the value of \code{NextMarker} from the current page's
response.}

\item{MaxItems}{The maximum number of distributions to return.}
}
\description{
Lists the CloudFront distributions that are associated with the specified resource that you own.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_list_distributions_by_owned_resource/} for full documentation.
}
\keyword{internal}
