% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_list_distributions_by_trust_store}
\alias{cloudfront_list_distributions_by_trust_store}
\title{Lists distributions by trust store}
\usage{
cloudfront_list_distributions_by_trust_store(
  TrustStoreIdentifier,
  Marker = NULL,
  MaxItems = NULL
)
}
\arguments{
\item{TrustStoreIdentifier}{[required] The distributions by trust store identifier.}

\item{Marker}{Use this field when paginating results to indicate where to begin in
your list. The response includes items in the list that occur after the
marker. To get the next page of the list, set this field's value to the
value of \code{NextMarker} from the current page's response.}

\item{MaxItems}{The maximum number of distributions that you want returned in the
response.}
}
\description{
Lists distributions by trust store.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_list_distributions_by_trust_store/} for full documentation.
}
\keyword{internal}
