% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_list_distributions_by_vpc_origin_id}
\alias{cloudfront_list_distributions_by_vpc_origin_id}
\title{List CloudFront distributions by their VPC origin ID}
\usage{
cloudfront_list_distributions_by_vpc_origin_id(
  Marker = NULL,
  MaxItems = NULL,
  VpcOriginId
)
}
\arguments{
\item{Marker}{The marker associated with the VPC origin distributions list.}

\item{MaxItems}{The maximum number of items included in the list.}

\item{VpcOriginId}{[required] The VPC origin ID.}
}
\description{
List CloudFront distributions by their VPC origin ID.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_list_distributions_by_vpc_origin_id/} for full documentation.
}
\keyword{internal}
