% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_list_domain_conflicts}
\alias{cloudfront_list_domain_conflicts}
\title{Lists existing domain associations that conflict with the domain that
you specify}
\usage{
cloudfront_list_domain_conflicts(
  Domain,
  DomainControlValidationResource,
  MaxItems = NULL,
  Marker = NULL
)
}
\arguments{
\item{Domain}{[required] The domain to check for conflicts.}

\item{DomainControlValidationResource}{[required] The distribution resource identifier. This can be the distribution or
distribution tenant that has a valid certificate, which covers the
domain that you specify.}

\item{MaxItems}{The maximum number of domain conflicts to return.}

\item{Marker}{The marker for the next set of domain conflicts.}
}
\description{
Lists existing domain associations that conflict with the domain that you specify.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_list_domain_conflicts/} for full documentation.
}
\keyword{internal}
