% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_list_functions}
\alias{cloudfront_list_functions}
\title{Gets a list of all CloudFront functions in your AWS account}
\usage{
cloudfront_list_functions(Marker, MaxItems, Stage)
}
\arguments{
\item{Marker}{Use this field when paginating results to indicate where to begin in
your list of functions. The response includes functions in the list that
occur after the marker. To get the next page of the list, set this
field’s value to the value of \code{NextMarker} from the current page’s
response.}

\item{MaxItems}{The maximum number of functions that you want in the response.}

\item{Stage}{An optional filter to return only the functions that are in the
specified stage, either \code{DEVELOPMENT} or \code{LIVE}.}
}
\value{
A list with the following syntax:\preformatted{list(
  FunctionList = list(
    NextMarker = "string",
    MaxItems = 123,
    Quantity = 123,
    Items = list(
      list(
        Name = "string",
        Status = "string",
        FunctionConfig = list(
          Comment = "string",
          Runtime = "cloudfront-js-1.0"
        ),
        FunctionMetadata = list(
          FunctionARN = "string",
          Stage = "DEVELOPMENT"|"LIVE",
          CreatedTime = as.POSIXct(
            "2015-01-01"
          ),
          LastModifiedTime = as.POSIXct(
            "2015-01-01"
          )
        )
      )
    )
  )
)
}
}
\description{
Gets a list of all CloudFront functions in your AWS account.

You can optionally apply a filter to return only the functions that are
in the specified stage, either \code{DEVELOPMENT} or \code{LIVE}.

You can optionally specify the maximum number of items to receive in the
response. If the total number of items in the list exceeds the maximum
that you specify, or the default maximum, the response is paginated. To
get the next page of items, send a subsequent request that specifies the
\code{NextMarker} value from the current response as the \code{Marker} value in
the subsequent request.
}
\section{Request syntax}{
\preformatted{svc$list_functions(
  Marker = "string",
  MaxItems = "string",
  Stage = "DEVELOPMENT"|"LIVE"
)
}
}

\keyword{internal}
