% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_list_invalidations_for_distribution_tenant}
\alias{cloudfront_list_invalidations_for_distribution_tenant}
\title{Lists the invalidations for a distribution tenant}
\usage{
cloudfront_list_invalidations_for_distribution_tenant(
  Id,
  Marker = NULL,
  MaxItems = NULL
)
}
\arguments{
\item{Id}{[required] The ID of the distribution tenant.}

\item{Marker}{Use this parameter when paginating results to indicate where to begin in
your list of invalidation batches. Because the results are returned in
decreasing order from most recent to oldest, the most recent results are
on the first page, the second page will contain earlier results, and so
on. To get the next page of results, set \code{Marker} to the value of the
\code{NextMarker} from the current page's response. This value is the same as
the ID of the last invalidation batch on that page.}

\item{MaxItems}{The maximum number of invalidations to return for the distribution
tenant.}
}
\description{
Lists the invalidations for a distribution tenant.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_list_invalidations_for_distribution_tenant/} for full documentation.
}
\keyword{internal}
