% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_list_origin_access_controls}
\alias{cloudfront_list_origin_access_controls}
\title{Gets the list of CloudFront origin access controls (OACs) in this Amazon
Web Services account}
\usage{
cloudfront_list_origin_access_controls(Marker = NULL, MaxItems = NULL)
}
\arguments{
\item{Marker}{Use this field when paginating results to indicate where to begin in
your list of origin access controls. The response includes the items in
the list that occur after the marker. To get the next page of the list,
set this field's value to the value of \code{NextMarker} from the current
page's response.}

\item{MaxItems}{The maximum number of origin access controls that you want in the
response.}
}
\description{
Gets the list of CloudFront origin access controls (OACs) in this Amazon Web Services account.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_list_origin_access_controls/} for full documentation.
}
\keyword{internal}
