% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_list_response_headers_policies}
\alias{cloudfront_list_response_headers_policies}
\title{Gets a list of response headers policies}
\usage{
cloudfront_list_response_headers_policies(
  Type = NULL,
  Marker = NULL,
  MaxItems = NULL
)
}
\arguments{
\item{Type}{A filter to get only the specified kind of response headers policies.
Valid values are:
\itemize{
\item \code{managed} – Gets only the managed policies created by Amazon Web
Services.
\item \code{custom} – Gets only the custom policies created in your Amazon Web
Services account.
}}

\item{Marker}{Use this field when paginating results to indicate where to begin in
your list of response headers policies. The response includes response
headers policies in the list that occur after the marker. To get the
next page of the list, set this field's value to the value of
\code{NextMarker} from the current page's response.}

\item{MaxItems}{The maximum number of response headers policies that you want to get in
the response.}
}
\description{
Gets a list of response headers policies.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_list_response_headers_policies/} for full documentation.
}
\keyword{internal}
