% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_list_tags_for_resource}
\alias{cloudfront_list_tags_for_resource}
\title{List tags for a CloudFront resource}
\usage{
cloudfront_list_tags_for_resource(Resource)
}
\arguments{
\item{Resource}{[required] An ARN of a CloudFront resource.}
}
\description{
List tags for a CloudFront resource. For more information, see \href{https://docs.aws.amazon.com/AmazonCloudFront/latest/DeveloperGuide/tagging.html}{Tagging a distribution} in the \emph{Amazon CloudFront Developer Guide}.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
