% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_list_trust_stores}
\alias{cloudfront_list_trust_stores}
\title{Lists trust stores}
\usage{
cloudfront_list_trust_stores(Marker = NULL, MaxItems = NULL)
}
\arguments{
\item{Marker}{Use this field when paginating results to indicate where to begin in
your list. The response includes items in the list that occur after the
marker. To get the next page of the list, set this field's value to the
value of \code{NextMarker} from the current page's response.}

\item{MaxItems}{The maximum number of trust stores that you want returned in the
response.}
}
\description{
Lists trust stores.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_list_trust_stores/} for full documentation.
}
\keyword{internal}
