% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_list_vpc_origins}
\alias{cloudfront_list_vpc_origins}
\title{List the CloudFront VPC origins in your account}
\usage{
cloudfront_list_vpc_origins(Marker = NULL, MaxItems = NULL)
}
\arguments{
\item{Marker}{The marker associated with the VPC origins list.}

\item{MaxItems}{The maximum number of items included in the list.}
}
\description{
List the CloudFront VPC origins in your account.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_list_vpc_origins/} for full documentation.
}
\keyword{internal}
