% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_publish_connection_function}
\alias{cloudfront_publish_connection_function}
\title{Publishes a connection function}
\usage{
cloudfront_publish_connection_function(Id, IfMatch)
}
\arguments{
\item{Id}{[required] The connection function ID.}

\item{IfMatch}{[required] The current version (\code{ETag} value) of the connection function.}
}
\description{
Publishes a connection function.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_publish_connection_function/} for full documentation.
}
\keyword{internal}
