% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_put_resource_policy}
\alias{cloudfront_put_resource_policy}
\title{Creates a resource control policy for a given CloudFront resource}
\usage{
cloudfront_put_resource_policy(ResourceArn, PolicyDocument)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the CloudFront resource for which the
policy is being created.}

\item{PolicyDocument}{[required] The JSON-formatted resource policy to create.}
}
\description{
Creates a resource control policy for a given CloudFront resource.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_put_resource_policy/} for full documentation.
}
\keyword{internal}
