% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_tag_resource}
\alias{cloudfront_tag_resource}
\title{Add tags to a CloudFront resource}
\usage{
cloudfront_tag_resource(Resource, Tags)
}
\arguments{
\item{Resource}{[required] An ARN of a CloudFront resource.}

\item{Tags}{[required] A complex type that contains zero or more \code{Tag} elements.}
}
\description{
Add tags to a CloudFront resource. For more information, see \href{https://docs.aws.amazon.com/AmazonCloudFront/latest/DeveloperGuide/tagging.html}{Tagging a distribution} in the \emph{Amazon CloudFront Developer Guide}.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_tag_resource/} for full documentation.
}
\keyword{internal}
