% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_test_function}
\alias{cloudfront_test_function}
\title{Tests a CloudFront function}
\usage{
cloudfront_test_function(Name, IfMatch, Stage, EventObject)
}
\arguments{
\item{Name}{[required] The name of the function that you are testing.}

\item{IfMatch}{[required] The current version (\code{ETag} value) of the function that you are testing,
which you can get using
\code{\link[=cloudfront_describe_function]{describe_function}}.}

\item{Stage}{The stage of the function that you are testing, either \code{DEVELOPMENT} or
\code{LIVE}.}

\item{EventObject}{[required] The event object to test the function with. For more information about
the structure of the event object, see \href{https://docs.aws.amazon.com/AmazonCloudFront/latest/DeveloperGuide/managing-functions.html#test-function}{Testing functions}
in the \emph{Amazon CloudFront Developer Guide}.}
}
\value{
A list with the following syntax:\preformatted{list(
  TestResult = list(
    FunctionSummary = list(
      Name = "string",
      Status = "string",
      FunctionConfig = list(
        Comment = "string",
        Runtime = "cloudfront-js-1.0"
      ),
      FunctionMetadata = list(
        FunctionARN = "string",
        Stage = "DEVELOPMENT"|"LIVE",
        CreatedTime = as.POSIXct(
          "2015-01-01"
        ),
        LastModifiedTime = as.POSIXct(
          "2015-01-01"
        )
      )
    ),
    ComputeUtilization = "string",
    FunctionExecutionLogs = list(
      "string"
    ),
    FunctionErrorMessage = "string",
    FunctionOutput = "string"
  )
)
}
}
\description{
Tests a CloudFront function.

To test a function, you provide an \emph{event object} that represents an
HTTP request or response that your CloudFront distribution could receive
in production. CloudFront runs the function, passing it the event object
that you provided, and returns the function’s result (the modified event
object) in the response. The response also contains function logs and
error messages, if any exist. For more information about testing
functions, see \href{https://docs.aws.amazon.com/AmazonCloudFront/latest/DeveloperGuide/managing-functions.html#test-function}{Testing functions}
in the \emph{Amazon CloudFront Developer Guide}.

To test a function, you provide the function’s name and version (\code{ETag}
value) along with the event object. To get the function’s name and
version, you can use \code{\link[=cloudfront_list_functions]{list_functions}} and
\code{\link[=cloudfront_describe_function]{describe_function}}.
}
\section{Request syntax}{
\preformatted{svc$test_function(
  Name = "string",
  IfMatch = "string",
  Stage = "DEVELOPMENT"|"LIVE",
  EventObject = raw
)
}
}

\keyword{internal}
