% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_untag_resource}
\alias{cloudfront_untag_resource}
\title{Remove tags from a CloudFront resource}
\usage{
cloudfront_untag_resource(Resource, TagKeys)
}
\arguments{
\item{Resource}{[required] An ARN of a CloudFront resource.}

\item{TagKeys}{[required] A complex type that contains zero or more \code{Tag} key elements.}
}
\description{
Remove tags from a CloudFront resource. For more information, see \href{https://docs.aws.amazon.com/AmazonCloudFront/latest/DeveloperGuide/tagging.html}{Tagging a distribution} in the \emph{Amazon CloudFront Developer Guide}.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_untag_resource/} for full documentation.
}
\keyword{internal}
