% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_update_anycast_ip_list}
\alias{cloudfront_update_anycast_ip_list}
\title{Updates an Anycast static IP list}
\usage{
cloudfront_update_anycast_ip_list(Id, IpAddressType = NULL, IfMatch)
}
\arguments{
\item{Id}{[required] The ID of the Anycast static IP list.}

\item{IpAddressType}{The IP address type for the Anycast static IP list. You can specify one
of the following options:
\itemize{
\item \code{ipv4} only
\item \code{ipv6} only
\item \code{dualstack} - Allocate a list of both IPv4 and IPv6 addresses
}}

\item{IfMatch}{[required] The current version (ETag value) of the Anycast static IP list that you
are updating.}
}
\description{
Updates an Anycast static IP list.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_update_anycast_ip_list/} for full documentation.
}
\keyword{internal}
