% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_update_cloud_front_origin_access_identity}
\alias{cloudfront_update_cloud_front_origin_access_identity}
\title{Update an origin access identity}
\usage{
cloudfront_update_cloud_front_origin_access_identity(
  CloudFrontOriginAccessIdentityConfig,
  Id,
  IfMatch = NULL
)
}
\arguments{
\item{CloudFrontOriginAccessIdentityConfig}{[required] The identity's configuration information.}

\item{Id}{[required] The identity's id.}

\item{IfMatch}{The value of the \code{ETag} header that you received when retrieving the
identity's configuration. For example: \code{E2QWRUHAPOMQZL}.}
}
\description{
Update an origin access identity.

See \url{https://paws-r.github.io/docs/cloudfront/update_cloud_front_origin_access_identity.html} for full documentation.
}
\keyword{internal}
