% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_update_connection_function}
\alias{cloudfront_update_connection_function}
\title{Updates a connection function}
\usage{
cloudfront_update_connection_function(
  Id,
  IfMatch,
  ConnectionFunctionConfig,
  ConnectionFunctionCode
)
}
\arguments{
\item{Id}{[required] The connection function ID.}

\item{IfMatch}{[required] The current version (\code{ETag} value) of the connection function you are
updating.}

\item{ConnectionFunctionConfig}{[required]}

\item{ConnectionFunctionCode}{[required] The connection function code.}
}
\description{
Updates a connection function.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_update_connection_function/} for full documentation.
}
\keyword{internal}
