% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_update_connection_group}
\alias{cloudfront_update_connection_group}
\title{Updates a connection group}
\usage{
cloudfront_update_connection_group(
  Id,
  Ipv6Enabled = NULL,
  IfMatch,
  AnycastIpListId = NULL,
  Enabled = NULL
)
}
\arguments{
\item{Id}{[required] The ID of the connection group.}

\item{Ipv6Enabled}{Enable IPv6 for the connection group. For more information, see \href{https://docs.aws.amazon.com/AmazonCloudFront/latest/DeveloperGuide/distribution-web-values-specify.html#DownloadDistValuesEnableIPv6}{Enable IPv6}
in the \emph{Amazon CloudFront Developer Guide}.}

\item{IfMatch}{[required] The value of the \code{ETag} header that you received when retrieving the
connection group that you're updating.}

\item{AnycastIpListId}{The ID of the Anycast static IP list.}

\item{Enabled}{Whether the connection group is enabled.}
}
\description{
Updates a connection group.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_update_connection_group/} for full documentation.
}
\keyword{internal}
