% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_update_domain_association}
\alias{cloudfront_update_domain_association}
\title{Moves a domain from its current distribution or distribution tenant to
another one}
\usage{
cloudfront_update_domain_association(Domain, TargetResource, IfMatch = NULL)
}
\arguments{
\item{Domain}{[required] The domain to update.}

\item{TargetResource}{[required] The target distribution resource for the domain. You can specify either
\code{DistributionId} or \code{DistributionTenantId}, but not both.}

\item{IfMatch}{The value of the \code{ETag} identifier for the distribution or distribution
tenant that will be associated with the domain.}
}
\description{
Moves a domain from its current distribution or distribution tenant to another one.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_update_domain_association/} for full documentation.
}
\keyword{internal}
