% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_update_function}
\alias{cloudfront_update_function}
\title{Updates a CloudFront function}
\usage{
cloudfront_update_function(Name, IfMatch, FunctionConfig, FunctionCode)
}
\arguments{
\item{Name}{[required] The name of the function that you are updating.}

\item{IfMatch}{[required] The current version (\code{ETag} value) of the function that you are
updating, which you can get using
\code{\link[=cloudfront_describe_function]{describe_function}}.}

\item{FunctionConfig}{[required] Configuration information about the function.}

\item{FunctionCode}{[required] The function code. For more information about writing a CloudFront
function, see \href{https://docs.aws.amazon.com/AmazonCloudFront/latest/DeveloperGuide/writing-function-code.html}{Writing function code for CloudFront Functions}
in the \emph{Amazon CloudFront Developer Guide}.}
}
\description{
Updates a CloudFront function.

See \url{https://paws-r.github.io/docs/cloudfront/update_function.html} for full documentation.
}
\keyword{internal}
