% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_update_key_group}
\alias{cloudfront_update_key_group}
\title{Updates a key group}
\usage{
cloudfront_update_key_group(KeyGroupConfig, Id, IfMatch = NULL)
}
\arguments{
\item{KeyGroupConfig}{[required] The key group configuration.}

\item{Id}{[required] The identifier of the key group that you are updating.}

\item{IfMatch}{The version of the key group that you are updating. The version is the
key group's \code{ETag} value.}
}
\description{
Updates a key group.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_update_key_group/} for full documentation.
}
\keyword{internal}
