% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_update_origin_access_control}
\alias{cloudfront_update_origin_access_control}
\title{Updates a CloudFront origin access control}
\usage{
cloudfront_update_origin_access_control(
  OriginAccessControlConfig,
  Id,
  IfMatch = NULL
)
}
\arguments{
\item{OriginAccessControlConfig}{[required] An origin access control.}

\item{Id}{[required] The unique identifier of the origin access control that you are
updating.}

\item{IfMatch}{The current version (\code{ETag} value) of the origin access control that you
are updating.}
}
\description{
Updates a CloudFront origin access control.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_update_origin_access_control/} for full documentation.
}
\keyword{internal}
