% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_update_origin_request_policy}
\alias{cloudfront_update_origin_request_policy}
\title{Updates an origin request policy configuration}
\usage{
cloudfront_update_origin_request_policy(
  OriginRequestPolicyConfig,
  Id,
  IfMatch = NULL
)
}
\arguments{
\item{OriginRequestPolicyConfig}{[required] An origin request policy configuration.}

\item{Id}{[required] The unique identifier for the origin request policy that you are
updating. The identifier is returned in a cache behavior's
\code{OriginRequestPolicyId} field in the response to
\code{\link[=cloudfront_get_distribution_config]{get_distribution_config}}.}

\item{IfMatch}{The version of the origin request policy that you are updating. The
version is returned in the origin request policy's \code{ETag} field in the
response to
\code{\link[=cloudfront_get_origin_request_policy_config]{get_origin_request_policy_config}}.}
}
\description{
Updates an origin request policy configuration.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_update_origin_request_policy/} for full documentation.
}
\keyword{internal}
