% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_update_realtime_log_config}
\alias{cloudfront_update_realtime_log_config}
\title{Updates a real-time log configuration}
\usage{
cloudfront_update_realtime_log_config(
  EndPoints = NULL,
  Fields = NULL,
  Name = NULL,
  ARN = NULL,
  SamplingRate = NULL
)
}
\arguments{
\item{EndPoints}{Contains information about the Amazon Kinesis data stream where you are
sending real-time log data.}

\item{Fields}{A list of fields to include in each real-time log record.

For more information about fields, see \href{https://docs.aws.amazon.com/AmazonCloudFront/latest/DeveloperGuide/real-time-logs.html#understand-real-time-log-config-fields}{Real-time log configuration fields}
in the \emph{Amazon CloudFront Developer Guide}.}

\item{Name}{The name for this real-time log configuration.}

\item{ARN}{The Amazon Resource Name (ARN) for this real-time log configuration.}

\item{SamplingRate}{The sampling rate for this real-time log configuration. The sampling
rate determines the percentage of viewer requests that are represented
in the real-time log data. You must provide an integer between 1 and
100, inclusive.}
}
\description{
Updates a real-time log configuration.

See \url{https://paws-r.github.io/docs/cloudfront/update_realtime_log_config.html} for full documentation.
}
\keyword{internal}
