% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_update_response_headers_policy}
\alias{cloudfront_update_response_headers_policy}
\title{Updates a response headers policy}
\usage{
cloudfront_update_response_headers_policy(
  ResponseHeadersPolicyConfig,
  Id,
  IfMatch = NULL
)
}
\arguments{
\item{ResponseHeadersPolicyConfig}{[required] A response headers policy configuration.}

\item{Id}{[required] The identifier for the response headers policy that you are updating.}

\item{IfMatch}{The version of the response headers policy that you are updating.

The version is returned in the cache policy's \code{ETag} field in the
response to
\code{\link[=cloudfront_get_response_headers_policy_config]{get_response_headers_policy_config}}.}
}
\description{
Updates a response headers policy.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_update_response_headers_policy/} for full documentation.
}
\keyword{internal}
