% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_update_streaming_distribution}
\alias{cloudfront_update_streaming_distribution}
\title{Update a streaming distribution}
\usage{
cloudfront_update_streaming_distribution(
  StreamingDistributionConfig,
  Id,
  IfMatch = NULL
)
}
\arguments{
\item{StreamingDistributionConfig}{[required] The streaming distribution's configuration information.}

\item{Id}{[required] The streaming distribution's id.}

\item{IfMatch}{The value of the \code{ETag} header that you received when retrieving the
streaming distribution's configuration. For example: \code{E2QWRUHAPOMQZL}.}
}
\description{
Update a streaming distribution.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_update_streaming_distribution/} for full documentation.
}
\keyword{internal}
