% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_update_vpc_origin}
\alias{cloudfront_update_vpc_origin}
\title{Update an Amazon CloudFront VPC origin in your account}
\usage{
cloudfront_update_vpc_origin(VpcOriginEndpointConfig, Id, IfMatch)
}
\arguments{
\item{VpcOriginEndpointConfig}{[required] The VPC origin endpoint configuration.}

\item{Id}{[required] The VPC origin ID.}

\item{IfMatch}{[required] The VPC origin to update, if a match occurs.}
}
\description{
Update an Amazon CloudFront VPC origin in your account.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_update_vpc_origin/} for full documentation.
}
\keyword{internal}
