% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_verify_dns_configuration}
\alias{cloudfront_verify_dns_configuration}
\title{Verify the DNS configuration for your domain names}
\usage{
cloudfront_verify_dns_configuration(Domain = NULL, Identifier)
}
\arguments{
\item{Domain}{The domain name that you're verifying.}

\item{Identifier}{[required] The identifier of the distribution tenant. You can specify the ARN, ID,
or name of the distribution tenant.}
}
\description{
Verify the DNS configuration for your domain names. This API operation checks whether your domain name points to the correct routing endpoint of the connection group, such as d111111abcdef8.cloudfront.net. You can use this API operation to troubleshoot and resolve DNS configuration issues.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_verify_dns_configuration/} for full documentation.
}
\keyword{internal}
