% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfrontkeyvaluestore_operations.R
\name{cloudfrontkeyvaluestore_delete_key}
\alias{cloudfrontkeyvaluestore_delete_key}
\title{Deletes the key value pair specified by the key}
\usage{
cloudfrontkeyvaluestore_delete_key(KvsARN, Key, IfMatch)
}
\arguments{
\item{KvsARN}{[required] The Amazon Resource Name (ARN) of the Key Value Store.}

\item{Key}{[required] The key to delete.}

\item{IfMatch}{[required] The current version (ETag) of the Key Value Store that you are deleting
keys from, which you can get using DescribeKeyValueStore.}
}
\description{
Deletes the key value pair specified by the key.

See \url{https://www.paws-r-sdk.com/docs/cloudfrontkeyvaluestore_delete_key/} for full documentation.
}
\keyword{internal}
