% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfrontkeyvaluestore_operations.R
\name{cloudfrontkeyvaluestore_describe_key_value_store}
\alias{cloudfrontkeyvaluestore_describe_key_value_store}
\title{Returns metadata information about Key Value Store}
\usage{
cloudfrontkeyvaluestore_describe_key_value_store(KvsARN)
}
\arguments{
\item{KvsARN}{[required] The Amazon Resource Name (ARN) of the Key Value Store.}
}
\description{
Returns metadata information about Key Value Store.

See \url{https://www.paws-r-sdk.com/docs/cloudfrontkeyvaluestore_describe_key_value_store/} for full documentation.
}
\keyword{internal}
