% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfrontkeyvaluestore_operations.R
\name{cloudfrontkeyvaluestore_get_key}
\alias{cloudfrontkeyvaluestore_get_key}
\title{Returns a key value pair}
\usage{
cloudfrontkeyvaluestore_get_key(KvsARN, Key)
}
\arguments{
\item{KvsARN}{[required] The Amazon Resource Name (ARN) of the Key Value Store.}

\item{Key}{[required] The key to get.}
}
\description{
Returns a key value pair.

See \url{https://www.paws-r-sdk.com/docs/cloudfrontkeyvaluestore_get_key/} for full documentation.
}
\keyword{internal}
