% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfrontkeyvaluestore_operations.R
\name{cloudfrontkeyvaluestore_list_keys}
\alias{cloudfrontkeyvaluestore_list_keys}
\title{Returns a list of key value pairs}
\usage{
cloudfrontkeyvaluestore_list_keys(KvsARN, NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{KvsARN}{[required] The Amazon Resource Name (ARN) of the Key Value Store.}

\item{NextToken}{If nextToken is returned in the response, there are more results
available. Make the next call using the returned token to retrieve the
next page.}

\item{MaxResults}{Maximum number of results that are returned per call. The default is 10
and maximum allowed page is 50.}
}
\description{
Returns a list of key value pairs.

See \url{https://www.paws-r-sdk.com/docs/cloudfrontkeyvaluestore_list_keys/} for full documentation.
}
\keyword{internal}
