% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfrontkeyvaluestore_operations.R
\name{cloudfrontkeyvaluestore_update_keys}
\alias{cloudfrontkeyvaluestore_update_keys}
\title{Puts or Deletes multiple key value pairs in a single, all-or-nothing
operation}
\usage{
cloudfrontkeyvaluestore_update_keys(
  KvsARN,
  IfMatch,
  Puts = NULL,
  Deletes = NULL
)
}
\arguments{
\item{KvsARN}{[required] The Amazon Resource Name (ARN) of the Key Value Store.}

\item{IfMatch}{[required] The current version (ETag) of the Key Value Store that you are updating
keys of, which you can get using DescribeKeyValueStore.}

\item{Puts}{List of key value pairs to put.}

\item{Deletes}{List of keys to delete.}
}
\description{
Puts or Deletes multiple key value pairs in a single, all-or-nothing operation.

See \url{https://www.paws-r-sdk.com/docs/cloudfrontkeyvaluestore_update_keys/} for full documentation.
}
\keyword{internal}
