% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_allocate_hosted_connection}
\alias{directconnect_allocate_hosted_connection}
\title{Creates a hosted connection on the specified interconnect or a link
aggregation group (LAG) of interconnects}
\usage{
directconnect_allocate_hosted_connection(connectionId, ownerAccount,
  bandwidth, connectionName, vlan, tags)
}
\arguments{
\item{connectionId}{[required] The ID of the interconnect or LAG.}

\item{ownerAccount}{[required] The ID of the AWS account ID of the customer for the connection.}

\item{bandwidth}{[required] The bandwidth of the connection. The possible values are 50Mbps,
100Mbps, 200Mbps, 300Mbps, 400Mbps, 500Mbps, 1Gbps, 2Gbps, 5Gbps, and
10Gbps. Note that only those AWS Direct Connect Partners who have met
specific requirements are allowed to create a 1Gbps, 2Gbps, 5Gbps or
10Gbps hosted connection.}

\item{connectionName}{[required] The name of the hosted connection.}

\item{vlan}{[required] The dedicated VLAN provisioned to the hosted connection.}

\item{tags}{The tags associated with the connection.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ownerAccount = "string",
  connectionId = "string",
  connectionName = "string",
  connectionState = "ordering"|"requested"|"pending"|"available"|"down"|"deleting"|"deleted"|"rejected"|"unknown",
  region = "string",
  location = "string",
  bandwidth = "string",
  vlan = 123,
  partnerName = "string",
  loaIssueTime = as.POSIXct(
    "2015-01-01"
  ),
  lagId = "string",
  awsDevice = "string",
  jumboFrameCapable = TRUE|FALSE,
  awsDeviceV2 = "string",
  hasLogicalRedundancy = "unknown"|"yes"|"no",
  tags = list(
    list(
      key = "string",
      value = "string"
    )
  ),
  providerName = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Creates a hosted connection on the specified interconnect or a link
aggregation group (LAG) of interconnects.

Allocates a VLAN number and a specified amount of capacity (bandwidth)
for use by a hosted connection on the specified interconnect or LAG of
interconnects. AWS polices the hosted connection for the specified
capacity and the AWS Direct Connect Partner must also police the hosted
connection for the specified capacity.

Intended for use by AWS Direct Connect Partners only.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$allocate_hosted_connection(
  connectionId = "string",
  ownerAccount = "string",
  bandwidth = "string",
  connectionName = "string",
  vlan = 123,
  tags = list(
    list(
      key = "string",
      value = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
