% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_allocate_transit_virtual_interface}
\alias{directconnect_allocate_transit_virtual_interface}
\title{Provisions a transit virtual interface to be owned by the specified
Amazon Web Services account}
\usage{
directconnect_allocate_transit_virtual_interface(
  connectionId,
  ownerAccount,
  newTransitVirtualInterfaceAllocation
)
}
\arguments{
\item{connectionId}{[required] The ID of the connection on which the transit virtual interface is
provisioned.}

\item{ownerAccount}{[required] The ID of the Amazon Web Services account that owns the transit virtual
interface.}

\item{newTransitVirtualInterfaceAllocation}{[required] Information about the transit virtual interface.}
}
\description{
Provisions a transit virtual interface to be owned by the specified Amazon Web Services account. Use this type of interface to connect a transit gateway to your Direct Connect gateway.

See \url{https://www.paws-r-sdk.com/docs/directconnect_allocate_transit_virtual_interface/} for full documentation.
}
\keyword{internal}
