% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_associate_connections_to_resiliency_group}
\alias{directconnect_associate_connections_to_resiliency_group}
\title{Associates one or more connections with the specified resiliency group}
\usage{
directconnect_associate_connections_to_resiliency_group(
  connectionIdentifiers,
  resiliencyGroupId,
  clientToken = NULL
)
}
\arguments{
\item{connectionIdentifiers}{[required] The IDs or ARNs of the connections to associate with the resiliency group.}

\item{resiliencyGroupId}{[required] The ID of the resiliency group.}

\item{clientToken}{A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.}
}
\description{
Associates one or more connections with the specified resiliency group. This operation is atomic: either all of the specified connections are associated, or the operation fails and no changes are made.

See \url{https://www.paws-r-sdk.com/docs/directconnect_associate_connections_to_resiliency_group/} for full documentation.
}
\keyword{internal}
