% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_associate_mac_sec_key}
\alias{directconnect_associate_mac_sec_key}
\title{Associates a MAC Security (MACsec) Connection Key Name (CKN)/
Connectivity Association Key (CAK) pair with an AWS Direct Connect
dedicated connection}
\usage{
directconnect_associate_mac_sec_key(connectionId, secretARN, ckn, cak)
}
\arguments{
\item{connectionId}{[required] The ID of the dedicated connection (dxcon-xxxx), or the ID of the LAG
(dxlag-xxxx).

You can use \code{\link[=directconnect_describe_connections]{describe_connections}}
or \code{\link[=directconnect_describe_lags]{describe_lags}} to retrieve connection
ID.}

\item{secretARN}{The Amazon Resource Name (ARN) of the MAC Security (MACsec) secret key
to associate with the dedicated connection.

You can use \code{\link[=directconnect_describe_connections]{describe_connections}}
or \code{\link[=directconnect_describe_lags]{describe_lags}} to retrieve the MAC
Security (MACsec) secret key.

If you use this request parameter, you do not use the \code{ckn} and \code{cak}
request parameters.}

\item{ckn}{The MAC Security (MACsec) CKN to associate with the dedicated
connection.

You can create the CKN/CAK pair using an industry standard tool.

The valid values are 64 hexadecimal characters (0-9, A-E).

If you use this request parameter, you must use the \code{cak} request
parameter and not use the \code{secretARN} request parameter.}

\item{cak}{The MAC Security (MACsec) CAK to associate with the dedicated
connection.

You can create the CKN/CAK pair using an industry standard tool.

The valid values are 64 hexadecimal characters (0-9, A-E).

If you use this request parameter, you must use the \code{ckn} request
parameter and not use the \code{secretARN} request parameter.}
}
\value{
A list with the following syntax:\preformatted{list(
  connectionId = "string",
  macSecKeys = list(
    list(
      secretARN = "string",
      ckn = "string",
      state = "string",
      startOn = "string"
    )
  )
)
}
}
\description{
Associates a MAC Security (MACsec) Connection Key Name (CKN)/
Connectivity Association Key (CAK) pair with an AWS Direct Connect
dedicated connection.

You must supply either the \verb{secretARN,} or the CKN/CAK (\code{ckn} and \code{cak})
pair in the request.

For information about MAC Security (MACsec) key considerations, see
\href{https://docs.aws.amazon.com/directconnect/latest/UserGuide/direct-connect-mac-sec-getting-started.html#mac-sec-key-consideration}{MACsec pre-shared CKN/CAK key considerations}
in the \emph{AWS Direct Connect User Guide}.
}
\section{Request syntax}{
\preformatted{svc$associate_mac_sec_key(
  connectionId = "string",
  secretARN = "string",
  ckn = "string",
  cak = "string"
)
}
}

\keyword{internal}
