% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_confirm_connection}
\alias{directconnect_confirm_connection}
\title{Confirms the creation of the specified hosted connection on an
interconnect}
\usage{
directconnect_confirm_connection(connectionId)
}
\arguments{
\item{connectionId}{[required] The ID of the hosted connection.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  connectionState = "ordering"|"requested"|"pending"|"available"|"down"|"deleting"|"deleted"|"rejected"|"unknown"
)
}\if{html}{\out{</div>}}
}
\description{
Confirms the creation of the specified hosted connection on an
interconnect.

Upon creation, the hosted connection is initially in the \code{Ordering}
state, and remains in this state until the owner confirms creation of
the hosted connection.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$confirm_connection(
  connectionId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
