% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_confirm_customer_agreement}
\alias{directconnect_confirm_customer_agreement}
\title{The confirmation of the terms of agreement when creating the
connection/link aggregation group (LAG)}
\usage{
directconnect_confirm_customer_agreement(agreementName = NULL)
}
\arguments{
\item{agreementName}{The name of the customer agreement.}
}
\description{
The confirmation of the terms of agreement when creating the connection/link aggregation group (LAG).

See \url{https://www.paws-r-sdk.com/docs/directconnect_confirm_customer_agreement/} for full documentation.
}
\keyword{internal}
