% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_confirm_public_virtual_interface}
\alias{directconnect_confirm_public_virtual_interface}
\title{Accepts ownership of a public virtual interface created by another
Amazon Web Services account}
\usage{
directconnect_confirm_public_virtual_interface(virtualInterfaceId)
}
\arguments{
\item{virtualInterfaceId}{[required] The ID of the virtual interface.}
}
\description{
Accepts ownership of a public virtual interface created by another Amazon Web Services account.

See \url{https://www.paws-r-sdk.com/docs/directconnect_confirm_public_virtual_interface/} for full documentation.
}
\keyword{internal}
