% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_create_public_virtual_interface}
\alias{directconnect_create_public_virtual_interface}
\title{Creates a public virtual interface}
\usage{
directconnect_create_public_virtual_interface(
  connectionId,
  newPublicVirtualInterface
)
}
\arguments{
\item{connectionId}{[required] The ID of the connection.}

\item{newPublicVirtualInterface}{[required] Information about the public virtual interface.}
}
\description{
Creates a public virtual interface. A virtual interface is the VLAN that transports Direct Connect traffic. A public virtual interface supports sending traffic to public services of Amazon Web Services such as Amazon S3.

See \url{https://paws-r.github.io/docs/directconnect/create_public_virtual_interface.html} for full documentation.
}
\keyword{internal}
