% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_create_resiliency_group}
\alias{directconnect_create_resiliency_group}
\title{Creates a resiliency group}
\usage{
directconnect_create_resiliency_group(
  resiliencyGroupName,
  intendedResiliencyModel,
  clientToken = NULL,
  tags = NULL
)
}
\arguments{
\item{resiliencyGroupName}{[required] The name of the resiliency group.}

\item{intendedResiliencyModel}{[required] The resiliency model that the resiliency group is intended to meet. The valid values are \code{maximum-resiliency}, \code{high-resiliency}, and \code{basic-resiliency}.}

\item{clientToken}{A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.}

\item{tags}{The tags to associate with the resiliency group.}
}
\description{
Creates a resiliency group. A resiliency group lets you group Direct Connect connections together and manage them as a single unit to meet a target resiliency model.

See \url{https://www.paws-r-sdk.com/docs/directconnect_create_resiliency_group/} for full documentation.
}
\keyword{internal}
