% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_create_transit_virtual_interface}
\alias{directconnect_create_transit_virtual_interface}
\title{Creates a transit virtual interface}
\usage{
directconnect_create_transit_virtual_interface(
  connectionId,
  newTransitVirtualInterface
)
}
\arguments{
\item{connectionId}{[required] The ID of the connection.}

\item{newTransitVirtualInterface}{[required] Information about the transit virtual interface.}
}
\description{
Creates a transit virtual interface. A transit virtual interface should be used to access one or more transit gateways associated with Direct Connect gateways. A transit virtual interface enables the connection of multiple VPCs attached to a transit gateway to a Direct Connect gateway.

See \url{https://paws-r.github.io/docs/directconnect/create_transit_virtual_interface.html} for full documentation.
}
\keyword{internal}
