% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_delete_bgp_peer}
\alias{directconnect_delete_bgp_peer}
\title{Deletes the specified BGP peer on the specified virtual interface with
the specified customer address and ASN}
\usage{
directconnect_delete_bgp_peer(
  virtualInterfaceId = NULL,
  asn = NULL,
  asnLong = NULL,
  customerAddress = NULL,
  bgpPeerId = NULL
)
}
\arguments{
\item{virtualInterfaceId}{The ID of the virtual interface.}

\item{asn}{The autonomous system number (ASN). The valid range is from 1 to 2147483646 for Border Gateway Protocol (BGP) configuration. If you provide a number greater than the maximum, an error is returned. Use \code{asnLong} instead.
\itemize{
\item You can use \code{asnLong} or \code{asn}, but not both. We recommend using \code{asnLong} as it supports a greater pool of numbers.
\item If you provide a value in the same API call for both \code{asn} and \code{asnLong}, the API will only accept the value for \code{asnLong}.
\item If you enter a 4-byte ASN for the \code{asn} parameter, the API returns an error.
\item If you are using a 2-byte ASN, the API response will include the 2-byte value for both the \code{asn} and \code{asnLong} fields.
}}

\item{asnLong}{The long ASN for the BGP peer to be deleted from a Direct Connect virtual interface. The valid range is from 1 to 4294967294 for BGP configuration.

Note the following limitations when using \code{asnLong}:
\itemize{
\item You can use \code{asnLong} or \code{asn}, but not both. We recommend using \code{asnLong} as it supports a greater pool of numbers.
\item \code{asnLong} accepts any valid ASN value, regardless if it's 2-byte or 4-byte.
\item When using a 4-byte \code{asnLong}, the API response returns \code{0} for the legacy \code{asn} attribute since 4-byte ASN values exceed the maximum supported value of 2,147,483,647.
\item If you are using a 2-byte ASN, the API response will include the 2-byte value for both the \code{asn} and \code{asnLong} fields.
\item If you provide a value in the same API call for both \code{asn} and \code{asnLong}, the API will only accept the value for \code{asnLong}.
}}

\item{customerAddress}{The IP address assigned to the customer interface.}

\item{bgpPeerId}{The ID of the BGP peer.}
}
\description{
Deletes the specified BGP peer on the specified virtual interface with the specified customer address and ASN.

See \url{https://www.paws-r-sdk.com/docs/directconnect_delete_bgp_peer/} for full documentation.
}
\keyword{internal}
