% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_delete_direct_connect_gateway_association}
\alias{directconnect_delete_direct_connect_gateway_association}
\title{Deletes the association between the specified Direct Connect gateway and
virtual private gateway}
\usage{
directconnect_delete_direct_connect_gateway_association(
  associationId = NULL,
  directConnectGatewayId = NULL,
  virtualGatewayId = NULL
)
}
\arguments{
\item{associationId}{The ID of the Direct Connect gateway association.}

\item{directConnectGatewayId}{The ID of the Direct Connect gateway.}

\item{virtualGatewayId}{The ID of the virtual private gateway.}
}
\description{
Deletes the association between the specified Direct Connect gateway and virtual private gateway.

See \url{https://paws-r.github.io/docs/directconnect/delete_direct_connect_gateway_association.html} for full documentation.
}
\keyword{internal}
