% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_delete_lag}
\alias{directconnect_delete_lag}
\title{Deletes the specified link aggregation group (LAG)}
\usage{
directconnect_delete_lag(lagId)
}
\arguments{
\item{lagId}{[required] The ID of the LAG.}
}
\description{
Deletes the specified link aggregation group (LAG). You cannot delete a LAG if it has active virtual interfaces or hosted connections.

See \url{https://www.paws-r-sdk.com/docs/directconnect_delete_lag/} for full documentation.
}
\keyword{internal}
