% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_describe_connection_loa}
\alias{directconnect_describe_connection_loa}
\title{Deprecated}
\usage{
directconnect_describe_connection_loa(connectionId, providerName,
  loaContentType)
}
\arguments{
\item{connectionId}{[required] The ID of the connection.}

\item{providerName}{The name of the APN partner or service provider who establishes
connectivity on your behalf. If you specify this parameter, the LOA-CFA
lists the provider name alongside your company name as the requester of
the cross connect.}

\item{loaContentType}{The standard media type for the LOA-CFA document. The only supported
value is application/pdf.}
}
\description{
Deprecated. Use \code{\link[=directconnect_describe_loa]{describe_loa}} instead.

Gets the LOA-CFA for a connection.

The Letter of Authorization - Connecting Facility Assignment (LOA-CFA)
is a document that your APN partner or service provider uses when
establishing your cross connect to AWS at the colocation facility. For
more information, see \href{https://docs.aws.amazon.com/directconnect/latest/UserGuide/Colocation.html}{Requesting Cross Connects at AWS Direct Connect Locations}
in the \emph{AWS Direct Connect User Guide}.
}
\section{Request syntax}{
\preformatted{svc$describe_connection_loa(
  connectionId = "string",
  providerName = "string",
  loaContentType = "application/pdf"
)
}
}

\keyword{internal}
