% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_describe_connections_on_interconnect}
\alias{directconnect_describe_connections_on_interconnect}
\title{Deprecated}
\usage{
directconnect_describe_connections_on_interconnect(interconnectId)
}
\arguments{
\item{interconnectId}{[required] The ID of the interconnect.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  connections = list(
    list(
      ownerAccount = "string",
      connectionId = "string",
      connectionName = "string",
      connectionState = "ordering"|"requested"|"pending"|"available"|"down"|"deleting"|"deleted"|"rejected"|"unknown",
      region = "string",
      location = "string",
      bandwidth = "string",
      vlan = 123,
      partnerName = "string",
      loaIssueTime = as.POSIXct(
        "2015-01-01"
      ),
      lagId = "string",
      awsDevice = "string",
      jumboFrameCapable = TRUE|FALSE,
      awsDeviceV2 = "string",
      hasLogicalRedundancy = "unknown"|"yes"|"no",
      tags = list(
        list(
          key = "string",
          value = "string"
        )
      ),
      providerName = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Deprecated. Use
\code{\link[=directconnect_describe_hosted_connections]{describe_hosted_connections}}
instead.

Lists the connections that have been provisioned on the specified
interconnect.

Intended for use by AWS Direct Connect Partners only.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_connections_on_interconnect(
  interconnectId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
