% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_describe_direct_connect_gateway_associations}
\alias{directconnect_describe_direct_connect_gateway_associations}
\title{Lists the associations between your Direct Connect gateways and virtual
private gateways and transit gateways}
\usage{
directconnect_describe_direct_connect_gateway_associations(
  associationId, associatedGatewayId, directConnectGatewayId, maxResults,
  nextToken, virtualGatewayId)
}
\arguments{
\item{associationId}{The ID of the Direct Connect gateway association.}

\item{associatedGatewayId}{The ID of the associated gateway.}

\item{directConnectGatewayId}{The ID of the Direct Connect gateway.}

\item{maxResults}{The maximum number of results to return with a single call. To retrieve
the remaining results, make another call with the returned \code{nextToken}
value.

If \code{MaxResults} is given a value larger than 100, only 100 results are
returned.}

\item{nextToken}{The token provided in the previous call to retrieve the next page.}

\item{virtualGatewayId}{The ID of the virtual private gateway or transit gateway.}
}
\value{
A list with the following syntax:\preformatted{list(
  directConnectGatewayAssociations = list(
    list(
      directConnectGatewayId = "string",
      directConnectGatewayOwnerAccount = "string",
      associationState = "associating"|"associated"|"disassociating"|"disassociated"|"updating",
      stateChangeError = "string",
      associatedGateway = list(
        id = "string",
        type = "virtualPrivateGateway"|"transitGateway",
        ownerAccount = "string",
        region = "string"
      ),
      associationId = "string",
      allowedPrefixesToDirectConnectGateway = list(
        list(
          cidr = "string"
        )
      ),
      virtualGatewayId = "string",
      virtualGatewayRegion = "string",
      virtualGatewayOwnerAccount = "string"
    )
  ),
  nextToken = "string"
)
}
}
\description{
Lists the associations between your Direct Connect gateways and virtual
private gateways and transit gateways. You must specify one of the
following:
\itemize{
\item A Direct Connect gateway

The response contains all virtual private gateways and transit
gateways associated with the Direct Connect gateway.
\item A virtual private gateway

The response contains the Direct Connect gateway.
\item A transit gateway

The response contains the Direct Connect gateway.
\item A Direct Connect gateway and a virtual private gateway

The response contains the association between the Direct Connect
gateway and virtual private gateway.
\item A Direct Connect gateway and a transit gateway

The response contains the association between the Direct Connect
gateway and transit gateway.
}
}
\section{Request syntax}{
\preformatted{svc$describe_direct_connect_gateway_associations(
  associationId = "string",
  associatedGatewayId = "string",
  directConnectGatewayId = "string",
  maxResults = 123,
  nextToken = "string",
  virtualGatewayId = "string"
)
}
}

\keyword{internal}
