% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_describe_interconnect_loa}
\alias{directconnect_describe_interconnect_loa}
\title{Deprecated}
\usage{
directconnect_describe_interconnect_loa(interconnectId, providerName,
  loaContentType)
}
\arguments{
\item{interconnectId}{[required] The ID of the interconnect.}

\item{providerName}{The name of the service provider who establishes connectivity on your
behalf. If you supply this parameter, the LOA-CFA lists the provider
name alongside your company name as the requester of the cross connect.}

\item{loaContentType}{The standard media type for the LOA-CFA document. The only supported
value is application/pdf.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  loa = list(
    loaContent = raw,
    loaContentType = "application/pdf"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Deprecated. Use \code{\link[=directconnect_describe_loa]{describe_loa}} instead.

Gets the LOA-CFA for the specified interconnect.

The Letter of Authorization - Connecting Facility Assignment (LOA-CFA)
is a document that is used when establishing your cross connect to AWS
at the colocation facility. For more information, see \href{https://docs.aws.amazon.com/directconnect/latest/UserGuide/Colocation.html}{Requesting Cross Connects at AWS Direct Connect Locations}
in the \emph{AWS Direct Connect User Guide}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_interconnect_loa(
  interconnectId = "string",
  providerName = "string",
  loaContentType = "application/pdf"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
