% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_describe_lags}
\alias{directconnect_describe_lags}
\title{Describes all your link aggregation groups (LAG) or the specified LAG}
\usage{
directconnect_describe_lags(lagId = NULL)
}
\arguments{
\item{lagId}{The ID of the LAG.}
}
\description{
Describes all your link aggregation groups (LAG) or the specified LAG.

See \url{https://www.paws-r-sdk.com/docs/directconnect_describe_lags/} for full documentation.
}
\keyword{internal}
