% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_describe_locations}
\alias{directconnect_describe_locations}
\title{Lists the Direct Connect locations in the current Amazon Web Services
Region}
\usage{
directconnect_describe_locations()
}
\description{
Lists the Direct Connect locations in the current Amazon Web Services Region. These are the locations that can be selected when calling \code{\link[=directconnect_create_connection]{create_connection}} or \code{\link[=directconnect_create_interconnect]{create_interconnect}}.

See \url{https://www.paws-r-sdk.com/docs/directconnect_describe_locations/} for full documentation.
}
\keyword{internal}
