% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_describe_router_configuration}
\alias{directconnect_describe_router_configuration}
\title{Details about the router}
\usage{
directconnect_describe_router_configuration(
  virtualInterfaceId,
  routerTypeIdentifier = NULL
)
}
\arguments{
\item{virtualInterfaceId}{[required] The ID of the virtual interface.}

\item{routerTypeIdentifier}{Identifies the router by a combination of vendor, platform, and software
version. For example, \verb{CiscoSystemsInc-2900SeriesRouters-IOS124}.}
}
\description{
Details about the router.

See \url{https://www.paws-r-sdk.com/docs/directconnect_describe_router_configuration/} for full documentation.
}
\keyword{internal}
