% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_describe_virtual_gateways}
\alias{directconnect_describe_virtual_gateways}
\title{Lists the virtual private gateways owned by the Amazon Web Services
account}
\usage{
directconnect_describe_virtual_gateways()
}
\description{
Lists the virtual private gateways owned by the Amazon Web Services account.

See \url{https://www.paws-r-sdk.com/docs/directconnect_describe_virtual_gateways/} for full documentation.
}
\keyword{internal}
