% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_disassociate_connection_from_lag}
\alias{directconnect_disassociate_connection_from_lag}
\title{Disassociates a connection from a link aggregation group (LAG)}
\usage{
directconnect_disassociate_connection_from_lag(connectionId, lagId)
}
\arguments{
\item{connectionId}{[required] The ID of the connection.}

\item{lagId}{[required] The ID of the LAG.}
}
\description{
Disassociates a connection from a link aggregation group (LAG). The connection is interrupted and re-established as a standalone connection (the connection is not deleted; to delete the connection, use the \code{\link[=directconnect_delete_connection]{delete_connection}} request). If the LAG has associated virtual interfaces or hosted connections, they remain associated with the LAG. A disassociated connection owned by an Direct Connect Partner is automatically converted to an interconnect.

See \url{https://paws-r.github.io/docs/directconnect/disassociate_connection_from_lag.html} for full documentation.
}
\keyword{internal}
