% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_get_resiliency_group}
\alias{directconnect_get_resiliency_group}
\title{Gets information about the specified resiliency group}
\usage{
directconnect_get_resiliency_group(resiliencyGroupId)
}
\arguments{
\item{resiliencyGroupId}{[required] The ID of the resiliency group.}
}
\description{
Gets information about the specified resiliency group.

See \url{https://www.paws-r-sdk.com/docs/directconnect_get_resiliency_group/} for full documentation.
}
\keyword{internal}
