% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_list_resiliency_groups}
\alias{directconnect_list_resiliency_groups}
\title{Lists the resiliency groups owned by your Amazon Web Services account in
the current Amazon Web Services Region}
\usage{
directconnect_list_resiliency_groups(maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{maxResults}{The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned \code{nextToken} value.

If \code{MaxResults} is given a value larger than 100, only 100 results are returned.}

\item{nextToken}{The token for the next page of results.}
}
\description{
Lists the resiliency groups owned by your Amazon Web Services account in the current Amazon Web Services Region.

See \url{https://www.paws-r-sdk.com/docs/directconnect_list_resiliency_groups/} for full documentation.
}
\keyword{internal}
