% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_list_virtual_interface_routes}
\alias{directconnect_list_virtual_interface_routes}
\title{Lists the routes for the specified virtual interface}
\usage{
directconnect_list_virtual_interface_routes(
  virtualInterfaceId = NULL,
  filters = NULL,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{virtualInterfaceId}{The ID of the virtual interface.}

\item{filters}{The filters to apply to the routes returned.}

\item{maxResults}{The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned \code{nextToken} value.

If \code{MaxResults} is given a value larger than 100, only 100 results are returned.}

\item{nextToken}{The token for the next page of results.}
}
\description{
Lists the routes for the specified virtual interface.

See \url{https://www.paws-r-sdk.com/docs/directconnect_list_virtual_interface_routes/} for full documentation.
}
\keyword{internal}
