% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_stop_bgp_failover_test}
\alias{directconnect_stop_bgp_failover_test}
\title{Stops the virtual interface failover test}
\usage{
directconnect_stop_bgp_failover_test(virtualInterfaceId)
}
\arguments{
\item{virtualInterfaceId}{[required] The ID of the virtual interface you no longer want to test.}
}
\description{
Stops the virtual interface failover test.

See \url{https://www.paws-r-sdk.com/docs/directconnect_stop_bgp_failover_test/} for full documentation.
}
\keyword{internal}
