% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_tag_resource}
\alias{directconnect_tag_resource}
\title{Adds the specified tags to the specified Direct Connect resource}
\usage{
directconnect_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource.}

\item{tags}{[required] The tags to add.}
}
\description{
Adds the specified tags to the specified Direct Connect resource. Each resource can have a maximum of 50 tags.

See \url{https://paws-r.github.io/docs/directconnect/tag_resource.html} for full documentation.
}
\keyword{internal}
