% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_untag_resource}
\alias{directconnect_untag_resource}
\title{Removes one or more tags from the specified Direct Connect resource}
\usage{
directconnect_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource.}

\item{tagKeys}{[required] The tag keys of the tags to remove.}
}
\description{
Removes one or more tags from the specified Direct Connect resource.

See \url{https://paws-r.github.io/docs/directconnect/untag_resource.html} for full documentation.
}
\keyword{internal}
