% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_update_connection}
\alias{directconnect_update_connection}
\title{Updates the AWS Direct Connect dedicated connection configuration}
\usage{
directconnect_update_connection(connectionId, connectionName,
  encryptionMode)
}
\arguments{
\item{connectionId}{[required] The ID of the dedicated connection.

You can use \code{\link[=directconnect_describe_connections]{describe_connections}}
to retrieve the connection ID.}

\item{connectionName}{The name of the connection.}

\item{encryptionMode}{The connection MAC Security (MACsec) encryption mode.

The valid values are \code{no_encrypt}, \code{should_encrypt}, and \code{must_encrypt}.}
}
\value{
A list with the following syntax:\preformatted{list(
  ownerAccount = "string",
  connectionId = "string",
  connectionName = "string",
  connectionState = "ordering"|"requested"|"pending"|"available"|"down"|"deleting"|"deleted"|"rejected"|"unknown",
  region = "string",
  location = "string",
  bandwidth = "string",
  vlan = 123,
  partnerName = "string",
  loaIssueTime = as.POSIXct(
    "2015-01-01"
  ),
  lagId = "string",
  awsDevice = "string",
  jumboFrameCapable = TRUE|FALSE,
  awsDeviceV2 = "string",
  hasLogicalRedundancy = "unknown"|"yes"|"no",
  tags = list(
    list(
      key = "string",
      value = "string"
    )
  ),
  providerName = "string",
  macSecCapable = TRUE|FALSE,
  portEncryptionStatus = "string",
  encryptionMode = "string",
  macSecKeys = list(
    list(
      secretARN = "string",
      ckn = "string",
      state = "string",
      startOn = "string"
    )
  )
)
}
}
\description{
Updates the AWS Direct Connect dedicated connection configuration.

You can update the following parameters for a connection:
\itemize{
\item The connection name
\item The connection's MAC Security (MACsec) encryption mode.
}
}
\section{Request syntax}{
\preformatted{svc$update_connection(
  connectionId = "string",
  connectionName = "string",
  encryptionMode = "string"
)
}
}

\keyword{internal}
