% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_update_connections_billing_mode}
\alias{directconnect_update_connections_billing_mode}
\title{Updates the billing mode for the specified Direct Connect connections}
\usage{
directconnect_update_connections_billing_mode(connectionIds, billingMode)
}
\arguments{
\item{connectionIds}{[required] The IDs of the connections to update. You can specify from 1 to 200 connections.}

\item{billingMode}{[required] The billing mode to apply to the specified connections. The valid values are \code{PayAsYouGo}, \code{FlatRateTier1}, \code{FlatRateTier2}, \code{FlatRateTier3}, \code{FlatRateTier4}, and \code{FlatRateTier5}.}
}
\description{
Updates the billing mode for the specified Direct Connect connections. You can update the billing mode for up to 200 connections in a single request.

See \url{https://www.paws-r-sdk.com/docs/directconnect_update_connections_billing_mode/} for full documentation.
}
\keyword{internal}
