% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_update_direct_connect_gateway}
\alias{directconnect_update_direct_connect_gateway}
\title{Updates the name of a current Direct Connect gateway}
\usage{
directconnect_update_direct_connect_gateway(
  directConnectGatewayId,
  newDirectConnectGatewayName
)
}
\arguments{
\item{directConnectGatewayId}{[required] The ID of the Direct Connect gateway to update.}

\item{newDirectConnectGatewayName}{[required] The new name for the Direct Connect gateway.}
}
\description{
Updates the name of a current Direct Connect gateway.

See \url{https://paws-r.github.io/docs/directconnect/update_direct_connect_gateway.html} for full documentation.
}
\keyword{internal}
