% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_update_direct_connect_gateway_association}
\alias{directconnect_update_direct_connect_gateway_association}
\title{Updates the specified attributes of the Direct Connect gateway
association}
\usage{
directconnect_update_direct_connect_gateway_association(
  associationId = NULL,
  addAllowedPrefixesToDirectConnectGateway = NULL,
  removeAllowedPrefixesToDirectConnectGateway = NULL
)
}
\arguments{
\item{associationId}{The ID of the Direct Connect gateway association.}

\item{addAllowedPrefixesToDirectConnectGateway}{The Amazon VPC prefixes to advertise to the Direct Connect gateway.}

\item{removeAllowedPrefixesToDirectConnectGateway}{The Amazon VPC prefixes to no longer advertise to the Direct Connect
gateway.}
}
\description{
Updates the specified attributes of the Direct Connect gateway association.

See \url{https://paws-r.github.io/docs/directconnect/update_direct_connect_gateway_association.html} for full documentation.
}
\keyword{internal}
