% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_update_resiliency_group}
\alias{directconnect_update_resiliency_group}
\title{Updates the name of the specified resiliency group}
\usage{
directconnect_update_resiliency_group(
  resiliencyGroupId,
  resiliencyGroupName,
  clientToken = NULL
)
}
\arguments{
\item{resiliencyGroupId}{[required] The ID of the resiliency group.}

\item{resiliencyGroupName}{[required] The new name of the resiliency group.}

\item{clientToken}{A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.}
}
\description{
Updates the name of the specified resiliency group.

See \url{https://www.paws-r-sdk.com/docs/directconnect_update_resiliency_group/} for full documentation.
}
\keyword{internal}
