% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_update_virtual_interface_attributes}
\alias{directconnect_update_virtual_interface_attributes}
\title{Updates the specified attributes of the specified virtual private
interface}
\usage{
directconnect_update_virtual_interface_attributes(
  virtualInterfaceId,
  mtu = NULL,
  enableSiteLink = NULL,
  virtualInterfaceName = NULL,
  prefixPoolAllocatedCountIpv4 = NULL,
  prefixPoolAllocatedCountIpv6 = NULL,
  rateLimit = NULL
)
}
\arguments{
\item{virtualInterfaceId}{[required] The ID of the virtual private interface.}

\item{mtu}{The maximum transmission unit (MTU), in bytes. The supported values are 1500 and 8500. The default value is 1500.}

\item{enableSiteLink}{Indicates whether to enable or disable SiteLink.}

\item{virtualInterfaceName}{The name of the virtual private interface.}

\item{prefixPoolAllocatedCountIpv4}{The number of inbound IPv4 route prefixes to allocate to the virtual interface. Not applicable to public virtual interfaces.}

\item{prefixPoolAllocatedCountIpv6}{The number of inbound IPv6 route prefixes to allocate to the virtual interface. Not applicable to public virtual interfaces.}

\item{rateLimit}{The rate limit (bandwidth allocation) to apply to the virtual interface. Use this to update the bandwidth allocation on an existing virtual interface.}
}
\description{
Updates the specified attributes of the specified virtual private interface.

See \url{https://www.paws-r-sdk.com/docs/directconnect_update_virtual_interface_attributes/} for full documentation.
}
\keyword{internal}
