% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elb_operations.R
\name{elb_attach_load_balancer_to_subnets}
\alias{elb_attach_load_balancer_to_subnets}
\title{Adds one or more subnets to the set of configured subnets for the
specified load balancer}
\usage{
elb_attach_load_balancer_to_subnets(LoadBalancerName, Subnets)
}
\arguments{
\item{LoadBalancerName}{[required] The name of the load balancer.}

\item{Subnets}{[required] The IDs of the subnets to add. You can add only one subnet per
Availability Zone.}
}
\description{
Adds one or more subnets to the set of configured subnets for the specified load balancer.

See \url{https://paws-r.github.io/docs/elb/attach_load_balancer_to_subnets.html} for full documentation.
}
\keyword{internal}
