% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elb_operations.R
\name{elb_configure_health_check}
\alias{elb_configure_health_check}
\title{Specifies the health check settings to use when evaluating the health
state of your EC2 instances}
\usage{
elb_configure_health_check(LoadBalancerName, HealthCheck)
}
\arguments{
\item{LoadBalancerName}{[required] The name of the load balancer.}

\item{HealthCheck}{[required] The configuration information.}
}
\description{
Specifies the health check settings to use when evaluating the health state of your EC2 instances.

See \url{https://paws-r.github.io/docs/elb/configure_health_check.html} for full documentation.
}
\keyword{internal}
