% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elb_operations.R
\name{elb_create_app_cookie_stickiness_policy}
\alias{elb_create_app_cookie_stickiness_policy}
\title{Generates a stickiness policy with sticky session lifetimes that follow
that of an application-generated cookie}
\usage{
elb_create_app_cookie_stickiness_policy(
  LoadBalancerName,
  PolicyName,
  CookieName
)
}
\arguments{
\item{LoadBalancerName}{[required] The name of the load balancer.}

\item{PolicyName}{[required] The name of the policy being created. Policy names must consist of
alphanumeric characters and dashes (-). This name must be unique within
the set of policies for this load balancer.}

\item{CookieName}{[required] The name of the application cookie used for stickiness.}
}
\description{
Generates a stickiness policy with sticky session lifetimes that follow that of an application-generated cookie. This policy can be associated only with HTTP/HTTPS listeners.

See \url{https://www.paws-r-sdk.com/docs/elb_create_app_cookie_stickiness_policy/} for full documentation.
}
\keyword{internal}
