% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elb_operations.R
\name{elb_create_load_balancer_listeners}
\alias{elb_create_load_balancer_listeners}
\title{Creates one or more listeners for the specified load balancer}
\usage{
elb_create_load_balancer_listeners(LoadBalancerName, Listeners)
}
\arguments{
\item{LoadBalancerName}{[required] The name of the load balancer.}

\item{Listeners}{[required] The listeners.}
}
\description{
Creates one or more listeners for the specified load balancer. If a listener with the specified port does not already exist, it is created; otherwise, the properties of the new listener must match the properties of the existing listener.

See \url{https://paws-r.github.io/docs/elb/create_load_balancer_listeners.html} for full documentation.
}
\keyword{internal}
