% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elb_operations.R
\name{elb_create_load_balancer_policy}
\alias{elb_create_load_balancer_policy}
\title{Creates a policy with the specified attributes for the specified load
balancer}
\usage{
elb_create_load_balancer_policy(
  LoadBalancerName,
  PolicyName,
  PolicyTypeName,
  PolicyAttributes = NULL
)
}
\arguments{
\item{LoadBalancerName}{[required] The name of the load balancer.}

\item{PolicyName}{[required] The name of the load balancer policy to be created. This name must be
unique within the set of policies for this load balancer.}

\item{PolicyTypeName}{[required] The name of the base policy type. To get the list of policy types, use
\code{\link[=elb_describe_load_balancer_policy_types]{describe_load_balancer_policy_types}}.}

\item{PolicyAttributes}{The policy attributes.}
}
\description{
Creates a policy with the specified attributes for the specified load balancer.

See \url{https://paws-r.github.io/docs/elb/create_load_balancer_policy.html} for full documentation.
}
\keyword{internal}
