% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elb_operations.R
\name{elb_delete_load_balancer}
\alias{elb_delete_load_balancer}
\title{Deletes the specified load balancer}
\usage{
elb_delete_load_balancer(LoadBalancerName)
}
\arguments{
\item{LoadBalancerName}{[required] The name of the load balancer.}
}
\description{
Deletes the specified load balancer.

See \url{https://www.paws-r-sdk.com/docs/elb_delete_load_balancer/} for full documentation.
}
\keyword{internal}
