% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elb_operations.R
\name{elb_delete_load_balancer_listeners}
\alias{elb_delete_load_balancer_listeners}
\title{Deletes the specified listeners from the specified load balancer}
\usage{
elb_delete_load_balancer_listeners(LoadBalancerName, LoadBalancerPorts)
}
\arguments{
\item{LoadBalancerName}{[required] The name of the load balancer.}

\item{LoadBalancerPorts}{[required] The client port numbers of the listeners.}
}
\description{
Deletes the specified listeners from the specified load balancer.

See \url{https://paws-r.github.io/docs/elb/delete_load_balancer_listeners.html} for full documentation.
}
\keyword{internal}
