% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elb_operations.R
\name{elb_deregister_instances_from_load_balancer}
\alias{elb_deregister_instances_from_load_balancer}
\title{Deregisters the specified instances from the specified load balancer}
\usage{
elb_deregister_instances_from_load_balancer(LoadBalancerName, Instances)
}
\arguments{
\item{LoadBalancerName}{[required] The name of the load balancer.}

\item{Instances}{[required] The IDs of the instances.}
}
\description{
Deregisters the specified instances from the specified load balancer. After the instance is deregistered, it no longer receives traffic from the load balancer.

See \url{https://paws-r.github.io/docs/elb/deregister_instances_from_load_balancer.html} for full documentation.
}
\keyword{internal}
