% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elb_operations.R
\name{elb_describe_load_balancer_policies}
\alias{elb_describe_load_balancer_policies}
\title{Describes the specified policies}
\usage{
elb_describe_load_balancer_policies(
  LoadBalancerName = NULL,
  PolicyNames = NULL
)
}
\arguments{
\item{LoadBalancerName}{The name of the load balancer.}

\item{PolicyNames}{The names of the policies.}
}
\description{
Describes the specified policies.

See \url{https://paws-r.github.io/docs/elb/describe_load_balancer_policies.html} for full documentation.
}
\keyword{internal}
