% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elb_operations.R
\name{elb_disable_availability_zones_for_load_balancer}
\alias{elb_disable_availability_zones_for_load_balancer}
\title{Removes the specified Availability Zones from the set of Availability
Zones for the specified load balancer in EC2-Classic or a default VPC}
\usage{
elb_disable_availability_zones_for_load_balancer(
  LoadBalancerName,
  AvailabilityZones
)
}
\arguments{
\item{LoadBalancerName}{[required] The name of the load balancer.}

\item{AvailabilityZones}{[required] The Availability Zones.}
}
\description{
Removes the specified Availability Zones from the set of Availability Zones for the specified load balancer in EC2-Classic or a default VPC.

See \url{https://www.paws-r-sdk.com/docs/elb_disable_availability_zones_for_load_balancer/} for full documentation.
}
\keyword{internal}
