% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elb_operations.R
\name{elb_register_instances_with_load_balancer}
\alias{elb_register_instances_with_load_balancer}
\title{Adds the specified instances to the specified load balancer}
\usage{
elb_register_instances_with_load_balancer(LoadBalancerName, Instances)
}
\arguments{
\item{LoadBalancerName}{[required] The name of the load balancer.}

\item{Instances}{[required] The IDs of the instances.}
}
\description{
Adds the specified instances to the specified load balancer.

See \url{https://paws-r.github.io/docs/elb/register_instances_with_load_balancer.html} for full documentation.
}
\keyword{internal}
