% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elb_operations.R
\name{elb_set_load_balancer_policies_for_backend_server}
\alias{elb_set_load_balancer_policies_for_backend_server}
\title{Replaces the set of policies associated with the specified port on which
the EC2 instance is listening with a new set of policies}
\usage{
elb_set_load_balancer_policies_for_backend_server(
  LoadBalancerName,
  InstancePort,
  PolicyNames
)
}
\arguments{
\item{LoadBalancerName}{[required] The name of the load balancer.}

\item{InstancePort}{[required] The port number associated with the EC2 instance.}

\item{PolicyNames}{[required] The names of the policies. If the list is empty, then all current
polices are removed from the EC2 instance.}
}
\description{
Replaces the set of policies associated with the specified port on which the EC2 instance is listening with a new set of policies. At this time, only the back-end server authentication policy type can be applied to the instance ports; this policy type is composed of multiple public key policies.

See \url{https://www.paws-r-sdk.com/docs/elb_set_load_balancer_policies_for_backend_server/} for full documentation.
}
\keyword{internal}
