% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elb_operations.R
\name{elb_set_load_balancer_policies_of_listener}
\alias{elb_set_load_balancer_policies_of_listener}
\title{Replaces the current set of policies for the specified load balancer
port with the specified set of policies}
\usage{
elb_set_load_balancer_policies_of_listener(
  LoadBalancerName,
  LoadBalancerPort,
  PolicyNames
)
}
\arguments{
\item{LoadBalancerName}{[required] The name of the load balancer.}

\item{LoadBalancerPort}{[required] The external port of the load balancer.}

\item{PolicyNames}{[required] The names of the policies. This list must include all policies to be
enabled. If you omit a policy that is currently enabled, it is disabled.
If the list is empty, all current policies are disabled.}
}
\description{
Replaces the current set of policies for the specified load balancer port with the specified set of policies.

See \url{https://www.paws-r-sdk.com/docs/elb_set_load_balancer_policies_of_listener/} for full documentation.
}
\keyword{internal}
