% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbv2_operations.R
\name{elbv2_add_tags}
\alias{elbv2_add_tags}
\title{Adds the specified tags to the specified Elastic Load Balancing resource}
\usage{
elbv2_add_tags(ResourceArns, Tags)
}
\arguments{
\item{ResourceArns}{[required] The Amazon Resource Name (ARN) of the resource.}

\item{Tags}{[required] The tags.}
}
\description{
Adds the specified tags to the specified Elastic Load Balancing resource. You can tag your Application Load Balancers, Network Load Balancers, Gateway Load Balancers, target groups, trust stores, listeners, and rules.

See \url{https://www.paws-r-sdk.com/docs/elbv2_add_tags/} for full documentation.
}
\keyword{internal}
