% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbv2_operations.R
\name{elbv2_add_trust_store_revocations}
\alias{elbv2_add_trust_store_revocations}
\title{Adds the specified revocation file to the specified trust store}
\usage{
elbv2_add_trust_store_revocations(TrustStoreArn, RevocationContents = NULL)
}
\arguments{
\item{TrustStoreArn}{[required] The Amazon Resource Name (ARN) of the trust store.}

\item{RevocationContents}{The revocation file to add.}
}
\description{
Adds the specified revocation file to the specified trust store.

See \url{https://www.paws-r-sdk.com/docs/elbv2_add_trust_store_revocations/} for full documentation.
}
\keyword{internal}
