% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbv2_operations.R
\name{elbv2_create_trust_store}
\alias{elbv2_create_trust_store}
\title{Creates a trust store}
\usage{
elbv2_create_trust_store(
  Name,
  CaCertificatesBundleS3Bucket,
  CaCertificatesBundleS3Key,
  CaCertificatesBundleS3ObjectVersion = NULL,
  Tags = NULL
)
}
\arguments{
\item{Name}{[required] The name of the trust store.

This name must be unique per region and can't be changed after creation.}

\item{CaCertificatesBundleS3Bucket}{[required] The Amazon S3 bucket for the ca certificates bundle.}

\item{CaCertificatesBundleS3Key}{[required] The Amazon S3 path for the ca certificates bundle.}

\item{CaCertificatesBundleS3ObjectVersion}{The Amazon S3 object version for the ca certificates bundle. If
undefined the current version is used.}

\item{Tags}{The tags to assign to the trust store.}
}
\description{
Creates a trust store.

See \url{https://www.paws-r-sdk.com/docs/elbv2_create_trust_store/} for full documentation.
}
\keyword{internal}
