% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbv2_operations.R
\name{elbv2_delete_load_balancer}
\alias{elbv2_delete_load_balancer}
\title{Deletes the specified Application Load Balancer, Network Load Balancer,
or Gateway Load Balancer}
\usage{
elbv2_delete_load_balancer(LoadBalancerArn)
}
\arguments{
\item{LoadBalancerArn}{[required] The Amazon Resource Name (ARN) of the load balancer.}
}
\description{
Deletes the specified Application Load Balancer, Network Load Balancer, or Gateway Load Balancer. Deleting a load balancer also deletes its listeners.

See \url{https://paws-r.github.io/docs/elbv2/delete_load_balancer.html} for full documentation.
}
\keyword{internal}
