% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbv2_operations.R
\name{elbv2_delete_rule}
\alias{elbv2_delete_rule}
\title{Deletes the specified rule}
\usage{
elbv2_delete_rule(RuleArn)
}
\arguments{
\item{RuleArn}{[required] The Amazon Resource Name (ARN) of the rule.}
}
\description{
Deletes the specified rule.

See \url{https://www.paws-r-sdk.com/docs/elbv2_delete_rule/} for full documentation.
}
\keyword{internal}
