% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbv2_operations.R
\name{elbv2_delete_shared_trust_store_association}
\alias{elbv2_delete_shared_trust_store_association}
\title{Deletes a shared trust store association}
\usage{
elbv2_delete_shared_trust_store_association(TrustStoreArn, ResourceArn)
}
\arguments{
\item{TrustStoreArn}{[required] The Amazon Resource Name (ARN) of the trust store.}

\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource.}
}
\description{
Deletes a shared trust store association.

See \url{https://www.paws-r-sdk.com/docs/elbv2_delete_shared_trust_store_association/} for full documentation.
}
\keyword{internal}
