% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbv2_operations.R
\name{elbv2_delete_target_group}
\alias{elbv2_delete_target_group}
\title{Deletes the specified target group}
\usage{
elbv2_delete_target_group(TargetGroupArn)
}
\arguments{
\item{TargetGroupArn}{[required] The Amazon Resource Name (ARN) of the target group.}
}
\description{
Deletes the specified target group.

See \url{https://www.paws-r-sdk.com/docs/elbv2_delete_target_group/} for full documentation.
}
\keyword{internal}
