% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbv2_operations.R
\name{elbv2_delete_trust_store}
\alias{elbv2_delete_trust_store}
\title{Deletes a trust store}
\usage{
elbv2_delete_trust_store(TrustStoreArn)
}
\arguments{
\item{TrustStoreArn}{[required] The Amazon Resource Name (ARN) of the trust store.}
}
\description{
Deletes a trust store.

See \url{https://www.paws-r-sdk.com/docs/elbv2_delete_trust_store/} for full documentation.
}
\keyword{internal}
