% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbv2_operations.R
\name{elbv2_deregister_targets}
\alias{elbv2_deregister_targets}
\title{Deregisters the specified targets from the specified target group}
\usage{
elbv2_deregister_targets(TargetGroupArn, Targets)
}
\arguments{
\item{TargetGroupArn}{[required] The Amazon Resource Name (ARN) of the target group.}

\item{Targets}{[required] The targets. If you specified a port override when you registered a
target, you must specify both the target ID and the port when you
deregister it.}
}
\description{
Deregisters the specified targets from the specified target group. After the targets are deregistered, they no longer receive traffic from the load balancer.

See \url{https://paws-r.github.io/docs/elbv2/deregister_targets.html} for full documentation.
}
\keyword{internal}
