% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbv2_operations.R
\name{elbv2_describe_capacity_reservation}
\alias{elbv2_describe_capacity_reservation}
\title{Describes the capacity reservation status for the specified load
balancer}
\usage{
elbv2_describe_capacity_reservation(LoadBalancerArn)
}
\arguments{
\item{LoadBalancerArn}{[required] The Amazon Resource Name (ARN) of the load balancer.}
}
\description{
Describes the capacity reservation status for the specified load balancer.

See \url{https://www.paws-r-sdk.com/docs/elbv2_describe_capacity_reservation/} for full documentation.
}
\keyword{internal}
