% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbv2_operations.R
\name{elbv2_describe_listener_attributes}
\alias{elbv2_describe_listener_attributes}
\title{Describes the attributes for the specified listener}
\usage{
elbv2_describe_listener_attributes(ListenerArn)
}
\arguments{
\item{ListenerArn}{[required] The Amazon Resource Name (ARN) of the listener.}
}
\description{
Describes the attributes for the specified listener.

See \url{https://www.paws-r-sdk.com/docs/elbv2_describe_listener_attributes/} for full documentation.
}
\keyword{internal}
