% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbv2_operations.R
\name{elbv2_describe_load_balancer_attributes}
\alias{elbv2_describe_load_balancer_attributes}
\title{Describes the attributes for the specified Application Load Balancer,
Network Load Balancer, or Gateway Load Balancer}
\usage{
elbv2_describe_load_balancer_attributes(LoadBalancerArn)
}
\arguments{
\item{LoadBalancerArn}{[required] The Amazon Resource Name (ARN) of the load balancer.}
}
\description{
Describes the attributes for the specified Application Load Balancer, Network Load Balancer, or Gateway Load Balancer.

See \url{https://paws-r.github.io/docs/elbv2/describe_load_balancer_attributes.html} for full documentation.
}
\keyword{internal}
