% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbv2_operations.R
\name{elbv2_describe_trust_store_associations}
\alias{elbv2_describe_trust_store_associations}
\title{Describes all resources associated with the specified trust store}
\usage{
elbv2_describe_trust_store_associations(
  TrustStoreArn,
  Marker = NULL,
  PageSize = NULL
)
}
\arguments{
\item{TrustStoreArn}{[required] The Amazon Resource Name (ARN) of the trust store.}

\item{Marker}{The marker for the next set of results. (You received this marker from a
previous call.)}

\item{PageSize}{The maximum number of results to return with this call.}
}
\description{
Describes all resources associated with the specified trust store.

See \url{https://www.paws-r-sdk.com/docs/elbv2_describe_trust_store_associations/} for full documentation.
}
\keyword{internal}
