% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbv2_operations.R
\name{elbv2_describe_trust_store_revocations}
\alias{elbv2_describe_trust_store_revocations}
\title{Describes the revocation files in use by the specified trust store or
revocation files}
\usage{
elbv2_describe_trust_store_revocations(
  TrustStoreArn,
  RevocationIds = NULL,
  Marker = NULL,
  PageSize = NULL
)
}
\arguments{
\item{TrustStoreArn}{[required] The Amazon Resource Name (ARN) of the trust store.}

\item{RevocationIds}{The revocation IDs of the revocation files you want to describe.}

\item{Marker}{The marker for the next set of results. (You received this marker from a
previous call.)}

\item{PageSize}{The maximum number of results to return with this call.}
}
\description{
Describes the revocation files in use by the specified trust store or revocation files.

See \url{https://www.paws-r-sdk.com/docs/elbv2_describe_trust_store_revocations/} for full documentation.
}
\keyword{internal}
