% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbv2_operations.R
\name{elbv2_describe_trust_stores}
\alias{elbv2_describe_trust_stores}
\title{Describes all trust stores for the specified account}
\usage{
elbv2_describe_trust_stores(
  TrustStoreArns = NULL,
  Names = NULL,
  Marker = NULL,
  PageSize = NULL
)
}
\arguments{
\item{TrustStoreArns}{The Amazon Resource Name (ARN) of the trust store.}

\item{Names}{The names of the trust stores.}

\item{Marker}{The marker for the next set of results. (You received this marker from a
previous call.)}

\item{PageSize}{The maximum number of results to return with this call.}
}
\description{
Describes all trust stores for the specified account.

See \url{https://www.paws-r-sdk.com/docs/elbv2_describe_trust_stores/} for full documentation.
}
\keyword{internal}
