% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbv2_operations.R
\name{elbv2_get_resource_policy}
\alias{elbv2_get_resource_policy}
\title{Retrieves the resource policy for a specified resource}
\usage{
elbv2_get_resource_policy(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource.}
}
\description{
Retrieves the resource policy for a specified resource.

See \url{https://www.paws-r-sdk.com/docs/elbv2_get_resource_policy/} for full documentation.
}
\keyword{internal}
