% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbv2_operations.R
\name{elbv2_get_trust_store_ca_certificates_bundle}
\alias{elbv2_get_trust_store_ca_certificates_bundle}
\title{Retrieves the ca certificate bundle}
\usage{
elbv2_get_trust_store_ca_certificates_bundle(TrustStoreArn)
}
\arguments{
\item{TrustStoreArn}{[required] The Amazon Resource Name (ARN) of the trust store.}
}
\description{
Retrieves the ca certificate bundle.

See \url{https://www.paws-r-sdk.com/docs/elbv2_get_trust_store_ca_certificates_bundle/} for full documentation.
}
\keyword{internal}
