% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbv2_operations.R
\name{elbv2_get_trust_store_revocation_content}
\alias{elbv2_get_trust_store_revocation_content}
\title{Retrieves the specified revocation file}
\usage{
elbv2_get_trust_store_revocation_content(TrustStoreArn, RevocationId)
}
\arguments{
\item{TrustStoreArn}{[required] The Amazon Resource Name (ARN) of the trust store.}

\item{RevocationId}{[required] The revocation ID of the revocation file.}
}
\description{
Retrieves the specified revocation file.

See \url{https://www.paws-r-sdk.com/docs/elbv2_get_trust_store_revocation_content/} for full documentation.
}
\keyword{internal}
