% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbv2_operations.R
\name{elbv2_modify_capacity_reservation}
\alias{elbv2_modify_capacity_reservation}
\title{Modifies the capacity reservation of the specified load balancer}
\usage{
elbv2_modify_capacity_reservation(
  LoadBalancerArn,
  MinimumLoadBalancerCapacity = NULL,
  ResetCapacityReservation = NULL
)
}
\arguments{
\item{LoadBalancerArn}{[required] The Amazon Resource Name (ARN) of the load balancer.}

\item{MinimumLoadBalancerCapacity}{The minimum load balancer capacity reserved.}

\item{ResetCapacityReservation}{Resets the capacity reservation.}
}
\description{
Modifies the capacity reservation of the specified load balancer.

See \url{https://www.paws-r-sdk.com/docs/elbv2_modify_capacity_reservation/} for full documentation.
}
\keyword{internal}
