% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbv2_operations.R
\name{elbv2_modify_ip_pools}
\alias{elbv2_modify_ip_pools}
\title{[Application Load Balancers] Modify the IP pool associated to a
load balancer}
\usage{
elbv2_modify_ip_pools(
  LoadBalancerArn,
  IpamPools = NULL,
  RemoveIpamPools = NULL
)
}
\arguments{
\item{LoadBalancerArn}{[required] The Amazon Resource Name (ARN) of the load balancer.}

\item{IpamPools}{The IPAM pools to be modified.}

\item{RemoveIpamPools}{Remove the IP pools in use by the load balancer.}
}
\description{
[Application Load Balancers] Modify the IP pool associated to a load balancer.

See \url{https://www.paws-r-sdk.com/docs/elbv2_modify_ip_pools/} for full documentation.
}
\keyword{internal}
