% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbv2_operations.R
\name{elbv2_modify_listener_attributes}
\alias{elbv2_modify_listener_attributes}
\title{Modifies the specified attributes of the specified listener}
\usage{
elbv2_modify_listener_attributes(ListenerArn, Attributes)
}
\arguments{
\item{ListenerArn}{[required] The Amazon Resource Name (ARN) of the listener.}

\item{Attributes}{[required] The listener attributes.}
}
\description{
Modifies the specified attributes of the specified listener.

See \url{https://www.paws-r-sdk.com/docs/elbv2_modify_listener_attributes/} for full documentation.
}
\keyword{internal}
