% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbv2_operations.R
\name{elbv2_modify_target_group_attributes}
\alias{elbv2_modify_target_group_attributes}
\title{Modifies the specified attributes of the specified target group}
\usage{
elbv2_modify_target_group_attributes(TargetGroupArn, Attributes)
}
\arguments{
\item{TargetGroupArn}{[required] The Amazon Resource Name (ARN) of the target group.}

\item{Attributes}{[required] The attributes.}
}
\description{
Modifies the specified attributes of the specified target group.

See \url{https://www.paws-r-sdk.com/docs/elbv2_modify_target_group_attributes/} for full documentation.
}
\keyword{internal}
