% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbv2_operations.R
\name{elbv2_modify_trust_store}
\alias{elbv2_modify_trust_store}
\title{Update the ca certificate bundle for the specified trust store}
\usage{
elbv2_modify_trust_store(
  TrustStoreArn,
  CaCertificatesBundleS3Bucket,
  CaCertificatesBundleS3Key,
  CaCertificatesBundleS3ObjectVersion = NULL
)
}
\arguments{
\item{TrustStoreArn}{[required] The Amazon Resource Name (ARN) of the trust store.}

\item{CaCertificatesBundleS3Bucket}{[required] The Amazon S3 bucket for the ca certificates bundle.}

\item{CaCertificatesBundleS3Key}{[required] The Amazon S3 path for the ca certificates bundle.}

\item{CaCertificatesBundleS3ObjectVersion}{The Amazon S3 object version for the ca certificates bundle. If
undefined the current version is used.}
}
\description{
Update the ca certificate bundle for the specified trust store.

See \url{https://www.paws-r-sdk.com/docs/elbv2_modify_trust_store/} for full documentation.
}
\keyword{internal}
