% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbv2_operations.R
\name{elbv2_remove_trust_store_revocations}
\alias{elbv2_remove_trust_store_revocations}
\title{Removes the specified revocation file from the specified trust store}
\usage{
elbv2_remove_trust_store_revocations(TrustStoreArn, RevocationIds)
}
\arguments{
\item{TrustStoreArn}{[required] The Amazon Resource Name (ARN) of the trust store.}

\item{RevocationIds}{[required] The revocation IDs of the revocation files you want to remove.}
}
\description{
Removes the specified revocation file from the specified trust store.

See \url{https://www.paws-r-sdk.com/docs/elbv2_remove_trust_store_revocations/} for full documentation.
}
\keyword{internal}
