% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbv2_operations.R
\name{elbv2_set_rule_priorities}
\alias{elbv2_set_rule_priorities}
\title{Sets the priorities of the specified rules}
\usage{
elbv2_set_rule_priorities(RulePriorities)
}
\arguments{
\item{RulePriorities}{[required] The rule priorities.}
}
\description{
Sets the priorities of the specified rules.

See \url{https://www.paws-r-sdk.com/docs/elbv2_set_rule_priorities/} for full documentation.
}
\keyword{internal}
