% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbv2_operations.R
\name{elbv2_set_security_groups}
\alias{elbv2_set_security_groups}
\title{Associates the specified security groups with the specified Application
Load Balancer}
\usage{
elbv2_set_security_groups(LoadBalancerArn, SecurityGroups)
}
\arguments{
\item{LoadBalancerArn}{[required] The Amazon Resource Name (ARN) of the load balancer.}

\item{SecurityGroups}{[required] The IDs of the security groups.}
}
\description{
Associates the specified security groups with the specified Application Load Balancer. The specified security groups override the previously associated security groups.

See \url{https://paws-r.github.io/docs/elbv2/set_security_groups.html} for full documentation.
}
\keyword{internal}
