% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalaccelerator_operations.R
\name{globalaccelerator_add_custom_routing_endpoints}
\alias{globalaccelerator_add_custom_routing_endpoints}
\title{Associate a virtual private cloud (VPC) subnet endpoint with your custom
routing accelerator}
\usage{
globalaccelerator_add_custom_routing_endpoints(
  EndpointConfigurations,
  EndpointGroupArn
)
}
\arguments{
\item{EndpointConfigurations}{[required] The list of endpoint objects to add to a custom routing accelerator.}

\item{EndpointGroupArn}{[required] The Amazon Resource Name (ARN) of the endpoint group for the custom
routing endpoint.}
}
\description{
Associate a virtual private cloud (VPC) subnet endpoint with your custom routing accelerator.

See \url{https://www.paws-r-sdk.com/docs/globalaccelerator_add_custom_routing_endpoints/} for full documentation.
}
\keyword{internal}
