% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalaccelerator_operations.R
\name{globalaccelerator_add_endpoints}
\alias{globalaccelerator_add_endpoints}
\title{Add endpoints to an endpoint group}
\usage{
globalaccelerator_add_endpoints(EndpointConfigurations, EndpointGroupArn)
}
\arguments{
\item{EndpointConfigurations}{[required] The list of endpoint objects.}

\item{EndpointGroupArn}{[required] The Amazon Resource Name (ARN) of the endpoint group.}
}
\description{
Add endpoints to an endpoint group. The \code{\link[=globalaccelerator_add_endpoints]{add_endpoints}} API operation is the recommended option for adding endpoints. The alternative options are to add endpoints when you create an endpoint group (with the \code{\link[=globalaccelerator_create_endpoint_group]{create_endpoint_group}} API) or when you update an endpoint group (with the \code{\link[=globalaccelerator_update_endpoint_group]{update_endpoint_group}} API).

See \url{https://paws-r.github.io/docs/globalaccelerator/add_endpoints.html} for full documentation.
}
\keyword{internal}
