% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalaccelerator_operations.R
\name{globalaccelerator_advertise_byoip_cidr}
\alias{globalaccelerator_advertise_byoip_cidr}
\title{Advertises an IPv4 address range that is provisioned for use with your
Amazon Web Services resources through bring your own IP addresses
(BYOIP)}
\usage{
globalaccelerator_advertise_byoip_cidr(Cidr)
}
\arguments{
\item{Cidr}{[required] The address range, in CIDR notation. This must be the exact range that
you provisioned. You can't advertise only a portion of the provisioned
range.

For more information, see \href{https://docs.aws.amazon.com/global-accelerator/latest/dg/using-byoip.html}{Bring your own IP addresses (BYOIP)}
in the Global Accelerator Developer Guide.}
}
\description{
Advertises an IPv4 address range that is provisioned for use with your Amazon Web Services resources through bring your own IP addresses (BYOIP). It can take a few minutes before traffic to the specified addresses starts routing to Amazon Web Services because of propagation delays.

See \url{https://www.paws-r-sdk.com/docs/globalaccelerator_advertise_byoip_cidr/} for full documentation.
}
\keyword{internal}
