% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalaccelerator_operations.R
\name{globalaccelerator_delete_accelerator}
\alias{globalaccelerator_delete_accelerator}
\title{Delete an accelerator}
\usage{
globalaccelerator_delete_accelerator(AcceleratorArn)
}
\arguments{
\item{AcceleratorArn}{[required] The Amazon Resource Name (ARN) of an accelerator.}
}
\description{
Delete an accelerator. Before you can delete an accelerator, you must disable it and remove all dependent resources (listeners and endpoint groups). To disable the accelerator, update the accelerator to set \code{Enabled} to false.

See \url{https://paws-r.github.io/docs/globalaccelerator/delete_accelerator.html} for full documentation.
}
\keyword{internal}
