% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalaccelerator_operations.R
\name{globalaccelerator_delete_cross_account_attachment}
\alias{globalaccelerator_delete_cross_account_attachment}
\title{Delete a cross-account attachment}
\usage{
globalaccelerator_delete_cross_account_attachment(AttachmentArn)
}
\arguments{
\item{AttachmentArn}{[required] The Amazon Resource Name (ARN) for the cross-account attachment to
delete.}
}
\description{
Delete a cross-account attachment. When you delete an attachment, Global Accelerator revokes the permission to use the resources in the attachment from all principals in the list of principals. Global Accelerator revokes the permission for specific resources by doing the following:

See \url{https://www.paws-r-sdk.com/docs/globalaccelerator_delete_cross_account_attachment/} for full documentation.
}
\keyword{internal}
