% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalaccelerator_operations.R
\name{globalaccelerator_delete_custom_routing_accelerator}
\alias{globalaccelerator_delete_custom_routing_accelerator}
\title{Delete a custom routing accelerator}
\usage{
globalaccelerator_delete_custom_routing_accelerator(AcceleratorArn)
}
\arguments{
\item{AcceleratorArn}{[required] The Amazon Resource Name (ARN) of the custom routing accelerator to
delete.}
}
\description{
Delete a custom routing accelerator. Before you can delete an accelerator, you must disable it and remove all dependent resources (listeners and endpoint groups). To disable the accelerator, update the accelerator to set \code{Enabled} to false.

See \url{https://www.paws-r-sdk.com/docs/globalaccelerator_delete_custom_routing_accelerator/} for full documentation.
}
\keyword{internal}
