% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalaccelerator_operations.R
\name{globalaccelerator_delete_custom_routing_endpoint_group}
\alias{globalaccelerator_delete_custom_routing_endpoint_group}
\title{Delete an endpoint group from a listener for a custom routing
accelerator}
\usage{
globalaccelerator_delete_custom_routing_endpoint_group(EndpointGroupArn)
}
\arguments{
\item{EndpointGroupArn}{[required] The Amazon Resource Name (ARN) of the endpoint group to delete.}
}
\description{
Delete an endpoint group from a listener for a custom routing accelerator.

See \url{https://paws-r.github.io/docs/globalaccelerator/delete_custom_routing_endpoint_group.html} for full documentation.
}
\keyword{internal}
