% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalaccelerator_operations.R
\name{globalaccelerator_delete_endpoint_group}
\alias{globalaccelerator_delete_endpoint_group}
\title{Delete an endpoint group from a listener}
\usage{
globalaccelerator_delete_endpoint_group(EndpointGroupArn)
}
\arguments{
\item{EndpointGroupArn}{[required] The Amazon Resource Name (ARN) of the endpoint group to delete.}
}
\description{
Delete an endpoint group from a listener.

See \url{https://www.paws-r-sdk.com/docs/globalaccelerator_delete_endpoint_group/} for full documentation.
}
\keyword{internal}
