% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalaccelerator_operations.R
\name{globalaccelerator_delete_listener}
\alias{globalaccelerator_delete_listener}
\title{Delete a listener from an accelerator}
\usage{
globalaccelerator_delete_listener(ListenerArn)
}
\arguments{
\item{ListenerArn}{[required] The Amazon Resource Name (ARN) of the listener.}
}
\value{
An empty list.
}
\description{
Delete a listener from an accelerator.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_listener(
  ListenerArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
