% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalaccelerator_operations.R
\name{globalaccelerator_deprovision_byoip_cidr}
\alias{globalaccelerator_deprovision_byoip_cidr}
\title{Releases the specified address range that you provisioned to use with
your Amazon Web Services resources through bring your own IP addresses
(BYOIP) and deletes the corresponding address pool}
\usage{
globalaccelerator_deprovision_byoip_cidr(Cidr)
}
\arguments{
\item{Cidr}{[required] The address range, in CIDR notation. The prefix must be the same prefix
that you specified when you provisioned the address range.}
}
\description{
Releases the specified address range that you provisioned to use with your Amazon Web Services resources through bring your own IP addresses (BYOIP) and deletes the corresponding address pool.

See \url{https://paws-r.github.io/docs/globalaccelerator/deprovision_byoip_cidr.html} for full documentation.
}
\keyword{internal}
