% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalaccelerator_operations.R
\name{globalaccelerator_describe_custom_routing_accelerator_attributes}
\alias{globalaccelerator_describe_custom_routing_accelerator_attributes}
\title{Describe the attributes of a custom routing accelerator}
\usage{
globalaccelerator_describe_custom_routing_accelerator_attributes(
  AcceleratorArn
)
}
\arguments{
\item{AcceleratorArn}{[required] The Amazon Resource Name (ARN) of the custom routing accelerator to
describe the attributes for.}
}
\description{
Describe the attributes of a custom routing accelerator.

See \url{https://paws-r.github.io/docs/globalaccelerator/describe_custom_routing_accelerator_attributes.html} for full documentation.
}
\keyword{internal}
