% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalaccelerator_operations.R
\name{globalaccelerator_describe_accelerator}
\alias{globalaccelerator_describe_accelerator}
\title{Describe an accelerator}
\usage{
globalaccelerator_describe_accelerator(AcceleratorArn)
}
\arguments{
\item{AcceleratorArn}{[required] The Amazon Resource Name (ARN) of the accelerator to describe.}
}
\description{
Describe an accelerator.

See \url{https://paws-r.github.io/docs/globalaccelerator/describe_accelerator.html} for full documentation.
}
\keyword{internal}
