% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalaccelerator_operations.R
\name{globalaccelerator_describe_cross_account_attachment}
\alias{globalaccelerator_describe_cross_account_attachment}
\title{Gets configuration information about a cross-account attachment}
\usage{
globalaccelerator_describe_cross_account_attachment(AttachmentArn)
}
\arguments{
\item{AttachmentArn}{[required] The Amazon Resource Name (ARN) for the cross-account attachment to
describe.}
}
\description{
Gets configuration information about a cross-account attachment.

See \url{https://www.paws-r-sdk.com/docs/globalaccelerator_describe_cross_account_attachment/} for full documentation.
}
\keyword{internal}
