% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalaccelerator_operations.R
\name{globalaccelerator_describe_custom_routing_accelerator}
\alias{globalaccelerator_describe_custom_routing_accelerator}
\title{Describe a custom routing accelerator}
\usage{
globalaccelerator_describe_custom_routing_accelerator(AcceleratorArn)
}
\arguments{
\item{AcceleratorArn}{[required] The Amazon Resource Name (ARN) of the accelerator to describe.}
}
\description{
Describe a custom routing accelerator.

See \url{https://www.paws-r-sdk.com/docs/globalaccelerator_describe_custom_routing_accelerator/} for full documentation.
}
\keyword{internal}
