% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalaccelerator_operations.R
\name{globalaccelerator_describe_custom_routing_accelerator_attributes}
\alias{globalaccelerator_describe_custom_routing_accelerator_attributes}
\title{Describe the attributes of a custom routing accelerator}
\usage{
globalaccelerator_describe_custom_routing_accelerator_attributes(
  AcceleratorArn)
}
\arguments{
\item{AcceleratorArn}{[required] The Amazon Resource Name (ARN) of the custom routing accelerator to
describe the attributes for.}
}
\value{
A list with the following syntax:\preformatted{list(
  AcceleratorAttributes = list(
    FlowLogsEnabled = TRUE|FALSE,
    FlowLogsS3Bucket = "string",
    FlowLogsS3Prefix = "string"
  )
)
}
}
\description{
Describe the attributes of a custom routing accelerator.
}
\section{Request syntax}{
\preformatted{svc$describe_custom_routing_accelerator_attributes(
  AcceleratorArn = "string"
)
}
}

\keyword{internal}
