% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalaccelerator_operations.R
\name{globalaccelerator_describe_custom_routing_endpoint_group}
\alias{globalaccelerator_describe_custom_routing_endpoint_group}
\title{Describe an endpoint group for a custom routing accelerator}
\usage{
globalaccelerator_describe_custom_routing_endpoint_group(
  EndpointGroupArn)
}
\arguments{
\item{EndpointGroupArn}{[required] The Amazon Resource Name (ARN) of the endpoint group to describe.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  EndpointGroup = list(
    EndpointGroupArn = "string",
    EndpointGroupRegion = "string",
    DestinationDescriptions = list(
      list(
        FromPort = 123,
        ToPort = 123,
        Protocols = list(
          "TCP"|"UDP"
        )
      )
    ),
    EndpointDescriptions = list(
      list(
        EndpointId = "string"
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Describe an endpoint group for a custom routing accelerator.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_custom_routing_endpoint_group(
  EndpointGroupArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
