% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalaccelerator_operations.R
\name{globalaccelerator_list_cross_account_attachments}
\alias{globalaccelerator_list_cross_account_attachments}
\title{List the cross-account attachments that have been created in Global
Accelerator}
\usage{
globalaccelerator_list_cross_account_attachments(
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{MaxResults}{The number of cross-account attachment objects that you want to return
with this call. The default value is 10.}

\item{NextToken}{The token for the next set of results. You receive this token from a
previous call.}
}
\description{
List the cross-account attachments that have been created in Global Accelerator.

See \url{https://www.paws-r-sdk.com/docs/globalaccelerator_list_cross_account_attachments/} for full documentation.
}
\keyword{internal}
