% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalaccelerator_operations.R
\name{globalaccelerator_list_custom_routing_endpoint_groups}
\alias{globalaccelerator_list_custom_routing_endpoint_groups}
\title{List the endpoint groups that are associated with a listener for a
custom routing accelerator}
\usage{
globalaccelerator_list_custom_routing_endpoint_groups(
  ListenerArn,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{ListenerArn}{[required] The Amazon Resource Name (ARN) of the listener to list endpoint groups
for.}

\item{MaxResults}{The number of endpoint group objects that you want to return with this
call. The default value is 10.}

\item{NextToken}{The token for the next set of results. You receive this token from a
previous call.}
}
\description{
List the endpoint groups that are associated with a listener for a custom routing accelerator.

See \url{https://www.paws-r-sdk.com/docs/globalaccelerator_list_custom_routing_endpoint_groups/} for full documentation.
}
\keyword{internal}
