% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalaccelerator_operations.R
\name{globalaccelerator_list_custom_routing_listeners}
\alias{globalaccelerator_list_custom_routing_listeners}
\title{List the listeners for a custom routing accelerator}
\usage{
globalaccelerator_list_custom_routing_listeners(AcceleratorArn,
  MaxResults, NextToken)
}
\arguments{
\item{AcceleratorArn}{[required] The Amazon Resource Name (ARN) of the accelerator to list listeners for.}

\item{MaxResults}{The number of listener objects that you want to return with this call.
The default value is 10.}

\item{NextToken}{The token for the next set of results. You receive this token from a
previous call.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Listeners = list(
    list(
      ListenerArn = "string",
      PortRanges = list(
        list(
          FromPort = 123,
          ToPort = 123
        )
      )
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
List the listeners for a custom routing accelerator.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_custom_routing_listeners(
  AcceleratorArn = "string",
  MaxResults = 123,
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
