% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalaccelerator_operations.R
\name{globalaccelerator_remove_endpoints}
\alias{globalaccelerator_remove_endpoints}
\title{Remove endpoints from an endpoint group}
\usage{
globalaccelerator_remove_endpoints(EndpointIdentifiers, EndpointGroupArn)
}
\arguments{
\item{EndpointIdentifiers}{[required] The identifiers of the endpoints that you want to remove.}

\item{EndpointGroupArn}{[required] The Amazon Resource Name (ARN) of the endpoint group.}
}
\description{
Remove endpoints from an endpoint group.

See \url{https://paws-r.github.io/docs/globalaccelerator/remove_endpoints.html} for full documentation.
}
\keyword{internal}
