% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalaccelerator_operations.R
\name{globalaccelerator_tag_resource}
\alias{globalaccelerator_tag_resource}
\title{Add tags to an accelerator resource}
\usage{
globalaccelerator_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the Global Accelerator resource to add
tags to. An ARN uniquely identifies a resource.}

\item{Tags}{[required] The tags to add to a resource. A tag consists of a key and a value that
you define.}
}
\description{
Add tags to an accelerator resource.

See \url{https://www.paws-r-sdk.com/docs/globalaccelerator_tag_resource/} for full documentation.
}
\keyword{internal}
