% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalaccelerator_operations.R
\name{globalaccelerator_untag_resource}
\alias{globalaccelerator_untag_resource}
\title{Remove tags from a Global Accelerator resource}
\usage{
globalaccelerator_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the Global Accelerator resource to
remove tags from. An ARN uniquely identifies a resource.}

\item{TagKeys}{[required] The tag key pairs that you want to remove from the specified resources.}
}
\description{
Remove tags from a Global Accelerator resource. When you specify a tag key, the action removes both that key and its associated value. The operation succeeds even if you attempt to remove tags from an accelerator that was already removed.

See \url{https://paws-r.github.io/docs/globalaccelerator/untag_resource.html} for full documentation.
}
\keyword{internal}
