% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalaccelerator_operations.R
\name{globalaccelerator_update_endpoint_group}
\alias{globalaccelerator_update_endpoint_group}
\title{Update an endpoint group}
\usage{
globalaccelerator_update_endpoint_group(
  EndpointGroupArn,
  EndpointConfigurations = NULL,
  TrafficDialPercentage = NULL,
  HealthCheckPort = NULL,
  HealthCheckProtocol = NULL,
  HealthCheckPath = NULL,
  HealthCheckIntervalSeconds = NULL,
  ThresholdCount = NULL,
  PortOverrides = NULL
)
}
\arguments{
\item{EndpointGroupArn}{[required] The Amazon Resource Name (ARN) of the endpoint group.}

\item{EndpointConfigurations}{The list of endpoint objects. A resource must be valid and active when
you add it as an endpoint.}

\item{TrafficDialPercentage}{The percentage of traffic to send to an Amazon Web Services Region.
Additional traffic is distributed to other endpoint groups for this
listener.

Use this action to increase (dial up) or decrease (dial down) traffic to
a specific Region. The percentage is applied to the traffic that would
otherwise have been routed to the Region based on optimal routing.

The default value is 100.}

\item{HealthCheckPort}{The port that Global Accelerator uses to check the health of endpoints
that are part of this endpoint group. The default port is the listener
port that this endpoint group is associated with. If the listener port
is a list of ports, Global Accelerator uses the first port in the list.}

\item{HealthCheckProtocol}{The protocol that Global Accelerator uses to check the health of
endpoints that are part of this endpoint group. The default value is
TCP.}

\item{HealthCheckPath}{If the protocol is HTTP/S, then this specifies the path that is the
destination for health check targets. The default value is slash (/).}

\item{HealthCheckIntervalSeconds}{The time—10 seconds or 30 seconds—between each health check for an
endpoint. The default value is 30.}

\item{ThresholdCount}{The number of consecutive health checks required to set the state of a
healthy endpoint to unhealthy, or to set an unhealthy endpoint to
healthy. The default value is 3.}

\item{PortOverrides}{Override specific listener ports used to route traffic to endpoints that
are part of this endpoint group. For example, you can create a port
override in which the listener receives user traffic on ports 80 and
443, but your accelerator routes that traffic to ports 1080 and 1443,
respectively, on the endpoints.

For more information, see \href{https://docs.aws.amazon.com/global-accelerator/latest/dg/about-endpoint-groups-port-override.html}{Overriding listener ports}
in the \emph{Global Accelerator Developer Guide}.}
}
\description{
Update an endpoint group. A resource must be valid and active when you add it as an endpoint.

See \url{https://www.paws-r-sdk.com/docs/globalaccelerator_update_endpoint_group/} for full documentation.
}
\keyword{internal}
