% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalaccelerator_operations.R
\name{globalaccelerator_withdraw_byoip_cidr}
\alias{globalaccelerator_withdraw_byoip_cidr}
\title{Stops advertising an address range that is provisioned as an address
pool}
\usage{
globalaccelerator_withdraw_byoip_cidr(Cidr)
}
\arguments{
\item{Cidr}{[required] The address range, in CIDR notation.}
}
\description{
Stops advertising an address range that is provisioned as an address pool. You can perform this operation at most once every 10 seconds, even if you specify different address ranges each time.

See \url{https://paws-r.github.io/docs/globalaccelerator/withdraw_byoip_cidr.html} for full documentation.
}
\keyword{internal}
