% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkfirewall_operations.R
\name{networkfirewall_accept_network_firewall_transit_gateway_attachment}
\alias{networkfirewall_accept_network_firewall_transit_gateway_attachment}
\title{Accepts a transit gateway attachment request for Network Firewall}
\usage{
networkfirewall_accept_network_firewall_transit_gateway_attachment(
  TransitGatewayAttachmentId
)
}
\arguments{
\item{TransitGatewayAttachmentId}{[required] Required. The unique identifier of the transit gateway attachment to
accept. This ID is returned in the response when creating a transit
gateway-attached firewall.}
}
\description{
Accepts a transit gateway attachment request for Network Firewall. When you accept the attachment request, Network Firewall creates the necessary routing components to enable traffic flow between the transit gateway and firewall endpoints.

See \url{https://www.paws-r-sdk.com/docs/networkfirewall_accept_network_firewall_transit_gateway_attachment/} for full documentation.
}
\keyword{internal}
