% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkfirewall_operations.R
\name{networkfirewall_associate_availability_zones}
\alias{networkfirewall_associate_availability_zones}
\title{Associates the specified Availability Zones with a transit
gateway-attached firewall}
\usage{
networkfirewall_associate_availability_zones(
  UpdateToken = NULL,
  FirewallArn = NULL,
  FirewallName = NULL,
  AvailabilityZoneMappings
)
}
\arguments{
\item{UpdateToken}{An optional token that you can use for optimistic locking. Network
Firewall returns a token to your requests that access the firewall. The
token marks the state of the firewall resource at the time of the
request.

To make an unconditional change to the firewall, omit the token in your
update request. Without the token, Network Firewall performs your
updates regardless of whether the firewall has changed since you last
retrieved it.

To make a conditional change to the firewall, provide the token in your
update request. Network Firewall uses the token to ensure that the
firewall hasn't changed since you last retrieved it. If it has changed,
the operation fails with an \code{InvalidTokenException}. If this happens,
retrieve the firewall again to get a current copy of it with a new
token. Reapply your changes as needed, then try the operation again
using the new token.}

\item{FirewallArn}{The Amazon Resource Name (ARN) of the firewall.

You must specify the ARN or the name, and you can specify both.}

\item{FirewallName}{The descriptive name of the firewall. You can't change the name of a
firewall after you create it.

You must specify the ARN or the name, and you can specify both.}

\item{AvailabilityZoneMappings}{[required] Required. The Availability Zones where you want to create firewall
endpoints. You must specify at least one Availability Zone.}
}
\description{
Associates the specified Availability Zones with a transit gateway-attached firewall. For each Availability Zone, Network Firewall creates a firewall endpoint to process traffic. You can specify one or more Availability Zones where you want to deploy the firewall.

See \url{https://www.paws-r-sdk.com/docs/networkfirewall_associate_availability_zones/} for full documentation.
}
\keyword{internal}
