% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkfirewall_operations.R
\name{networkfirewall_create_container_association}
\alias{networkfirewall_create_container_association}
\title{Creates a Network Firewall container association}
\usage{
networkfirewall_create_container_association(
  ContainerAssociationName,
  Description = NULL,
  Type,
  ContainerMonitoringConfigurations,
  Tags = NULL
)
}
\arguments{
\item{ContainerAssociationName}{[required] The descriptive name of the container association. You can't change the name of a container association after you create it.}

\item{Description}{A description of the container association.}

\item{Type}{[required] The type of containers to monitor. You can't change the container type after creation. Valid values:
\itemize{
\item \code{ECS} - Amazon Elastic Container Service
\item \code{EKS} - Amazon Elastic Kubernetes Service
}}

\item{ContainerMonitoringConfigurations}{[required] The monitoring configurations for the container association. Each configuration specifies an Amazon ECS or Amazon EKS cluster to monitor and optional attribute filters to narrow which containers are tracked.}

\item{Tags}{The key:value pairs to associate with the resource.}
}
\description{
Creates a Network Firewall container association. The association monitors container lifecycle events in your Amazon ECS or Amazon EKS clusters and resolves running container addresses for use in firewall rules.

See \url{https://www.paws-r-sdk.com/docs/networkfirewall_create_container_association/} for full documentation.
}
\keyword{internal}
