% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkfirewall_operations.R
\name{networkfirewall_create_firewall}
\alias{networkfirewall_create_firewall}
\title{Creates an Network Firewall Firewall and accompanying FirewallStatus for
a VPC}
\usage{
networkfirewall_create_firewall(
  FirewallName,
  FirewallPolicyArn,
  VpcId = NULL,
  SubnetMappings = NULL,
  DeleteProtection = NULL,
  SubnetChangeProtection = NULL,
  FirewallPolicyChangeProtection = NULL,
  Description = NULL,
  Tags = NULL,
  EncryptionConfiguration = NULL,
  EnabledAnalysisTypes = NULL,
  TransitGatewayId = NULL,
  AvailabilityZoneMappings = NULL,
  AvailabilityZoneChangeProtection = NULL,
  NatGatewayMappings = NULL,
  ProxySettings = NULL,
  NoSourcePreservation = NULL,
  VpcEndpoint = NULL
)
}
\arguments{
\item{FirewallName}{[required] The descriptive name of the firewall. You can't change the name of a firewall after you create it.}

\item{FirewallPolicyArn}{[required] The Amazon Resource Name (ARN) of the FirewallPolicy that you want to use for the firewall.}

\item{VpcId}{The unique identifier of the VPC where Network Firewall should create the firewall.

You can't change this setting after you create the firewall.}

\item{SubnetMappings}{The public subnets to use for your Network Firewall firewalls. Each subnet must belong to a different Availability Zone in the VPC. Network Firewall creates a firewall endpoint in each subnet.}

\item{DeleteProtection}{A flag indicating whether it is possible to delete the firewall. A setting of \code{TRUE} indicates that the firewall is protected against deletion. Use this setting to protect against accidentally deleting a firewall that is in use. When you create a firewall, the operation initializes this flag to \code{TRUE}.}

\item{SubnetChangeProtection}{A setting indicating whether the firewall is protected against changes to the subnet associations. Use this setting to protect against accidentally modifying the subnet associations for a firewall that is in use. When you create a firewall, the operation initializes this setting to \code{TRUE}.}

\item{FirewallPolicyChangeProtection}{A setting indicating whether the firewall is protected against a change to the firewall policy association. Use this setting to protect against accidentally modifying the firewall policy for a firewall that is in use. When you create a firewall, the operation initializes this setting to \code{TRUE}.}

\item{Description}{A description of the firewall.}

\item{Tags}{The key:value pairs to associate with the resource.}

\item{EncryptionConfiguration}{A complex type that contains settings for encryption of your firewall resources.}

\item{EnabledAnalysisTypes}{An optional setting indicating the specific traffic analysis types to enable on the firewall.}

\item{TransitGatewayId}{Required when creating a transit gateway-attached firewall. The unique identifier of the transit gateway to attach to this firewall. You can provide either a transit gateway from your account or one that has been shared with you through Resource Access Manager.

After creating the firewall, you cannot change the transit gateway association. To use a different transit gateway, you must create a new firewall.

For information about creating firewalls, see \code{\link[=networkfirewall_create_firewall]{create_firewall}}. For specific guidance about transit gateway-attached firewalls, see \href{https://docs.aws.amazon.com/network-firewall/latest/developerguide/tgw-firewall-considerations.html}{Considerations for transit gateway-attached firewalls} in the \emph{Network Firewall Developer Guide}.}

\item{AvailabilityZoneMappings}{Required. The Availability Zones where you want to create firewall endpoints for a transit gateway-attached firewall. You must specify at least one Availability Zone. Consider enabling the firewall in every Availability Zone where you have workloads to maintain Availability Zone isolation.

You can modify Availability Zones later using \code{\link[=networkfirewall_associate_availability_zones]{associate_availability_zones}} or \code{\link[=networkfirewall_disassociate_availability_zones]{disassociate_availability_zones}}, but this may briefly disrupt traffic. The \code{AvailabilityZoneChangeProtection} setting controls whether you can make these modifications.}

\item{AvailabilityZoneChangeProtection}{Optional. A setting indicating whether the firewall is protected against changes to its Availability Zone configuration. When set to \code{TRUE}, you cannot add or remove Availability Zones without first disabling this protection using \code{\link[=networkfirewall_update_availability_zone_change_protection]{update_availability_zone_change_protection}}.

Default value: \code{FALSE}}

\item{NatGatewayMappings}{The NAT gateways that the firewall uses to proxy traffic when \code{NoSourcePreservation} is \code{TRUE}. Network Firewall attaches the firewall to each NAT gateway that you specify, so that egress traffic is proxied through the NAT gateway.}

\item{ProxySettings}{The listener configuration for a proxy mode firewall, used when \code{NoSourcePreservation} is \code{TRUE}. This specifies the ports and protocols on which the firewall's proxy listens for traffic.}

\item{NoSourcePreservation}{Optional. Indicates whether the firewall operates in proxy mode, in which the source IP address of the traffic is not preserved. When set to \code{TRUE}, the firewall proxies traffic through a NAT gateway and the traffic reaching the destination uses the NAT gateway's IP address as the source.

When you set this to \code{TRUE}, you must specify \code{NatGatewayMappings} and \code{VpcEndpoint} instead of a top-level \code{VpcId} and \code{SubnetMappings}.

You can't change this setting after you create the firewall.

Default value: \code{FALSE}}

\item{VpcEndpoint}{The VPC and subnets for the firewall endpoint, used when \code{NoSourcePreservation} is \code{TRUE}. Network Firewall creates the firewall endpoint in the subnets that you specify here.

For proxy mode firewalls, provide the firewall's VPC and endpoint subnets through this parameter instead of the top-level \code{VpcId} and \code{SubnetMappings}.}
}
\description{
Creates an Network Firewall Firewall and accompanying FirewallStatus for a VPC.

See \url{https://www.paws-r-sdk.com/docs/networkfirewall_create_firewall/} for full documentation.
}
\keyword{internal}
