% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkfirewall_operations.R
\name{networkfirewall_create_proxy}
\alias{networkfirewall_create_proxy}
\title{Creates an Network Firewall Proxy}
\usage{
networkfirewall_create_proxy(
  ProxyName,
  NatGatewayId,
  ProxyConfigurationName = NULL,
  ProxyConfigurationArn = NULL,
  ListenerProperties = NULL,
  TlsInterceptProperties,
  Tags = NULL
)
}
\arguments{
\item{ProxyName}{[required] The descriptive name of the proxy. You can't change the name of a proxy
after you create it.}

\item{NatGatewayId}{[required] A unique identifier for the NAT gateway to use with proxy resources.}

\item{ProxyConfigurationName}{The descriptive name of the proxy configuration. You can't change the
name of a proxy configuration after you create it.

You must specify the ARN or the name, and you can specify both.}

\item{ProxyConfigurationArn}{The Amazon Resource Name (ARN) of a proxy configuration.

You must specify the ARN or the name, and you can specify both.}

\item{ListenerProperties}{Listener properties for HTTP and HTTPS traffic.}

\item{TlsInterceptProperties}{[required] TLS decryption on traffic to filter on attributes in the HTTP header.}

\item{Tags}{The key:value pairs to associate with the resource.}
}
\description{
Creates an Network Firewall Proxy

See \url{https://www.paws-r-sdk.com/docs/networkfirewall_create_proxy/} for full documentation.
}
\keyword{internal}
