% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkfirewall_operations.R
\name{networkfirewall_create_proxy_configuration}
\alias{networkfirewall_create_proxy_configuration}
\title{Creates an Network Firewall ProxyConfiguration}
\usage{
networkfirewall_create_proxy_configuration(
  ProxyConfigurationName,
  Description = NULL,
  RuleGroupNames = NULL,
  RuleGroupArns = NULL,
  DefaultRulePhaseActions,
  Tags = NULL
)
}
\arguments{
\item{ProxyConfigurationName}{[required] The descriptive name of the proxy configuration. You can't change the
name of a proxy configuration after you create it.}

\item{Description}{A description of the proxy configuration.}

\item{RuleGroupNames}{The proxy rule group name(s) to attach to the proxy configuration.

You must specify the ARNs or the names, and you can specify both.}

\item{RuleGroupArns}{The proxy rule group arn(s) to attach to the proxy configuration.

You must specify the ARNs or the names, and you can specify both.}

\item{DefaultRulePhaseActions}{[required] Evaluation points in the traffic flow where rules are applied. There are
three phases in a traffic where the rule match is applied.}

\item{Tags}{The key:value pairs to associate with the resource.}
}
\description{
Creates an Network Firewall ProxyConfiguration

See \url{https://www.paws-r-sdk.com/docs/networkfirewall_create_proxy_configuration/} for full documentation.
}
\keyword{internal}
