% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkfirewall_operations.R
\name{networkfirewall_create_proxy_rule_group}
\alias{networkfirewall_create_proxy_rule_group}
\title{Creates an Network Firewall ProxyRuleGroup}
\usage{
networkfirewall_create_proxy_rule_group(
  ProxyRuleGroupName,
  Description = NULL,
  Rules = NULL,
  Tags = NULL
)
}
\arguments{
\item{ProxyRuleGroupName}{[required] The descriptive name of the proxy rule group. You can't change the name
of a proxy rule group after you create it.}

\item{Description}{A description of the proxy rule group.}

\item{Rules}{Individual rules that define match conditions and actions for
application-layer traffic. Rules specify what to inspect (domains,
headers, methods) and what action to take (allow, deny, alert).}

\item{Tags}{The key:value pairs to associate with the resource.}
}
\description{
Creates an Network Firewall ProxyRuleGroup

See \url{https://www.paws-r-sdk.com/docs/networkfirewall_create_proxy_rule_group/} for full documentation.
}
\keyword{internal}
