% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkfirewall_operations.R
\name{networkfirewall_create_proxy_rules}
\alias{networkfirewall_create_proxy_rules}
\title{Creates Network Firewall ProxyRule resources}
\usage{
networkfirewall_create_proxy_rules(
  ProxyRuleGroupArn = NULL,
  ProxyRuleGroupName = NULL,
  Rules
)
}
\arguments{
\item{ProxyRuleGroupArn}{The Amazon Resource Name (ARN) of a proxy rule group.

You must specify the ARN or the name, and you can specify both.}

\item{ProxyRuleGroupName}{The descriptive name of the proxy rule group. You can't change the name
of a proxy rule group after you create it.

You must specify the ARN or the name, and you can specify both.}

\item{Rules}{[required] Individual rules that define match conditions and actions for
application-layer traffic. Rules specify what to inspect (domains,
headers, methods) and what action to take (allow, deny, alert).}
}
\description{
Creates Network Firewall ProxyRule resources.

See \url{https://www.paws-r-sdk.com/docs/networkfirewall_create_proxy_rules/} for full documentation.
}
\keyword{internal}
