% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkfirewall_operations.R
\name{networkfirewall_create_vpc_endpoint_association}
\alias{networkfirewall_create_vpc_endpoint_association}
\title{Creates a firewall endpoint for an Network Firewall firewall}
\usage{
networkfirewall_create_vpc_endpoint_association(
  FirewallArn,
  VpcId,
  SubnetMapping,
  Description = NULL,
  Tags = NULL
)
}
\arguments{
\item{FirewallArn}{[required] The Amazon Resource Name (ARN) of the firewall.}

\item{VpcId}{[required] The unique identifier of the VPC where you want to create a firewall
endpoint.}

\item{SubnetMapping}{[required]}

\item{Description}{A description of the VPC endpoint association.}

\item{Tags}{The key:value pairs to associate with the resource.}
}
\description{
Creates a firewall endpoint for an Network Firewall firewall. This type of firewall endpoint is independent of the firewall endpoints that you specify in the \code{Firewall} itself, and you define it in addition to those endpoints after the firewall has been created. You can define a VPC endpoint association using a different VPC than the one you used in the firewall specifications.

See \url{https://www.paws-r-sdk.com/docs/networkfirewall_create_vpc_endpoint_association/} for full documentation.
}
\keyword{internal}
