% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkfirewall_operations.R
\name{networkfirewall_delete_network_firewall_transit_gateway_attachment}
\alias{networkfirewall_delete_network_firewall_transit_gateway_attachment}
\title{Deletes a transit gateway attachment from a Network Firewall}
\usage{
networkfirewall_delete_network_firewall_transit_gateway_attachment(
  TransitGatewayAttachmentId
)
}
\arguments{
\item{TransitGatewayAttachmentId}{[required] Required. The unique identifier of the transit gateway attachment to
delete.}
}
\description{
Deletes a transit gateway attachment from a Network Firewall. Either the firewall owner or the transit gateway owner can delete the attachment.

See \url{https://www.paws-r-sdk.com/docs/networkfirewall_delete_network_firewall_transit_gateway_attachment/} for full documentation.
}
\keyword{internal}
