% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkfirewall_operations.R
\name{networkfirewall_delete_container_association}
\alias{networkfirewall_delete_container_association}
\title{Deletes a container association}
\usage{
networkfirewall_delete_container_association(
  ContainerAssociationName = NULL,
  ContainerAssociationArn = NULL
)
}
\arguments{
\item{ContainerAssociationName}{The descriptive name of the container association.

You must specify the ARN or the name, and you can specify both.}

\item{ContainerAssociationArn}{The Amazon Resource Name (ARN) of the container association.

You must specify the ARN or the name, and you can specify both.}
}
\description{
Deletes a container association. The resource transitions to a \code{DELETING} state. Deletion is asynchronous - Network Firewall returns immediately while cleanup proceeds in the background. You can't delete a container association while a rule group references it.

See \url{https://www.paws-r-sdk.com/docs/networkfirewall_delete_container_association/} for full documentation.
}
\keyword{internal}
