% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkfirewall_operations.R
\name{networkfirewall_delete_firewall}
\alias{networkfirewall_delete_firewall}
\title{Deletes the specified Firewall and its FirewallStatus}
\usage{
networkfirewall_delete_firewall(FirewallName = NULL, FirewallArn = NULL)
}
\arguments{
\item{FirewallName}{The descriptive name of the firewall. You can't change the name of a
firewall after you create it.

You must specify the ARN or the name, and you can specify both.}

\item{FirewallArn}{The Amazon Resource Name (ARN) of the firewall.

You must specify the ARN or the name, and you can specify both.}
}
\description{
Deletes the specified Firewall and its FirewallStatus. This operation requires the firewall's \code{DeleteProtection} flag to be \code{FALSE}. You can't revert this operation.

See \url{https://www.paws-r-sdk.com/docs/networkfirewall_delete_firewall/} for full documentation.
}
\keyword{internal}
