% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkfirewall_operations.R
\name{networkfirewall_delete_firewall_policy}
\alias{networkfirewall_delete_firewall_policy}
\title{Deletes the specified FirewallPolicy}
\usage{
networkfirewall_delete_firewall_policy(
  FirewallPolicyName = NULL,
  FirewallPolicyArn = NULL
)
}
\arguments{
\item{FirewallPolicyName}{The descriptive name of the firewall policy. You can't change the name
of a firewall policy after you create it.

You must specify the ARN or the name, and you can specify both.}

\item{FirewallPolicyArn}{The Amazon Resource Name (ARN) of the firewall policy.

You must specify the ARN or the name, and you can specify both.}
}
\description{
Deletes the specified FirewallPolicy.

See \url{https://paws-r.github.io/docs/networkfirewall/delete_firewall_policy.html} for full documentation.
}
\keyword{internal}
