% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkfirewall_operations.R
\name{networkfirewall_delete_proxy}
\alias{networkfirewall_delete_proxy}
\title{Deletes the specified Proxy}
\usage{
networkfirewall_delete_proxy(NatGatewayId, ProxyName = NULL, ProxyArn = NULL)
}
\arguments{
\item{NatGatewayId}{[required] The NAT Gateway the proxy is attached to.}

\item{ProxyName}{The descriptive name of the proxy. You can't change the name of a proxy
after you create it.

You must specify the ARN or the name, and you can specify both.}

\item{ProxyArn}{The Amazon Resource Name (ARN) of a proxy.

You must specify the ARN or the name, and you can specify both.}
}
\description{
Deletes the specified Proxy.

See \url{https://www.paws-r-sdk.com/docs/networkfirewall_delete_proxy/} for full documentation.
}
\keyword{internal}
