% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkfirewall_operations.R
\name{networkfirewall_delete_proxy_configuration}
\alias{networkfirewall_delete_proxy_configuration}
\title{Deletes the specified ProxyConfiguration}
\usage{
networkfirewall_delete_proxy_configuration(
  ProxyConfigurationName = NULL,
  ProxyConfigurationArn = NULL
)
}
\arguments{
\item{ProxyConfigurationName}{The descriptive name of the proxy configuration. You can't change the
name of a proxy configuration after you create it.

You must specify the ARN or the name, and you can specify both.}

\item{ProxyConfigurationArn}{The Amazon Resource Name (ARN) of a proxy configuration.

You must specify the ARN or the name, and you can specify both.}
}
\description{
Deletes the specified ProxyConfiguration.

See \url{https://www.paws-r-sdk.com/docs/networkfirewall_delete_proxy_configuration/} for full documentation.
}
\keyword{internal}
