% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkfirewall_operations.R
\name{networkfirewall_delete_proxy_rule_group}
\alias{networkfirewall_delete_proxy_rule_group}
\title{Deletes the specified ProxyRuleGroup}
\usage{
networkfirewall_delete_proxy_rule_group(
  ProxyRuleGroupName = NULL,
  ProxyRuleGroupArn = NULL
)
}
\arguments{
\item{ProxyRuleGroupName}{The descriptive name of the proxy rule group. You can't change the name
of a proxy rule group after you create it.

You must specify the ARN or the name, and you can specify both.}

\item{ProxyRuleGroupArn}{The Amazon Resource Name (ARN) of a proxy rule group.

You must specify the ARN or the name, and you can specify both.}
}
\description{
Deletes the specified ProxyRuleGroup.

See \url{https://www.paws-r-sdk.com/docs/networkfirewall_delete_proxy_rule_group/} for full documentation.
}
\keyword{internal}
