% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkfirewall_operations.R
\name{networkfirewall_delete_proxy_rules}
\alias{networkfirewall_delete_proxy_rules}
\title{Deletes the specified ProxyRule(s)}
\usage{
networkfirewall_delete_proxy_rules(
  ProxyRuleGroupArn = NULL,
  ProxyRuleGroupName = NULL,
  Rules
)
}
\arguments{
\item{ProxyRuleGroupArn}{The Amazon Resource Name (ARN) of a proxy rule group.

You must specify the ARN or the name, and you can specify both.}

\item{ProxyRuleGroupName}{The descriptive name of the proxy rule group. You can't change the name
of a proxy rule group after you create it.

You must specify the ARN or the name, and you can specify both.}

\item{Rules}{[required] The proxy rule(s) to remove from the existing proxy rule group.}
}
\description{
Deletes the specified ProxyRule(s). currently attached to a ProxyRuleGroup

See \url{https://www.paws-r-sdk.com/docs/networkfirewall_delete_proxy_rules/} for full documentation.
}
\keyword{internal}
