% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkfirewall_operations.R
\name{networkfirewall_describe_container_association}
\alias{networkfirewall_describe_container_association}
\title{Retrieves the configuration and status of a container association}
\usage{
networkfirewall_describe_container_association(
  ContainerAssociationName = NULL,
  ContainerAssociationArn = NULL
)
}
\arguments{
\item{ContainerAssociationName}{The descriptive name of the container association.

You must specify the ARN or the name, and you can specify both.}

\item{ContainerAssociationArn}{The Amazon Resource Name (ARN) of the container association.

You must specify the ARN or the name, and you can specify both.}
}
\description{
Retrieves the configuration and status of a container association.

See \url{https://www.paws-r-sdk.com/docs/networkfirewall_describe_container_association/} for full documentation.
}
\keyword{internal}
