% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkfirewall_operations.R
\name{networkfirewall_describe_firewall}
\alias{networkfirewall_describe_firewall}
\title{Returns the data objects for the specified firewall}
\usage{
networkfirewall_describe_firewall(FirewallName = NULL, FirewallArn = NULL)
}
\arguments{
\item{FirewallName}{The descriptive name of the firewall. You can't change the name of a
firewall after you create it.

You must specify the ARN or the name, and you can specify both.}

\item{FirewallArn}{The Amazon Resource Name (ARN) of the firewall.

You must specify the ARN or the name, and you can specify both.}
}
\description{
Returns the data objects for the specified firewall.

See \url{https://paws-r.github.io/docs/networkfirewall/describe_firewall.html} for full documentation.
}
\keyword{internal}
