% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkfirewall_operations.R
\name{networkfirewall_describe_firewall_metadata}
\alias{networkfirewall_describe_firewall_metadata}
\title{Returns the high-level information about a firewall, including the
Availability Zones where the Firewall is currently in use}
\usage{
networkfirewall_describe_firewall_metadata(FirewallArn = NULL)
}
\arguments{
\item{FirewallArn}{The Amazon Resource Name (ARN) of the firewall.}
}
\description{
Returns the high-level information about a firewall, including the Availability Zones where the Firewall is currently in use.

See \url{https://www.paws-r-sdk.com/docs/networkfirewall_describe_firewall_metadata/} for full documentation.
}
\keyword{internal}
