% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkfirewall_operations.R
\name{networkfirewall_describe_flow_operation}
\alias{networkfirewall_describe_flow_operation}
\title{Returns key information about a specific flow operation}
\usage{
networkfirewall_describe_flow_operation(
  FirewallArn,
  AvailabilityZone = NULL,
  VpcEndpointAssociationArn = NULL,
  VpcEndpointId = NULL,
  FlowOperationId
)
}
\arguments{
\item{FirewallArn}{[required] The Amazon Resource Name (ARN) of the firewall.}

\item{AvailabilityZone}{The ID of the Availability Zone where the firewall is located. For
example, \verb{us-east-2a}.

Defines the scope a flow operation. You can use up to 20 filters to
configure a single flow operation.}

\item{VpcEndpointAssociationArn}{The Amazon Resource Name (ARN) of a VPC endpoint association.}

\item{VpcEndpointId}{A unique identifier for the primary endpoint associated with a firewall.}

\item{FlowOperationId}{[required] A unique identifier for the flow operation. This ID is returned in the
responses to start and list commands. You provide to describe commands.}
}
\description{
Returns key information about a specific flow operation.

See \url{https://www.paws-r-sdk.com/docs/networkfirewall_describe_flow_operation/} for full documentation.
}
\keyword{internal}
