% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkfirewall_operations.R
\name{networkfirewall_describe_logging_configuration}
\alias{networkfirewall_describe_logging_configuration}
\title{Returns the logging configuration for the specified firewall}
\usage{
networkfirewall_describe_logging_configuration(
  FirewallArn = NULL,
  FirewallName = NULL
)
}
\arguments{
\item{FirewallArn}{The Amazon Resource Name (ARN) of the firewall.

You must specify the ARN or the name, and you can specify both.}

\item{FirewallName}{The descriptive name of the firewall. You can't change the name of a
firewall after you create it.

You must specify the ARN or the name, and you can specify both.}
}
\description{
Returns the logging configuration for the specified firewall.

See \url{https://www.paws-r-sdk.com/docs/networkfirewall_describe_logging_configuration/} for full documentation.
}
\keyword{internal}
