% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkfirewall_operations.R
\name{networkfirewall_describe_proxy}
\alias{networkfirewall_describe_proxy}
\title{Returns the data objects for the specified proxy}
\usage{
networkfirewall_describe_proxy(ProxyName = NULL, ProxyArn = NULL)
}
\arguments{
\item{ProxyName}{The descriptive name of the proxy. You can't change the name of a proxy
after you create it.

You must specify the ARN or the name, and you can specify both.}

\item{ProxyArn}{The Amazon Resource Name (ARN) of a proxy.

You must specify the ARN or the name, and you can specify both.}
}
\description{
Returns the data objects for the specified proxy.

See \url{https://www.paws-r-sdk.com/docs/networkfirewall_describe_proxy/} for full documentation.
}
\keyword{internal}
