% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkfirewall_operations.R
\name{networkfirewall_describe_proxy_rule}
\alias{networkfirewall_describe_proxy_rule}
\title{Returns the data objects for the specified proxy configuration for the
specified proxy rule group}
\usage{
networkfirewall_describe_proxy_rule(
  ProxyRuleName,
  ProxyRuleGroupName = NULL,
  ProxyRuleGroupArn = NULL
)
}
\arguments{
\item{ProxyRuleName}{[required] The descriptive name of the proxy rule. You can't change the name of a
proxy rule after you create it.}

\item{ProxyRuleGroupName}{The descriptive name of the proxy rule group. You can't change the name
of a proxy rule group after you create it.

You must specify the ARN or the name, and you can specify both.}

\item{ProxyRuleGroupArn}{The Amazon Resource Name (ARN) of a proxy rule group.

You must specify the ARN or the name, and you can specify both.}
}
\description{
Returns the data objects for the specified proxy configuration for the specified proxy rule group.

See \url{https://www.paws-r-sdk.com/docs/networkfirewall_describe_proxy_rule/} for full documentation.
}
\keyword{internal}
