% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkfirewall_operations.R
\name{networkfirewall_describe_rule_group}
\alias{networkfirewall_describe_rule_group}
\title{Returns the data objects for the specified rule group}
\usage{
networkfirewall_describe_rule_group(
  RuleGroupName = NULL,
  RuleGroupArn = NULL,
  Type = NULL,
  AnalyzeRuleGroup = NULL
)
}
\arguments{
\item{RuleGroupName}{The descriptive name of the rule group. You can't change the name of a
rule group after you create it.

You must specify the ARN or the name, and you can specify both.}

\item{RuleGroupArn}{The Amazon Resource Name (ARN) of the rule group.

You must specify the ARN or the name, and you can specify both.}

\item{Type}{Indicates whether the rule group is stateless or stateful. If the rule
group is stateless, it contains stateless rules. If it is stateful, it
contains stateful rules.

This setting is required for requests that do not include the
\code{RuleGroupARN}.}

\item{AnalyzeRuleGroup}{Indicates whether you want Network Firewall to analyze the stateless
rules in the rule group for rule behavior such as asymmetric routing. If
set to \code{TRUE}, Network Firewall runs the analysis.}
}
\description{
Returns the data objects for the specified rule group.

See \url{https://www.paws-r-sdk.com/docs/networkfirewall_describe_rule_group/} for full documentation.
}
\keyword{internal}
