% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkfirewall_operations.R
\name{networkfirewall_describe_rule_group_summary}
\alias{networkfirewall_describe_rule_group_summary}
\title{Returns detailed information for a stateful rule group}
\usage{
networkfirewall_describe_rule_group_summary(
  RuleGroupName = NULL,
  RuleGroupArn = NULL,
  Type = NULL
)
}
\arguments{
\item{RuleGroupName}{The descriptive name of the rule group. You can't change the name of a
rule group after you create it.

You must specify the ARN or the name, and you can specify both.}

\item{RuleGroupArn}{Required. The Amazon Resource Name (ARN) of the rule group.

You must specify the ARN or the name, and you can specify both.}

\item{Type}{The type of rule group you want a summary for. This is a required field.

Valid value: \code{STATEFUL}

Note that \code{STATELESS} exists but is not currently supported. If you
provide \code{STATELESS}, an exception is returned.}
}
\description{
Returns detailed information for a stateful rule group.

See \url{https://www.paws-r-sdk.com/docs/networkfirewall_describe_rule_group_summary/} for full documentation.
}
\keyword{internal}
