% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkfirewall_operations.R
\name{networkfirewall_detach_rule_groups_from_proxy_configuration}
\alias{networkfirewall_detach_rule_groups_from_proxy_configuration}
\title{Detaches ProxyRuleGroup resources from a ProxyConfiguration}
\usage{
networkfirewall_detach_rule_groups_from_proxy_configuration(
  ProxyConfigurationName = NULL,
  ProxyConfigurationArn = NULL,
  RuleGroupNames = NULL,
  RuleGroupArns = NULL,
  UpdateToken
)
}
\arguments{
\item{ProxyConfigurationName}{The descriptive name of the proxy configuration. You can't change the
name of a proxy configuration after you create it.

You must specify the ARN or the name, and you can specify both.}

\item{ProxyConfigurationArn}{The Amazon Resource Name (ARN) of a proxy configuration.

You must specify the ARN or the name, and you can specify both.}

\item{RuleGroupNames}{The proxy rule group names to detach from the proxy configuration}

\item{RuleGroupArns}{The proxy rule group arns to detach from the proxy configuration}

\item{UpdateToken}{[required] A token used for optimistic locking. Network Firewall returns a token to
your requests that access the proxy configuration. The token marks the
state of the proxy configuration resource at the time of the request.

To make changes to the proxy configuration, you provide the token in
your request. Network Firewall uses the token to ensure that the proxy
configuration hasn't changed since you last retrieved it. If it has
changed, the operation fails with an \code{InvalidTokenException}. If this
happens, retrieve the proxy configuration again to get a current copy of
it with a current token. Reapply your changes as needed, then try the
operation again using the new token.}
}
\description{
Detaches ProxyRuleGroup resources from a ProxyConfiguration

See \url{https://www.paws-r-sdk.com/docs/networkfirewall_detach_rule_groups_from_proxy_configuration/} for full documentation.
}
\keyword{internal}
