% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkfirewall_operations.R
\name{networkfirewall_disassociate_availability_zones}
\alias{networkfirewall_disassociate_availability_zones}
\title{Removes the specified Availability Zone associations from a transit
gateway-attached firewall}
\usage{
networkfirewall_disassociate_availability_zones(
  UpdateToken = NULL,
  FirewallArn = NULL,
  FirewallName = NULL,
  AvailabilityZoneMappings
)
}
\arguments{
\item{UpdateToken}{An optional token that you can use for optimistic locking. Network
Firewall returns a token to your requests that access the firewall. The
token marks the state of the firewall resource at the time of the
request.

To make an unconditional change to the firewall, omit the token in your
update request. Without the token, Network Firewall performs your
updates regardless of whether the firewall has changed since you last
retrieved it.

To make a conditional change to the firewall, provide the token in your
update request. Network Firewall uses the token to ensure that the
firewall hasn't changed since you last retrieved it. If it has changed,
the operation fails with an \code{InvalidTokenException}. If this happens,
retrieve the firewall again to get a current copy of it with a new
token. Reapply your changes as needed, then try the operation again
using the new token.}

\item{FirewallArn}{The Amazon Resource Name (ARN) of the firewall.

You must specify the ARN or the name, and you can specify both.}

\item{FirewallName}{The descriptive name of the firewall. You can't change the name of a
firewall after you create it.

You must specify the ARN or the name, and you can specify both.}

\item{AvailabilityZoneMappings}{[required] Required. The Availability Zones to remove from the firewall's
configuration.}
}
\description{
Removes the specified Availability Zone associations from a transit gateway-attached firewall. This removes the firewall endpoints from these Availability Zones and stops traffic filtering in those zones. Before removing an Availability Zone, ensure you've updated your transit gateway route tables to redirect traffic appropriately.

See \url{https://www.paws-r-sdk.com/docs/networkfirewall_disassociate_availability_zones/} for full documentation.
}
\keyword{internal}
