% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkfirewall_operations.R
\name{networkfirewall_list_container_associations}
\alias{networkfirewall_list_container_associations}
\title{Lists the container associations in your account and Region}
\usage{
networkfirewall_list_container_associations(
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{MaxResults}{The maximum number of objects that you want Network Firewall to return for this request. If more objects are available, in the response, Network Firewall provides a \code{NextToken} value that you can use in a subsequent call to get the next batch of objects.}

\item{NextToken}{When you request a list of objects with a \code{MaxResults} setting, if the number of objects that are still available for retrieval exceeds the maximum you requested, Network Firewall returns a \code{NextToken} value in the response. To retrieve the next batch of objects, use the token returned from the prior request in your next request.}
}
\description{
Lists the container associations in your account and Region. Use the \code{NextToken} parameter in subsequent requests to retrieve additional results.

See \url{https://www.paws-r-sdk.com/docs/networkfirewall_list_container_associations/} for full documentation.
}
\keyword{internal}
