% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkfirewall_operations.R
\name{networkfirewall_list_flow_operation_results}
\alias{networkfirewall_list_flow_operation_results}
\title{Returns the results of a specific flow operation}
\usage{
networkfirewall_list_flow_operation_results(
  FirewallArn,
  FlowOperationId,
  NextToken = NULL,
  MaxResults = NULL,
  AvailabilityZone = NULL,
  VpcEndpointId = NULL,
  VpcEndpointAssociationArn = NULL
)
}
\arguments{
\item{FirewallArn}{[required] The Amazon Resource Name (ARN) of the firewall.}

\item{FlowOperationId}{[required] A unique identifier for the flow operation. This ID is returned in the
responses to start and list commands. You provide to describe commands.}

\item{NextToken}{When you request a list of objects with a \code{MaxResults} setting, if the
number of objects that are still available for retrieval exceeds the
maximum you requested, Network Firewall returns a \code{NextToken} value in
the response. To retrieve the next batch of objects, use the token
returned from the prior request in your next request.}

\item{MaxResults}{The maximum number of objects that you want Network Firewall to return
for this request. If more objects are available, in the response,
Network Firewall provides a \code{NextToken} value that you can use in a
subsequent call to get the next batch of objects.}

\item{AvailabilityZone}{The ID of the Availability Zone where the firewall is located. For
example, \verb{us-east-2a}.

Defines the scope a flow operation. You can use up to 20 filters to
configure a single flow operation.}

\item{VpcEndpointId}{A unique identifier for the primary endpoint associated with a firewall.}

\item{VpcEndpointAssociationArn}{The Amazon Resource Name (ARN) of a VPC endpoint association.}
}
\description{
Returns the results of a specific flow operation.

See \url{https://www.paws-r-sdk.com/docs/networkfirewall_list_flow_operation_results/} for full documentation.
}
\keyword{internal}
