% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkfirewall_operations.R
\name{networkfirewall_list_proxy_configurations}
\alias{networkfirewall_list_proxy_configurations}
\title{Retrieves the metadata for the proxy configuration that you have defined}
\usage{
networkfirewall_list_proxy_configurations(NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{NextToken}{When you request a list of objects with a \code{MaxResults} setting, if the
number of objects that are still available for retrieval exceeds the
maximum you requested, Network Firewall returns a \code{NextToken} value in
the response. To retrieve the next batch of objects, use the token
returned from the prior request in your next request.}

\item{MaxResults}{The maximum number of objects that you want Network Firewall to return
for this request. If more objects are available, in the response,
Network Firewall provides a \code{NextToken} value that you can use in a
subsequent call to get the next batch of objects.}
}
\description{
Retrieves the metadata for the proxy configuration that you have defined. Depending on your setting for max results and the number of proxy configurations, a single call might not return the full list.

See \url{https://www.paws-r-sdk.com/docs/networkfirewall_list_proxy_configurations/} for full documentation.
}
\keyword{internal}
