% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkfirewall_operations.R
\name{networkfirewall_list_vpc_endpoint_associations}
\alias{networkfirewall_list_vpc_endpoint_associations}
\title{Retrieves the metadata for the VPC endpoint associations that you have
defined}
\usage{
networkfirewall_list_vpc_endpoint_associations(
  NextToken = NULL,
  MaxResults = NULL,
  FirewallArn = NULL
)
}
\arguments{
\item{NextToken}{When you request a list of objects with a \code{MaxResults} setting, if the
number of objects that are still available for retrieval exceeds the
maximum you requested, Network Firewall returns a \code{NextToken} value in
the response. To retrieve the next batch of objects, use the token
returned from the prior request in your next request.}

\item{MaxResults}{The maximum number of objects that you want Network Firewall to return
for this request. If more objects are available, in the response,
Network Firewall provides a \code{NextToken} value that you can use in a
subsequent call to get the next batch of objects.}

\item{FirewallArn}{The Amazon Resource Name (ARN) of the firewall.

If you don't specify this, Network Firewall retrieves all VPC endpoint
associations that you have defined.}
}
\description{
Retrieves the metadata for the VPC endpoint associations that you have defined. If you specify a fireawll, this returns only the endpoint associations for that firewall.

See \url{https://www.paws-r-sdk.com/docs/networkfirewall_list_vpc_endpoint_associations/} for full documentation.
}
\keyword{internal}
