% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkfirewall_operations.R
\name{networkfirewall_reject_network_firewall_transit_gateway_attachment}
\alias{networkfirewall_reject_network_firewall_transit_gateway_attachment}
\title{Rejects a transit gateway attachment request for Network Firewall}
\usage{
networkfirewall_reject_network_firewall_transit_gateway_attachment(
  TransitGatewayAttachmentId
)
}
\arguments{
\item{TransitGatewayAttachmentId}{[required] Required. The unique identifier of the transit gateway attachment to
reject. This ID is returned in the response when creating a transit
gateway-attached firewall.}
}
\description{
Rejects a transit gateway attachment request for Network Firewall. When you reject the attachment request, Network Firewall cancels the creation of routing components between the transit gateway and firewall endpoints.

See \url{https://www.paws-r-sdk.com/docs/networkfirewall_reject_network_firewall_transit_gateway_attachment/} for full documentation.
}
\keyword{internal}
