% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkfirewall_operations.R
\name{networkfirewall_start_analysis_report}
\alias{networkfirewall_start_analysis_report}
\title{Generates a traffic analysis report for the timeframe and traffic type
you specify}
\usage{
networkfirewall_start_analysis_report(
  FirewallName = NULL,
  FirewallArn = NULL,
  AnalysisType
)
}
\arguments{
\item{FirewallName}{The descriptive name of the firewall. You can't change the name of a
firewall after you create it.

You must specify the ARN or the name, and you can specify both.}

\item{FirewallArn}{The Amazon Resource Name (ARN) of the firewall.

You must specify the ARN or the name, and you can specify both.}

\item{AnalysisType}{[required] The type of traffic that will be used to generate a report.}
}
\description{
Generates a traffic analysis report for the timeframe and traffic type you specify.

See \url{https://www.paws-r-sdk.com/docs/networkfirewall_start_analysis_report/} for full documentation.
}
\keyword{internal}
