% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkfirewall_operations.R
\name{networkfirewall_untag_resource}
\alias{networkfirewall_untag_resource}
\title{Removes the tags with the specified keys from the specified resource}
\usage{
networkfirewall_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource.}

\item{TagKeys}{[required]}
}
\description{
Removes the tags with the specified keys from the specified resource. Tags are key:value pairs that you can use to categorize and manage your resources, for purposes like billing. For example, you might set the tag key to "customer" and the value to the customer name or ID. You can specify one or more tags to add to each Amazon Web Services resource, up to 50 tags for a resource.

See \url{https://paws-r.github.io/docs/networkfirewall/untag_resource.html} for full documentation.
}
\keyword{internal}
