% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkfirewall_operations.R
\name{networkfirewall_update_availability_zone_change_protection}
\alias{networkfirewall_update_availability_zone_change_protection}
\title{Modifies the AvailabilityZoneChangeProtection setting for a transit
gateway-attached firewall}
\usage{
networkfirewall_update_availability_zone_change_protection(
  UpdateToken = NULL,
  FirewallArn = NULL,
  FirewallName = NULL,
  AvailabilityZoneChangeProtection
)
}
\arguments{
\item{UpdateToken}{An optional token that you can use for optimistic locking. Network
Firewall returns a token to your requests that access the firewall. The
token marks the state of the firewall resource at the time of the
request.

To make an unconditional change to the firewall, omit the token in your
update request. Without the token, Network Firewall performs your
updates regardless of whether the firewall has changed since you last
retrieved it.

To make a conditional change to the firewall, provide the token in your
update request. Network Firewall uses the token to ensure that the
firewall hasn't changed since you last retrieved it. If it has changed,
the operation fails with an \code{InvalidTokenException}. If this happens,
retrieve the firewall again to get a current copy of it with a new
token. Reapply your changes as needed, then try the operation again
using the new token.}

\item{FirewallArn}{The Amazon Resource Name (ARN) of the firewall.

You must specify the ARN or the name, and you can specify both.}

\item{FirewallName}{The descriptive name of the firewall. You can't change the name of a
firewall after you create it.

You must specify the ARN or the name, and you can specify both.}

\item{AvailabilityZoneChangeProtection}{[required] A setting indicating whether the firewall is protected against changes
to the subnet associations. Use this setting to protect against
accidentally modifying the subnet associations for a firewall that is in
use. When you create a firewall, the operation initializes this setting
to \code{TRUE}.}
}
\description{
Modifies the \code{AvailabilityZoneChangeProtection} setting for a transit gateway-attached firewall. When enabled, this setting prevents accidental changes to the firewall's Availability Zone configuration. This helps protect against disrupting traffic flow in production environments.

See \url{https://www.paws-r-sdk.com/docs/networkfirewall_update_availability_zone_change_protection/} for full documentation.
}
\keyword{internal}
