% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkfirewall_operations.R
\name{networkfirewall_update_container_association}
\alias{networkfirewall_update_container_association}
\title{Updates the monitoring configurations and description of a container
association}
\usage{
networkfirewall_update_container_association(
  ContainerAssociationName = NULL,
  ContainerAssociationArn = NULL,
  Description = NULL,
  Type,
  ContainerMonitoringConfigurations,
  Tags = NULL,
  UpdateToken
)
}
\arguments{
\item{ContainerAssociationName}{The descriptive name of the container association.

You must specify the ARN or the name, and you can specify both.}

\item{ContainerAssociationArn}{The Amazon Resource Name (ARN) of the container association.

You must specify the ARN or the name, and you can specify both.}

\item{Description}{A description of the container association. When omitted, the existing description remains unchanged. To clear the description, pass an empty string.}

\item{Type}{[required] The container type. This value must match the existing type and can't be changed. Valid values:
\itemize{
\item \code{ECS} - Amazon Elastic Container Service
\item \code{EKS} - Amazon Elastic Kubernetes Service
}}

\item{ContainerMonitoringConfigurations}{[required] The updated monitoring configurations for the container association. Each configuration specifies an Amazon ECS or Amazon EKS cluster to monitor and optional attribute filters.}

\item{Tags}{The key:value pairs to associate with the resource.}

\item{UpdateToken}{[required] A token used for optimistic locking. Network Firewall returns a token to your requests that access the container association. The token marks the state of the container association resource at the time of the request.

To make changes to the container association, you provide the token in your request. Network Firewall uses the token to ensure that the container association hasn't changed since you last retrieved it. If it has changed, the operation fails with an \code{InvalidTokenException}. If this happens, retrieve the container association again to get a current copy of it with a current token. Reapply your changes as needed, then try the operation again using the new token.}
}
\description{
Updates the monitoring configurations and description of a container association. You can't change the container type after creation. Provide an update token to enable optimistic concurrency control.

See \url{https://www.paws-r-sdk.com/docs/networkfirewall_update_container_association/} for full documentation.
}
\keyword{internal}
