% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkfirewall_operations.R
\name{networkfirewall_update_logging_configuration}
\alias{networkfirewall_update_logging_configuration}
\title{Sets the logging configuration for the specified firewall}
\usage{
networkfirewall_update_logging_configuration(
  FirewallArn = NULL,
  FirewallName = NULL,
  LoggingConfiguration = NULL
)
}
\arguments{
\item{FirewallArn}{The Amazon Resource Name (ARN) of the firewall.

You must specify the ARN or the name, and you can specify both.}

\item{FirewallName}{The descriptive name of the firewall. You can't change the name of a
firewall after you create it.

You must specify the ARN or the name, and you can specify both.}

\item{LoggingConfiguration}{Defines how Network Firewall performs logging for a firewall. If you
omit this setting, Network Firewall disables logging for the firewall.}
}
\description{
Sets the logging configuration for the specified firewall.

See \url{https://paws-r.github.io/docs/networkfirewall/update_logging_configuration.html} for full documentation.
}
\keyword{internal}
