% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkfirewall_operations.R
\name{networkfirewall_update_proxy}
\alias{networkfirewall_update_proxy}
\title{Updates the properties of the specified proxy}
\usage{
networkfirewall_update_proxy(
  NatGatewayId,
  ProxyName = NULL,
  ProxyArn = NULL,
  ListenerPropertiesToAdd = NULL,
  ListenerPropertiesToRemove = NULL,
  TlsInterceptProperties = NULL,
  UpdateToken
)
}
\arguments{
\item{NatGatewayId}{[required] The NAT Gateway the proxy is attached to.}

\item{ProxyName}{The descriptive name of the proxy. You can't change the name of a proxy
after you create it.

You must specify the ARN or the name, and you can specify both.}

\item{ProxyArn}{The Amazon Resource Name (ARN) of a proxy.

You must specify the ARN or the name, and you can specify both.}

\item{ListenerPropertiesToAdd}{Listener properties for HTTP and HTTPS traffic to add.}

\item{ListenerPropertiesToRemove}{Listener properties for HTTP and HTTPS traffic to remove.}

\item{TlsInterceptProperties}{TLS decryption on traffic to filter on attributes in the HTTP header.}

\item{UpdateToken}{[required] A token used for optimistic locking. Network Firewall returns a token to
your requests that access the proxy. The token marks the state of the
proxy resource at the time of the request.

To make changes to the proxy, you provide the token in your request.
Network Firewall uses the token to ensure that the proxy hasn't changed
since you last retrieved it. If it has changed, the operation fails with
an \code{InvalidTokenException}. If this happens, retrieve the proxy again to
get a current copy of it with a current token. Reapply your changes as
needed, then try the operation again using the new token.}
}
\description{
Updates the properties of the specified proxy.

See \url{https://www.paws-r-sdk.com/docs/networkfirewall_update_proxy/} for full documentation.
}
\keyword{internal}
