% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkfirewall_operations.R
\name{networkfirewall_update_proxy_configuration}
\alias{networkfirewall_update_proxy_configuration}
\title{Updates the properties of the specified proxy configuration}
\usage{
networkfirewall_update_proxy_configuration(
  ProxyConfigurationName = NULL,
  ProxyConfigurationArn = NULL,
  DefaultRulePhaseActions,
  UpdateToken
)
}
\arguments{
\item{ProxyConfigurationName}{The descriptive name of the proxy configuration. You can't change the
name of a proxy configuration after you create it.

You must specify the ARN or the name, and you can specify both.}

\item{ProxyConfigurationArn}{The Amazon Resource Name (ARN) of a proxy configuration.

You must specify the ARN or the name, and you can specify both.}

\item{DefaultRulePhaseActions}{[required] Evaluation points in the traffic flow where rules are applied. There are
three phases in a traffic where the rule match is applied.}

\item{UpdateToken}{[required] A token used for optimistic locking. Network Firewall returns a token to
your requests that access the proxy configuration. The token marks the
state of the proxy configuration resource at the time of the request.

To make changes to the proxy configuration, you provide the token in
your request. Network Firewall uses the token to ensure that the proxy
configuration hasn't changed since you last retrieved it. If it has
changed, the operation fails with an \code{InvalidTokenException}. If this
happens, retrieve the proxy configuration again to get a current copy of
it with a current token. Reapply your changes as needed, then try the
operation again using the new token.}
}
\description{
Updates the properties of the specified proxy configuration.

See \url{https://www.paws-r-sdk.com/docs/networkfirewall_update_proxy_configuration/} for full documentation.
}
\keyword{internal}
