% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkfirewall_operations.R
\name{networkfirewall_update_proxy_rule}
\alias{networkfirewall_update_proxy_rule}
\title{Updates the properties of the specified proxy rule}
\usage{
networkfirewall_update_proxy_rule(
  ProxyRuleGroupName = NULL,
  ProxyRuleGroupArn = NULL,
  ProxyRuleName,
  Description = NULL,
  Action = NULL,
  AddConditions = NULL,
  RemoveConditions = NULL,
  UpdateToken
)
}
\arguments{
\item{ProxyRuleGroupName}{The descriptive name of the proxy rule group. You can't change the name
of a proxy rule group after you create it.

You must specify the ARN or the name, and you can specify both.}

\item{ProxyRuleGroupArn}{The Amazon Resource Name (ARN) of a proxy rule group.

You must specify the ARN or the name, and you can specify both.}

\item{ProxyRuleName}{[required] The descriptive name of the proxy rule. You can't change the name of a
proxy rule after you create it.}

\item{Description}{A description of the proxy rule.}

\item{Action}{Depending on the match action, the proxy either stops the evaluation (if
the action is terminal - allow or deny), or continues it (if the action
is alert) until it matches a rule with a terminal action.}

\item{AddConditions}{Proxy rule conditions to add. Match criteria that specify what traffic
attributes to examine. Conditions include operators (StringEquals,
StringLike) and values to match against.}

\item{RemoveConditions}{Proxy rule conditions to remove. Match criteria that specify what
traffic attributes to examine. Conditions include operators
(StringEquals, StringLike) and values to match against.}

\item{UpdateToken}{[required] A token used for optimistic locking. Network Firewall returns a token to
your requests that access the proxy rule. The token marks the state of
the proxy rule resource at the time of the request.

To make changes to the proxy rule, you provide the token in your
request. Network Firewall uses the token to ensure that the proxy rule
hasn't changed since you last retrieved it. If it has changed, the
operation fails with an \code{InvalidTokenException}. If this happens,
retrieve the proxy rule again to get a current copy of it with a current
token. Reapply your changes as needed, then try the operation again
using the new token.}
}
\description{
Updates the properties of the specified proxy rule.

See \url{https://www.paws-r-sdk.com/docs/networkfirewall_update_proxy_rule/} for full documentation.
}
\keyword{internal}
