% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkfirewall_operations.R
\name{networkfirewall_update_proxy_rule_priorities}
\alias{networkfirewall_update_proxy_rule_priorities}
\title{Updates proxy rule priorities within a proxy rule group}
\usage{
networkfirewall_update_proxy_rule_priorities(
  ProxyRuleGroupName = NULL,
  ProxyRuleGroupArn = NULL,
  RuleGroupRequestPhase,
  Rules,
  UpdateToken
)
}
\arguments{
\item{ProxyRuleGroupName}{The descriptive name of the proxy rule group. You can't change the name
of a proxy rule group after you create it.

You must specify the ARN or the name, and you can specify both.}

\item{ProxyRuleGroupArn}{The Amazon Resource Name (ARN) of a proxy rule group.

You must specify the ARN or the name, and you can specify both.}

\item{RuleGroupRequestPhase}{[required] Evaluation points in the traffic flow where rules are applied. There are
three phases in a traffic where the rule match is applied.}

\item{Rules}{[required] proxy rule resources to update to new positions.}

\item{UpdateToken}{[required] A token used for optimistic locking. Network Firewall returns a token to
your requests that access the proxy rule group. The token marks the
state of the proxy rule group resource at the time of the request.

To make changes to the proxy rule group, you provide the token in your
request. Network Firewall uses the token to ensure that the proxy rule
group hasn't changed since you last retrieved it. If it has changed, the
operation fails with an \code{InvalidTokenException}. If this happens,
retrieve the proxy rule group again to get a current copy of it with a
current token. Reapply your changes as needed, then try the operation
again using the new token.}
}
\description{
Updates proxy rule priorities within a proxy rule group.

See \url{https://www.paws-r-sdk.com/docs/networkfirewall_update_proxy_rule_priorities/} for full documentation.
}
\keyword{internal}
