% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_associate_connect_peer}
\alias{networkmanager_associate_connect_peer}
\title{Associates a core network Connect peer with a device and optionally,
with a link}
\usage{
networkmanager_associate_connect_peer(
  GlobalNetworkId,
  ConnectPeerId,
  DeviceId,
  LinkId = NULL
)
}
\arguments{
\item{GlobalNetworkId}{[required] The ID of your global network.}

\item{ConnectPeerId}{[required] The ID of the Connect peer.}

\item{DeviceId}{[required] The ID of the device.}

\item{LinkId}{The ID of the link.}
}
\description{
Associates a core network Connect peer with a device and optionally, with a link.

See \url{https://www.paws-r-sdk.com/docs/networkmanager_associate_connect_peer/} for full documentation.
}
\keyword{internal}
