% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_associate_customer_gateway}
\alias{networkmanager_associate_customer_gateway}
\title{Associates a customer gateway with a device and optionally, with a link}
\usage{
networkmanager_associate_customer_gateway(
  CustomerGatewayArn,
  GlobalNetworkId,
  DeviceId,
  LinkId = NULL
)
}
\arguments{
\item{CustomerGatewayArn}{[required] The Amazon Resource Name (ARN) of the customer gateway.}

\item{GlobalNetworkId}{[required] The ID of the global network.}

\item{DeviceId}{[required] The ID of the device.}

\item{LinkId}{The ID of the link.}
}
\description{
Associates a customer gateway with a device and optionally, with a link. If you specify a link, it must be associated with the specified device.

See \url{https://www.paws-r-sdk.com/docs/networkmanager_associate_customer_gateway/} for full documentation.
}
\keyword{internal}
