% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_associate_link}
\alias{networkmanager_associate_link}
\title{Associates a link to a device}
\usage{
networkmanager_associate_link(GlobalNetworkId, DeviceId, LinkId)
}
\arguments{
\item{GlobalNetworkId}{[required] The ID of the global network.}

\item{DeviceId}{[required] The ID of the device.}

\item{LinkId}{[required] The ID of the link.}
}
\description{
Associates a link to a device. A device can be associated to multiple links and a link can be associated to multiple devices. The device and link must be in the same global network and the same site.

See \url{https://www.paws-r-sdk.com/docs/networkmanager_associate_link/} for full documentation.
}
\keyword{internal}
