% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_associate_transit_gateway_connect_peer}
\alias{networkmanager_associate_transit_gateway_connect_peer}
\title{Associates a transit gateway Connect peer with a device, and optionally,
with a link}
\usage{
networkmanager_associate_transit_gateway_connect_peer(
  GlobalNetworkId,
  TransitGatewayConnectPeerArn,
  DeviceId,
  LinkId = NULL
)
}
\arguments{
\item{GlobalNetworkId}{[required] The ID of the global network.}

\item{TransitGatewayConnectPeerArn}{[required] The Amazon Resource Name (ARN) of the Connect peer.}

\item{DeviceId}{[required] The ID of the device.}

\item{LinkId}{The ID of the link.}
}
\description{
Associates a transit gateway Connect peer with a device, and optionally, with a link. If you specify a link, it must be associated with the specified device.

See \url{https://paws-r.github.io/docs/networkmanager/associate_transit_gateway_connect_peer.html} for full documentation.
}
\keyword{internal}
