% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_create_connect_attachment}
\alias{networkmanager_create_connect_attachment}
\title{Creates a core network Connect attachment from a specified core network
attachment}
\usage{
networkmanager_create_connect_attachment(
  CoreNetworkId,
  EdgeLocation,
  TransportAttachmentId,
  Options,
  Tags = NULL,
  ClientToken = NULL
)
}
\arguments{
\item{CoreNetworkId}{[required] The ID of a core network where you want to create the attachment.}

\item{EdgeLocation}{[required] The Region where the edge is located.}

\item{TransportAttachmentId}{[required] The ID of the attachment between the two connections.}

\item{Options}{[required] Options for creating an attachment.}

\item{Tags}{The list of key-value tags associated with the request.}

\item{ClientToken}{The client token associated with the request.}
}
\description{
Creates a core network Connect attachment from a specified core network attachment.

See \url{https://paws-r.github.io/docs/networkmanager/create_connect_attachment.html} for full documentation.
}
\keyword{internal}
