% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_create_core_network}
\alias{networkmanager_create_core_network}
\title{Creates a core network as part of your global network, and optionally,
with a core network policy}
\usage{
networkmanager_create_core_network(
  GlobalNetworkId,
  Description = NULL,
  Tags = NULL,
  PolicyDocument = NULL,
  ClientToken = NULL
)
}
\arguments{
\item{GlobalNetworkId}{[required] The ID of the global network that a core network will be a part of.}

\item{Description}{The description of a core network.}

\item{Tags}{Key-value tags associated with a core network request.}

\item{PolicyDocument}{The policy document for creating a core network.}

\item{ClientToken}{The client token associated with a core network request.}
}
\description{
Creates a core network as part of your global network, and optionally, with a core network policy.

See \url{https://www.paws-r-sdk.com/docs/networkmanager_create_core_network/} for full documentation.
}
\keyword{internal}
