% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_create_core_network_prefix_list_association}
\alias{networkmanager_create_core_network_prefix_list_association}
\title{Creates an association between a core network and a prefix list for
routing control}
\usage{
networkmanager_create_core_network_prefix_list_association(
  CoreNetworkId,
  PrefixListArn,
  PrefixListAlias,
  ClientToken = NULL
)
}
\arguments{
\item{CoreNetworkId}{[required] The ID of the core network to associate with the prefix list.}

\item{PrefixListArn}{[required] The ARN of the prefix list to associate with the core network.}

\item{PrefixListAlias}{[required] An optional alias for the prefix list association.}

\item{ClientToken}{A unique, case-sensitive identifier that you provide to ensure the
idempotency of the request.}
}
\description{
Creates an association between a core network and a prefix list for routing control.

See \url{https://www.paws-r-sdk.com/docs/networkmanager_create_core_network_prefix_list_association/} for full documentation.
}
\keyword{internal}
