% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_create_link}
\alias{networkmanager_create_link}
\title{Creates a new link for a specified site}
\usage{
networkmanager_create_link(
  GlobalNetworkId,
  Description = NULL,
  Type = NULL,
  Bandwidth,
  Provider = NULL,
  SiteId,
  Tags = NULL
)
}
\arguments{
\item{GlobalNetworkId}{[required] The ID of the global network.}

\item{Description}{A description of the link.

Constraints: Maximum length of 256 characters.}

\item{Type}{The type of the link.

Constraints: Maximum length of 128 characters. Cannot include the
following characters: | \\ ^}

\item{Bandwidth}{[required] The upload speed and download speed in Mbps.}

\item{Provider}{The provider of the link.

Constraints: Maximum length of 128 characters. Cannot include the
following characters: | \\ ^}

\item{SiteId}{[required] The ID of the site.}

\item{Tags}{The tags to apply to the resource during creation.}
}
\description{
Creates a new link for a specified site.

See \url{https://paws-r.github.io/docs/networkmanager/create_link.html} for full documentation.
}
\keyword{internal}
