% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_create_site}
\alias{networkmanager_create_site}
\title{Creates a new site in a global network}
\usage{
networkmanager_create_site(
  GlobalNetworkId,
  Description = NULL,
  Location = NULL,
  Tags = NULL
)
}
\arguments{
\item{GlobalNetworkId}{[required] The ID of the global network.}

\item{Description}{A description of your site.

Constraints: Maximum length of 256 characters.}

\item{Location}{The site location. This information is used for visualization in the
Network Manager console. If you specify the address, the latitude and
longitude are automatically calculated.
\itemize{
\item \code{Address}: The physical address of the site.
\item \code{Latitude}: The latitude of the site.
\item \code{Longitude}: The longitude of the site.
}}

\item{Tags}{The tags to apply to the resource during creation.}
}
\description{
Creates a new site in a global network.

See \url{https://www.paws-r-sdk.com/docs/networkmanager_create_site/} for full documentation.
}
\keyword{internal}
