% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_create_transit_gateway_peering}
\alias{networkmanager_create_transit_gateway_peering}
\title{Creates a transit gateway peering connection}
\usage{
networkmanager_create_transit_gateway_peering(
  CoreNetworkId,
  TransitGatewayArn,
  Tags = NULL,
  ClientToken = NULL
)
}
\arguments{
\item{CoreNetworkId}{[required] The ID of a core network.}

\item{TransitGatewayArn}{[required] The ARN of the transit gateway for the peering request.}

\item{Tags}{The list of key-value tags associated with the request.}

\item{ClientToken}{The client token associated with the request.}
}
\description{
Creates a transit gateway peering connection.

See \url{https://www.paws-r-sdk.com/docs/networkmanager_create_transit_gateway_peering/} for full documentation.
}
\keyword{internal}
