% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_create_transit_gateway_route_table_attachment}
\alias{networkmanager_create_transit_gateway_route_table_attachment}
\title{Creates a transit gateway route table attachment}
\usage{
networkmanager_create_transit_gateway_route_table_attachment(
  PeeringId,
  TransitGatewayRouteTableArn,
  Tags = NULL,
  ClientToken = NULL
)
}
\arguments{
\item{PeeringId}{[required] The ID of the peer for the}

\item{TransitGatewayRouteTableArn}{[required] The ARN of the transit gateway route table for the attachment request.}

\item{Tags}{The list of key-value tags associated with the request.}

\item{ClientToken}{The client token associated with the request.}
}
\description{
Creates a transit gateway route table attachment.

See \url{https://paws-r.github.io/docs/networkmanager/create_transit_gateway_route_table_attachment.html} for full documentation.
}
\keyword{internal}
