% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_create_vpc_attachment}
\alias{networkmanager_create_vpc_attachment}
\title{Creates a VPC attachment on an edge location of a core network}
\usage{
networkmanager_create_vpc_attachment(
  CoreNetworkId,
  VpcArn,
  SubnetArns,
  Options = NULL,
  RoutingPolicyLabel = NULL,
  Tags = NULL,
  ClientToken = NULL
)
}
\arguments{
\item{CoreNetworkId}{[required] The ID of a core network for the VPC attachment.}

\item{VpcArn}{[required] The ARN of the VPC.}

\item{SubnetArns}{[required] The subnet ARN of the VPC attachment.}

\item{Options}{Options for the VPC attachment.}

\item{RoutingPolicyLabel}{The routing policy label to apply to the VPC attachment for traffic
routing decisions.}

\item{Tags}{The key-value tags associated with the request.}

\item{ClientToken}{The client token associated with the request.}
}
\description{
Creates a VPC attachment on an edge location of a core network.

See \url{https://www.paws-r-sdk.com/docs/networkmanager_create_vpc_attachment/} for full documentation.
}
\keyword{internal}
