% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_delete_connect_peer}
\alias{networkmanager_delete_connect_peer}
\title{Deletes a Connect peer}
\usage{
networkmanager_delete_connect_peer(ConnectPeerId)
}
\arguments{
\item{ConnectPeerId}{[required] The ID of the deleted Connect peer.}
}
\description{
Deletes a Connect peer.

See \url{https://paws-r.github.io/docs/networkmanager/delete_connect_peer.html} for full documentation.
}
\keyword{internal}
