% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_delete_connection}
\alias{networkmanager_delete_connection}
\title{Deletes the specified connection in your global network}
\usage{
networkmanager_delete_connection(GlobalNetworkId, ConnectionId)
}
\arguments{
\item{GlobalNetworkId}{[required] The ID of the global network.}

\item{ConnectionId}{[required] The ID of the connection.}
}
\description{
Deletes the specified connection in your global network.

See \url{https://paws-r.github.io/docs/networkmanager/delete_connection.html} for full documentation.
}
\keyword{internal}
