% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_delete_core_network}
\alias{networkmanager_delete_core_network}
\title{Deletes a core network along with all core network policies}
\usage{
networkmanager_delete_core_network(CoreNetworkId)
}
\arguments{
\item{CoreNetworkId}{[required] The network ID of the deleted core network.}
}
\description{
Deletes a core network along with all core network policies. This can only be done if there are no attachments on a core network.

See \url{https://www.paws-r-sdk.com/docs/networkmanager_delete_core_network/} for full documentation.
}
\keyword{internal}
