% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_delete_core_network_policy_version}
\alias{networkmanager_delete_core_network_policy_version}
\title{Deletes a policy version from a core network}
\usage{
networkmanager_delete_core_network_policy_version(
  CoreNetworkId,
  PolicyVersionId
)
}
\arguments{
\item{CoreNetworkId}{[required] The ID of a core network for the deleted policy.}

\item{PolicyVersionId}{[required] The version ID of the deleted policy.}
}
\description{
Deletes a policy version from a core network. You can't delete the current LIVE policy.

See \url{https://paws-r.github.io/docs/networkmanager/delete_core_network_policy_version.html} for full documentation.
}
\keyword{internal}
