% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_delete_core_network_prefix_list_association}
\alias{networkmanager_delete_core_network_prefix_list_association}
\title{Deletes an association between a core network and a prefix list}
\usage{
networkmanager_delete_core_network_prefix_list_association(
  CoreNetworkId,
  PrefixListArn
)
}
\arguments{
\item{CoreNetworkId}{[required] The ID of the core network from which to delete the prefix list
association.}

\item{PrefixListArn}{[required] The ARN of the prefix list to disassociate from the core network.}
}
\description{
Deletes an association between a core network and a prefix list.

See \url{https://www.paws-r-sdk.com/docs/networkmanager_delete_core_network_prefix_list_association/} for full documentation.
}
\keyword{internal}
