% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_delete_device}
\alias{networkmanager_delete_device}
\title{Deletes an existing device}
\usage{
networkmanager_delete_device(GlobalNetworkId, DeviceId)
}
\arguments{
\item{GlobalNetworkId}{[required] The ID of the global network.}

\item{DeviceId}{[required] The ID of the device.}
}
\description{
Deletes an existing device. You must first disassociate the device from any links and customer gateways.

See \url{https://www.paws-r-sdk.com/docs/networkmanager_delete_device/} for full documentation.
}
\keyword{internal}
