% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_delete_resource_policy}
\alias{networkmanager_delete_resource_policy}
\title{Deletes a resource policy for the specified resource}
\usage{
networkmanager_delete_resource_policy(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The ARN of the policy to delete.}
}
\description{
Deletes a resource policy for the specified resource. This revokes the access of the principals specified in the resource policy.

See \url{https://www.paws-r-sdk.com/docs/networkmanager_delete_resource_policy/} for full documentation.
}
\keyword{internal}
