% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_delete_site}
\alias{networkmanager_delete_site}
\title{Deletes an existing site}
\usage{
networkmanager_delete_site(GlobalNetworkId, SiteId)
}
\arguments{
\item{GlobalNetworkId}{[required] The ID of the global network.}

\item{SiteId}{[required] The ID of the site.}
}
\description{
Deletes an existing site. The site cannot be associated with any device or link.

See \url{https://paws-r.github.io/docs/networkmanager/delete_site.html} for full documentation.
}
\keyword{internal}
