% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_describe_global_networks}
\alias{networkmanager_describe_global_networks}
\title{Describes one or more global networks}
\usage{
networkmanager_describe_global_networks(
  GlobalNetworkIds = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{GlobalNetworkIds}{The IDs of one or more global networks. The maximum is 10.}

\item{MaxResults}{The maximum number of results to return.}

\item{NextToken}{The token for the next page of results.}
}
\description{
Describes one or more global networks. By default, all global networks are described. To describe the objects in your global network, you must use the appropriate \verb{Get*} action. For example, to list the transit gateways in your global network, use \code{\link[=networkmanager_get_transit_gateway_registrations]{get_transit_gateway_registrations}}.

See \url{https://paws-r.github.io/docs/networkmanager/describe_global_networks.html} for full documentation.
}
\keyword{internal}
