% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_disassociate_customer_gateway}
\alias{networkmanager_disassociate_customer_gateway}
\title{Disassociates a customer gateway from a device and a link}
\usage{
networkmanager_disassociate_customer_gateway(
  GlobalNetworkId,
  CustomerGatewayArn
)
}
\arguments{
\item{GlobalNetworkId}{[required] The ID of the global network.}

\item{CustomerGatewayArn}{[required] The Amazon Resource Name (ARN) of the customer gateway.}
}
\description{
Disassociates a customer gateway from a device and a link.

See \url{https://paws-r.github.io/docs/networkmanager/disassociate_customer_gateway.html} for full documentation.
}
\keyword{internal}
