% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_disassociate_transit_gateway_connect_peer}
\alias{networkmanager_disassociate_transit_gateway_connect_peer}
\title{Disassociates a transit gateway Connect peer from a device and link}
\usage{
networkmanager_disassociate_transit_gateway_connect_peer(
  GlobalNetworkId,
  TransitGatewayConnectPeerArn
)
}
\arguments{
\item{GlobalNetworkId}{[required] The ID of the global network.}

\item{TransitGatewayConnectPeerArn}{[required] The Amazon Resource Name (ARN) of the transit gateway Connect peer.}
}
\description{
Disassociates a transit gateway Connect peer from a device and link.

See \url{https://www.paws-r-sdk.com/docs/networkmanager_disassociate_transit_gateway_connect_peer/} for full documentation.
}
\keyword{internal}
