% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_get_connect_peer_associations}
\alias{networkmanager_get_connect_peer_associations}
\title{Returns information about a core network Connect peer associations}
\usage{
networkmanager_get_connect_peer_associations(
  GlobalNetworkId,
  ConnectPeerIds = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{GlobalNetworkId}{[required] The ID of the global network.}

\item{ConnectPeerIds}{The IDs of the Connect peers.}

\item{MaxResults}{The maximum number of results to return.}

\item{NextToken}{The token for the next page of results.}
}
\description{
Returns information about a core network Connect peer associations.

See \url{https://paws-r.github.io/docs/networkmanager/get_connect_peer_associations.html} for full documentation.
}
\keyword{internal}
