% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_get_core_network}
\alias{networkmanager_get_core_network}
\title{Returns information about the LIVE policy for a core network}
\usage{
networkmanager_get_core_network(CoreNetworkId)
}
\arguments{
\item{CoreNetworkId}{[required] The ID of a core network.}
}
\description{
Returns information about the LIVE policy for a core network.

See \url{https://www.paws-r-sdk.com/docs/networkmanager_get_core_network/} for full documentation.
}
\keyword{internal}
