% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_get_core_network_change_set}
\alias{networkmanager_get_core_network_change_set}
\title{Returns a change set between the LIVE core network policy and a
submitted policy}
\usage{
networkmanager_get_core_network_change_set(
  CoreNetworkId,
  PolicyVersionId,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{CoreNetworkId}{[required] The ID of a core network.}

\item{PolicyVersionId}{[required] The ID of the policy version.}

\item{MaxResults}{The maximum number of results to return.}

\item{NextToken}{The token for the next page of results.}
}
\description{
Returns a change set between the LIVE core network policy and a submitted policy.

See \url{https://www.paws-r-sdk.com/docs/networkmanager_get_core_network_change_set/} for full documentation.
}
\keyword{internal}
