% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_get_core_network_policy}
\alias{networkmanager_get_core_network_policy}
\title{Returns details about a core network policy}
\usage{
networkmanager_get_core_network_policy(
  CoreNetworkId,
  PolicyVersionId = NULL,
  Alias = NULL
)
}
\arguments{
\item{CoreNetworkId}{[required] The ID of a core network.}

\item{PolicyVersionId}{The ID of a core network policy version.}

\item{Alias}{The alias of a core network policy}
}
\description{
Returns details about a core network policy. You can get details about your current live policy or any previous policy version.

See \url{https://www.paws-r-sdk.com/docs/networkmanager_get_core_network_policy/} for full documentation.
}
\keyword{internal}
