% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_get_customer_gateway_associations}
\alias{networkmanager_get_customer_gateway_associations}
\title{Gets the association information for customer gateways that are
associated with devices and links in your global network}
\usage{
networkmanager_get_customer_gateway_associations(
  GlobalNetworkId,
  CustomerGatewayArns = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{GlobalNetworkId}{[required] The ID of the global network.}

\item{CustomerGatewayArns}{One or more customer gateway Amazon Resource Names (ARNs). The maximum
is 10.}

\item{MaxResults}{The maximum number of results to return.}

\item{NextToken}{The token for the next page of results.}
}
\description{
Gets the association information for customer gateways that are associated with devices and links in your global network.

See \url{https://paws-r.github.io/docs/networkmanager/get_customer_gateway_associations.html} for full documentation.
}
\keyword{internal}
