% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_get_devices}
\alias{networkmanager_get_devices}
\title{Gets information about one or more of your devices in a global network}
\usage{
networkmanager_get_devices(
  GlobalNetworkId,
  DeviceIds = NULL,
  SiteId = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{GlobalNetworkId}{[required] The ID of the global network.}

\item{DeviceIds}{One or more device IDs. The maximum is 10.}

\item{SiteId}{The ID of the site.}

\item{MaxResults}{The maximum number of results to return.}

\item{NextToken}{The token for the next page of results.}
}
\description{
Gets information about one or more of your devices in a global network.

See \url{https://www.paws-r-sdk.com/docs/networkmanager_get_devices/} for full documentation.
}
\keyword{internal}
