% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_get_direct_connect_gateway_attachment}
\alias{networkmanager_get_direct_connect_gateway_attachment}
\title{Returns information about a specific Amazon Web Services Direct Connect
gateway attachment}
\usage{
networkmanager_get_direct_connect_gateway_attachment(AttachmentId)
}
\arguments{
\item{AttachmentId}{[required] The ID of the Direct Connect gateway attachment that you want to see
details about.}
}
\description{
Returns information about a specific Amazon Web Services Direct Connect gateway attachment.

See \url{https://www.paws-r-sdk.com/docs/networkmanager_get_direct_connect_gateway_attachment/} for full documentation.
}
\keyword{internal}
