% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_get_links}
\alias{networkmanager_get_links}
\title{Gets information about one or more links in a specified global network}
\usage{
networkmanager_get_links(
  GlobalNetworkId,
  LinkIds = NULL,
  SiteId = NULL,
  Type = NULL,
  Provider = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{GlobalNetworkId}{[required] The ID of the global network.}

\item{LinkIds}{One or more link IDs. The maximum is 10.}

\item{SiteId}{The ID of the site.}

\item{Type}{The link type.}

\item{Provider}{The link provider.}

\item{MaxResults}{The maximum number of results to return.}

\item{NextToken}{The token for the next page of results.}
}
\description{
Gets information about one or more links in a specified global network.

See \url{https://www.paws-r-sdk.com/docs/networkmanager_get_links/} for full documentation.
}
\keyword{internal}
