% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_get_resource_policy}
\alias{networkmanager_get_resource_policy}
\title{Returns information about a resource policy}
\usage{
networkmanager_get_resource_policy(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The ARN of the resource.}
}
\description{
Returns information about a resource policy.

See \url{https://paws-r.github.io/docs/networkmanager/get_resource_policy.html} for full documentation.
}
\keyword{internal}
