% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_get_site_to_site_vpn_attachment}
\alias{networkmanager_get_site_to_site_vpn_attachment}
\title{Returns information about a site-to-site VPN attachment}
\usage{
networkmanager_get_site_to_site_vpn_attachment(AttachmentId)
}
\arguments{
\item{AttachmentId}{[required] The ID of the attachment.}
}
\description{
Returns information about a site-to-site VPN attachment.

See \url{https://paws-r.github.io/docs/networkmanager/get_site_to_site_vpn_attachment.html} for full documentation.
}
\keyword{internal}
