% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_get_transit_gateway_peering}
\alias{networkmanager_get_transit_gateway_peering}
\title{Returns information about a transit gateway peer}
\usage{
networkmanager_get_transit_gateway_peering(PeeringId)
}
\arguments{
\item{PeeringId}{[required] The ID of the peering request.}
}
\description{
Returns information about a transit gateway peer.

See \url{https://paws-r.github.io/docs/networkmanager/get_transit_gateway_peering.html} for full documentation.
}
\keyword{internal}
