% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_get_transit_gateway_registrations}
\alias{networkmanager_get_transit_gateway_registrations}
\title{Gets information about the transit gateway registrations in a specified
global network}
\usage{
networkmanager_get_transit_gateway_registrations(
  GlobalNetworkId,
  TransitGatewayArns = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{GlobalNetworkId}{[required] The ID of the global network.}

\item{TransitGatewayArns}{The Amazon Resource Names (ARNs) of one or more transit gateways. The
maximum is 10.}

\item{MaxResults}{The maximum number of results to return.}

\item{NextToken}{The token for the next page of results.}
}
\description{
Gets information about the transit gateway registrations in a specified global network.

See \url{https://paws-r.github.io/docs/networkmanager/get_transit_gateway_registrations.html} for full documentation.
}
\keyword{internal}
