% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_list_attachment_routing_policy_associations}
\alias{networkmanager_list_attachment_routing_policy_associations}
\title{Lists the routing policy associations for attachments in a core network}
\usage{
networkmanager_list_attachment_routing_policy_associations(
  CoreNetworkId,
  AttachmentId = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{CoreNetworkId}{[required] The ID of the core network to list attachment routing policy
associations for.}

\item{AttachmentId}{The ID of a specific attachment to filter the routing policy
associations.}

\item{MaxResults}{The maximum number of results to return in a single page.}

\item{NextToken}{The token for the next page of results.}
}
\description{
Lists the routing policy associations for attachments in a core network.

See \url{https://www.paws-r-sdk.com/docs/networkmanager_list_attachment_routing_policy_associations/} for full documentation.
}
\keyword{internal}
