% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_list_attachments}
\alias{networkmanager_list_attachments}
\title{Returns a list of core network attachments}
\usage{
networkmanager_list_attachments(
  CoreNetworkId = NULL,
  AttachmentType = NULL,
  EdgeLocation = NULL,
  State = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{CoreNetworkId}{The ID of a core network.}

\item{AttachmentType}{The type of attachment.}

\item{EdgeLocation}{The Region where the edge is located.}

\item{State}{The state of the attachment.}

\item{MaxResults}{The maximum number of results to return.}

\item{NextToken}{The token for the next page of results.}
}
\description{
Returns a list of core network attachments.

See \url{https://www.paws-r-sdk.com/docs/networkmanager_list_attachments/} for full documentation.
}
\keyword{internal}
