% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_list_connect_peers}
\alias{networkmanager_list_connect_peers}
\title{Returns a list of core network Connect peers}
\usage{
networkmanager_list_connect_peers(
  CoreNetworkId = NULL,
  ConnectAttachmentId = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{CoreNetworkId}{The ID of a core network.}

\item{ConnectAttachmentId}{The ID of the attachment.}

\item{MaxResults}{The maximum number of results to return.}

\item{NextToken}{The token for the next page of results.}
}
\description{
Returns a list of core network Connect peers.

See \url{https://www.paws-r-sdk.com/docs/networkmanager_list_connect_peers/} for full documentation.
}
\keyword{internal}
