% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_list_core_network_policy_versions}
\alias{networkmanager_list_core_network_policy_versions}
\title{Returns a list of core network policy versions}
\usage{
networkmanager_list_core_network_policy_versions(
  CoreNetworkId,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{CoreNetworkId}{[required] The ID of a core network.}

\item{MaxResults}{The maximum number of results to return.}

\item{NextToken}{The token for the next page of results.}
}
\description{
Returns a list of core network policy versions.

See \url{https://paws-r.github.io/docs/networkmanager/list_core_network_policy_versions.html} for full documentation.
}
\keyword{internal}
