% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_list_core_network_prefix_list_associations}
\alias{networkmanager_list_core_network_prefix_list_associations}
\title{Lists the prefix list associations for a core network}
\usage{
networkmanager_list_core_network_prefix_list_associations(
  CoreNetworkId,
  PrefixListArn = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{CoreNetworkId}{[required] The ID of the core network to list prefix list associations for.}

\item{PrefixListArn}{The ARN of a specific prefix list to filter the associations.}

\item{MaxResults}{The maximum number of results to return in a single page.}

\item{NextToken}{The token for the next page of results.}
}
\description{
Lists the prefix list associations for a core network.

See \url{https://www.paws-r-sdk.com/docs/networkmanager_list_core_network_prefix_list_associations/} for full documentation.
}
\keyword{internal}
