% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_list_core_network_routing_information}
\alias{networkmanager_list_core_network_routing_information}
\title{Lists routing information for a core network, including routes and their
attributes}
\usage{
networkmanager_list_core_network_routing_information(
  CoreNetworkId,
  SegmentName,
  EdgeLocation,
  NextHopFilters = NULL,
  LocalPreferenceMatches = NULL,
  ExactAsPathMatches = NULL,
  MedMatches = NULL,
  CommunityMatches = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{CoreNetworkId}{[required] The ID of the core network to retrieve routing information for.}

\item{SegmentName}{[required] The name of the segment to filter routing information by.}

\item{EdgeLocation}{[required] The edge location to filter routing information by.}

\item{NextHopFilters}{Filters to apply based on next hop information.}

\item{LocalPreferenceMatches}{Local preference values to match when filtering routing information.}

\item{ExactAsPathMatches}{Exact AS path values to match when filtering routing information.}

\item{MedMatches}{Multi-Exit Discriminator (MED) values to match when filtering routing
information.}

\item{CommunityMatches}{BGP community values to match when filtering routing information.}

\item{MaxResults}{The maximum number of routing information entries to return in a single
page.}

\item{NextToken}{The token for the next page of results.}
}
\description{
Lists routing information for a core network, including routes and their attributes.

See \url{https://www.paws-r-sdk.com/docs/networkmanager_list_core_network_routing_information/} for full documentation.
}
\keyword{internal}
