% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_list_core_networks}
\alias{networkmanager_list_core_networks}
\title{Returns a list of owned and shared core networks}
\usage{
networkmanager_list_core_networks(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{The maximum number of results to return.}

\item{NextToken}{The token for the next page of results.}
}
\description{
Returns a list of owned and shared core networks.

See \url{https://paws-r.github.io/docs/networkmanager/list_core_networks.html} for full documentation.
}
\keyword{internal}
