% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_list_organization_service_access_status}
\alias{networkmanager_list_organization_service_access_status}
\title{Gets the status of the Service Linked Role (SLR) deployment for the
accounts in a given Amazon Web Services Organization}
\usage{
networkmanager_list_organization_service_access_status(
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{MaxResults}{The maximum number of results to return.}

\item{NextToken}{The token for the next page of results.}
}
\description{
Gets the status of the Service Linked Role (SLR) deployment for the accounts in a given Amazon Web Services Organization.

See \url{https://www.paws-r-sdk.com/docs/networkmanager_list_organization_service_access_status/} for full documentation.
}
\keyword{internal}
