% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_list_peerings}
\alias{networkmanager_list_peerings}
\title{Lists the peerings for a core network}
\usage{
networkmanager_list_peerings(
  CoreNetworkId = NULL,
  PeeringType = NULL,
  EdgeLocation = NULL,
  State = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{CoreNetworkId}{The ID of a core network.}

\item{PeeringType}{Returns a list of a peering requests.}

\item{EdgeLocation}{Returns a list edge locations for the}

\item{State}{Returns a list of the peering request states.}

\item{MaxResults}{The maximum number of results to return.}

\item{NextToken}{The token for the next page of results.}
}
\description{
Lists the peerings for a core network.

See \url{https://www.paws-r-sdk.com/docs/networkmanager_list_peerings/} for full documentation.
}
\keyword{internal}
