% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_put_attachment_routing_policy_label}
\alias{networkmanager_put_attachment_routing_policy_label}
\title{Applies a routing policy label to an attachment for traffic routing
decisions}
\usage{
networkmanager_put_attachment_routing_policy_label(
  CoreNetworkId,
  AttachmentId,
  RoutingPolicyLabel,
  ClientToken = NULL
)
}
\arguments{
\item{CoreNetworkId}{[required] The ID of the core network containing the attachment.}

\item{AttachmentId}{[required] The ID of the attachment to apply the routing policy label to.}

\item{RoutingPolicyLabel}{[required] The routing policy label to apply to the attachment.}

\item{ClientToken}{A unique, case-sensitive identifier that you provide to ensure the
idempotency of the request.}
}
\description{
Applies a routing policy label to an attachment for traffic routing decisions.

See \url{https://www.paws-r-sdk.com/docs/networkmanager_put_attachment_routing_policy_label/} for full documentation.
}
\keyword{internal}
