% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_put_core_network_policy}
\alias{networkmanager_put_core_network_policy}
\title{Creates a new, immutable version of a core network policy}
\usage{
networkmanager_put_core_network_policy(
  CoreNetworkId,
  PolicyDocument,
  Description = NULL,
  LatestVersionId = NULL,
  ClientToken = NULL
)
}
\arguments{
\item{CoreNetworkId}{[required] The ID of a core network.}

\item{PolicyDocument}{[required] The policy document.}

\item{Description}{a core network policy description.}

\item{LatestVersionId}{The ID of a core network policy.}

\item{ClientToken}{The client token associated with the request.}
}
\description{
Creates a new, immutable version of a core network policy. A subsequent change set is created showing the differences between the LIVE policy and the submitted policy.

See \url{https://paws-r.github.io/docs/networkmanager/put_core_network_policy.html} for full documentation.
}
\keyword{internal}
