% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_put_resource_policy}
\alias{networkmanager_put_resource_policy}
\title{Creates or updates a resource policy}
\usage{
networkmanager_put_resource_policy(PolicyDocument, ResourceArn)
}
\arguments{
\item{PolicyDocument}{[required] The JSON resource policy document.}

\item{ResourceArn}{[required] The ARN of the resource policy.}
}
\description{
Creates or updates a resource policy.

See \url{https://paws-r.github.io/docs/networkmanager/put_resource_policy.html} for full documentation.
}
\keyword{internal}
