% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_register_transit_gateway}
\alias{networkmanager_register_transit_gateway}
\title{Registers a transit gateway in your global network}
\usage{
networkmanager_register_transit_gateway(GlobalNetworkId, TransitGatewayArn)
}
\arguments{
\item{GlobalNetworkId}{[required] The ID of the global network.}

\item{TransitGatewayArn}{[required] The Amazon Resource Name (ARN) of the transit gateway.}
}
\description{
Registers a transit gateway in your global network. The transit gateway can be in any Amazon Web Services Region, but it must be owned by the same Amazon Web Services account that owns the global network. You cannot register a transit gateway in more than one global network.

See \url{https://paws-r.github.io/docs/networkmanager/register_transit_gateway.html} for full documentation.
}
\keyword{internal}
