% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_reject_attachment}
\alias{networkmanager_reject_attachment}
\title{Rejects a core network attachment request}
\usage{
networkmanager_reject_attachment(AttachmentId)
}
\arguments{
\item{AttachmentId}{[required] The ID of the attachment.}
}
\description{
Rejects a core network attachment request.

See \url{https://www.paws-r-sdk.com/docs/networkmanager_reject_attachment/} for full documentation.
}
\keyword{internal}
