% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_remove_attachment_routing_policy_label}
\alias{networkmanager_remove_attachment_routing_policy_label}
\title{Removes a routing policy label from an attachment}
\usage{
networkmanager_remove_attachment_routing_policy_label(
  CoreNetworkId,
  AttachmentId
)
}
\arguments{
\item{CoreNetworkId}{[required] The ID of the core network containing the attachment.}

\item{AttachmentId}{[required] The ID of the attachment to remove the routing policy label from.}
}
\description{
Removes a routing policy label from an attachment.

See \url{https://www.paws-r-sdk.com/docs/networkmanager_remove_attachment_routing_policy_label/} for full documentation.
}
\keyword{internal}
