% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_restore_core_network_policy_version}
\alias{networkmanager_restore_core_network_policy_version}
\title{Restores a previous policy version as a new, immutable version of a core
network policy}
\usage{
networkmanager_restore_core_network_policy_version(
  CoreNetworkId,
  PolicyVersionId
)
}
\arguments{
\item{CoreNetworkId}{[required] The ID of a core network.}

\item{PolicyVersionId}{[required] The ID of the policy version to restore.}
}
\description{
Restores a previous policy version as a new, immutable version of a core network policy. A subsequent change set is created showing the differences between the LIVE policy and restored policy.

See \url{https://paws-r.github.io/docs/networkmanager/restore_core_network_policy_version.html} for full documentation.
}
\keyword{internal}
