% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_start_organization_service_access_update}
\alias{networkmanager_start_organization_service_access_update}
\title{Enables the Network Manager service for an Amazon Web Services
Organization}
\usage{
networkmanager_start_organization_service_access_update(Action)
}
\arguments{
\item{Action}{[required] The action to take for the update request. This can be either \code{ENABLE}
or \code{DISABLE}.}
}
\description{
Enables the Network Manager service for an Amazon Web Services Organization. This can only be called by a management account within the organization.

See \url{https://paws-r.github.io/docs/networkmanager/start_organization_service_access_update.html} for full documentation.
}
\keyword{internal}
