% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_tag_resource}
\alias{networkmanager_tag_resource}
\title{Tags a specified resource}
\usage{
networkmanager_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource.}

\item{Tags}{[required] The tags to apply to the specified resource.}
}
\description{
Tags a specified resource.

See \url{https://paws-r.github.io/docs/networkmanager/tag_resource.html} for full documentation.
}
\keyword{internal}
