% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_update_connection}
\alias{networkmanager_update_connection}
\title{Updates the information for an existing connection}
\usage{
networkmanager_update_connection(
  GlobalNetworkId,
  ConnectionId,
  LinkId = NULL,
  ConnectedLinkId = NULL,
  Description = NULL
)
}
\arguments{
\item{GlobalNetworkId}{[required] The ID of the global network.}

\item{ConnectionId}{[required] The ID of the connection.}

\item{LinkId}{The ID of the link for the first device in the connection.}

\item{ConnectedLinkId}{The ID of the link for the second device in the connection.}

\item{Description}{A description of the connection.

Length Constraints: Maximum length of 256 characters.}
}
\description{
Updates the information for an existing connection. To remove information for any of the parameters, specify an empty string.

See \url{https://www.paws-r-sdk.com/docs/networkmanager_update_connection/} for full documentation.
}
\keyword{internal}
