% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_update_core_network}
\alias{networkmanager_update_core_network}
\title{Updates the description of a core network}
\usage{
networkmanager_update_core_network(CoreNetworkId, Description = NULL)
}
\arguments{
\item{CoreNetworkId}{[required] The ID of a core network.}

\item{Description}{The description of the update.}
}
\description{
Updates the description of a core network.

See \url{https://www.paws-r-sdk.com/docs/networkmanager_update_core_network/} for full documentation.
}
\keyword{internal}
