% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_update_direct_connect_gateway_attachment}
\alias{networkmanager_update_direct_connect_gateway_attachment}
\title{Updates the edge locations associated with an Amazon Web Services Direct
Connect gateway attachment}
\usage{
networkmanager_update_direct_connect_gateway_attachment(
  AttachmentId,
  EdgeLocations = NULL
)
}
\arguments{
\item{AttachmentId}{[required] The ID of the Direct Connect gateway attachment for the updated edge
locations.}

\item{EdgeLocations}{One or more edge locations to update for the Direct Connect gateway
attachment. The updated array of edge locations overwrites the previous
array of locations. \code{EdgeLocations} is only used for Direct Connect
gateway attachments.}
}
\description{
Updates the edge locations associated with an Amazon Web Services Direct Connect gateway attachment.

See \url{https://www.paws-r-sdk.com/docs/networkmanager_update_direct_connect_gateway_attachment/} for full documentation.
}
\keyword{internal}
