% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_update_link}
\alias{networkmanager_update_link}
\title{Updates the details for an existing link}
\usage{
networkmanager_update_link(
  GlobalNetworkId,
  LinkId,
  Description = NULL,
  Type = NULL,
  Bandwidth = NULL,
  Provider = NULL
)
}
\arguments{
\item{GlobalNetworkId}{[required] The ID of the global network.}

\item{LinkId}{[required] The ID of the link.}

\item{Description}{A description of the link.

Constraints: Maximum length of 256 characters.}

\item{Type}{The type of the link.

Constraints: Maximum length of 128 characters.}

\item{Bandwidth}{The upload and download speed in Mbps.}

\item{Provider}{The provider of the link.

Constraints: Maximum length of 128 characters.}
}
\description{
Updates the details for an existing link. To remove information for any of the parameters, specify an empty string.

See \url{https://www.paws-r-sdk.com/docs/networkmanager_update_link/} for full documentation.
}
\keyword{internal}
