% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_update_network_resource_metadata}
\alias{networkmanager_update_network_resource_metadata}
\title{Updates the resource metadata for the specified global network}
\usage{
networkmanager_update_network_resource_metadata(
  GlobalNetworkId,
  ResourceArn,
  Metadata
)
}
\arguments{
\item{GlobalNetworkId}{[required] The ID of the global network.}

\item{ResourceArn}{[required] The ARN of the resource.}

\item{Metadata}{[required] The resource metadata.}
}
\description{
Updates the resource metadata for the specified global network.

See \url{https://www.paws-r-sdk.com/docs/networkmanager_update_network_resource_metadata/} for full documentation.
}
\keyword{internal}
