% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_update_vpc_attachment}
\alias{networkmanager_update_vpc_attachment}
\title{Updates a VPC attachment}
\usage{
networkmanager_update_vpc_attachment(
  AttachmentId,
  AddSubnetArns = NULL,
  RemoveSubnetArns = NULL,
  Options = NULL
)
}
\arguments{
\item{AttachmentId}{[required] The ID of the attachment.}

\item{AddSubnetArns}{Adds a subnet ARN to the VPC attachment.}

\item{RemoveSubnetArns}{Removes a subnet ARN from the attachment.}

\item{Options}{Additional options for updating the VPC attachment.}
}
\description{
Updates a VPC attachment.

See \url{https://www.paws-r-sdk.com/docs/networkmanager_update_vpc_attachment/} for full documentation.
}
\keyword{internal}
