% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53_operations.R
\name{route53_associate_vpc_with_hosted_zone}
\alias{route53_associate_vpc_with_hosted_zone}
\title{Associates an Amazon VPC with a private hosted zone}
\usage{
route53_associate_vpc_with_hosted_zone(HostedZoneId, VPC, Comment = NULL)
}
\arguments{
\item{HostedZoneId}{[required] The ID of the private hosted zone that you want to associate an Amazon
VPC with.

Note that you can't associate a VPC with a hosted zone that doesn't have
an existing VPC association.}

\item{VPC}{[required] A complex type that contains information about the VPC that you want to
associate with a private hosted zone.}

\item{Comment}{\emph{Optional:} A comment about the association request.}
}
\description{
Associates an Amazon VPC with a private hosted zone.

See \url{https://paws-r.github.io/docs/route53/associate_vpc_with_hosted_zone.html} for full documentation.
}
\keyword{internal}
