% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53_operations.R
\name{route53_change_resource_record_sets}
\alias{route53_change_resource_record_sets}
\title{Creates, changes, or deletes a resource record set, which contains
authoritative DNS information for a specified domain name or subdomain
name}
\usage{
route53_change_resource_record_sets(HostedZoneId, ChangeBatch)
}
\arguments{
\item{HostedZoneId}{[required] The ID of the hosted zone that contains the resource record sets that
you want to change.}

\item{ChangeBatch}{[required] A complex type that contains an optional comment and the \code{Changes}
element.}
}
\description{
Creates, changes, or deletes a resource record set, which contains authoritative DNS information for a specified domain name or subdomain name. For example, you can use \code{\link[=route53_change_resource_record_sets]{change_resource_record_sets}} to create a resource record set that routes traffic for test.example.com to a web server that has an IP address of 192.0.2.44.

See \url{https://paws-r.github.io/docs/route53/change_resource_record_sets.html} for full documentation.
}
\keyword{internal}
