% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53_operations.R
\name{route53_delete_cidr_collection}
\alias{route53_delete_cidr_collection}
\title{Deletes a CIDR collection in the current Amazon Web Services account}
\usage{
route53_delete_cidr_collection(Id)
}
\arguments{
\item{Id}{[required] The UUID of the collection to delete.}
}
\description{
Deletes a CIDR collection in the current Amazon Web Services account. The collection must be empty before it can be deleted.

See \url{https://www.paws-r-sdk.com/docs/route53_delete_cidr_collection/} for full documentation.
}
\keyword{internal}
