% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53_operations.R
\name{route53_delete_health_check}
\alias{route53_delete_health_check}
\title{Deletes a health check}
\usage{
route53_delete_health_check(HealthCheckId)
}
\arguments{
\item{HealthCheckId}{[required] The ID of the health check that you want to delete.}
}
\description{
Deletes a health check.

See \url{https://paws-r.github.io/docs/route53/delete_health_check.html} for full documentation.
}
\keyword{internal}
