% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53_operations.R
\name{route53_delete_hosted_zone}
\alias{route53_delete_hosted_zone}
\title{Deletes a hosted zone}
\usage{
route53_delete_hosted_zone(Id)
}
\arguments{
\item{Id}{[required] The ID of the hosted zone you want to delete.}
}
\description{
Deletes a hosted zone.

See \url{https://www.paws-r-sdk.com/docs/route53_delete_hosted_zone/} for full documentation.
}
\keyword{internal}
