% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53_operations.R
\name{route53_delete_query_logging_config}
\alias{route53_delete_query_logging_config}
\title{Deletes a configuration for DNS query logging}
\usage{
route53_delete_query_logging_config(Id)
}
\arguments{
\item{Id}{[required] The ID of the configuration that you want to delete.}
}
\description{
Deletes a configuration for DNS query logging. If you delete a configuration, Amazon Route 53 stops sending query logs to CloudWatch Logs. Route 53 doesn't delete any logs that are already in CloudWatch Logs.

See \url{https://paws-r.github.io/docs/route53/delete_query_logging_config.html} for full documentation.
}
\keyword{internal}
