% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53_operations.R
\name{route53_delete_traffic_policy_instance}
\alias{route53_delete_traffic_policy_instance}
\title{Deletes a traffic policy instance and all of the resource record sets
that Amazon Route 53 created when you created the instance}
\usage{
route53_delete_traffic_policy_instance(Id)
}
\arguments{
\item{Id}{[required] The ID of the traffic policy instance that you want to delete.

When you delete a traffic policy instance, Amazon Route 53 also deletes
all of the resource record sets that were created when you created the
traffic policy instance.}
}
\description{
Deletes a traffic policy instance and all of the resource record sets that Amazon Route 53 created when you created the instance.

See \url{https://paws-r.github.io/docs/route53/delete_traffic_policy_instance.html} for full documentation.
}
\keyword{internal}
