% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53_operations.R
\name{route53_disable_hosted_zone_dnssec}
\alias{route53_disable_hosted_zone_dnssec}
\title{Disables DNSSEC signing in a specific hosted zone}
\usage{
route53_disable_hosted_zone_dnssec(HostedZoneId)
}
\arguments{
\item{HostedZoneId}{[required] A unique string used to identify a hosted zone.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ChangeInfo = list(
    Id = "string",
    Status = "PENDING"|"INSYNC",
    SubmittedAt = as.POSIXct(
      "2015-01-01"
    ),
    Comment = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Disables DNSSEC signing in a specific hosted zone. This action does not
deactivate any key signing keys (KSKs) that are active in the hosted
zone.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$disable_hosted_zone_dnssec(
  HostedZoneId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
