% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53_operations.R
\name{route53_disassociate_vpc_from_hosted_zone}
\alias{route53_disassociate_vpc_from_hosted_zone}
\title{Disassociates an Amazon Virtual Private Cloud (Amazon VPC) from an
Amazon Route 53 private hosted zone}
\usage{
route53_disassociate_vpc_from_hosted_zone(HostedZoneId, VPC, Comment = NULL)
}
\arguments{
\item{HostedZoneId}{[required] The ID of the private hosted zone that you want to disassociate a VPC
from.}

\item{VPC}{[required] A complex type that contains information about the VPC that you're
disassociating from the specified hosted zone.}

\item{Comment}{\emph{Optional:} A comment about the disassociation request.}
}
\description{
Disassociates an Amazon Virtual Private Cloud (Amazon VPC) from an Amazon Route 53 private hosted zone. Note the following:

See \url{https://paws-r.github.io/docs/route53/disassociate_vpc_from_hosted_zone.html} for full documentation.
}
\keyword{internal}
