% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53_operations.R
\name{route53_enable_hosted_zone_dnssec}
\alias{route53_enable_hosted_zone_dnssec}
\title{Enables DNSSEC signing in a specific hosted zone}
\usage{
route53_enable_hosted_zone_dnssec(HostedZoneId)
}
\arguments{
\item{HostedZoneId}{[required] A unique string used to identify a hosted zone.}
}
\description{
Enables DNSSEC signing in a specific hosted zone.
}
\section{Request syntax}{
\preformatted{svc$enable_hosted_zone_dnssec(
  HostedZoneId = "string"
)
}
}

\keyword{internal}
