% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53_operations.R
\name{route53_get_change}
\alias{route53_get_change}
\title{Returns the current status of a change batch request}
\usage{
route53_get_change(Id)
}
\arguments{
\item{Id}{[required] The ID of the change batch request. The value that you specify here is
the value that
\code{\link[=route53_change_resource_record_sets]{change_resource_record_sets}}
returned in the \code{Id} element when you submitted the request.}
}
\description{
Returns the current status of a change batch request. The status is one of the following values:

See \url{https://paws-r.github.io/docs/route53/get_change.html} for full documentation.
}
\keyword{internal}
