% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53_operations.R
\name{route53_get_checker_ip_ranges}
\alias{route53_get_checker_ip_ranges}
\title{Route 53 does not perform authorization for this API because it
retrieves information that is already available to the public}
\usage{
route53_get_checker_ip_ranges()
}
\description{
Route 53 does not perform authorization for this API because it retrieves information that is already available to the public.

See \url{https://paws-r.github.io/docs/route53/get_checker_ip_ranges.html} for full documentation.
}
\keyword{internal}
