% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53_operations.R
\name{route53_get_health_check_count}
\alias{route53_get_health_check_count}
\title{Retrieves the number of health checks that are associated with the
current Amazon Web Services account}
\usage{
route53_get_health_check_count()
}
\description{
Retrieves the number of health checks that are associated with the current Amazon Web Services account.

See \url{https://paws-r.github.io/docs/route53/get_health_check_count.html} for full documentation.
}
\keyword{internal}
