% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53_operations.R
\name{route53_get_hosted_zone}
\alias{route53_get_hosted_zone}
\title{Gets information about a specified hosted zone including the four name
servers assigned to the hosted zone}
\usage{
route53_get_hosted_zone(Id)
}
\arguments{
\item{Id}{[required] The ID of the hosted zone that you want to get information about.}
}
\description{
Gets information about a specified hosted zone including the four name servers assigned to the hosted zone.

See \url{https://paws-r.github.io/docs/route53/get_hosted_zone.html} for full documentation.
}
\keyword{internal}
