% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53_operations.R
\name{route53_get_hosted_zone_count}
\alias{route53_get_hosted_zone_count}
\title{Retrieves the number of hosted zones that are associated with the
current AWS account}
\usage{
route53_get_hosted_zone_count()
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  HostedZoneCount = 123
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves the number of hosted zones that are associated with the
current AWS account.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_hosted_zone_count()
}\if{html}{\out{</div>}}
}

\keyword{internal}
