% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53_operations.R
\name{route53_get_query_logging_config}
\alias{route53_get_query_logging_config}
\title{Gets information about a specified configuration for DNS query logging}
\usage{
route53_get_query_logging_config(Id)
}
\arguments{
\item{Id}{[required] The ID of the configuration for DNS query logging that you want to get
information about.}
}
\description{
Gets information about a specified configuration for DNS query logging.

See \url{https://www.paws-r-sdk.com/docs/route53_get_query_logging_config/} for full documentation.
}
\keyword{internal}
