% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53_operations.R
\name{route53_get_reusable_delegation_set}
\alias{route53_get_reusable_delegation_set}
\title{Retrieves information about a specified reusable delegation set,
including the four name servers that are assigned to the delegation set}
\usage{
route53_get_reusable_delegation_set(Id)
}
\arguments{
\item{Id}{[required] The ID of the reusable delegation set that you want to get a list of
name servers for.}
}
\description{
Retrieves information about a specified reusable delegation set, including the four name servers that are assigned to the delegation set.

See \url{https://paws-r.github.io/docs/route53/get_reusable_delegation_set.html} for full documentation.
}
\keyword{internal}
