% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53_operations.R
\name{route53_get_traffic_policy}
\alias{route53_get_traffic_policy}
\title{Gets information about a specific traffic policy version}
\usage{
route53_get_traffic_policy(Id, Version)
}
\arguments{
\item{Id}{[required] The ID of the traffic policy that you want to get information about.}

\item{Version}{[required] The version number of the traffic policy that you want to get
information about.}
}
\description{
Gets information about a specific traffic policy version.

See \url{https://paws-r.github.io/docs/route53/get_traffic_policy.html} for full documentation.
}
\keyword{internal}
