% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53_operations.R
\name{route53_get_traffic_policy_instance}
\alias{route53_get_traffic_policy_instance}
\title{Gets information about a specified traffic policy instance}
\usage{
route53_get_traffic_policy_instance(Id)
}
\arguments{
\item{Id}{[required] The ID of the traffic policy instance that you want to get information
about.}
}
\description{
Gets information about a specified traffic policy instance.

See \url{https://paws-r.github.io/docs/route53/get_traffic_policy_instance.html} for full documentation.
}
\keyword{internal}
