% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53_operations.R
\name{route53_list_cidr_blocks}
\alias{route53_list_cidr_blocks}
\title{Returns a paginated list of location objects and their CIDR blocks}
\usage{
route53_list_cidr_blocks(
  CollectionId,
  LocationName = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{CollectionId}{[required] The UUID of the CIDR collection.}

\item{LocationName}{The name of the CIDR collection location.}

\item{NextToken}{An opaque pagination token to indicate where the service is to begin
enumerating results.}

\item{MaxResults}{Maximum number of results you want returned.}
}
\description{
Returns a paginated list of location objects and their CIDR blocks.

See \url{https://www.paws-r-sdk.com/docs/route53_list_cidr_blocks/} for full documentation.
}
\keyword{internal}
