% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53_operations.R
\name{route53_list_cidr_collections}
\alias{route53_list_cidr_collections}
\title{Returns a paginated list of CIDR collections in the Amazon Web Services
account (metadata only)}
\usage{
route53_list_cidr_collections(NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{NextToken}{An opaque pagination token to indicate where the service is to begin
enumerating results.

If no value is provided, the listing of results starts from the
beginning.}

\item{MaxResults}{The maximum number of CIDR collections to return in the response.}
}
\description{
Returns a paginated list of CIDR collections in the Amazon Web Services account (metadata only).

See \url{https://www.paws-r-sdk.com/docs/route53_list_cidr_collections/} for full documentation.
}
\keyword{internal}
