% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53_operations.R
\name{route53_list_traffic_policy_instances_by_hosted_zone}
\alias{route53_list_traffic_policy_instances_by_hosted_zone}
\title{Gets information about the traffic policy instances that you created in
a specified hosted zone}
\usage{
route53_list_traffic_policy_instances_by_hosted_zone(HostedZoneId,
  TrafficPolicyInstanceNameMarker, TrafficPolicyInstanceTypeMarker,
  MaxItems)
}
\arguments{
\item{HostedZoneId}{[required] The ID of the hosted zone that you want to list traffic policy instances
for.}

\item{TrafficPolicyInstanceNameMarker}{If the value of \code{IsTruncated} in the previous response is true, you have
more traffic policy instances. To get more traffic policy instances,
submit another
\code{\link[=route53_list_traffic_policy_instances]{list_traffic_policy_instances}}
request. For the value of \code{trafficpolicyinstancename}, specify the value
of \code{TrafficPolicyInstanceNameMarker} from the previous response, which
is the name of the first traffic policy instance in the next group of
traffic policy instances.

If the value of \code{IsTruncated} in the previous response was \code{false},
there are no more traffic policy instances to get.}

\item{TrafficPolicyInstanceTypeMarker}{If the value of \code{IsTruncated} in the previous response is true, you have
more traffic policy instances. To get more traffic policy instances,
submit another
\code{\link[=route53_list_traffic_policy_instances]{list_traffic_policy_instances}}
request. For the value of \code{trafficpolicyinstancetype}, specify the value
of \code{TrafficPolicyInstanceTypeMarker} from the previous response, which
is the type of the first traffic policy instance in the next group of
traffic policy instances.

If the value of \code{IsTruncated} in the previous response was \code{false},
there are no more traffic policy instances to get.}

\item{MaxItems}{The maximum number of traffic policy instances to be included in the
response body for this request. If you have more than \code{MaxItems} traffic
policy instances, the value of the \code{IsTruncated} element in the response
is \code{true}, and the values of \code{HostedZoneIdMarker},
\code{TrafficPolicyInstanceNameMarker}, and \code{TrafficPolicyInstanceTypeMarker}
represent the first traffic policy instance that Amazon Route 53 will
return if you submit another request.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  TrafficPolicyInstances = list(
    list(
      Id = "string",
      HostedZoneId = "string",
      Name = "string",
      TTL = 123,
      State = "string",
      Message = "string",
      TrafficPolicyId = "string",
      TrafficPolicyVersion = 123,
      TrafficPolicyType = "SOA"|"A"|"TXT"|"NS"|"CNAME"|"MX"|"NAPTR"|"PTR"|"SRV"|"SPF"|"AAAA"|"CAA"|"DS"
    )
  ),
  TrafficPolicyInstanceNameMarker = "string",
  TrafficPolicyInstanceTypeMarker = "SOA"|"A"|"TXT"|"NS"|"CNAME"|"MX"|"NAPTR"|"PTR"|"SRV"|"SPF"|"AAAA"|"CAA"|"DS",
  IsTruncated = TRUE|FALSE,
  MaxItems = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Gets information about the traffic policy instances that you created in
a specified hosted zone.

After you submit a
\code{\link[=route53_create_traffic_policy_instance]{create_traffic_policy_instance}}
or an
\code{\link[=route53_update_traffic_policy_instance]{update_traffic_policy_instance}}
request, there's a brief delay while Amazon Route 53 creates the
resource record sets that are specified in the traffic policy
definition. For more information, see the \code{State} response element.

Route 53 returns a maximum of 100 items in each response. If you have a
lot of traffic policy instances, you can use the \code{MaxItems} parameter to
list them in groups of up to 100.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_traffic_policy_instances_by_hosted_zone(
  HostedZoneId = "string",
  TrafficPolicyInstanceNameMarker = "string",
  TrafficPolicyInstanceTypeMarker = "SOA"|"A"|"TXT"|"NS"|"CNAME"|"MX"|"NAPTR"|"PTR"|"SRV"|"SPF"|"AAAA"|"CAA"|"DS",
  MaxItems = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
