% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53_operations.R
\name{route53_update_hosted_zone_features}
\alias{route53_update_hosted_zone_features}
\title{Updates the features configuration for a hosted zone}
\usage{
route53_update_hosted_zone_features(
  HostedZoneId,
  EnableAcceleratedRecovery = NULL
)
}
\arguments{
\item{HostedZoneId}{[required] The ID of the hosted zone for which you want to update features. This is
the unique identifier for your hosted zone.}

\item{EnableAcceleratedRecovery}{Specifies whether to enable accelerated recovery for the hosted zone.
Set to \code{true} to enable accelerated recovery, or \code{false} to disable it.}
}
\description{
Updates the features configuration for a hosted zone. This operation allows you to enable or disable specific features for your hosted zone, such as accelerated recovery.

See \url{https://www.paws-r-sdk.com/docs/route53_update_hosted_zone_features/} for full documentation.
}
\keyword{internal}
