% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53_operations.R
\name{route53_update_traffic_policy_comment}
\alias{route53_update_traffic_policy_comment}
\title{Updates the comment for a specified traffic policy version}
\usage{
route53_update_traffic_policy_comment(Id, Version, Comment)
}
\arguments{
\item{Id}{[required] The value of \code{Id} for the traffic policy that you want to update the
comment for.}

\item{Version}{[required] The value of \code{Version} for the traffic policy that you want to update
the comment for.}

\item{Comment}{[required] The new comment for the specified traffic policy and version.}
}
\value{
A list with the following syntax:\preformatted{list(
  TrafficPolicy = list(
    Id = "string",
    Version = 123,
    Name = "string",
    Type = "SOA"|"A"|"TXT"|"NS"|"CNAME"|"MX"|"NAPTR"|"PTR"|"SRV"|"SPF"|"AAAA"|"CAA"|"DS",
    Document = "string",
    Comment = "string"
  )
)
}
}
\description{
Updates the comment for a specified traffic policy version.
}
\section{Request syntax}{
\preformatted{svc$update_traffic_policy_comment(
  Id = "string",
  Version = 123,
  Comment = "string"
)
}
}

\keyword{internal}
