% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53domains_operations.R
\name{route53domains_accept_domain_transfer_from_another_aws_account}
\alias{route53domains_accept_domain_transfer_from_another_aws_account}
\title{Accepts the transfer of a domain from another Amazon Web Services
account to the currentAmazon Web Services account}
\usage{
route53domains_accept_domain_transfer_from_another_aws_account(
  DomainName,
  Password
)
}
\arguments{
\item{DomainName}{[required] The name of the domain that was specified when another Amazon Web
Services account submitted a
\code{\link[=route53domains_transfer_domain_to_another_aws_account]{transfer_domain_to_another_aws_account}}
request.}

\item{Password}{[required] The password that was returned by the
\code{\link[=route53domains_transfer_domain_to_another_aws_account]{transfer_domain_to_another_aws_account}}
request.}
}
\description{
Accepts the transfer of a domain from another Amazon Web Services account to the currentAmazon Web Services account. You initiate a transfer between Amazon Web Services accounts using \code{\link[=route53domains_transfer_domain_to_another_aws_account]{transfer_domain_to_another_aws_account}}.

See \url{https://www.paws-r-sdk.com/docs/route53domains_accept_domain_transfer_from_another_aws_account/} for full documentation.
}
\keyword{internal}
