% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53domains_operations.R
\name{route53domains_accept_domain_transfer_from_another_aws_account}
\alias{route53domains_accept_domain_transfer_from_another_aws_account}
\title{Accepts the transfer of a domain from another AWS account to the current
AWS account}
\usage{
route53domains_accept_domain_transfer_from_another_aws_account(
  DomainName, Password)
}
\arguments{
\item{DomainName}{[required] The name of the domain that was specified when another AWS account
submitted a
\code{\link[=route53domains_transfer_domain_to_another_aws_account]{transfer_domain_to_another_aws_account}}
request.}

\item{Password}{[required] The password that was returned by the
\code{\link[=route53domains_transfer_domain_to_another_aws_account]{transfer_domain_to_another_aws_account}}
request.}
}
\description{
Accepts the transfer of a domain from another AWS account to the current
AWS account. You initiate a transfer between AWS accounts using
\code{\link[=route53domains_transfer_domain_to_another_aws_account]{transfer_domain_to_another_aws_account}}.

Use either \code{\link[=route53domains_list_operations]{list_operations}} or
\code{\link[=route53domains_get_operation_detail]{get_operation_detail}} to
determine whether the operation succeeded.
\code{\link[=route53domains_get_operation_detail]{get_operation_detail}} provides
additional information, for example,
\verb{Domain Transfer from Aws Account 111122223333 has been cancelled}.
}
\section{Request syntax}{
\preformatted{svc$accept_domain_transfer_from_another_aws_account(
  DomainName = "string",
  Password = "string"
)
}
}

\keyword{internal}
