% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53domains_operations.R
\name{route53domains_delete_domain}
\alias{route53domains_delete_domain}
\title{This operation deletes the specified domain}
\usage{
route53domains_delete_domain(DomainName)
}
\arguments{
\item{DomainName}{[required] Name of the domain to be deleted.}
}
\description{
This operation deletes the specified domain. This action is permanent. For more information, see \href{https://docs.aws.amazon.com/Route53/latest/DeveloperGuide/domain-delete.html}{Deleting a domain name registration}.

See \url{https://paws-r.github.io/docs/route53domains/delete_domain.html} for full documentation.
}
\keyword{internal}
