% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53domains_operations.R
\name{route53domains_delete_tags_for_domain}
\alias{route53domains_delete_tags_for_domain}
\title{This operation deletes the specified tags for a domain}
\usage{
route53domains_delete_tags_for_domain(DomainName, TagsToDelete)
}
\arguments{
\item{DomainName}{[required] The domain for which you want to delete one or more tags.}

\item{TagsToDelete}{[required] A list of tag keys to delete.}
}
\description{
This operation deletes the specified tags for a domain.

See \url{https://www.paws-r-sdk.com/docs/route53domains_delete_tags_for_domain/} for full documentation.
}
\keyword{internal}
