% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53domains_operations.R
\name{route53domains_enable_domain_auto_renew}
\alias{route53domains_enable_domain_auto_renew}
\title{This operation configures Amazon Route 53 to automatically renew the
specified domain before the domain registration expires}
\usage{
route53domains_enable_domain_auto_renew(DomainName)
}
\arguments{
\item{DomainName}{[required] The name of the domain that you want to enable automatic renewal for.}
}
\value{
An empty list.
}
\description{
This operation configures Amazon Route 53 to automatically renew the
specified domain before the domain registration expires. The cost of
renewing your domain registration is billed to your AWS account.

The period during which you can renew a domain name varies by TLD. For a
list of TLDs and their renewal policies, see \href{https://docs.aws.amazon.com/Route53/latest/DeveloperGuide/registrar-tld-list.html}{Domains That You Can Register with Amazon Route 53}
in the \emph{Amazon Route 53 Developer Guide}. Route 53 requires that you
renew before the end of the renewal period so we can complete processing
before the deadline.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$enable_domain_auto_renew(
  DomainName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
