% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53domains_operations.R
\name{route53domains_list_tags_for_domain}
\alias{route53domains_list_tags_for_domain}
\title{This operation returns all of the tags that are associated with the
specified domain}
\usage{
route53domains_list_tags_for_domain(DomainName)
}
\arguments{
\item{DomainName}{[required] The domain for which you want to get a list of tags.}
}
\value{
A list with the following syntax:\preformatted{list(
  TagList = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}
\description{
This operation returns all of the tags that are associated with the
specified domain.

All tag operations are eventually consistent; subsequent operations
might not immediately represent all issued operations.
}
\section{Request syntax}{
\preformatted{svc$list_tags_for_domain(
  DomainName = "string"
)
}
}

\keyword{internal}
