% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53domains_operations.R
\name{route53domains_push_domain}
\alias{route53domains_push_domain}
\title{Moves a domain from Amazon Web Services to another registrar}
\usage{
route53domains_push_domain(DomainName, Target)
}
\arguments{
\item{DomainName}{[required] Name of the domain.}

\item{Target}{[required] New IPS tag for the domain.}
}
\description{
Moves a domain from Amazon Web Services to another registrar.

See \url{https://www.paws-r-sdk.com/docs/route53domains_push_domain/} for full documentation.
}
\keyword{internal}
