% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53domains_operations.R
\name{route53domains_reject_domain_transfer_from_another_aws_account}
\alias{route53domains_reject_domain_transfer_from_another_aws_account}
\title{Rejects the transfer of a domain from another Amazon Web Services
account to the current Amazon Web Services account}
\usage{
route53domains_reject_domain_transfer_from_another_aws_account(DomainName)
}
\arguments{
\item{DomainName}{[required] The name of the domain that was specified when another Amazon Web
Services account submitted a
\code{\link[=route53domains_transfer_domain_to_another_aws_account]{transfer_domain_to_another_aws_account}}
request.}
}
\description{
Rejects the transfer of a domain from another Amazon Web Services account to the current Amazon Web Services account. You initiate a transfer betweenAmazon Web Services accounts using \code{\link[=route53domains_transfer_domain_to_another_aws_account]{transfer_domain_to_another_aws_account}}.

See \url{https://paws-r.github.io/docs/route53domains/reject_domain_transfer_from_another_aws_account.html} for full documentation.
}
\keyword{internal}
