% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53domains_operations.R
\name{route53domains_resend_contact_reachability_email}
\alias{route53domains_resend_contact_reachability_email}
\title{For operations that require confirmation that the email address for the
registrant contact is valid, such as registering a new domain, this
operation resends the confirmation email to the current email address
for the registrant contact}
\usage{
route53domains_resend_contact_reachability_email(domainName)
}
\arguments{
\item{domainName}{The name of the domain for which you want Route 53 to resend a
confirmation email to the registrant contact.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  domainName = "string",
  emailAddress = "string",
  isAlreadyVerified = TRUE|FALSE
)
}\if{html}{\out{</div>}}
}
\description{
For operations that require confirmation that the email address for the
registrant contact is valid, such as registering a new domain, this
operation resends the confirmation email to the current email address
for the registrant contact.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$resend_contact_reachability_email(
  domainName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
