% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53domains_operations.R
\name{route53domains_transfer_domain_to_another_aws_account}
\alias{route53domains_transfer_domain_to_another_aws_account}
\title{Transfers a domain from the current Amazon Web Services account to
another Amazon Web Services account}
\usage{
route53domains_transfer_domain_to_another_aws_account(DomainName, AccountId)
}
\arguments{
\item{DomainName}{[required] The name of the domain that you want to transfer from the current Amazon
Web Services account to another account.}

\item{AccountId}{[required] The account ID of the Amazon Web Services account that you want to
transfer the domain to, for example, \code{111122223333}.}
}
\description{
Transfers a domain from the current Amazon Web Services account to another Amazon Web Services account. Note the following:

See \url{https://paws-r.github.io/docs/route53domains/transfer_domain_to_another_aws_account.html} for full documentation.
}
\keyword{internal}
