% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53domains_operations.R
\name{route53domains_update_domain_contact}
\alias{route53domains_update_domain_contact}
\title{This operation updates the contact information for a particular domain}
\usage{
route53domains_update_domain_contact(
  DomainName,
  AdminContact = NULL,
  RegistrantContact = NULL,
  TechContact = NULL,
  Consent = NULL
)
}
\arguments{
\item{DomainName}{[required] The name of the domain that you want to update contact information for.}

\item{AdminContact}{Provides detailed contact information.}

\item{RegistrantContact}{Provides detailed contact information.}

\item{TechContact}{Provides detailed contact information.}

\item{Consent}{Customer's consent for the owner change request. Required if the domain
is not free (consent price is more than $0.00).}
}
\description{
This operation updates the contact information for a particular domain. You must specify information for at least one contact: registrant, administrator, or technical.

See \url{https://www.paws-r-sdk.com/docs/route53domains_update_domain_contact/} for full documentation.
}
\keyword{internal}
