% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53domains_operations.R
\name{route53domains_update_domain_nameservers}
\alias{route53domains_update_domain_nameservers}
\title{This operation replaces the current set of name servers for the domain
with the specified set of name servers}
\usage{
route53domains_update_domain_nameservers(
  DomainName,
  FIAuthKey = NULL,
  Nameservers
)
}
\arguments{
\item{DomainName}{[required] The name of the domain that you want to change name servers for.}

\item{FIAuthKey}{The authorization key for .fi domains}

\item{Nameservers}{[required] A list of new name servers for the domain.}
}
\description{
This operation replaces the current set of name servers for the domain with the specified set of name servers. If you use Amazon Route 53 as your DNS service, specify the four name servers in the delegation set for the hosted zone for the domain.

See \url{https://paws-r.github.io/docs/route53domains/update_domain_nameservers.html} for full documentation.
}
\keyword{internal}
