% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53domains_operations.R
\name{route53domains_update_tags_for_domain}
\alias{route53domains_update_tags_for_domain}
\title{This operation adds or updates tags for a specified domain}
\usage{
route53domains_update_tags_for_domain(DomainName, TagsToUpdate)
}
\arguments{
\item{DomainName}{[required] The domain for which you want to add or update tags.}

\item{TagsToUpdate}{A list of the tag keys and values that you want to add or update. If you
specify a key that already exists, the corresponding value will be
replaced.}
}
\value{
An empty list.
}
\description{
This operation adds or updates tags for a specified domain.

All tag operations are eventually consistent; subsequent operations
might not immediately represent all issued operations.
}
\section{Request syntax}{
\preformatted{svc$update_tags_for_domain(
  DomainName = "string",
  TagsToUpdate = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
