% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53profiles_operations.R
\name{route53profiles_associate_profile}
\alias{route53profiles_associate_profile}
\title{Associates a Route 53 Profiles profile with a VPC}
\usage{
route53profiles_associate_profile(Name, ProfileId, ResourceId, Tags = NULL)
}
\arguments{
\item{Name}{[required] A name for the association.}

\item{ProfileId}{[required] ID of the Profile.}

\item{ResourceId}{[required] The ID of the VPC.}

\item{Tags}{A list of the tag keys and values that you want to identify the Profile
association.}
}
\description{
Associates a Route 53 Profiles profile with a VPC. A VPC can have only one Profile associated with it, but a Profile can be associated with 1000 of VPCs (and you can request a higher quota). For more information, see \url{https://docs.aws.amazon.com/Route53/latest/DeveloperGuide/DNSLimitations.html#limits-api-entities}.

See \url{https://www.paws-r-sdk.com/docs/route53profiles_associate_profile/} for full documentation.
}
\keyword{internal}
