% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53profiles_operations.R
\name{route53profiles_associate_resource_to_profile}
\alias{route53profiles_associate_resource_to_profile}
\title{Associates a DNS reource configuration to a Route 53 Profile}
\usage{
route53profiles_associate_resource_to_profile(
  Name,
  ProfileId,
  ResourceArn,
  ResourceProperties = NULL
)
}
\arguments{
\item{Name}{[required] Name for the resource association.}

\item{ProfileId}{[required] ID of the Profile.}

\item{ResourceArn}{[required] Amazon resource number, ARN, of the DNS resource.}

\item{ResourceProperties}{If you are adding a DNS Firewall rule group, include also a priority.
The priority indicates the processing order for the rule groups,
starting with the priority assinged the lowest value.

The allowed values for priority are between 100 and 9900.}
}
\description{
Associates a DNS reource configuration to a Route 53 Profile.

See \url{https://www.paws-r-sdk.com/docs/route53profiles_associate_resource_to_profile/} for full documentation.
}
\keyword{internal}
