% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53profiles_operations.R
\name{route53profiles_get_profile}
\alias{route53profiles_get_profile}
\title{Returns information about a specified Route 53 Profile, such as whether
whether the Profile is shared, and the current status of the Profile}
\usage{
route53profiles_get_profile(ProfileId)
}
\arguments{
\item{ProfileId}{[required] ID of the Profile.}
}
\description{
Returns information about a specified Route 53 Profile, such as whether whether the Profile is shared, and the current status of the Profile.

See \url{https://www.paws-r-sdk.com/docs/route53profiles_get_profile/} for full documentation.
}
\keyword{internal}
