% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53profiles_operations.R
\name{route53profiles_get_profile_association}
\alias{route53profiles_get_profile_association}
\title{Retrieves a Route 53 Profile association for a VPC}
\usage{
route53profiles_get_profile_association(ProfileAssociationId)
}
\arguments{
\item{ProfileAssociationId}{[required] The identifier of the association you want to get information about.}
}
\description{
Retrieves a Route 53 Profile association for a VPC. A VPC can have only one Profile association, but a Profile can be associated with up to 5000 VPCs.

See \url{https://www.paws-r-sdk.com/docs/route53profiles_get_profile_association/} for full documentation.
}
\keyword{internal}
