% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53profiles_operations.R
\name{route53profiles_get_profile_resource_association}
\alias{route53profiles_get_profile_resource_association}
\title{Returns information about a specified Route 53 Profile resource
association}
\usage{
route53profiles_get_profile_resource_association(ProfileResourceAssociationId)
}
\arguments{
\item{ProfileResourceAssociationId}{[required] The ID of the profile resource association that you want to get
information about.}
}
\description{
Returns information about a specified Route 53 Profile resource association.

See \url{https://www.paws-r-sdk.com/docs/route53profiles_get_profile_resource_association/} for full documentation.
}
\keyword{internal}
