% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53profiles_operations.R
\name{route53profiles_list_profile_associations}
\alias{route53profiles_list_profile_associations}
\title{Lists all the VPCs that the specified Route 53 Profile is associated
with}
\usage{
route53profiles_list_profile_associations(
  MaxResults = NULL,
  NextToken = NULL,
  ProfileId = NULL,
  ResourceId = NULL
)
}
\arguments{
\item{MaxResults}{The maximum number of objects that you want to return for this request.
If more objects are available, in the response, a \code{NextToken} value,
which you can use in a subsequent call to get the next batch of objects,
is provided.

If you don't specify a value for \code{MaxResults}, up to 100 objects are
returned.}

\item{NextToken}{For the first call to this list request, omit this value.

When you request a list of objects, at most the number of objects
specified by \code{MaxResults} is returned. If more objects are available for
retrieval, a \code{NextToken} value is returned in the response. To retrieve
the next batch of objects, use the token that was returned for the prior
request in your next request.}

\item{ProfileId}{ID of the Profile.}

\item{ResourceId}{ID of the VPC.}
}
\description{
Lists all the VPCs that the specified Route 53 Profile is associated with.

See \url{https://www.paws-r-sdk.com/docs/route53profiles_list_profile_associations/} for full documentation.
}
\keyword{internal}
