% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53profiles_operations.R
\name{route53profiles_list_tags_for_resource}
\alias{route53profiles_list_tags_for_resource}
\title{Lists the tags that you associated with the specified resource}
\usage{
route53profiles_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) for the resource that you want to list
the tags for.}
}
\description{
Lists the tags that you associated with the specified resource.

See \url{https://www.paws-r-sdk.com/docs/route53profiles_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
