% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53profiles_operations.R
\name{route53profiles_tag_resource}
\alias{route53profiles_tag_resource}
\title{Adds one or more tags to a specified resource}
\usage{
route53profiles_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) for the resource that you want to add
tags to.}

\item{Tags}{[required] The tags that you want to add to the specified resource.}
}
\description{
Adds one or more tags to a specified resource.

See \url{https://www.paws-r-sdk.com/docs/route53profiles_tag_resource/} for full documentation.
}
\keyword{internal}
