% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53profiles_operations.R
\name{route53profiles_untag_resource}
\alias{route53profiles_untag_resource}
\title{Removes one or more tags from a specified resource}
\usage{
route53profiles_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) for the resource that you want to remove
tags from.}

\item{TagKeys}{[required] The tags that you want to remove to the specified resource.}
}
\description{
Removes one or more tags from a specified resource.

See \url{https://www.paws-r-sdk.com/docs/route53profiles_untag_resource/} for full documentation.
}
\keyword{internal}
