% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53profiles_operations.R
\name{route53profiles_update_profile_resource_association}
\alias{route53profiles_update_profile_resource_association}
\title{Updates the specified Route 53 Profile resourse association}
\usage{
route53profiles_update_profile_resource_association(
  Name = NULL,
  ProfileResourceAssociationId,
  ResourceProperties = NULL
)
}
\arguments{
\item{Name}{Name of the resource association.}

\item{ProfileResourceAssociationId}{[required] ID of the resource association.}

\item{ResourceProperties}{If you are adding a DNS Firewall rule group, include also a priority.
The priority indicates the processing order for the rule groups,
starting with the priority assinged the lowest value.

The allowed values for priority are between 100 and 9900.}
}
\description{
Updates the specified Route 53 Profile resourse association.

See \url{https://www.paws-r-sdk.com/docs/route53profiles_update_profile_resource_association/} for full documentation.
}
\keyword{internal}
