% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53recoverycontrolconfig_operations.R
\name{route53recoverycontrolconfig_create_cluster}
\alias{route53recoverycontrolconfig_create_cluster}
\title{Create a new cluster}
\usage{
route53recoverycontrolconfig_create_cluster(
  ClientToken = NULL,
  ClusterName,
  Tags = NULL,
  NetworkType = NULL
)
}
\arguments{
\item{ClientToken}{A unique, case-sensitive string of up to 64 ASCII characters. To make an
idempotent API request with an action, specify a client token in the
request.}

\item{ClusterName}{[required] The name of the cluster.}

\item{Tags}{The tags associated with the cluster.}

\item{NetworkType}{The network type of the cluster. NetworkType can be one of the
following: IPV4, DUALSTACK.}
}
\description{
Create a new cluster. A cluster is a set of redundant Regional endpoints against which you can run API calls to update or get the state of one or more routing controls. Each cluster has a name, status, Amazon Resource Name (ARN), and an array of the five cluster endpoints (one for each supported Amazon Web Services Region) that you can use with API calls to the cluster data plane.

See \url{https://www.paws-r-sdk.com/docs/route53recoverycontrolconfig_create_cluster/} for full documentation.
}
\keyword{internal}
