% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53recoverycontrolconfig_operations.R
\name{route53recoverycontrolconfig_create_routing_control}
\alias{route53recoverycontrolconfig_create_routing_control}
\title{Creates a new routing control}
\usage{
route53recoverycontrolconfig_create_routing_control(
  ClientToken = NULL,
  ClusterArn,
  ControlPanelArn = NULL,
  RoutingControlName
)
}
\arguments{
\item{ClientToken}{A unique, case-sensitive string of up to 64 ASCII characters. To make an
idempotent API request with an action, specify a client token in the
request.}

\item{ClusterArn}{[required] The Amazon Resource Name (ARN) of the cluster that includes the routing
control.}

\item{ControlPanelArn}{The Amazon Resource Name (ARN) of the control panel that includes the
routing control.}

\item{RoutingControlName}{[required] The name of the routing control.}
}
\description{
Creates a new routing control.

See \url{https://www.paws-r-sdk.com/docs/route53recoverycontrolconfig_create_routing_control/} for full documentation.
}
\keyword{internal}
