% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53recoverycontrolconfig_operations.R
\name{route53recoverycontrolconfig_create_safety_rule}
\alias{route53recoverycontrolconfig_create_safety_rule}
\title{Creates a safety rule in a control panel}
\usage{
route53recoverycontrolconfig_create_safety_rule(
  AssertionRule = NULL,
  ClientToken = NULL,
  GatingRule = NULL,
  Tags = NULL
)
}
\arguments{
\item{AssertionRule}{The assertion rule requested.}

\item{ClientToken}{A unique, case-sensitive string of up to 64 ASCII characters. To make an
idempotent API request with an action, specify a client token in the
request.}

\item{GatingRule}{The gating rule requested.}

\item{Tags}{The tags associated with the safety rule.}
}
\description{
Creates a safety rule in a control panel. Safety rules let you add safeguards around changing routing control states, and for enabling and disabling routing controls, to help prevent unexpected outcomes.

See \url{https://paws-r.github.io/docs/route53recoverycontrolconfig/create_safety_rule.html} for full documentation.
}
\keyword{internal}
