% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53recoverycontrolconfig_operations.R
\name{route53recoverycontrolconfig_delete_control_panel}
\alias{route53recoverycontrolconfig_delete_control_panel}
\title{Deletes a control panel}
\usage{
route53recoverycontrolconfig_delete_control_panel(ControlPanelArn)
}
\arguments{
\item{ControlPanelArn}{[required] The Amazon Resource Name (ARN) of the control panel.}
}
\description{
Deletes a control panel.

See \url{https://paws-r.github.io/docs/route53recoverycontrolconfig/delete_control_panel.html} for full documentation.
}
\keyword{internal}
