% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53recoverycontrolconfig_operations.R
\name{route53recoverycontrolconfig_delete_routing_control}
\alias{route53recoverycontrolconfig_delete_routing_control}
\title{Deletes a routing control}
\usage{
route53recoverycontrolconfig_delete_routing_control(RoutingControlArn)
}
\arguments{
\item{RoutingControlArn}{[required] The Amazon Resource Name (ARN) of the routing control that you're
deleting.}
}
\description{
Deletes a routing control.

See \url{https://paws-r.github.io/docs/route53recoverycontrolconfig/delete_routing_control.html} for full documentation.
}
\keyword{internal}
