% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53recoverycontrolconfig_operations.R
\name{route53recoverycontrolconfig_delete_safety_rule}
\alias{route53recoverycontrolconfig_delete_safety_rule}
\title{Deletes a safety rule}
\usage{
route53recoverycontrolconfig_delete_safety_rule(SafetyRuleArn)
}
\arguments{
\item{SafetyRuleArn}{[required] The ARN of the safety rule.}
}
\description{
Deletes a safety rule.

See \url{https://paws-r.github.io/docs/route53recoverycontrolconfig/delete_safety_rule.html} for full documentation.
}
\keyword{internal}
