% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53recoverycontrolconfig_operations.R
\name{route53recoverycontrolconfig_describe_cluster}
\alias{route53recoverycontrolconfig_describe_cluster}
\title{Display the details about a cluster}
\usage{
route53recoverycontrolconfig_describe_cluster(ClusterArn)
}
\arguments{
\item{ClusterArn}{[required] The Amazon Resource Name (ARN) of the cluster.}
}
\description{
Display the details about a cluster. The response includes the cluster name, endpoints, status, and Amazon Resource Name (ARN).

See \url{https://paws-r.github.io/docs/route53recoverycontrolconfig/describe_cluster.html} for full documentation.
}
\keyword{internal}
