% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53recoverycontrolconfig_operations.R
\name{route53recoverycontrolconfig_describe_control_panel}
\alias{route53recoverycontrolconfig_describe_control_panel}
\title{Displays details about a control panel}
\usage{
route53recoverycontrolconfig_describe_control_panel(ControlPanelArn)
}
\arguments{
\item{ControlPanelArn}{[required] The Amazon Resource Name (ARN) of the control panel.}
}
\description{
Displays details about a control panel.

See \url{https://www.paws-r-sdk.com/docs/route53recoverycontrolconfig_describe_control_panel/} for full documentation.
}
\keyword{internal}
