% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53recoverycontrolconfig_operations.R
\name{route53recoverycontrolconfig_describe_safety_rule}
\alias{route53recoverycontrolconfig_describe_safety_rule}
\title{Returns information about a safety rule}
\usage{
route53recoverycontrolconfig_describe_safety_rule(SafetyRuleArn)
}
\arguments{
\item{SafetyRuleArn}{[required] The ARN of the safety rule.}
}
\description{
Returns information about a safety rule.

See \url{https://www.paws-r-sdk.com/docs/route53recoverycontrolconfig_describe_safety_rule/} for full documentation.
}
\keyword{internal}
