% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53recoverycontrolconfig_operations.R
\name{route53recoverycontrolconfig_get_resource_policy}
\alias{route53recoverycontrolconfig_get_resource_policy}
\title{Get information about the resource policy for a cluster}
\usage{
route53recoverycontrolconfig_get_resource_policy(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource.}
}
\description{
Get information about the resource policy for a cluster.

See \url{https://www.paws-r-sdk.com/docs/route53recoverycontrolconfig_get_resource_policy/} for full documentation.
}
\keyword{internal}
