% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53recoverycontrolconfig_operations.R
\name{route53recoverycontrolconfig_list_clusters}
\alias{route53recoverycontrolconfig_list_clusters}
\title{Returns an array of all the clusters in an account}
\usage{
route53recoverycontrolconfig_list_clusters(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{The number of objects that you want to return with this call.}

\item{NextToken}{The token that identifies which batch of results you want to see.}
}
\description{
Returns an array of all the clusters in an account.

See \url{https://www.paws-r-sdk.com/docs/route53recoverycontrolconfig_list_clusters/} for full documentation.
}
\keyword{internal}
