% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53recoverycontrolconfig_operations.R
\name{route53recoverycontrolconfig_list_control_panels}
\alias{route53recoverycontrolconfig_list_control_panels}
\title{Returns an array of control panels in an account or in a cluster}
\usage{
route53recoverycontrolconfig_list_control_panels(
  ClusterArn = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{ClusterArn}{The Amazon Resource Name (ARN) of a cluster.}

\item{MaxResults}{The number of objects that you want to return with this call.}

\item{NextToken}{The token that identifies which batch of results you want to see.}
}
\description{
Returns an array of control panels in an account or in a cluster.

See \url{https://paws-r.github.io/docs/route53recoverycontrolconfig/list_control_panels.html} for full documentation.
}
\keyword{internal}
