% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53recoverycontrolconfig_operations.R
\name{route53recoverycontrolconfig_list_routing_controls}
\alias{route53recoverycontrolconfig_list_routing_controls}
\title{Returns an array of routing controls for a control panel}
\usage{
route53recoverycontrolconfig_list_routing_controls(
  ControlPanelArn,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{ControlPanelArn}{[required] The Amazon Resource Name (ARN) of the control panel.}

\item{MaxResults}{The number of objects that you want to return with this call.}

\item{NextToken}{The token that identifies which batch of results you want to see.}
}
\description{
Returns an array of routing controls for a control panel. A routing control is an Amazon Route 53 Application Recovery Controller construct that has one of two states: ON and OFF. You can map the routing control state to the state of an Amazon Route 53 health check, which can be used to control routing.

See \url{https://paws-r.github.io/docs/route53recoverycontrolconfig/list_routing_controls.html} for full documentation.
}
\keyword{internal}
