% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53recoverycontrolconfig_operations.R
\name{route53recoverycontrolconfig_list_safety_rules}
\alias{route53recoverycontrolconfig_list_safety_rules}
\title{List the safety rules (the assertion rules and gating rules) that you've
defined for the routing controls in a control panel}
\usage{
route53recoverycontrolconfig_list_safety_rules(
  ControlPanelArn,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{ControlPanelArn}{[required] The Amazon Resource Name (ARN) of the control panel.}

\item{MaxResults}{The number of objects that you want to return with this call.}

\item{NextToken}{The token that identifies which batch of results you want to see.}
}
\description{
List the safety rules (the assertion rules and gating rules) that you've defined for the routing controls in a control panel.

See \url{https://www.paws-r-sdk.com/docs/route53recoverycontrolconfig_list_safety_rules/} for full documentation.
}
\keyword{internal}
