% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53recoverycontrolconfig_operations.R
\name{route53recoverycontrolconfig_tag_resource}
\alias{route53recoverycontrolconfig_tag_resource}
\title{Adds a tag to a resource}
\usage{
route53recoverycontrolconfig_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) for the resource that's tagged.}

\item{Tags}{[required] The tags associated with the resource.}
}
\description{
Adds a tag to a resource.

See \url{https://www.paws-r-sdk.com/docs/route53recoverycontrolconfig_tag_resource/} for full documentation.
}
\keyword{internal}
