% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53recoverycontrolconfig_operations.R
\name{route53recoverycontrolconfig_untag_resource}
\alias{route53recoverycontrolconfig_untag_resource}
\title{Removes a tag from a resource}
\usage{
route53recoverycontrolconfig_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) for the resource that's tagged.}

\item{TagKeys}{[required] Keys for the tags to be removed.}
}
\description{
Removes a tag from a resource.

See \url{https://www.paws-r-sdk.com/docs/route53recoverycontrolconfig_untag_resource/} for full documentation.
}
\keyword{internal}
