% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53recoverycontrolconfig_operations.R
\name{route53recoverycontrolconfig_update_cluster}
\alias{route53recoverycontrolconfig_update_cluster}
\title{Updates an existing cluster}
\usage{
route53recoverycontrolconfig_update_cluster(ClusterArn, NetworkType)
}
\arguments{
\item{ClusterArn}{[required] The Amazon Resource Name (ARN) of the cluster.}

\item{NetworkType}{[required] The network type of the cluster. NetworkType can be one of the
following: IPV4, DUALSTACK.}
}
\description{
Updates an existing cluster. You can only update the network type of a cluster.

See \url{https://www.paws-r-sdk.com/docs/route53recoverycontrolconfig_update_cluster/} for full documentation.
}
\keyword{internal}
