% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53recoverycontrolconfig_operations.R
\name{route53recoverycontrolconfig_update_control_panel}
\alias{route53recoverycontrolconfig_update_control_panel}
\title{Updates a control panel}
\usage{
route53recoverycontrolconfig_update_control_panel(
  ControlPanelArn,
  ControlPanelName
)
}
\arguments{
\item{ControlPanelArn}{[required] The Amazon Resource Name (ARN) of the control panel.}

\item{ControlPanelName}{[required] The name of the control panel.}
}
\description{
Updates a control panel. The only update you can make to a control panel is to change the name of the control panel.

See \url{https://www.paws-r-sdk.com/docs/route53recoverycontrolconfig_update_control_panel/} for full documentation.
}
\keyword{internal}
