% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53recoverycontrolconfig_operations.R
\name{route53recoverycontrolconfig_update_safety_rule}
\alias{route53recoverycontrolconfig_update_safety_rule}
\title{Update a safety rule (an assertion rule or gating rule)}
\usage{
route53recoverycontrolconfig_update_safety_rule(
  AssertionRuleUpdate = NULL,
  GatingRuleUpdate = NULL
)
}
\arguments{
\item{AssertionRuleUpdate}{The assertion rule to update.}

\item{GatingRuleUpdate}{The gating rule to update.}
}
\description{
Update a safety rule (an assertion rule or gating rule). You can only update the name and the waiting period for a safety rule. To make other updates, delete the safety rule and create a new one.

See \url{https://paws-r.github.io/docs/route53recoverycontrolconfig/update_safety_rule.html} for full documentation.
}
\keyword{internal}
