% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53recoveryreadiness_operations.R
\name{route53recoveryreadiness_create_cell}
\alias{route53recoveryreadiness_create_cell}
\title{Creates a cell in an account}
\usage{
route53recoveryreadiness_create_cell(CellName, Cells = NULL, Tags = NULL)
}
\arguments{
\item{CellName}{[required] The name of the cell to create.}

\item{Cells}{A list of cell Amazon Resource Names (ARNs) contained within this cell,
for use in nested cells. For example, Availability Zones within specific
Amazon Web Services Regions.}

\item{Tags}{}
}
\description{
Creates a cell in an account.

See \url{https://paws-r.github.io/docs/route53recoveryreadiness/create_cell.html} for full documentation.
}
\keyword{internal}
