% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53recoveryreadiness_operations.R
\name{route53recoveryreadiness_create_cross_account_authorization}
\alias{route53recoveryreadiness_create_cross_account_authorization}
\title{Creates a cross-account readiness authorization}
\usage{
route53recoveryreadiness_create_cross_account_authorization(
  CrossAccountAuthorization
)
}
\arguments{
\item{CrossAccountAuthorization}{[required] The cross-account authorization.}
}
\description{
Creates a cross-account readiness authorization. This lets you authorize another account to work with Route 53 Application Recovery Controller, for example, to check the readiness status of resources in a separate account.

See \url{https://paws-r.github.io/docs/route53recoveryreadiness/create_cross_account_authorization.html} for full documentation.
}
\keyword{internal}
