% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53recoveryreadiness_operations.R
\name{route53recoveryreadiness_create_readiness_check}
\alias{route53recoveryreadiness_create_readiness_check}
\title{Creates a readiness check in an account}
\usage{
route53recoveryreadiness_create_readiness_check(
  ReadinessCheckName,
  ResourceSetName,
  Tags = NULL
)
}
\arguments{
\item{ReadinessCheckName}{[required] The name of the readiness check to create.}

\item{ResourceSetName}{[required] The name of the resource set to check.}

\item{Tags}{}
}
\description{
Creates a readiness check in an account. A readiness check monitors a resource set in your application, such as a set of Amazon Aurora instances, that Application Recovery Controller is auditing recovery readiness for. The audits run once every minute on every resource that's associated with a readiness check.

See \url{https://paws-r.github.io/docs/route53recoveryreadiness/create_readiness_check.html} for full documentation.
}
\keyword{internal}
