% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53recoveryreadiness_operations.R
\name{route53recoveryreadiness_create_recovery_group}
\alias{route53recoveryreadiness_create_recovery_group}
\title{Creates a recovery group in an account}
\usage{
route53recoveryreadiness_create_recovery_group(
  Cells = NULL,
  RecoveryGroupName,
  Tags = NULL
)
}
\arguments{
\item{Cells}{A list of the cell Amazon Resource Names (ARNs) in the recovery group.}

\item{RecoveryGroupName}{[required] The name of the recovery group to create.}

\item{Tags}{}
}
\description{
Creates a recovery group in an account. A recovery group corresponds to an application and includes a list of the cells that make up the application.

See \url{https://paws-r.github.io/docs/route53recoveryreadiness/create_recovery_group.html} for full documentation.
}
\keyword{internal}
