% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53recoveryreadiness_operations.R
\name{route53recoveryreadiness_delete_cell}
\alias{route53recoveryreadiness_delete_cell}
\title{Delete a cell}
\usage{
route53recoveryreadiness_delete_cell(CellName)
}
\arguments{
\item{CellName}{[required] The name of the cell.}
}
\description{
Delete a cell. When successful, the response code is 204, with no response body.

See \url{https://www.paws-r-sdk.com/docs/route53recoveryreadiness_delete_cell/} for full documentation.
}
\keyword{internal}
