% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53recoveryreadiness_operations.R
\name{route53recoveryreadiness_delete_readiness_check}
\alias{route53recoveryreadiness_delete_readiness_check}
\title{Deletes a readiness check}
\usage{
route53recoveryreadiness_delete_readiness_check(ReadinessCheckName)
}
\arguments{
\item{ReadinessCheckName}{[required] Name of a readiness check.}
}
\description{
Deletes a readiness check.

See \url{https://www.paws-r-sdk.com/docs/route53recoveryreadiness_delete_readiness_check/} for full documentation.
}
\keyword{internal}
