% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53recoveryreadiness_operations.R
\name{route53recoveryreadiness_delete_recovery_group}
\alias{route53recoveryreadiness_delete_recovery_group}
\title{Deletes a recovery group}
\usage{
route53recoveryreadiness_delete_recovery_group(RecoveryGroupName)
}
\arguments{
\item{RecoveryGroupName}{[required] The name of a recovery group.}
}
\description{
Deletes a recovery group.

See \url{https://www.paws-r-sdk.com/docs/route53recoveryreadiness_delete_recovery_group/} for full documentation.
}
\keyword{internal}
