% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53recoveryreadiness_operations.R
\name{route53recoveryreadiness_get_architecture_recommendations}
\alias{route53recoveryreadiness_get_architecture_recommendations}
\title{Gets recommendations about architecture designs for improving resiliency
for an application, based on a recovery group}
\usage{
route53recoveryreadiness_get_architecture_recommendations(
  MaxResults = NULL,
  NextToken = NULL,
  RecoveryGroupName
)
}
\arguments{
\item{MaxResults}{The number of objects that you want to return with this call.}

\item{NextToken}{The token that identifies which batch of results you want to see.}

\item{RecoveryGroupName}{[required] The name of a recovery group.}
}
\description{
Gets recommendations about architecture designs for improving resiliency for an application, based on a recovery group.

See \url{https://paws-r.github.io/docs/route53recoveryreadiness/get_architecture_recommendations.html} for full documentation.
}
\keyword{internal}
