% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53recoveryreadiness_operations.R
\name{route53recoveryreadiness_get_cell}
\alias{route53recoveryreadiness_get_cell}
\title{Gets information about a cell including cell name, cell Amazon Resource
Name (ARN), ARNs of nested cells for this cell, and a list of those cell
ARNs with their associated recovery group ARNs}
\usage{
route53recoveryreadiness_get_cell(CellName)
}
\arguments{
\item{CellName}{[required] The name of the cell.}
}
\description{
Gets information about a cell including cell name, cell Amazon Resource Name (ARN), ARNs of nested cells for this cell, and a list of those cell ARNs with their associated recovery group ARNs.

See \url{https://paws-r.github.io/docs/route53recoveryreadiness/get_cell.html} for full documentation.
}
\keyword{internal}
