% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53recoveryreadiness_operations.R
\name{route53recoveryreadiness_get_cell_readiness_summary}
\alias{route53recoveryreadiness_get_cell_readiness_summary}
\title{Gets readiness for a cell}
\usage{
route53recoveryreadiness_get_cell_readiness_summary(
  CellName,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{CellName}{[required] The name of the cell.}

\item{MaxResults}{The number of objects that you want to return with this call.}

\item{NextToken}{The token that identifies which batch of results you want to see.}
}
\description{
Gets readiness for a cell. Aggregates the readiness of all the resources that are associated with the cell into a single value.

See \url{https://paws-r.github.io/docs/route53recoveryreadiness/get_cell_readiness_summary.html} for full documentation.
}
\keyword{internal}
