% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53recoveryreadiness_operations.R
\name{route53recoveryreadiness_get_readiness_check_status}
\alias{route53recoveryreadiness_get_readiness_check_status}
\title{Gets the readiness status for an individual readiness check}
\usage{
route53recoveryreadiness_get_readiness_check_status(
  MaxResults = NULL,
  NextToken = NULL,
  ReadinessCheckName
)
}
\arguments{
\item{MaxResults}{The number of objects that you want to return with this call.}

\item{NextToken}{The token that identifies which batch of results you want to see.}

\item{ReadinessCheckName}{[required] Name of a readiness check.}
}
\description{
Gets the readiness status for an individual readiness check. To see the overall readiness status for a recovery group, that considers the readiness status for all the readiness checks in a recovery group, use GetRecoveryGroupReadinessSummary.

See \url{https://paws-r.github.io/docs/route53recoveryreadiness/get_readiness_check_status.html} for full documentation.
}
\keyword{internal}
