% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53recoveryreadiness_operations.R
\name{route53recoveryreadiness_get_recovery_group}
\alias{route53recoveryreadiness_get_recovery_group}
\title{Gets details about a recovery group, including a list of the cells that
are included in it}
\usage{
route53recoveryreadiness_get_recovery_group(RecoveryGroupName)
}
\arguments{
\item{RecoveryGroupName}{[required] The name of a recovery group.}
}
\description{
Gets details about a recovery group, including a list of the cells that are included in it.

See \url{https://www.paws-r-sdk.com/docs/route53recoveryreadiness_get_recovery_group/} for full documentation.
}
\keyword{internal}
