% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53recoveryreadiness_operations.R
\name{route53recoveryreadiness_get_resource_set}
\alias{route53recoveryreadiness_get_resource_set}
\title{Displays the details about a resource set, including a list of the
resources in the set}
\usage{
route53recoveryreadiness_get_resource_set(ResourceSetName)
}
\arguments{
\item{ResourceSetName}{[required] Name of a resource set.}
}
\description{
Displays the details about a resource set, including a list of the resources in the set.

See \url{https://paws-r.github.io/docs/route53recoveryreadiness/get_resource_set.html} for full documentation.
}
\keyword{internal}
