% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53recoveryreadiness_operations.R
\name{route53recoveryreadiness_list_cells}
\alias{route53recoveryreadiness_list_cells}
\title{Lists the cells for an account}
\usage{
route53recoveryreadiness_list_cells(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{The number of objects that you want to return with this call.}

\item{NextToken}{The token that identifies which batch of results you want to see.}
}
\description{
Lists the cells for an account.

See \url{https://www.paws-r-sdk.com/docs/route53recoveryreadiness_list_cells/} for full documentation.
}
\keyword{internal}
