% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53recoveryreadiness_operations.R
\name{route53recoveryreadiness_list_cross_account_authorizations}
\alias{route53recoveryreadiness_list_cross_account_authorizations}
\title{Lists the cross-account readiness authorizations that are in place for
an account}
\usage{
route53recoveryreadiness_list_cross_account_authorizations(
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{MaxResults}{The number of objects that you want to return with this call.}

\item{NextToken}{The token that identifies which batch of results you want to see.}
}
\description{
Lists the cross-account readiness authorizations that are in place for an account.

See \url{https://www.paws-r-sdk.com/docs/route53recoveryreadiness_list_cross_account_authorizations/} for full documentation.
}
\keyword{internal}
