% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53recoveryreadiness_operations.R
\name{route53recoveryreadiness_list_readiness_checks}
\alias{route53recoveryreadiness_list_readiness_checks}
\title{Lists the readiness checks for an account}
\usage{
route53recoveryreadiness_list_readiness_checks(
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{MaxResults}{The number of objects that you want to return with this call.}

\item{NextToken}{The token that identifies which batch of results you want to see.}
}
\description{
Lists the readiness checks for an account.

See \url{https://www.paws-r-sdk.com/docs/route53recoveryreadiness_list_readiness_checks/} for full documentation.
}
\keyword{internal}
