% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53recoveryreadiness_operations.R
\name{route53recoveryreadiness_list_recovery_groups}
\alias{route53recoveryreadiness_list_recovery_groups}
\title{Lists the recovery groups in an account}
\usage{
route53recoveryreadiness_list_recovery_groups(
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{MaxResults}{The number of objects that you want to return with this call.}

\item{NextToken}{The token that identifies which batch of results you want to see.}
}
\description{
Lists the recovery groups in an account.

See \url{https://paws-r.github.io/docs/route53recoveryreadiness/list_recovery_groups.html} for full documentation.
}
\keyword{internal}
