% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53recoveryreadiness_operations.R
\name{route53recoveryreadiness_list_resource_sets}
\alias{route53recoveryreadiness_list_resource_sets}
\title{Lists the resource sets in an account}
\usage{
route53recoveryreadiness_list_resource_sets(
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{MaxResults}{The number of objects that you want to return with this call.}

\item{NextToken}{The token that identifies which batch of results you want to see.}
}
\description{
Lists the resource sets in an account.

See \url{https://paws-r.github.io/docs/route53recoveryreadiness/list_resource_sets.html} for full documentation.
}
\keyword{internal}
