% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53recoveryreadiness_operations.R
\name{route53recoveryreadiness_list_rules}
\alias{route53recoveryreadiness_list_rules}
\title{Lists all readiness rules, or lists the readiness rules for a specific
resource type}
\usage{
route53recoveryreadiness_list_rules(
  MaxResults = NULL,
  NextToken = NULL,
  ResourceType = NULL
)
}
\arguments{
\item{MaxResults}{The number of objects that you want to return with this call.}

\item{NextToken}{The token that identifies which batch of results you want to see.}

\item{ResourceType}{The resource type that a readiness rule applies to.}
}
\description{
Lists all readiness rules, or lists the readiness rules for a specific resource type.

See \url{https://paws-r.github.io/docs/route53recoveryreadiness/list_rules.html} for full documentation.
}
\keyword{internal}
