% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53recoveryreadiness_operations.R
\name{route53recoveryreadiness_list_tags_for_resources}
\alias{route53recoveryreadiness_list_tags_for_resources}
\title{Lists the tags for a resource}
\usage{
route53recoveryreadiness_list_tags_for_resources(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) for a resource.}
}
\description{
Lists the tags for a resource.

See \url{https://www.paws-r-sdk.com/docs/route53recoveryreadiness_list_tags_for_resources/} for full documentation.
}
\keyword{internal}
