% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53recoveryreadiness_operations.R
\name{route53recoveryreadiness_untag_resource}
\alias{route53recoveryreadiness_untag_resource}
\title{Removes a tag from a resource}
\usage{
route53recoveryreadiness_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) for a resource.}

\item{TagKeys}{[required] The keys for tags you add to resources.}
}
\description{
Removes a tag from a resource.

See \url{https://www.paws-r-sdk.com/docs/route53recoveryreadiness_untag_resource/} for full documentation.
}
\keyword{internal}
