% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53recoveryreadiness_operations.R
\name{route53recoveryreadiness_update_cell}
\alias{route53recoveryreadiness_update_cell}
\title{Updates a cell to replace the list of nested cells with a new list of
nested cells}
\usage{
route53recoveryreadiness_update_cell(CellName, Cells)
}
\arguments{
\item{CellName}{[required] The name of the cell.}

\item{Cells}{[required] A list of cell Amazon Resource Names (ARNs), which completely replaces
the previous list.}
}
\description{
Updates a cell to replace the list of nested cells with a new list of nested cells.

See \url{https://paws-r.github.io/docs/route53recoveryreadiness/update_cell.html} for full documentation.
}
\keyword{internal}
