% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53recoveryreadiness_operations.R
\name{route53recoveryreadiness_update_readiness_check}
\alias{route53recoveryreadiness_update_readiness_check}
\title{Updates a readiness check}
\usage{
route53recoveryreadiness_update_readiness_check(
  ReadinessCheckName,
  ResourceSetName
)
}
\arguments{
\item{ReadinessCheckName}{[required] Name of a readiness check.}

\item{ResourceSetName}{[required] The name of the resource set to be checked.}
}
\description{
Updates a readiness check.

See \url{https://paws-r.github.io/docs/route53recoveryreadiness/update_readiness_check.html} for full documentation.
}
\keyword{internal}
