% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53recoveryreadiness_operations.R
\name{route53recoveryreadiness_update_resource_set}
\alias{route53recoveryreadiness_update_resource_set}
\title{Updates a resource set}
\usage{
route53recoveryreadiness_update_resource_set(
  ResourceSetName,
  ResourceSetType,
  Resources
)
}
\arguments{
\item{ResourceSetName}{[required] Name of a resource set.}

\item{ResourceSetType}{[required] The resource type of the resources in the resource set. Enter one of the
following values for resource type:

AWS::ApiGateway::Stage, AWS::ApiGatewayV2::Stage,
AWS::AutoScaling::AutoScalingGroup, AWS::CloudWatch::Alarm,
AWS::EC2::CustomerGateway, AWS::DynamoDB::Table, AWS::EC2::Volume,
AWS::ElasticLoadBalancing::LoadBalancer,
AWS::ElasticLoadBalancingV2::LoadBalancer, AWS::Lambda::Function,
AWS::MSK::Cluster, AWS::RDS::DBCluster, AWS::Route53::HealthCheck,
AWS::SQS::Queue, AWS::SNS::Topic, AWS::SNS::Subscription, AWS::EC2::VPC,
AWS::EC2::VPNConnection, AWS::EC2::VPNGateway,
AWS::Route53RecoveryReadiness::DNSTargetResource}

\item{Resources}{[required] A list of resource objects.}
}
\description{
Updates a resource set.

See \url{https://www.paws-r-sdk.com/docs/route53recoveryreadiness_update_resource_set/} for full documentation.
}
\keyword{internal}
