% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53resolver_operations.R
\name{route53resolver_batch_create_firewall_rule}
\alias{route53resolver_batch_create_firewall_rule}
\title{Creates multiple DNS Firewall rules in the specified rule group}
\usage{
route53resolver_batch_create_firewall_rule(CreateFirewallRuleEntries)
}
\arguments{
\item{CreateFirewallRuleEntries}{[required] The list of firewall rules to create.}
}
\description{
Creates multiple DNS Firewall rules in the specified rule group.

See \url{https://www.paws-r-sdk.com/docs/route53resolver_batch_create_firewall_rule/} for full documentation.
}
\keyword{internal}
