% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53resolver_operations.R
\name{route53resolver_batch_delete_firewall_rule}
\alias{route53resolver_batch_delete_firewall_rule}
\title{Deletes multiple DNS Firewall rules from the specified rule group}
\usage{
route53resolver_batch_delete_firewall_rule(DeleteFirewallRuleEntries)
}
\arguments{
\item{DeleteFirewallRuleEntries}{[required] The list of firewall rules to delete.}
}
\description{
Deletes multiple DNS Firewall rules from the specified rule group.

See \url{https://www.paws-r-sdk.com/docs/route53resolver_batch_delete_firewall_rule/} for full documentation.
}
\keyword{internal}
