% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53resolver_operations.R
\name{route53resolver_batch_update_firewall_rule}
\alias{route53resolver_batch_update_firewall_rule}
\title{Updates multiple DNS Firewall rules in the specified rule group}
\usage{
route53resolver_batch_update_firewall_rule(UpdateFirewallRuleEntries)
}
\arguments{
\item{UpdateFirewallRuleEntries}{[required] The list of firewall rules to update.}
}
\description{
Updates multiple DNS Firewall rules in the specified rule group.

See \url{https://www.paws-r-sdk.com/docs/route53resolver_batch_update_firewall_rule/} for full documentation.
}
\keyword{internal}
