% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53resolver_operations.R
\name{route53resolver_delete_outpost_resolver}
\alias{route53resolver_delete_outpost_resolver}
\title{Deletes a Resolver on the Outpost}
\usage{
route53resolver_delete_outpost_resolver(Id)
}
\arguments{
\item{Id}{[required] A unique string that identifies the Resolver on the Outpost.}
}
\description{
Deletes a Resolver on the Outpost.

See \url{https://www.paws-r-sdk.com/docs/route53resolver_delete_outpost_resolver/} for full documentation.
}
\keyword{internal}
