% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53resolver_operations.R
\name{route53resolver_disassociate_resolver_query_log_config}
\alias{route53resolver_disassociate_resolver_query_log_config}
\title{Disassociates a VPC from a query logging configuration}
\usage{
route53resolver_disassociate_resolver_query_log_config(
  ResolverQueryLogConfigId,
  ResourceId
)
}
\arguments{
\item{ResolverQueryLogConfigId}{[required] The ID of the query logging configuration that you want to disassociate
a specified VPC from.}

\item{ResourceId}{[required] The ID of the Amazon VPC that you want to disassociate from a specified
query logging configuration.}
}
\description{
Disassociates a VPC from a query logging configuration.

See \url{https://paws-r.github.io/docs/route53resolver/disassociate_resolver_query_log_config.html} for full documentation.
}
\keyword{internal}
