% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53resolver_operations.R
\name{route53resolver_disassociate_resolver_rule}
\alias{route53resolver_disassociate_resolver_rule}
\title{Removes the association between a specified Resolver rule and a
specified VPC}
\usage{
route53resolver_disassociate_resolver_rule(VPCId, ResolverRuleId)
}
\arguments{
\item{VPCId}{[required] The ID of the VPC that you want to disassociate the Resolver rule from.}

\item{ResolverRuleId}{[required] The ID of the Resolver rule that you want to disassociate from the
specified VPC.}
}
\description{
Removes the association between a specified Resolver rule and a specified VPC.

See \url{https://paws-r.github.io/docs/route53resolver/disassociate_resolver_rule.html} for full documentation.
}
\keyword{internal}
