% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53resolver_operations.R
\name{route53resolver_get_firewall_config}
\alias{route53resolver_get_firewall_config}
\title{Retrieves the configuration of the firewall behavior provided by DNS
Firewall for a single Amazon virtual private cloud (VPC)}
\usage{
route53resolver_get_firewall_config(ResourceId)
}
\arguments{
\item{ResourceId}{[required] The ID of the Amazon virtual private cloud (VPC) that the configuration
is for.}
}
\value{
A list with the following syntax:\preformatted{list(
  FirewallConfig = list(
    Id = "string",
    ResourceId = "string",
    OwnerId = "string",
    FirewallFailOpen = "ENABLED"|"DISABLED"
  )
)
}
}
\description{
Retrieves the configuration of the firewall behavior provided by DNS
Firewall for a single Amazon virtual private cloud (VPC).
}
\section{Request syntax}{
\preformatted{svc$get_firewall_config(
  ResourceId = "string"
)
}
}

\keyword{internal}
