% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53resolver_operations.R
\name{route53resolver_get_firewall_rule_group}
\alias{route53resolver_get_firewall_rule_group}
\title{Retrieves the specified firewall rule group}
\usage{
route53resolver_get_firewall_rule_group(FirewallRuleGroupId)
}
\arguments{
\item{FirewallRuleGroupId}{[required] The unique identifier of the firewall rule group.}
}
\value{
A list with the following syntax:\preformatted{list(
  FirewallRuleGroup = list(
    Id = "string",
    Arn = "string",
    Name = "string",
    RuleCount = 123,
    Status = "COMPLETE"|"DELETING"|"UPDATING",
    StatusMessage = "string",
    OwnerId = "string",
    CreatorRequestId = "string",
    ShareStatus = "NOT_SHARED"|"SHARED_WITH_ME"|"SHARED_BY_ME",
    CreationTime = "string",
    ModificationTime = "string"
  )
)
}
}
\description{
Retrieves the specified firewall rule group.
}
\section{Request syntax}{
\preformatted{svc$get_firewall_rule_group(
  FirewallRuleGroupId = "string"
)
}
}

\keyword{internal}
