% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53resolver_operations.R
\name{route53resolver_get_firewall_rule_group_association}
\alias{route53resolver_get_firewall_rule_group_association}
\title{Retrieves a firewall rule group association, which enables DNS filtering
for a VPC with one rule group}
\usage{
route53resolver_get_firewall_rule_group_association(
  FirewallRuleGroupAssociationId)
}
\arguments{
\item{FirewallRuleGroupAssociationId}{[required] The identifier of the FirewallRuleGroupAssociation.}
}
\value{
A list with the following syntax:\preformatted{list(
  FirewallRuleGroupAssociation = list(
    Id = "string",
    Arn = "string",
    FirewallRuleGroupId = "string",
    VpcId = "string",
    Name = "string",
    Priority = 123,
    MutationProtection = "ENABLED"|"DISABLED",
    ManagedOwnerName = "string",
    Status = "COMPLETE"|"DELETING"|"UPDATING",
    StatusMessage = "string",
    CreatorRequestId = "string",
    CreationTime = "string",
    ModificationTime = "string"
  )
)
}
}
\description{
Retrieves a firewall rule group association, which enables DNS filtering
for a VPC with one rule group. A VPC can have more than one firewall
rule group association, and a rule group can be associated with more
than one VPC.
}
\section{Request syntax}{
\preformatted{svc$get_firewall_rule_group_association(
  FirewallRuleGroupAssociationId = "string"
)
}
}

\keyword{internal}
