% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53resolver_operations.R
\name{route53resolver_get_firewall_rule_group_policy}
\alias{route53resolver_get_firewall_rule_group_policy}
\title{Returns the AWS Identity and Access Management (AWS IAM) policy for
sharing the specified rule group}
\usage{
route53resolver_get_firewall_rule_group_policy(Arn)
}
\arguments{
\item{Arn}{[required] The ARN (Amazon Resource Name) for the rule group.}
}
\value{
A list with the following syntax:\preformatted{list(
  FirewallRuleGroupPolicy = "string"
)
}
}
\description{
Returns the AWS Identity and Access Management (AWS IAM) policy for
sharing the specified rule group. You can use the policy to share the
rule group using AWS Resource Access Manager (RAM).
}
\section{Request syntax}{
\preformatted{svc$get_firewall_rule_group_policy(
  Arn = "string"
)
}
}

\keyword{internal}
