% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53resolver_operations.R
\name{route53resolver_get_outpost_resolver}
\alias{route53resolver_get_outpost_resolver}
\title{Gets information about a specified Resolver on the Outpost, such as its
instance count and type, name, and the current status of the Resolver}
\usage{
route53resolver_get_outpost_resolver(Id)
}
\arguments{
\item{Id}{[required] The ID of the Resolver on the Outpost.}
}
\description{
Gets information about a specified Resolver on the Outpost, such as its instance count and type, name, and the current status of the Resolver.

See \url{https://www.paws-r-sdk.com/docs/route53resolver_get_outpost_resolver/} for full documentation.
}
\keyword{internal}
