% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53resolver_operations.R
\name{route53resolver_get_resolver_dnssec_config}
\alias{route53resolver_get_resolver_dnssec_config}
\title{Gets DNSSEC validation information for a specified resource}
\usage{
route53resolver_get_resolver_dnssec_config(ResourceId)
}
\arguments{
\item{ResourceId}{[required] The ID of the virtual private cloud (VPC) for the DNSSEC validation
status.}
}
\description{
Gets DNSSEC validation information for a specified resource.

See \url{https://www.paws-r-sdk.com/docs/route53resolver_get_resolver_dnssec_config/} for full documentation.
}
\keyword{internal}
