% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53resolver_operations.R
\name{route53resolver_get_resolver_endpoint}
\alias{route53resolver_get_resolver_endpoint}
\title{Gets information about a specified Resolver endpoint, such as whether
it's an inbound or an outbound Resolver endpoint, and the current status
of the endpoint}
\usage{
route53resolver_get_resolver_endpoint(ResolverEndpointId)
}
\arguments{
\item{ResolverEndpointId}{[required] The ID of the Resolver endpoint that you want to get information about.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ResolverEndpoint = list(
    Id = "string",
    CreatorRequestId = "string",
    Arn = "string",
    Name = "string",
    SecurityGroupIds = list(
      "string"
    ),
    Direction = "INBOUND"|"OUTBOUND",
    IpAddressCount = 123,
    HostVPCId = "string",
    Status = "CREATING"|"OPERATIONAL"|"UPDATING"|"AUTO_RECOVERING"|"ACTION_NEEDED"|"DELETING",
    StatusMessage = "string",
    CreationTime = "string",
    ModificationTime = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Gets information about a specified Resolver endpoint, such as whether
it's an inbound or an outbound Resolver endpoint, and the current status
of the endpoint.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_resolver_endpoint(
  ResolverEndpointId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
