% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53resolver_operations.R
\name{route53resolver_get_resolver_rule_association}
\alias{route53resolver_get_resolver_rule_association}
\title{Gets information about an association between a specified Resolver rule
and a VPC}
\usage{
route53resolver_get_resolver_rule_association(ResolverRuleAssociationId)
}
\arguments{
\item{ResolverRuleAssociationId}{[required] The ID of the Resolver rule association that you want to get information
about.}
}
\description{
Gets information about an association between a specified Resolver rule and a VPC. You associate a Resolver rule and a VPC using \code{\link[=route53resolver_associate_resolver_rule]{associate_resolver_rule}}.

See \url{https://paws-r.github.io/docs/route53resolver/get_resolver_rule_association.html} for full documentation.
}
\keyword{internal}
