% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53resolver_operations.R
\name{route53resolver_list_firewall_rule_groups}
\alias{route53resolver_list_firewall_rule_groups}
\title{Retrieves the minimal high-level information for the rule groups that
you have defined}
\usage{
route53resolver_list_firewall_rule_groups(MaxResults, NextToken)
}
\arguments{
\item{MaxResults}{The maximum number of objects that you want Resolver to return for this
request. If more objects are available, in the response, Resolver
provides a \code{NextToken} value that you can use in a subsequent call to
get the next batch of objects.

If you don't specify a value for \code{MaxResults}, Resolver returns up to
100 objects.}

\item{NextToken}{For the first call to this list request, omit this value.

When you request a list of objects, Resolver returns at most the number
of objects specified in \code{MaxResults}. If more objects are available for
retrieval, Resolver returns a \code{NextToken} value in the response. To
retrieve the next batch of objects, use the token that was returned for
the prior request in your next request.}
}
\value{
A list with the following syntax:\preformatted{list(
  NextToken = "string",
  FirewallRuleGroups = list(
    list(
      Id = "string",
      Arn = "string",
      Name = "string",
      OwnerId = "string",
      CreatorRequestId = "string",
      ShareStatus = "NOT_SHARED"|"SHARED_WITH_ME"|"SHARED_BY_ME"
    )
  )
)
}
}
\description{
Retrieves the minimal high-level information for the rule groups that
you have defined.

A single call might return only a partial list of the rule groups. For
information, see \code{MaxResults}.
}
\section{Request syntax}{
\preformatted{svc$list_firewall_rule_groups(
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
