% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53resolver_operations.R
\name{route53resolver_list_outpost_resolvers}
\alias{route53resolver_list_outpost_resolvers}
\title{Lists all the Resolvers on Outposts that were created using the current
Amazon Web Services account}
\usage{
route53resolver_list_outpost_resolvers(
  OutpostArn = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{OutpostArn}{The Amazon Resource Name (ARN) of the Outpost.}

\item{MaxResults}{The maximum number of Resolvers on the Outpost that you want to return
in the response to a \code{ListOutpostResolver} request. If you don't specify
a value for \code{MaxResults}, the request returns up to 100 Resolvers.}

\item{NextToken}{For the first \code{ListOutpostResolver} request, omit this value.}
}
\description{
Lists all the Resolvers on Outposts that were created using the current Amazon Web Services account.

See \url{https://www.paws-r-sdk.com/docs/route53resolver_list_outpost_resolvers/} for full documentation.
}
\keyword{internal}
