% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53resolver_operations.R
\name{route53resolver_put_firewall_rule_group_policy}
\alias{route53resolver_put_firewall_rule_group_policy}
\title{Attaches an AWS Identity and Access Management (AWS IAM) policy for
sharing the rule group}
\usage{
route53resolver_put_firewall_rule_group_policy(Arn,
  FirewallRuleGroupPolicy)
}
\arguments{
\item{Arn}{[required] The ARN (Amazon Resource Name) for the rule group that you want to
share.}

\item{FirewallRuleGroupPolicy}{[required] The AWS Identity and Access Management (AWS IAM) policy to attach to the
rule group.}
}
\value{
A list with the following syntax:\preformatted{list(
  ReturnValue = TRUE|FALSE
)
}
}
\description{
Attaches an AWS Identity and Access Management (AWS IAM) policy for
sharing the rule group. You can use the policy to share the rule group
using AWS Resource Access Manager (RAM).
}
\section{Request syntax}{
\preformatted{svc$put_firewall_rule_group_policy(
  Arn = "string",
  FirewallRuleGroupPolicy = "string"
)
}
}

\keyword{internal}
