% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53resolver_operations.R
\name{route53resolver_untag_resource}
\alias{route53resolver_untag_resource}
\title{Removes one or more tags from a specified resource}
\usage{
route53resolver_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) for the resource that you want to remove
tags from. To get the ARN for a resource, use the applicable \code{Get} or
\code{List} command:
\itemize{
\item \code{\link[=route53resolver_get_resolver_endpoint]{get_resolver_endpoint}}
\item \code{\link[=route53resolver_get_resolver_rule]{get_resolver_rule}}
\item \code{\link[=route53resolver_get_resolver_rule_association]{get_resolver_rule_association}}
\item \code{\link[=route53resolver_list_resolver_endpoints]{list_resolver_endpoints}}
\item \code{\link[=route53resolver_list_resolver_rule_associations]{list_resolver_rule_associations}}
\item \code{\link[=route53resolver_list_resolver_rules]{list_resolver_rules}}
}}

\item{TagKeys}{[required] The tags that you want to remove to the specified resource.}
}
\value{
An empty list.
}
\description{
Removes one or more tags from a specified resource.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  ResourceArn = "string",
  TagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
