% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicediscovery_operations.R
\name{servicediscovery_delete_namespace}
\alias{servicediscovery_delete_namespace}
\title{Deletes a namespace from the current account}
\usage{
servicediscovery_delete_namespace(Id)
}
\arguments{
\item{Id}{[required] The ID of the namespace that you want to delete.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  OperationId = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Deletes a namespace from the current account. If the namespace still
contains one or more services, the request fails.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_namespace(
  Id = "string"
)
}\if{html}{\out{</div>}}
}

\examples{
\dontrun{
# Example: Delete namespace
svc$delete_namespace(
  Id = "ns-ylexjili4cdxy3xm"
)
}

}
\keyword{internal}
