% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicediscovery_operations.R
\name{servicediscovery_delete_service}
\alias{servicediscovery_delete_service}
\title{Deletes a specified service}
\usage{
servicediscovery_delete_service(Id)
}
\arguments{
\item{Id}{[required] The ID of the service that you want to delete.}
}
\value{
An empty list.
}
\description{
Deletes a specified service. If the service still contains one or more
registered instances, the request fails.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_service(
  Id = "string"
)
}\if{html}{\out{</div>}}
}

\examples{
\dontrun{
# Example: Delete service
svc$delete_service(
  Id = "srv-p5zdwlg5uvvzjita"
)
}

}
\keyword{internal}
