% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicediscovery_operations.R
\name{servicediscovery_delete_service_attributes}
\alias{servicediscovery_delete_service_attributes}
\title{Deletes specific attributes associated with a service}
\usage{
servicediscovery_delete_service_attributes(ServiceId, Attributes)
}
\arguments{
\item{ServiceId}{[required] The ID of the service from which the attributes will be deleted.}

\item{Attributes}{[required] A list of keys corresponding to each attribute that you want to delete.}
}
\description{
Deletes specific attributes associated with a service.

See \url{https://www.paws-r-sdk.com/docs/servicediscovery_delete_service_attributes/} for full documentation.
}
\keyword{internal}
