% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicediscovery_operations.R
\name{servicediscovery_discover_instances_revision}
\alias{servicediscovery_discover_instances_revision}
\title{Discovers the increasing revision associated with an instance}
\usage{
servicediscovery_discover_instances_revision(NamespaceName, ServiceName)
}
\arguments{
\item{NamespaceName}{[required] The \code{HttpName} name of the namespace. It's found in the \code{HttpProperties}
member of the \code{Properties} member of the namespace.}

\item{ServiceName}{[required] The name of the service that you specified when you registered the
instance.}
}
\description{
Discovers the increasing revision associated with an instance.

See \url{https://www.paws-r-sdk.com/docs/servicediscovery_discover_instances_revision/} for full documentation.
}
\keyword{internal}
