% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicediscovery_operations.R
\name{servicediscovery_get_service_attributes}
\alias{servicediscovery_get_service_attributes}
\title{Returns the attributes associated with a specified service}
\usage{
servicediscovery_get_service_attributes(ServiceId)
}
\arguments{
\item{ServiceId}{[required] The ID of the service that you want to get attributes for.}
}
\description{
Returns the attributes associated with a specified service.

See \url{https://www.paws-r-sdk.com/docs/servicediscovery_get_service_attributes/} for full documentation.
}
\keyword{internal}
