% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicediscovery_operations.R
\name{servicediscovery_untag_resource}
\alias{servicediscovery_untag_resource}
\title{Removes one or more tags from the specified resource}
\usage{
servicediscovery_untag_resource(ResourceARN, TagKeys)
}
\arguments{
\item{ResourceARN}{[required] The Amazon Resource Name (ARN) of the resource that you want to retrieve
tags for.}

\item{TagKeys}{[required] The tag keys to remove from the specified resource.}
}
\description{
Removes one or more tags from the specified resource.

See \url{https://www.paws-r-sdk.com/docs/servicediscovery_untag_resource/} for full documentation.
}
\keyword{internal}
