% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicediscovery_operations.R
\name{servicediscovery_update_http_namespace}
\alias{servicediscovery_update_http_namespace}
\title{Updates an HTTP namespace}
\usage{
servicediscovery_update_http_namespace(Id, UpdaterRequestId = NULL, Namespace)
}
\arguments{
\item{Id}{[required] The ID or Amazon Resource Name (ARN) of the namespace that you want to
update.}

\item{UpdaterRequestId}{A unique string that identifies the request and that allows failed
\code{\link[=servicediscovery_update_http_namespace]{update_http_namespace}}
requests to be retried without the risk of running the operation twice.
\code{UpdaterRequestId} can be any unique string (for example, a
date/timestamp).}

\item{Namespace}{[required] Updated properties for the the HTTP namespace.}
}
\description{
Updates an HTTP namespace.

See \url{https://www.paws-r-sdk.com/docs/servicediscovery_update_http_namespace/} for full documentation.
}
\keyword{internal}
