% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicediscovery_operations.R
\name{servicediscovery_update_instance_custom_health_status}
\alias{servicediscovery_update_instance_custom_health_status}
\title{Submits a request to change the health status of a custom health check
to healthy or unhealthy}
\usage{
servicediscovery_update_instance_custom_health_status(
  ServiceId,
  InstanceId,
  Status
)
}
\arguments{
\item{ServiceId}{[required] The ID of the service that includes the configuration for the custom
health check that you want to change the status for.}

\item{InstanceId}{[required] The ID of the instance that you want to change the health status for.}

\item{Status}{[required] The new status of the instance, \code{HEALTHY} or \code{UNHEALTHY}.}
}
\description{
Submits a request to change the health status of a custom health check to healthy or unhealthy.

See \url{https://paws-r.github.io/docs/servicediscovery/update_instance_custom_health_status.html} for full documentation.
}
\keyword{internal}
