% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicediscovery_operations.R
\name{servicediscovery_update_service}
\alias{servicediscovery_update_service}
\title{Submits a request to perform the following operations:}
\usage{
servicediscovery_update_service(Id, Service)
}
\arguments{
\item{Id}{[required] The ID of the service that you want to update.}

\item{Service}{[required] A complex type that contains the new settings for the service. You can
specify a maximum of 30 attributes (key-value pairs).}
}
\description{
Submits a request to perform the following operations:

See \url{https://www.paws-r-sdk.com/docs/servicediscovery_update_service/} for full documentation.
}
\keyword{internal}
