% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicediscovery_operations.R
\name{servicediscovery_update_service_attributes}
\alias{servicediscovery_update_service_attributes}
\title{Submits a request to update a specified service to add service-level
attributes}
\usage{
servicediscovery_update_service_attributes(ServiceId, Attributes)
}
\arguments{
\item{ServiceId}{[required] The ID or Amazon Resource Name (ARN) of the service that you want to
update. For services created in a namespace shared with your Amazon Web
Services account, specify the service ARN.}

\item{Attributes}{[required] A string map that contains attribute key-value pairs.}
}
\description{
Submits a request to update a specified service to add service-level attributes.

See \url{https://www.paws-r-sdk.com/docs/servicediscovery_update_service_attributes/} for full documentation.
}
\keyword{internal}
