% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telconetworkbuilder_operations.R
\name{telconetworkbuilder_cancel_sol_network_operation}
\alias{telconetworkbuilder_cancel_sol_network_operation}
\title{Cancels a network operation}
\usage{
telconetworkbuilder_cancel_sol_network_operation(nsLcmOpOccId)
}
\arguments{
\item{nsLcmOpOccId}{[required] The identifier of the network operation.}
}
\description{
Cancels a network operation.

See \url{https://www.paws-r-sdk.com/docs/telconetworkbuilder_cancel_sol_network_operation/} for full documentation.
}
\keyword{internal}
