% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telconetworkbuilder_operations.R
\name{telconetworkbuilder_create_sol_function_package}
\alias{telconetworkbuilder_create_sol_function_package}
\title{Creates a function package}
\usage{
telconetworkbuilder_create_sol_function_package(tags = NULL)
}
\arguments{
\item{tags}{A tag is a label that you assign to an Amazon Web Services resource.
Each tag consists of a key and an optional value. You can use tags to
search and filter your resources or track your Amazon Web Services
costs.}
}
\description{
Creates a function package.

See \url{https://www.paws-r-sdk.com/docs/telconetworkbuilder_create_sol_function_package/} for full documentation.
}
\keyword{internal}
