% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telconetworkbuilder_operations.R
\name{telconetworkbuilder_create_sol_network_instance}
\alias{telconetworkbuilder_create_sol_network_instance}
\title{Creates a network instance}
\usage{
telconetworkbuilder_create_sol_network_instance(
  nsDescription = NULL,
  nsName,
  nsdInfoId,
  tags = NULL
)
}
\arguments{
\item{nsDescription}{Network instance description.}

\item{nsName}{[required] Network instance name.}

\item{nsdInfoId}{[required] ID for network service descriptor.}

\item{tags}{A tag is a label that you assign to an Amazon Web Services resource.
Each tag consists of a key and an optional value. You can use tags to
search and filter your resources or track your Amazon Web Services
costs.}
}
\description{
Creates a network instance.

See \url{https://www.paws-r-sdk.com/docs/telconetworkbuilder_create_sol_network_instance/} for full documentation.
}
\keyword{internal}
