% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telconetworkbuilder_operations.R
\name{telconetworkbuilder_delete_sol_function_package}
\alias{telconetworkbuilder_delete_sol_function_package}
\title{Deletes a function package}
\usage{
telconetworkbuilder_delete_sol_function_package(vnfPkgId)
}
\arguments{
\item{vnfPkgId}{[required] ID of the function package.}
}
\description{
Deletes a function package.

See \url{https://www.paws-r-sdk.com/docs/telconetworkbuilder_delete_sol_function_package/} for full documentation.
}
\keyword{internal}
