% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telconetworkbuilder_operations.R
\name{telconetworkbuilder_delete_sol_network_instance}
\alias{telconetworkbuilder_delete_sol_network_instance}
\title{Deletes a network instance}
\usage{
telconetworkbuilder_delete_sol_network_instance(nsInstanceId)
}
\arguments{
\item{nsInstanceId}{[required] Network instance ID.}
}
\description{
Deletes a network instance.

See \url{https://www.paws-r-sdk.com/docs/telconetworkbuilder_delete_sol_network_instance/} for full documentation.
}
\keyword{internal}
