% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telconetworkbuilder_operations.R
\name{telconetworkbuilder_get_sol_function_instance}
\alias{telconetworkbuilder_get_sol_function_instance}
\title{Gets the details of a network function instance, including the
instantiation state and metadata from the function package descriptor in
the network function package}
\usage{
telconetworkbuilder_get_sol_function_instance(vnfInstanceId)
}
\arguments{
\item{vnfInstanceId}{[required] ID of the network function.}
}
\description{
Gets the details of a network function instance, including the instantiation state and metadata from the function package descriptor in the network function package.

See \url{https://www.paws-r-sdk.com/docs/telconetworkbuilder_get_sol_function_instance/} for full documentation.
}
\keyword{internal}
