% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telconetworkbuilder_operations.R
\name{telconetworkbuilder_get_sol_function_package}
\alias{telconetworkbuilder_get_sol_function_package}
\title{Gets the details of an individual function package, such as the
operational state and whether the package is in use}
\usage{
telconetworkbuilder_get_sol_function_package(vnfPkgId)
}
\arguments{
\item{vnfPkgId}{[required] ID of the function package.}
}
\description{
Gets the details of an individual function package, such as the operational state and whether the package is in use.

See \url{https://www.paws-r-sdk.com/docs/telconetworkbuilder_get_sol_function_package/} for full documentation.
}
\keyword{internal}
