% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telconetworkbuilder_operations.R
\name{telconetworkbuilder_get_sol_function_package_content}
\alias{telconetworkbuilder_get_sol_function_package_content}
\title{Gets the contents of a function package}
\usage{
telconetworkbuilder_get_sol_function_package_content(accept, vnfPkgId)
}
\arguments{
\item{accept}{[required] The format of the package that you want to download from the function
packages.}

\item{vnfPkgId}{[required] ID of the function package.}
}
\description{
Gets the contents of a function package.

See \url{https://www.paws-r-sdk.com/docs/telconetworkbuilder_get_sol_function_package_content/} for full documentation.
}
\keyword{internal}
