% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telconetworkbuilder_operations.R
\name{telconetworkbuilder_get_sol_function_package_descriptor}
\alias{telconetworkbuilder_get_sol_function_package_descriptor}
\title{Gets a function package descriptor in a function package}
\usage{
telconetworkbuilder_get_sol_function_package_descriptor(accept, vnfPkgId)
}
\arguments{
\item{accept}{[required] Indicates which content types, expressed as MIME types, the client is
able to understand.}

\item{vnfPkgId}{[required] ID of the function package.}
}
\description{
Gets a function package descriptor in a function package.

See \url{https://www.paws-r-sdk.com/docs/telconetworkbuilder_get_sol_function_package_descriptor/} for full documentation.
}
\keyword{internal}
