% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telconetworkbuilder_operations.R
\name{telconetworkbuilder_get_sol_network_instance}
\alias{telconetworkbuilder_get_sol_network_instance}
\title{Gets the details of the network instance}
\usage{
telconetworkbuilder_get_sol_network_instance(nsInstanceId)
}
\arguments{
\item{nsInstanceId}{[required] ID of the network instance.}
}
\description{
Gets the details of the network instance.

See \url{https://www.paws-r-sdk.com/docs/telconetworkbuilder_get_sol_network_instance/} for full documentation.
}
\keyword{internal}
