% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telconetworkbuilder_operations.R
\name{telconetworkbuilder_get_sol_network_operation}
\alias{telconetworkbuilder_get_sol_network_operation}
\title{Gets the details of a network operation, including the tasks involved in
the network operation and the status of the tasks}
\usage{
telconetworkbuilder_get_sol_network_operation(nsLcmOpOccId)
}
\arguments{
\item{nsLcmOpOccId}{[required] The identifier of the network operation.}
}
\description{
Gets the details of a network operation, including the tasks involved in the network operation and the status of the tasks.

See \url{https://www.paws-r-sdk.com/docs/telconetworkbuilder_get_sol_network_operation/} for full documentation.
}
\keyword{internal}
