% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telconetworkbuilder_operations.R
\name{telconetworkbuilder_get_sol_network_package}
\alias{telconetworkbuilder_get_sol_network_package}
\title{Gets the details of a network package}
\usage{
telconetworkbuilder_get_sol_network_package(nsdInfoId)
}
\arguments{
\item{nsdInfoId}{[required] ID of the network service descriptor in the network package.}
}
\description{
Gets the details of a network package.

See \url{https://www.paws-r-sdk.com/docs/telconetworkbuilder_get_sol_network_package/} for full documentation.
}
\keyword{internal}
