% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telconetworkbuilder_operations.R
\name{telconetworkbuilder_get_sol_network_package_content}
\alias{telconetworkbuilder_get_sol_network_package_content}
\title{Gets the contents of a network package}
\usage{
telconetworkbuilder_get_sol_network_package_content(accept, nsdInfoId)
}
\arguments{
\item{accept}{[required] The format of the package you want to download from the network package.}

\item{nsdInfoId}{[required] ID of the network service descriptor in the network package.}
}
\description{
Gets the contents of a network package.

See \url{https://www.paws-r-sdk.com/docs/telconetworkbuilder_get_sol_network_package_content/} for full documentation.
}
\keyword{internal}
