% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telconetworkbuilder_operations.R
\name{telconetworkbuilder_get_sol_network_package_descriptor}
\alias{telconetworkbuilder_get_sol_network_package_descriptor}
\title{Gets the content of the network service descriptor}
\usage{
telconetworkbuilder_get_sol_network_package_descriptor(nsdInfoId)
}
\arguments{
\item{nsdInfoId}{[required] ID of the network service descriptor in the network package.}
}
\description{
Gets the content of the network service descriptor.

See \url{https://www.paws-r-sdk.com/docs/telconetworkbuilder_get_sol_network_package_descriptor/} for full documentation.
}
\keyword{internal}
