% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telconetworkbuilder_operations.R
\name{telconetworkbuilder_list_sol_network_instances}
\alias{telconetworkbuilder_list_sol_network_instances}
\title{Lists your network instances}
\usage{
telconetworkbuilder_list_sol_network_instances(
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{maxResults}{The maximum number of results to include in the response.}

\item{nextToken}{The token for the next page of results.}
}
\description{
Lists your network instances.

See \url{https://www.paws-r-sdk.com/docs/telconetworkbuilder_list_sol_network_instances/} for full documentation.
}
\keyword{internal}
