% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telconetworkbuilder_operations.R
\name{telconetworkbuilder_list_sol_network_packages}
\alias{telconetworkbuilder_list_sol_network_packages}
\title{Lists network packages}
\usage{
telconetworkbuilder_list_sol_network_packages(
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{maxResults}{The maximum number of results to include in the response.}

\item{nextToken}{The token for the next page of results.}
}
\description{
Lists network packages.

See \url{https://www.paws-r-sdk.com/docs/telconetworkbuilder_list_sol_network_packages/} for full documentation.
}
\keyword{internal}
