% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telconetworkbuilder_operations.R
\name{telconetworkbuilder_list_tags_for_resource}
\alias{telconetworkbuilder_list_tags_for_resource}
\title{Lists tags for AWS TNB resources}
\usage{
telconetworkbuilder_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] Resource ARN.}
}
\description{
Lists tags for AWS TNB resources.

See \url{https://www.paws-r-sdk.com/docs/telconetworkbuilder_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
