% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telconetworkbuilder_operations.R
\name{telconetworkbuilder_put_sol_function_package_content}
\alias{telconetworkbuilder_put_sol_function_package_content}
\title{Uploads the contents of a function package}
\usage{
telconetworkbuilder_put_sol_function_package_content(
  contentType = NULL,
  file,
  vnfPkgId
)
}
\arguments{
\item{contentType}{Function package content type.}

\item{file}{[required] Function package file.}

\item{vnfPkgId}{[required] Function package ID.}
}
\description{
Uploads the contents of a function package.

See \url{https://www.paws-r-sdk.com/docs/telconetworkbuilder_put_sol_function_package_content/} for full documentation.
}
\keyword{internal}
