% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telconetworkbuilder_operations.R
\name{telconetworkbuilder_put_sol_network_package_content}
\alias{telconetworkbuilder_put_sol_network_package_content}
\title{Uploads the contents of a network package}
\usage{
telconetworkbuilder_put_sol_network_package_content(
  contentType = NULL,
  file,
  nsdInfoId
)
}
\arguments{
\item{contentType}{Network package content type.}

\item{file}{[required] Network package file.}

\item{nsdInfoId}{[required] Network service descriptor info ID.}
}
\description{
Uploads the contents of a network package.

See \url{https://www.paws-r-sdk.com/docs/telconetworkbuilder_put_sol_network_package_content/} for full documentation.
}
\keyword{internal}
