% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telconetworkbuilder_operations.R
\name{telconetworkbuilder_tag_resource}
\alias{telconetworkbuilder_tag_resource}
\title{Tags an AWS TNB resource}
\usage{
telconetworkbuilder_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] Resource ARN.}

\item{tags}{[required] A tag is a label that you assign to an Amazon Web Services resource.
Each tag consists of a key and an optional value. You can use tags to
search and filter your resources or track your Amazon Web Services
costs.}
}
\description{
Tags an AWS TNB resource.

See \url{https://www.paws-r-sdk.com/docs/telconetworkbuilder_tag_resource/} for full documentation.
}
\keyword{internal}
