% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telconetworkbuilder_operations.R
\name{telconetworkbuilder_terminate_sol_network_instance}
\alias{telconetworkbuilder_terminate_sol_network_instance}
\title{Terminates a network instance}
\usage{
telconetworkbuilder_terminate_sol_network_instance(nsInstanceId, tags = NULL)
}
\arguments{
\item{nsInstanceId}{[required] ID of the network instance.}

\item{tags}{A tag is a label that you assign to an Amazon Web Services resource.
Each tag consists of a key and an optional value. When you use this API,
the tags are only applied to the network operation that is created.
These tags are not applied to the network instance. Use tags to search
and filter your resources or track your Amazon Web Services costs.}
}
\description{
Terminates a network instance.

See \url{https://www.paws-r-sdk.com/docs/telconetworkbuilder_terminate_sol_network_instance/} for full documentation.
}
\keyword{internal}
