% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telconetworkbuilder_operations.R
\name{telconetworkbuilder_untag_resource}
\alias{telconetworkbuilder_untag_resource}
\title{Untags an AWS TNB resource}
\usage{
telconetworkbuilder_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] Resource ARN.}

\item{tagKeys}{[required] Tag keys.}
}
\description{
Untags an AWS TNB resource.

See \url{https://www.paws-r-sdk.com/docs/telconetworkbuilder_untag_resource/} for full documentation.
}
\keyword{internal}
