% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telconetworkbuilder_operations.R
\name{telconetworkbuilder_update_sol_function_package}
\alias{telconetworkbuilder_update_sol_function_package}
\title{Updates the operational state of function package}
\usage{
telconetworkbuilder_update_sol_function_package(operationalState, vnfPkgId)
}
\arguments{
\item{operationalState}{[required] Operational state of the function package.}

\item{vnfPkgId}{[required] ID of the function package.}
}
\description{
Updates the operational state of function package.

See \url{https://www.paws-r-sdk.com/docs/telconetworkbuilder_update_sol_function_package/} for full documentation.
}
\keyword{internal}
