% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telconetworkbuilder_operations.R
\name{telconetworkbuilder_update_sol_network_package}
\alias{telconetworkbuilder_update_sol_network_package}
\title{Updates the operational state of a network package}
\usage{
telconetworkbuilder_update_sol_network_package(nsdInfoId, nsdOperationalState)
}
\arguments{
\item{nsdInfoId}{[required] ID of the network service descriptor in the network package.}

\item{nsdOperationalState}{[required] Operational state of the network service descriptor in the network
package.}
}
\description{
Updates the operational state of a network package.

See \url{https://www.paws-r-sdk.com/docs/telconetworkbuilder_update_sol_network_package/} for full documentation.
}
\keyword{internal}
