% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telconetworkbuilder_operations.R
\name{telconetworkbuilder_validate_sol_function_package_content}
\alias{telconetworkbuilder_validate_sol_function_package_content}
\title{Validates function package content}
\usage{
telconetworkbuilder_validate_sol_function_package_content(
  contentType = NULL,
  file,
  vnfPkgId
)
}
\arguments{
\item{contentType}{Function package content type.}

\item{file}{[required] Function package file.}

\item{vnfPkgId}{[required] Function package ID.}
}
\description{
Validates function package content. This can be used as a dry run before uploading function package content with \code{\link[=telconetworkbuilder_put_sol_function_package_content]{put_sol_function_package_content}}.

See \url{https://www.paws-r-sdk.com/docs/telconetworkbuilder_validate_sol_function_package_content/} for full documentation.
}
\keyword{internal}
