% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telconetworkbuilder_operations.R
\name{telconetworkbuilder_validate_sol_network_package_content}
\alias{telconetworkbuilder_validate_sol_network_package_content}
\title{Validates network package content}
\usage{
telconetworkbuilder_validate_sol_network_package_content(
  contentType = NULL,
  file,
  nsdInfoId
)
}
\arguments{
\item{contentType}{Network package content type.}

\item{file}{[required] Network package file.}

\item{nsdInfoId}{[required] Network service descriptor file.}
}
\description{
Validates network package content. This can be used as a dry run before uploading network package content with \code{\link[=telconetworkbuilder_put_sol_network_package_content]{put_sol_network_package_content}}.

See \url{https://www.paws-r-sdk.com/docs/telconetworkbuilder_validate_sol_network_package_content/} for full documentation.
}
\keyword{internal}
