% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpclattice_operations.R
\name{vpclattice_create_service_network_resource_association}
\alias{vpclattice_create_service_network_resource_association}
\title{Associates the specified service network with the specified resource
configuration}
\usage{
vpclattice_create_service_network_resource_association(
  clientToken = NULL,
  resourceConfigurationIdentifier,
  serviceNetworkIdentifier,
  privateDnsEnabled = NULL,
  tags = NULL
)
}
\arguments{
\item{clientToken}{A unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. If you retry a request that completed
successfully using the same client token and parameters, the retry
succeeds without performing any actions. If the parameters aren't
identical, the retry fails.}

\item{resourceConfigurationIdentifier}{[required] The ID of the resource configuration to associate with the service
network.}

\item{serviceNetworkIdentifier}{[required] The ID of the service network to associate with the resource
configuration.}

\item{privateDnsEnabled}{Indicates if private DNS is enabled for the service network resource
association.}

\item{tags}{A key-value pair to associate with a resource.}
}
\description{
Associates the specified service network with the specified resource configuration. This allows the resource configuration to receive connections through the service network, including through a service network VPC endpoint.

See \url{https://www.paws-r-sdk.com/docs/vpclattice_create_service_network_resource_association/} for full documentation.
}
\keyword{internal}
