% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpclattice_operations.R
\name{vpclattice_delete_domain_verification}
\alias{vpclattice_delete_domain_verification}
\title{Deletes the specified domain verification}
\usage{
vpclattice_delete_domain_verification(domainVerificationIdentifier)
}
\arguments{
\item{domainVerificationIdentifier}{[required] The ID of the domain verification to delete.}
}
\description{
Deletes the specified domain verification.

See \url{https://www.paws-r-sdk.com/docs/vpclattice_delete_domain_verification/} for full documentation.
}
\keyword{internal}
