% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpclattice_operations.R
\name{vpclattice_delete_listener}
\alias{vpclattice_delete_listener}
\title{Deletes the specified listener}
\usage{
vpclattice_delete_listener(listenerIdentifier, serviceIdentifier)
}
\arguments{
\item{listenerIdentifier}{[required] The ID or ARN of the listener.}

\item{serviceIdentifier}{[required] The ID or ARN of the service.}
}
\description{
Deletes the specified listener.

See \url{https://www.paws-r-sdk.com/docs/vpclattice_delete_listener/} for full documentation.
}
\keyword{internal}
