% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpclattice_operations.R
\name{vpclattice_delete_resource_configuration}
\alias{vpclattice_delete_resource_configuration}
\title{Deletes the specified resource configuration}
\usage{
vpclattice_delete_resource_configuration(resourceConfigurationIdentifier)
}
\arguments{
\item{resourceConfigurationIdentifier}{[required] The ID or ARN of the resource configuration.}
}
\description{
Deletes the specified resource configuration.

See \url{https://www.paws-r-sdk.com/docs/vpclattice_delete_resource_configuration/} for full documentation.
}
\keyword{internal}
