% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpclattice_operations.R
\name{vpclattice_delete_resource_gateway}
\alias{vpclattice_delete_resource_gateway}
\title{Deletes the specified resource gateway}
\usage{
vpclattice_delete_resource_gateway(resourceGatewayIdentifier)
}
\arguments{
\item{resourceGatewayIdentifier}{[required] The ID or ARN of the resource gateway.}
}
\description{
Deletes the specified resource gateway.

See \url{https://www.paws-r-sdk.com/docs/vpclattice_delete_resource_gateway/} for full documentation.
}
\keyword{internal}
