% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpclattice_operations.R
\name{vpclattice_delete_resource_policy}
\alias{vpclattice_delete_resource_policy}
\title{Deletes the specified resource policy}
\usage{
vpclattice_delete_resource_policy(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource.}
}
\description{
Deletes the specified resource policy.

See \url{https://www.paws-r-sdk.com/docs/vpclattice_delete_resource_policy/} for full documentation.
}
\keyword{internal}
