% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpclattice_operations.R
\name{vpclattice_delete_rule}
\alias{vpclattice_delete_rule}
\title{Deletes a listener rule}
\usage{
vpclattice_delete_rule(serviceIdentifier, listenerIdentifier, ruleIdentifier)
}
\arguments{
\item{serviceIdentifier}{[required] The ID or ARN of the service.}

\item{listenerIdentifier}{[required] The ID or ARN of the listener.}

\item{ruleIdentifier}{[required] The ID or ARN of the rule.}
}
\description{
Deletes a listener rule. Each listener has a default rule for checking connection requests, but you can define additional rules. Each rule consists of a priority, one or more actions, and one or more conditions. You can delete additional listener rules, but you cannot delete the default rule.

See \url{https://www.paws-r-sdk.com/docs/vpclattice_delete_rule/} for full documentation.
}
\keyword{internal}
