% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpclattice_operations.R
\name{vpclattice_delete_service}
\alias{vpclattice_delete_service}
\title{Deletes a service}
\usage{
vpclattice_delete_service(serviceIdentifier)
}
\arguments{
\item{serviceIdentifier}{[required] The ID or Amazon Resource Name (ARN) of the service.}
}
\description{
Deletes a service. A service can't be deleted if it's associated with a service network. If you delete a service, all resources related to the service, such as the resource policy, auth policy, listeners, listener rules, and access log subscriptions, are also deleted. For more information, see \href{https://docs.aws.amazon.com/vpc-lattice/latest/ug/services.html#delete-service}{Delete a service} in the \emph{Amazon VPC Lattice User Guide}.

See \url{https://www.paws-r-sdk.com/docs/vpclattice_delete_service/} for full documentation.
}
\keyword{internal}
