% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpclattice_operations.R
\name{vpclattice_delete_service_network}
\alias{vpclattice_delete_service_network}
\title{Deletes a service network}
\usage{
vpclattice_delete_service_network(serviceNetworkIdentifier)
}
\arguments{
\item{serviceNetworkIdentifier}{[required] The Amazon Resource Name (ARN) or ID of the service network.}
}
\description{
Deletes a service network. You can only delete the service network if there is no service or VPC associated with it. If you delete a service network, all resources related to the service network, such as the resource policy, auth policy, and access log subscriptions, are also deleted. For more information, see \href{https://docs.aws.amazon.com/vpc-lattice/latest/ug/service-networks.html#delete-service-network}{Delete a service network} in the \emph{Amazon VPC Lattice User Guide}.

See \url{https://www.paws-r-sdk.com/docs/vpclattice_delete_service_network/} for full documentation.
}
\keyword{internal}
