% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpclattice_operations.R
\name{vpclattice_delete_target_group}
\alias{vpclattice_delete_target_group}
\title{Deletes a target group}
\usage{
vpclattice_delete_target_group(targetGroupIdentifier)
}
\arguments{
\item{targetGroupIdentifier}{[required] The ID or Amazon Resource Name (ARN) of the target group.}
}
\description{
Deletes a target group. You can't delete a target group if it is used in a listener rule or if the target group creation is in progress.

See \url{https://www.paws-r-sdk.com/docs/vpclattice_delete_target_group/} for full documentation.
}
\keyword{internal}
