% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpclattice_operations.R
\name{vpclattice_deregister_targets}
\alias{vpclattice_deregister_targets}
\title{Deregisters the specified targets from the specified target group}
\usage{
vpclattice_deregister_targets(targetGroupIdentifier, targets)
}
\arguments{
\item{targetGroupIdentifier}{[required] The ID or Amazon Resource Name (ARN) of the target group.}

\item{targets}{[required] The targets to deregister.}
}
\description{
Deregisters the specified targets from the specified target group.

See \url{https://www.paws-r-sdk.com/docs/vpclattice_deregister_targets/} for full documentation.
}
\keyword{internal}
