% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpclattice_operations.R
\name{vpclattice_get_access_log_subscription}
\alias{vpclattice_get_access_log_subscription}
\title{Retrieves information about the specified access log subscription}
\usage{
vpclattice_get_access_log_subscription(accessLogSubscriptionIdentifier)
}
\arguments{
\item{accessLogSubscriptionIdentifier}{[required] The ID or ARN of the access log subscription.}
}
\description{
Retrieves information about the specified access log subscription.

See \url{https://www.paws-r-sdk.com/docs/vpclattice_get_access_log_subscription/} for full documentation.
}
\keyword{internal}
