% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpclattice_operations.R
\name{vpclattice_get_auth_policy}
\alias{vpclattice_get_auth_policy}
\title{Retrieves information about the auth policy for the specified service or
service network}
\usage{
vpclattice_get_auth_policy(resourceIdentifier)
}
\arguments{
\item{resourceIdentifier}{[required] The ID or Amazon Resource Name (ARN) of the service network or service.}
}
\description{
Retrieves information about the auth policy for the specified service or service network.

See \url{https://www.paws-r-sdk.com/docs/vpclattice_get_auth_policy/} for full documentation.
}
\keyword{internal}
