% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpclattice_operations.R
\name{vpclattice_get_domain_verification}
\alias{vpclattice_get_domain_verification}
\title{Retrieves information about a domain verification}
\usage{
vpclattice_get_domain_verification(domainVerificationIdentifier)
}
\arguments{
\item{domainVerificationIdentifier}{[required] The ID or ARN of the domain verification to retrieve.}
}
\description{
Retrieves information about a domain verification.ß

See \url{https://www.paws-r-sdk.com/docs/vpclattice_get_domain_verification/} for full documentation.
}
\keyword{internal}
