% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpclattice_operations.R
\name{vpclattice_get_resource_configuration}
\alias{vpclattice_get_resource_configuration}
\title{Retrieves information about the specified resource configuration}
\usage{
vpclattice_get_resource_configuration(resourceConfigurationIdentifier)
}
\arguments{
\item{resourceConfigurationIdentifier}{[required] The ID of the resource configuration.}
}
\description{
Retrieves information about the specified resource configuration.

See \url{https://www.paws-r-sdk.com/docs/vpclattice_get_resource_configuration/} for full documentation.
}
\keyword{internal}
