% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpclattice_operations.R
\name{vpclattice_get_resource_gateway}
\alias{vpclattice_get_resource_gateway}
\title{Retrieves information about the specified resource gateway}
\usage{
vpclattice_get_resource_gateway(resourceGatewayIdentifier)
}
\arguments{
\item{resourceGatewayIdentifier}{[required] The ID of the resource gateway.}
}
\description{
Retrieves information about the specified resource gateway.

See \url{https://www.paws-r-sdk.com/docs/vpclattice_get_resource_gateway/} for full documentation.
}
\keyword{internal}
