% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpclattice_operations.R
\name{vpclattice_get_rule}
\alias{vpclattice_get_rule}
\title{Retrieves information about the specified listener rules}
\usage{
vpclattice_get_rule(serviceIdentifier, listenerIdentifier, ruleIdentifier)
}
\arguments{
\item{serviceIdentifier}{[required] The ID or ARN of the service.}

\item{listenerIdentifier}{[required] The ID or ARN of the listener.}

\item{ruleIdentifier}{[required] The ID or ARN of the listener rule.}
}
\description{
Retrieves information about the specified listener rules. You can also retrieve information about the default listener rule. For more information, see \href{https://docs.aws.amazon.com/vpc-lattice/latest/ug/listeners.html#listener-rules}{Listener rules} in the \emph{Amazon VPC Lattice User Guide}.

See \url{https://www.paws-r-sdk.com/docs/vpclattice_get_rule/} for full documentation.
}
\keyword{internal}
