% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpclattice_operations.R
\name{vpclattice_get_service}
\alias{vpclattice_get_service}
\title{Retrieves information about the specified service}
\usage{
vpclattice_get_service(serviceIdentifier)
}
\arguments{
\item{serviceIdentifier}{[required] The ID or Amazon Resource Name (ARN) of the service.}
}
\description{
Retrieves information about the specified service.

See \url{https://www.paws-r-sdk.com/docs/vpclattice_get_service/} for full documentation.
}
\keyword{internal}
