% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpclattice_operations.R
\name{vpclattice_get_service_network}
\alias{vpclattice_get_service_network}
\title{Retrieves information about the specified service network}
\usage{
vpclattice_get_service_network(serviceNetworkIdentifier)
}
\arguments{
\item{serviceNetworkIdentifier}{[required] The ID or Amazon Resource Name (ARN) of the service network.}
}
\description{
Retrieves information about the specified service network.

See \url{https://www.paws-r-sdk.com/docs/vpclattice_get_service_network/} for full documentation.
}
\keyword{internal}
