% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpclattice_operations.R
\name{vpclattice_get_service_network_resource_association}
\alias{vpclattice_get_service_network_resource_association}
\title{Retrieves information about the specified association between a service
network and a resource configuration}
\usage{
vpclattice_get_service_network_resource_association(
  serviceNetworkResourceAssociationIdentifier
)
}
\arguments{
\item{serviceNetworkResourceAssociationIdentifier}{[required] The ID of the association.}
}
\description{
Retrieves information about the specified association between a service network and a resource configuration.

See \url{https://www.paws-r-sdk.com/docs/vpclattice_get_service_network_resource_association/} for full documentation.
}
\keyword{internal}
