% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpclattice_operations.R
\name{vpclattice_get_service_network_service_association}
\alias{vpclattice_get_service_network_service_association}
\title{Retrieves information about the specified association between a service
network and a service}
\usage{
vpclattice_get_service_network_service_association(
  serviceNetworkServiceAssociationIdentifier
)
}
\arguments{
\item{serviceNetworkServiceAssociationIdentifier}{[required] The ID or Amazon Resource Name (ARN) of the association.}
}
\description{
Retrieves information about the specified association between a service network and a service.

See \url{https://www.paws-r-sdk.com/docs/vpclattice_get_service_network_service_association/} for full documentation.
}
\keyword{internal}
