% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpclattice_operations.R
\name{vpclattice_get_service_network_vpc_association}
\alias{vpclattice_get_service_network_vpc_association}
\title{Retrieves information about the specified association between a service
network and a VPC}
\usage{
vpclattice_get_service_network_vpc_association(
  serviceNetworkVpcAssociationIdentifier
)
}
\arguments{
\item{serviceNetworkVpcAssociationIdentifier}{[required] The ID or ARN of the association.}
}
\description{
Retrieves information about the specified association between a service network and a VPC.

See \url{https://www.paws-r-sdk.com/docs/vpclattice_get_service_network_vpc_association/} for full documentation.
}
\keyword{internal}
