% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpclattice_operations.R
\name{vpclattice_get_target_group}
\alias{vpclattice_get_target_group}
\title{Retrieves information about the specified target group}
\usage{
vpclattice_get_target_group(targetGroupIdentifier)
}
\arguments{
\item{targetGroupIdentifier}{[required] The ID or ARN of the target group.}
}
\description{
Retrieves information about the specified target group.

See \url{https://www.paws-r-sdk.com/docs/vpclattice_get_target_group/} for full documentation.
}
\keyword{internal}
