% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpclattice_operations.R
\name{vpclattice_list_access_log_subscriptions}
\alias{vpclattice_list_access_log_subscriptions}
\title{Lists the access log subscriptions for the specified service network or
service}
\usage{
vpclattice_list_access_log_subscriptions(
  resourceIdentifier,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{resourceIdentifier}{[required] The ID or ARN of the service network or service.}

\item{maxResults}{The maximum number of results to return.}

\item{nextToken}{A pagination token for the next page of results.}
}
\description{
Lists the access log subscriptions for the specified service network or service.

See \url{https://www.paws-r-sdk.com/docs/vpclattice_list_access_log_subscriptions/} for full documentation.
}
\keyword{internal}
