% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpclattice_operations.R
\name{vpclattice_list_domain_verifications}
\alias{vpclattice_list_domain_verifications}
\title{Lists the domain verifications}
\usage{
vpclattice_list_domain_verifications(maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{maxResults}{The maximum number of results to return.}

\item{nextToken}{A pagination token for the next page of results.}
}
\description{
Lists the domain verifications.

See \url{https://www.paws-r-sdk.com/docs/vpclattice_list_domain_verifications/} for full documentation.
}
\keyword{internal}
