% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpclattice_operations.R
\name{vpclattice_list_resource_configurations}
\alias{vpclattice_list_resource_configurations}
\title{Lists the resource configurations owned by or shared with this account}
\usage{
vpclattice_list_resource_configurations(
  maxResults = NULL,
  nextToken = NULL,
  resourceConfigurationGroupIdentifier = NULL,
  resourceGatewayIdentifier = NULL
)
}
\arguments{
\item{maxResults}{The maximum page size.}

\item{nextToken}{A pagination token for the next page of results.}

\item{resourceConfigurationGroupIdentifier}{The ID of the group resource configuration.}

\item{resourceGatewayIdentifier}{The ID of the resource gateway for the resource configuration.}
}
\description{
Lists the resource configurations owned by or shared with this account.

See \url{https://www.paws-r-sdk.com/docs/vpclattice_list_resource_configurations/} for full documentation.
}
\keyword{internal}
