% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpclattice_operations.R
\name{vpclattice_list_resource_endpoint_associations}
\alias{vpclattice_list_resource_endpoint_associations}
\title{Lists the associations for the specified VPC endpoint}
\usage{
vpclattice_list_resource_endpoint_associations(
  resourceConfigurationIdentifier,
  resourceEndpointAssociationIdentifier = NULL,
  vpcEndpointId = NULL,
  vpcEndpointOwner = NULL,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{resourceConfigurationIdentifier}{[required] The ID for the resource configuration associated with the VPC endpoint.}

\item{resourceEndpointAssociationIdentifier}{The ID of the association.}

\item{vpcEndpointId}{The ID of the VPC endpoint in the association.}

\item{vpcEndpointOwner}{The owner of the VPC endpoint in the association.}

\item{maxResults}{The maximum page size.}

\item{nextToken}{A pagination token for the next page of results.}
}
\description{
Lists the associations for the specified VPC endpoint.

See \url{https://www.paws-r-sdk.com/docs/vpclattice_list_resource_endpoint_associations/} for full documentation.
}
\keyword{internal}
