% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpclattice_operations.R
\name{vpclattice_list_resource_gateways}
\alias{vpclattice_list_resource_gateways}
\title{Lists the resource gateways that you own or that were shared with you}
\usage{
vpclattice_list_resource_gateways(maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{maxResults}{The maximum page size.}

\item{nextToken}{If there are additional results, a pagination token for the next page of
results.}
}
\description{
Lists the resource gateways that you own or that were shared with you.

See \url{https://www.paws-r-sdk.com/docs/vpclattice_list_resource_gateways/} for full documentation.
}
\keyword{internal}
