% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpclattice_operations.R
\name{vpclattice_list_service_network_resource_associations}
\alias{vpclattice_list_service_network_resource_associations}
\title{Lists the associations between a service network and a resource
configuration}
\usage{
vpclattice_list_service_network_resource_associations(
  serviceNetworkIdentifier = NULL,
  resourceConfigurationIdentifier = NULL,
  maxResults = NULL,
  nextToken = NULL,
  includeChildren = NULL
)
}
\arguments{
\item{serviceNetworkIdentifier}{The ID of the service network.}

\item{resourceConfigurationIdentifier}{The ID of the resource configuration.}

\item{maxResults}{The maximum page size.}

\item{nextToken}{If there are additional results, a pagination token for the next page of
results.}

\item{includeChildren}{Include service network resource associations of the child resource
configuration with the grouped resource configuration.

The type is boolean and the default value is false.}
}
\description{
Lists the associations between a service network and a resource configuration.

See \url{https://www.paws-r-sdk.com/docs/vpclattice_list_service_network_resource_associations/} for full documentation.
}
\keyword{internal}
