% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpclattice_operations.R
\name{vpclattice_list_service_network_vpc_associations}
\alias{vpclattice_list_service_network_vpc_associations}
\title{Lists the service network and VPC associations}
\usage{
vpclattice_list_service_network_vpc_associations(
  maxResults = NULL,
  nextToken = NULL,
  serviceNetworkIdentifier = NULL,
  vpcIdentifier = NULL
)
}
\arguments{
\item{maxResults}{The maximum number of results to return.}

\item{nextToken}{A pagination token for the next page of results.}

\item{serviceNetworkIdentifier}{The ID or Amazon Resource Name (ARN) of the service network.}

\item{vpcIdentifier}{The ID or Amazon Resource Name (ARN) of the VPC.}
}
\description{
Lists the service network and VPC associations. You can filter the list either by VPC or service network. You must provide either the service network identifier or the VPC identifier.

See \url{https://www.paws-r-sdk.com/docs/vpclattice_list_service_network_vpc_associations/} for full documentation.
}
\keyword{internal}
