% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpclattice_operations.R
\name{vpclattice_list_services}
\alias{vpclattice_list_services}
\title{Lists the services owned by the caller account or shared with the caller
account}
\usage{
vpclattice_list_services(maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{maxResults}{The maximum number of results to return.}

\item{nextToken}{A pagination token for the next page of results.}
}
\description{
Lists the services owned by the caller account or shared with the caller account.

See \url{https://www.paws-r-sdk.com/docs/vpclattice_list_services/} for full documentation.
}
\keyword{internal}
