% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpclattice_operations.R
\name{vpclattice_list_target_groups}
\alias{vpclattice_list_target_groups}
\title{Lists your target groups}
\usage{
vpclattice_list_target_groups(
  maxResults = NULL,
  nextToken = NULL,
  vpcIdentifier = NULL,
  targetGroupType = NULL
)
}
\arguments{
\item{maxResults}{The maximum number of results to return.}

\item{nextToken}{A pagination token for the next page of results.}

\item{vpcIdentifier}{The ID or ARN of the VPC.}

\item{targetGroupType}{The target group type.}
}
\description{
Lists your target groups. You can narrow your search by using the filters below in your request.

See \url{https://www.paws-r-sdk.com/docs/vpclattice_list_target_groups/} for full documentation.
}
\keyword{internal}
