% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpclattice_operations.R
\name{vpclattice_put_auth_policy}
\alias{vpclattice_put_auth_policy}
\title{Creates or updates the auth policy}
\usage{
vpclattice_put_auth_policy(policy, resourceIdentifier)
}
\arguments{
\item{policy}{[required] The auth policy.}

\item{resourceIdentifier}{[required] The ID or Amazon Resource Name (ARN) of the service network or service
for which the policy is created.}
}
\description{
Creates or updates the auth policy.

See \url{https://www.paws-r-sdk.com/docs/vpclattice_put_auth_policy/} for full documentation.
}
\keyword{internal}
