% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpclattice_operations.R
\name{vpclattice_put_resource_policy}
\alias{vpclattice_put_resource_policy}
\title{Attaches a resource-based permission policy to a service or service
network}
\usage{
vpclattice_put_resource_policy(policy, resourceArn)
}
\arguments{
\item{policy}{[required] An IAM policy.}

\item{resourceArn}{[required] The ID or Amazon Resource Name (ARN) of the service network or service
for which the policy is created.}
}
\description{
Attaches a resource-based permission policy to a service or service network. The policy must contain the same actions and condition statements as the Amazon Web Services Resource Access Manager permission for sharing services and service networks.

See \url{https://www.paws-r-sdk.com/docs/vpclattice_put_resource_policy/} for full documentation.
}
\keyword{internal}
