% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpclattice_operations.R
\name{vpclattice_register_targets}
\alias{vpclattice_register_targets}
\title{Registers the targets with the target group}
\usage{
vpclattice_register_targets(targetGroupIdentifier, targets)
}
\arguments{
\item{targetGroupIdentifier}{[required] The ID or ARN of the target group.}

\item{targets}{[required] The targets.}
}
\description{
Registers the targets with the target group. If it's a Lambda target, you can only have one target in a target group.

See \url{https://www.paws-r-sdk.com/docs/vpclattice_register_targets/} for full documentation.
}
\keyword{internal}
