% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpclattice_operations.R
\name{vpclattice_start_domain_verification}
\alias{vpclattice_start_domain_verification}
\title{Starts the domain verification process for a custom domain name}
\usage{
vpclattice_start_domain_verification(
  clientToken = NULL,
  domainName,
  tags = NULL
)
}
\arguments{
\item{clientToken}{A unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. If you retry a request that completed
successfully using the same client token and parameters, the retry
succeeds without performing any actions. If the parameters aren't
identical, the retry fails.}

\item{domainName}{[required] The domain name to verify ownership for.}

\item{tags}{The tags for the domain verification.}
}
\description{
Starts the domain verification process for a custom domain name.

See \url{https://www.paws-r-sdk.com/docs/vpclattice_start_domain_verification/} for full documentation.
}
\keyword{internal}
