% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpclattice_operations.R
\name{vpclattice_tag_resource}
\alias{vpclattice_tag_resource}
\title{Adds the specified tags to the specified resource}
\usage{
vpclattice_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource.}

\item{tags}{[required] The tags for the resource.}
}
\description{
Adds the specified tags to the specified resource.

See \url{https://www.paws-r-sdk.com/docs/vpclattice_tag_resource/} for full documentation.
}
\keyword{internal}
