% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpclattice_operations.R
\name{vpclattice_untag_resource}
\alias{vpclattice_untag_resource}
\title{Removes the specified tags from the specified resource}
\usage{
vpclattice_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource.}

\item{tagKeys}{[required] The tag keys of the tags to remove.}
}
\description{
Removes the specified tags from the specified resource.

See \url{https://www.paws-r-sdk.com/docs/vpclattice_untag_resource/} for full documentation.
}
\keyword{internal}
