% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpclattice_operations.R
\name{vpclattice_update_access_log_subscription}
\alias{vpclattice_update_access_log_subscription}
\title{Updates the specified access log subscription}
\usage{
vpclattice_update_access_log_subscription(
  accessLogSubscriptionIdentifier,
  destinationArn
)
}
\arguments{
\item{accessLogSubscriptionIdentifier}{[required] The ID or ARN of the access log subscription.}

\item{destinationArn}{[required] The Amazon Resource Name (ARN) of the access log destination.}
}
\description{
Updates the specified access log subscription.

See \url{https://www.paws-r-sdk.com/docs/vpclattice_update_access_log_subscription/} for full documentation.
}
\keyword{internal}
