% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpclattice_operations.R
\name{vpclattice_update_listener}
\alias{vpclattice_update_listener}
\title{Updates the specified listener for the specified service}
\usage{
vpclattice_update_listener(
  serviceIdentifier,
  listenerIdentifier,
  defaultAction
)
}
\arguments{
\item{serviceIdentifier}{[required] The ID or ARN of the service.}

\item{listenerIdentifier}{[required] The ID or ARN of the listener.}

\item{defaultAction}{[required] The action for the default rule.}
}
\description{
Updates the specified listener for the specified service.

See \url{https://www.paws-r-sdk.com/docs/vpclattice_update_listener/} for full documentation.
}
\keyword{internal}
