% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpclattice_operations.R
\name{vpclattice_update_resource_configuration}
\alias{vpclattice_update_resource_configuration}
\title{Updates the specified resource configuration}
\usage{
vpclattice_update_resource_configuration(
  allowAssociationToShareableServiceNetwork = NULL,
  portRanges = NULL,
  resourceConfigurationDefinition = NULL,
  resourceConfigurationIdentifier
)
}
\arguments{
\item{allowAssociationToShareableServiceNetwork}{Indicates whether to add the resource configuration to service networks
that are shared with other accounts.}

\item{portRanges}{The TCP port ranges that a consumer can use to access a resource
configuration. You can separate port ranges with a comma. Example:
1-65535 or 1,2,22-30}

\item{resourceConfigurationDefinition}{The resource configuration.}

\item{resourceConfigurationIdentifier}{[required] The ID of the resource configuration.}
}
\description{
Updates the specified resource configuration.

See \url{https://www.paws-r-sdk.com/docs/vpclattice_update_resource_configuration/} for full documentation.
}
\keyword{internal}
