% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpclattice_operations.R
\name{vpclattice_update_resource_gateway}
\alias{vpclattice_update_resource_gateway}
\title{Updates the specified resource gateway}
\usage{
vpclattice_update_resource_gateway(
  resourceGatewayIdentifier,
  securityGroupIds = NULL
)
}
\arguments{
\item{resourceGatewayIdentifier}{[required] The ID or ARN of the resource gateway.}

\item{securityGroupIds}{The IDs of the security groups associated with the resource gateway.}
}
\description{
Updates the specified resource gateway.

See \url{https://www.paws-r-sdk.com/docs/vpclattice_update_resource_gateway/} for full documentation.
}
\keyword{internal}
