% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpclattice_operations.R
\name{vpclattice_update_service_network}
\alias{vpclattice_update_service_network}
\title{Updates the specified service network}
\usage{
vpclattice_update_service_network(serviceNetworkIdentifier, authType)
}
\arguments{
\item{serviceNetworkIdentifier}{[required] The ID or ARN of the service network.}

\item{authType}{[required] The type of IAM policy.
\itemize{
\item \code{NONE}: The resource does not use an IAM policy. This is the
default.
\item \code{AWS_IAM}: The resource uses an IAM policy. When this type is used,
auth is enabled and an auth policy is required.
}}
}
\description{
Updates the specified service network.

See \url{https://www.paws-r-sdk.com/docs/vpclattice_update_service_network/} for full documentation.
}
\keyword{internal}
