% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_create_documentation_part}
\alias{apigateway_create_documentation_part}
\title{Creates a documentation part}
\usage{
apigateway_create_documentation_part(restApiId, location, properties)
}
\arguments{
\item{restApiId}{[required] The string identifier of the associated RestApi.}

\item{location}{[required] The location of the targeted API entity of the to-be-created
documentation part.}

\item{properties}{[required] The new documentation content map of the targeted API entity. Enclosed
key-value pairs are API-specific, but only OpenAPI-compliant key-value
pairs can be exported and, hence, published.}
}
\description{
Creates a documentation part.

See \url{https://www.paws-r-sdk.com/docs/apigateway_create_documentation_part/} for full documentation.
}
\keyword{internal}
