% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_create_documentation_version}
\alias{apigateway_create_documentation_version}
\title{Creates a documentation version}
\usage{
apigateway_create_documentation_version(
  restApiId,
  documentationVersion,
  stageName = NULL,
  description = NULL
)
}
\arguments{
\item{restApiId}{[required] The string identifier of the associated RestApi.}

\item{documentationVersion}{[required] The version identifier of the new snapshot.}

\item{stageName}{The stage name to be associated with the new documentation snapshot.}

\item{description}{A description about the new documentation snapshot.}
}
\description{
Creates a documentation version

See \url{https://paws-r.github.io/docs/apigateway/create_documentation_version.html} for full documentation.
}
\keyword{internal}
