% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_create_model}
\alias{apigateway_create_model}
\title{Adds a new Model resource to an existing RestApi resource}
\usage{
apigateway_create_model(
  restApiId,
  name,
  description = NULL,
  schema = NULL,
  contentType
)
}
\arguments{
\item{restApiId}{[required] The RestApi identifier under which the Model will be created.}

\item{name}{[required] The name of the model. Must be alphanumeric.}

\item{description}{The description of the model.}

\item{schema}{The schema for the model. For \code{application/json} models, this should be
JSON schema draft 4 model.}

\item{contentType}{[required] The content-type for the model.}
}
\description{
Adds a new Model resource to an existing RestApi resource.

See \url{https://paws-r.github.io/docs/apigateway/create_model.html} for full documentation.
}
\keyword{internal}
