% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_create_request_validator}
\alias{apigateway_create_request_validator}
\title{Creates a ReqeustValidator of a given RestApi}
\usage{
apigateway_create_request_validator(restApiId, name,
  validateRequestBody, validateRequestParameters)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated RestApi.}

\item{name}{The name of the to-be-created RequestValidator.}

\item{validateRequestBody}{A Boolean flag to indicate whether to validate request body according to
the configured model schema for the method (\code{true}) or not (\code{false}).}

\item{validateRequestParameters}{A Boolean flag to indicate whether to validate request parameters,
\code{true}, or not \code{false}.}
}
\value{
A list with the following syntax:\preformatted{list(
  id = "string",
  name = "string",
  validateRequestBody = TRUE|FALSE,
  validateRequestParameters = TRUE|FALSE
)
}
}
\description{
Creates a ReqeustValidator of a given RestApi.
}
\section{Request syntax}{
\preformatted{svc$create_request_validator(
  restApiId = "string",
  name = "string",
  validateRequestBody = TRUE|FALSE,
  validateRequestParameters = TRUE|FALSE
)
}
}

\keyword{internal}
