% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_create_request_validator}
\alias{apigateway_create_request_validator}
\title{Creates a RequestValidator of a given RestApi}
\usage{
apigateway_create_request_validator(
  restApiId,
  name = NULL,
  validateRequestBody = NULL,
  validateRequestParameters = NULL
)
}
\arguments{
\item{restApiId}{[required] The string identifier of the associated RestApi.}

\item{name}{The name of the to-be-created RequestValidator.}

\item{validateRequestBody}{A Boolean flag to indicate whether to validate request body according to
the configured model schema for the method (\code{true}) or not (\code{false}).}

\item{validateRequestParameters}{A Boolean flag to indicate whether to validate request parameters,
\code{true}, or not \code{false}.}
}
\description{
Creates a RequestValidator of a given RestApi.

See \url{https://paws-r.github.io/docs/apigateway/create_request_validator.html} for full documentation.
}
\keyword{internal}
