% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_create_resource}
\alias{apigateway_create_resource}
\title{Creates a Resource resource}
\usage{
apigateway_create_resource(restApiId, parentId, pathPart)
}
\arguments{
\item{restApiId}{[required] The string identifier of the associated RestApi.}

\item{parentId}{[required] The parent resource's identifier.}

\item{pathPart}{[required] The last path segment for this resource.}
}
\description{
Creates a Resource resource.

See \url{https://www.paws-r-sdk.com/docs/apigateway_create_resource/} for full documentation.
}
\keyword{internal}
