% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_delete_authorizer}
\alias{apigateway_delete_authorizer}
\title{Deletes an existing Authorizer resource}
\usage{
apigateway_delete_authorizer(restApiId, authorizerId)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated RestApi.}

\item{authorizerId}{[required] [Required] The identifier of the Authorizer resource.}
}
\description{
Deletes an existing Authorizer resource.
}
\details{
::: \{.seeAlso\}
\href{https://docs.aws.amazon.com/cli/latest/reference/apigateway/delete-authorizer.html}{AWS CLI}
:::
}
\section{Request syntax}{
\preformatted{svc$delete_authorizer(
  restApiId = "string",
  authorizerId = "string"
)
}
}

\keyword{internal}
