% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_delete_authorizer}
\alias{apigateway_delete_authorizer}
\title{Deletes an existing Authorizer resource}
\usage{
apigateway_delete_authorizer(restApiId, authorizerId)
}
\arguments{
\item{restApiId}{[required] The string identifier of the associated RestApi.}

\item{authorizerId}{[required] The identifier of the Authorizer resource.}
}
\description{
Deletes an existing Authorizer resource.

See \url{https://paws-r.github.io/docs/apigateway/delete_authorizer.html} for full documentation.
}
\keyword{internal}
