% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_delete_base_path_mapping}
\alias{apigateway_delete_base_path_mapping}
\title{Deletes the BasePathMapping resource}
\usage{
apigateway_delete_base_path_mapping(domainName, basePath)
}
\arguments{
\item{domainName}{[required] [Required] The domain name of the BasePathMapping resource to delete.}

\item{basePath}{[required] [Required] The base path name of the BasePathMapping resource to
delete.

To specify an empty base path, set this parameter to \code{'(none)'}.}
}
\value{
An empty list.
}
\description{
Deletes the BasePathMapping resource.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_base_path_mapping(
  domainName = "string",
  basePath = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
