% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_delete_deployment}
\alias{apigateway_delete_deployment}
\title{Deletes a Deployment resource}
\usage{
apigateway_delete_deployment(restApiId, deploymentId)
}
\arguments{
\item{restApiId}{[required] The string identifier of the associated RestApi.}

\item{deploymentId}{[required] The identifier of the Deployment resource to delete.}
}
\description{
Deletes a Deployment resource. Deleting a deployment will only succeed if there are no Stage resources associated with it.

See \url{https://paws-r.github.io/docs/apigateway/delete_deployment.html} for full documentation.
}
\keyword{internal}
