% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_delete_domain_name}
\alias{apigateway_delete_domain_name}
\title{Deletes the DomainName resource}
\usage{
apigateway_delete_domain_name(domainName)
}
\arguments{
\item{domainName}{[required] [Required] The name of the DomainName resource to be deleted.}
}
\value{
An empty list.
}
\description{
Deletes the DomainName resource.
}
\section{Request syntax}{
\preformatted{svc$delete_domain_name(
  domainName = "string"
)
}
}

\keyword{internal}
