% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_delete_gateway_response}
\alias{apigateway_delete_gateway_response}
\title{Clears any customization of a GatewayResponse of a specified response
type on the given RestApi and resets it with the default settings}
\usage{
apigateway_delete_gateway_response(restApiId, responseType)
}
\arguments{
\item{restApiId}{[required] The string identifier of the associated RestApi.}

\item{responseType}{[required] The response type of the associated GatewayResponse.}
}
\description{
Clears any customization of a GatewayResponse of a specified response type on the given RestApi and resets it with the default settings.

See \url{https://paws-r.github.io/docs/apigateway/delete_gateway_response.html} for full documentation.
}
\keyword{internal}
