% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_delete_integration}
\alias{apigateway_delete_integration}
\title{Represents a delete integration}
\usage{
apigateway_delete_integration(restApiId, resourceId, httpMethod)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated RestApi.}

\item{resourceId}{[required] [Required] Specifies a delete integration request's resource
identifier.}

\item{httpMethod}{[required] [Required] Specifies a delete integration request's HTTP method.}
}
\value{
An empty list.
}
\description{
Represents a delete integration.
}
\section{Request syntax}{
\preformatted{svc$delete_integration(
  restApiId = "string",
  resourceId = "string",
  httpMethod = "string"
)
}
}

\keyword{internal}
