% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_delete_integration_response}
\alias{apigateway_delete_integration_response}
\title{Represents a delete integration response}
\usage{
apigateway_delete_integration_response(
  restApiId,
  resourceId,
  httpMethod,
  statusCode
)
}
\arguments{
\item{restApiId}{[required] The string identifier of the associated RestApi.}

\item{resourceId}{[required] Specifies a delete integration response request's resource identifier.}

\item{httpMethod}{[required] Specifies a delete integration response request's HTTP method.}

\item{statusCode}{[required] Specifies a delete integration response request's status code.}
}
\description{
Represents a delete integration response.

See \url{https://paws-r.github.io/docs/apigateway/delete_integration_response.html} for full documentation.
}
\keyword{internal}
