% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_delete_request_validator}
\alias{apigateway_delete_request_validator}
\title{Deletes a RequestValidator of a given RestApi}
\usage{
apigateway_delete_request_validator(restApiId, requestValidatorId)
}
\arguments{
\item{restApiId}{[required] The string identifier of the associated RestApi.}

\item{requestValidatorId}{[required] The identifier of the RequestValidator to be deleted.}
}
\description{
Deletes a RequestValidator of a given RestApi.

See \url{https://paws-r.github.io/docs/apigateway/delete_request_validator.html} for full documentation.
}
\keyword{internal}
