% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_delete_rest_api}
\alias{apigateway_delete_rest_api}
\title{Deletes the specified API}
\usage{
apigateway_delete_rest_api(restApiId)
}
\arguments{
\item{restApiId}{[required] The string identifier of the associated RestApi.}
}
\description{
Deletes the specified API.

See \url{https://paws-r.github.io/docs/apigateway/delete_rest_api.html} for full documentation.
}
\keyword{internal}
