% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_delete_rest_api}
\alias{apigateway_delete_rest_api}
\title{Deletes the specified API}
\usage{
apigateway_delete_rest_api(restApiId)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated RestApi.}
}
\value{
An empty list.
}
\description{
Deletes the specified API.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_rest_api(
  restApiId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
