% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_delete_stage}
\alias{apigateway_delete_stage}
\title{Deletes a Stage resource}
\usage{
apigateway_delete_stage(restApiId, stageName)
}
\arguments{
\item{restApiId}{[required] The string identifier of the associated RestApi.}

\item{stageName}{[required] The name of the Stage resource to delete.}
}
\description{
Deletes a Stage resource.

See \url{https://paws-r.github.io/docs/apigateway/delete_stage.html} for full documentation.
}
\keyword{internal}
