% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_flush_stage_cache}
\alias{apigateway_flush_stage_cache}
\title{Flushes a stage's cache}
\usage{
apigateway_flush_stage_cache(restApiId, stageName)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated RestApi.}

\item{stageName}{[required] [Required] The name of the stage to flush its cache.}
}
\description{
Flushes a stage's cache.
}
\section{Request syntax}{
\preformatted{svc$flush_stage_cache(
  restApiId = "string",
  stageName = "string"
)
}
}

\keyword{internal}
