% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_generate_client_certificate}
\alias{apigateway_generate_client_certificate}
\title{Generates a ClientCertificate resource}
\usage{
apigateway_generate_client_certificate(description)
}
\arguments{
\item{description}{The description of the ClientCertificate.}
}
\description{
Generates a ClientCertificate resource.
}
\section{Request syntax}{
\preformatted{svc$generate_client_certificate(
  description = "string"
)
}
}

\keyword{internal}
