% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_account}
\alias{apigateway_get_account}
\title{Gets information about the current Account resource}
\usage{
apigateway_get_account()
}
\value{
A list with the following syntax:\preformatted{list(
  cloudwatchRoleArn = "string",
  throttleSettings = list(
    burstLimit = 123,
    rateLimit = 123.0
  ),
  features = list(
    "string"
  ),
  apiKeyVersion = "string"
)
}
}
\description{
Gets information about the current Account resource.
}
\section{Request syntax}{
\preformatted{svc$get_account()
}
}

\keyword{internal}
