% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_account}
\alias{apigateway_get_account}
\title{Gets information about the current Account resource}
\usage{
apigateway_get_account()
}
\description{
Gets information about the current Account resource.

See \url{https://www.paws-r-sdk.com/docs/apigateway_get_account/} for full documentation.
}
\keyword{internal}
