% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_authorizer}
\alias{apigateway_get_authorizer}
\title{Describe an existing Authorizer resource}
\usage{
apigateway_get_authorizer(restApiId, authorizerId)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated RestApi.}

\item{authorizerId}{[required] [Required] The identifier of the Authorizer resource.}
}
\value{
A list with the following syntax:\preformatted{list(
  id = "string",
  name = "string",
  type = "TOKEN"|"REQUEST"|"COGNITO_USER_POOLS",
  providerARNs = list(
    "string"
  ),
  authType = "string",
  authorizerUri = "string",
  authorizerCredentials = "string",
  identitySource = "string",
  identityValidationExpression = "string",
  authorizerResultTtlInSeconds = 123
)
}
}
\description{
Describe an existing Authorizer resource.<div class="seeAlso">


\href{https://docs.aws.amazon.com/cli/latest/reference/apigateway/get-authorizer.html}{AWS CLI}</div>
}
\section{Request syntax}{
\preformatted{svc$get_authorizer(
  restApiId = "string",
  authorizerId = "string"
)
}
}

\keyword{internal}
