% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_authorizer}
\alias{apigateway_get_authorizer}
\title{Describe an existing Authorizer resource}
\usage{
apigateway_get_authorizer(restApiId, authorizerId)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated RestApi.}

\item{authorizerId}{[required] [Required] The identifier of the Authorizer resource.}
}
\description{
Describe an existing Authorizer resource.
}
\details{
::: {.seeAlso}
\href{https://docs.aws.amazon.com/cli/latest/reference/apigateway/get-authorizer.html}{AWS CLI}
:::
}
\section{Request syntax}{
\preformatted{svc$get_authorizer(
  restApiId = "string",
  authorizerId = "string"
)
}
}

\keyword{internal}
