% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_base_path_mappings}
\alias{apigateway_get_base_path_mappings}
\title{Represents a collection of BasePathMapping resources}
\usage{
apigateway_get_base_path_mappings(domainName, position = NULL, limit = NULL)
}
\arguments{
\item{domainName}{[required] The domain name of a BasePathMapping resource.}

\item{position}{The current pagination position in the paged result set.}

\item{limit}{The maximum number of returned results per page. The default value is 25
and the maximum value is 500.}
}
\description{
Represents a collection of BasePathMapping resources.

See \url{https://www.paws-r-sdk.com/docs/apigateway_get_base_path_mappings/} for full documentation.
}
\keyword{internal}
