% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_client_certificate}
\alias{apigateway_get_client_certificate}
\title{Gets information about the current ClientCertificate resource}
\usage{
apigateway_get_client_certificate(clientCertificateId)
}
\arguments{
\item{clientCertificateId}{[required] [Required] The identifier of the ClientCertificate resource to be
described.}
}
\value{
A list with the following syntax:\preformatted{list(
  clientCertificateId = "string",
  description = "string",
  pemEncodedCertificate = "string",
  createdDate = as.POSIXct(
    "2015-01-01"
  ),
  expirationDate = as.POSIXct(
    "2015-01-01"
  ),
  tags = list(
    "string"
  )
)
}
}
\description{
Gets information about the current ClientCertificate resource.
}
\section{Request syntax}{
\preformatted{svc$get_client_certificate(
  clientCertificateId = "string"
)
}
}

\keyword{internal}
