% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_client_certificate}
\alias{apigateway_get_client_certificate}
\title{Gets information about the current ClientCertificate resource}
\usage{
apigateway_get_client_certificate(clientCertificateId)
}
\arguments{
\item{clientCertificateId}{[required] The identifier of the ClientCertificate resource to be described.}
}
\description{
Gets information about the current ClientCertificate resource.

See \url{https://www.paws-r-sdk.com/docs/apigateway_get_client_certificate/} for full documentation.
}
\keyword{internal}
