% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_deployments}
\alias{apigateway_get_deployments}
\title{Gets information about a Deployments collection}
\usage{
apigateway_get_deployments(restApiId, position, limit)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated RestApi.}

\item{position}{The current pagination position in the paged result set.}

\item{limit}{The maximum number of returned results per page. The default value is 25
and the maximum value is 500.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  position = "string",
  items = list(
    list(
      id = "string",
      description = "string",
      createdDate = as.POSIXct(
        "2015-01-01"
      ),
      apiSummary = list(
        list(
          list(
            authorizationType = "string",
            apiKeyRequired = TRUE|FALSE
          )
        )
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Gets information about a Deployments collection.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_deployments(
  restApiId = "string",
  position = "string",
  limit = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
