% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_documentation_part}
\alias{apigateway_get_documentation_part}
\title{Get documentation part}
\usage{
apigateway_get_documentation_part(restApiId, documentationPartId)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated RestApi.}

\item{documentationPartId}{[required] [Required] The string identifier of the associated RestApi.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  id = "string",
  location = list(
    type = "API"|"AUTHORIZER"|"MODEL"|"RESOURCE"|"METHOD"|"PATH_PARAMETER"|"QUERY_PARAMETER"|"REQUEST_HEADER"|"REQUEST_BODY"|"RESPONSE"|"RESPONSE_HEADER"|"RESPONSE_BODY",
    path = "string",
    method = "string",
    statusCode = "string",
    name = "string"
  ),
  properties = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Get documentation part
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_documentation_part(
  restApiId = "string",
  documentationPartId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
