% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_documentation_part}
\alias{apigateway_get_documentation_part}
\title{Gets a documentation part}
\usage{
apigateway_get_documentation_part(restApiId, documentationPartId)
}
\arguments{
\item{restApiId}{[required] The string identifier of the associated RestApi.}

\item{documentationPartId}{[required] The string identifier of the associated RestApi.}
}
\description{
Gets a documentation part.

See \url{https://paws-r.github.io/docs/apigateway/get_documentation_part.html} for full documentation.
}
\keyword{internal}
