% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_documentation_part}
\alias{apigateway_get_documentation_part}
\title{Get documentation part}
\usage{
apigateway_get_documentation_part(restApiId, documentationPartId)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated RestApi.}

\item{documentationPartId}{[required] [Required] The string identifier of the associated RestApi.}
}
\description{
Get documentation part
}
\section{Request syntax}{
\preformatted{svc$get_documentation_part(
  restApiId = "string",
  documentationPartId = "string"
)
}
}

\keyword{internal}
