% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_domain_name}
\alias{apigateway_get_domain_name}
\title{Represents a domain name that is contained in a simpler, more intuitive
URL that can be called}
\usage{
apigateway_get_domain_name(domainName, domainNameId = NULL)
}
\arguments{
\item{domainName}{[required] The name of the DomainName resource.}

\item{domainNameId}{The identifier for the domain name resource. Supported only for private
custom domain names.}
}
\description{
Represents a domain name that is contained in a simpler, more intuitive URL that can be called.

See \url{https://www.paws-r-sdk.com/docs/apigateway_get_domain_name/} for full documentation.
}
\keyword{internal}
