% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_domain_name}
\alias{apigateway_get_domain_name}
\title{Represents a domain name that is contained in a simpler, more intuitive
URL that can be called}
\usage{
apigateway_get_domain_name(domainName)
}
\arguments{
\item{domainName}{[required] The name of the DomainName resource.}
}
\description{
Represents a domain name that is contained in a simpler, more intuitive URL that can be called.

See \url{https://paws-r.github.io/docs/apigateway/get_domain_name.html} for full documentation.
}
\keyword{internal}
