% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_domain_names}
\alias{apigateway_get_domain_names}
\title{Represents a collection of DomainName resources}
\usage{
apigateway_get_domain_names(position = NULL, limit = NULL)
}
\arguments{
\item{position}{The current pagination position in the paged result set.}

\item{limit}{The maximum number of returned results per page. The default value is 25
and the maximum value is 500.}
}
\description{
Represents a collection of DomainName resources.

See \url{https://paws-r.github.io/docs/apigateway/get_domain_names.html} for full documentation.
}
\keyword{internal}
