% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_integration}
\alias{apigateway_get_integration}
\title{Get the integration settings}
\usage{
apigateway_get_integration(restApiId, resourceId, httpMethod)
}
\arguments{
\item{restApiId}{[required] The string identifier of the associated RestApi.}

\item{resourceId}{[required] Specifies a get integration request's resource identifier}

\item{httpMethod}{[required] Specifies a get integration request's HTTP method.}
}
\description{
Get the integration settings.

See \url{https://paws-r.github.io/docs/apigateway/get_integration.html} for full documentation.
}
\keyword{internal}
