% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_method_response}
\alias{apigateway_get_method_response}
\title{Describes a MethodResponse resource}
\usage{
apigateway_get_method_response(restApiId, resourceId, httpMethod,
  statusCode)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated RestApi.}

\item{resourceId}{[required] [Required] The Resource identifier for the MethodResponse resource.}

\item{httpMethod}{[required] [Required] The HTTP verb of the Method resource.}

\item{statusCode}{[required] [Required] The status code for the MethodResponse resource.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  statusCode = "string",
  responseParameters = list(
    TRUE|FALSE
  ),
  responseModels = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Describes a MethodResponse resource.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_method_response(
  restApiId = "string",
  resourceId = "string",
  httpMethod = "string",
  statusCode = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
