% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_method_response}
\alias{apigateway_get_method_response}
\title{Describes a MethodResponse resource}
\usage{
apigateway_get_method_response(restApiId, resourceId, httpMethod, statusCode)
}
\arguments{
\item{restApiId}{[required] The string identifier of the associated RestApi.}

\item{resourceId}{[required] The Resource identifier for the MethodResponse resource.}

\item{httpMethod}{[required] The HTTP verb of the Method resource.}

\item{statusCode}{[required] The status code for the MethodResponse resource.}
}
\description{
Describes a MethodResponse resource.

See \url{https://paws-r.github.io/docs/apigateway/get_method_response.html} for full documentation.
}
\keyword{internal}
