% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_models}
\alias{apigateway_get_models}
\title{Describes existing Models defined for a RestApi resource}
\usage{
apigateway_get_models(restApiId, position, limit)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated RestApi.}

\item{position}{The current pagination position in the paged result set.}

\item{limit}{The maximum number of returned results per page. The default value is 25
and the maximum value is 500.}
}
\description{
Describes existing Models defined for a RestApi resource.
}
\section{Request syntax}{
\preformatted{svc$get_models(
  restApiId = "string",
  position = "string",
  limit = 123
)
}
}

\keyword{internal}
