% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_rest_api}
\alias{apigateway_get_rest_api}
\title{Lists the RestApi resource in the collection}
\usage{
apigateway_get_rest_api(restApiId)
}
\arguments{
\item{restApiId}{[required] The string identifier of the associated RestApi.}
}
\description{
Lists the RestApi resource in the collection.

See \url{https://www.paws-r-sdk.com/docs/apigateway_get_rest_api/} for full documentation.
}
\keyword{internal}
