% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_sdk_type}
\alias{apigateway_get_sdk_type}
\title{Gets an SDK type}
\usage{
apigateway_get_sdk_type(id)
}
\arguments{
\item{id}{[required] The identifier of the queried SdkType instance.}
}
\description{
Gets an SDK type.

See \url{https://paws-r.github.io/docs/apigateway/get_sdk_type.html} for full documentation.
}
\keyword{internal}
