% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_sdk_type}
\alias{apigateway_get_sdk_type}
\title{Get sdk type}
\usage{
apigateway_get_sdk_type(id)
}
\arguments{
\item{id}{[required] [Required] The identifier of the queried SdkType instance.}
}
\value{
A list with the following syntax:\preformatted{list(
  id = "string",
  friendlyName = "string",
  description = "string",
  configurationProperties = list(
    list(
      name = "string",
      friendlyName = "string",
      description = "string",
      required = TRUE|FALSE,
      defaultValue = "string"
    )
  )
)
}
}
\description{
Get sdk type
}
\section{Request syntax}{
\preformatted{svc$get_sdk_type(
  id = "string"
)
}
}

\keyword{internal}
