% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_stage}
\alias{apigateway_get_stage}
\title{Gets information about a Stage resource}
\usage{
apigateway_get_stage(restApiId, stageName)
}
\arguments{
\item{restApiId}{[required] The string identifier of the associated RestApi.}

\item{stageName}{[required] The name of the Stage resource to get information about.}
}
\description{
Gets information about a Stage resource.

See \url{https://www.paws-r-sdk.com/docs/apigateway_get_stage/} for full documentation.
}
\keyword{internal}
