% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_stages}
\alias{apigateway_get_stages}
\title{Gets information about one or more Stage resources}
\usage{
apigateway_get_stages(restApiId, deploymentId = NULL)
}
\arguments{
\item{restApiId}{[required] The string identifier of the associated RestApi.}

\item{deploymentId}{The stages' deployment identifiers.}
}
\description{
Gets information about one or more Stage resources.

See \url{https://paws-r.github.io/docs/apigateway/get_stages.html} for full documentation.
}
\keyword{internal}
