% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_tags}
\alias{apigateway_get_tags}
\title{Gets the Tags collection for a given resource}
\usage{
apigateway_get_tags(resourceArn, position, limit)
}
\arguments{
\item{resourceArn}{[required] [Required] The ARN of a resource that can be tagged.}

\item{position}{(Not currently supported) The current pagination position in the paged
result set.}

\item{limit}{(Not currently supported) The maximum number of returned results per
page. The default value is 25 and the maximum value is 500.}
}
\value{
A list with the following syntax:\preformatted{list(
  tags = list(
    "string"
  )
)
}
}
\description{
Gets the Tags collection for a given resource.
}
\section{Request syntax}{
\preformatted{svc$get_tags(
  resourceArn = "string",
  position = "string",
  limit = 123
)
}
}

\keyword{internal}
