% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_tags}
\alias{apigateway_get_tags}
\title{Gets the Tags collection for a given resource}
\usage{
apigateway_get_tags(resourceArn, position = NULL, limit = NULL)
}
\arguments{
\item{resourceArn}{[required] The ARN of a resource that can be tagged.}

\item{position}{(Not currently supported) The current pagination position in the paged
result set.}

\item{limit}{(Not currently supported) The maximum number of returned results per
page. The default value is 25 and the maximum value is 500.}
}
\description{
Gets the Tags collection for a given resource.

See \url{https://paws-r.github.io/docs/apigateway/get_tags.html} for full documentation.
}
\keyword{internal}
