% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_usage}
\alias{apigateway_get_usage}
\title{Gets the usage data of a usage plan in a specified time interval}
\usage{
apigateway_get_usage(usagePlanId, keyId, startDate, endDate, position,
  limit)
}
\arguments{
\item{usagePlanId}{[required] [Required] The Id of the usage plan associated with the usage data.}

\item{keyId}{The Id of the API key associated with the resultant usage data.}

\item{startDate}{[required] [Required] The starting date (e.g., 2016-01-01) of the usage data.}

\item{endDate}{[required] [Required] The ending date (e.g., 2016-12-31) of the usage data.}

\item{position}{The current pagination position in the paged result set.}

\item{limit}{The maximum number of returned results per page. The default value is 25
and the maximum value is 500.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  usagePlanId = "string",
  startDate = "string",
  endDate = "string",
  position = "string",
  items = list(
    list(
      list(
        123
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Gets the usage data of a usage plan in a specified time interval.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_usage(
  usagePlanId = "string",
  keyId = "string",
  startDate = "string",
  endDate = "string",
  position = "string",
  limit = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
