% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_usage_plan}
\alias{apigateway_get_usage_plan}
\title{Gets a usage plan of a given plan identifier}
\usage{
apigateway_get_usage_plan(usagePlanId)
}
\arguments{
\item{usagePlanId}{[required] [Required] The identifier of the UsagePlan resource to be retrieved.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  id = "string",
  name = "string",
  description = "string",
  apiStages = list(
    list(
      apiId = "string",
      stage = "string",
      throttle = list(
        list(
          burstLimit = 123,
          rateLimit = 123.0
        )
      )
    )
  ),
  throttle = list(
    burstLimit = 123,
    rateLimit = 123.0
  ),
  quota = list(
    limit = 123,
    offset = 123,
    period = "DAY"|"WEEK"|"MONTH"
  ),
  productCode = "string",
  tags = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Gets a usage plan of a given plan identifier.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_usage_plan(
  usagePlanId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
