% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_vpc_link}
\alias{apigateway_get_vpc_link}
\title{Gets a specified VPC link under the caller's account in a region}
\usage{
apigateway_get_vpc_link(vpcLinkId)
}
\arguments{
\item{vpcLinkId}{[required] [Required] The identifier of the VpcLink. It is used in an Integration
to reference this VpcLink.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  id = "string",
  name = "string",
  description = "string",
  targetArns = list(
    "string"
  ),
  status = "AVAILABLE"|"PENDING"|"DELETING"|"FAILED",
  statusMessage = "string",
  tags = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Gets a specified VPC link under the caller's account in a region.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_vpc_link(
  vpcLinkId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
