% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_untag_resource}
\alias{apigateway_untag_resource}
\title{Removes a tag from a given resource}
\usage{
apigateway_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The ARN of a resource that can be tagged.}

\item{tagKeys}{[required] The Tag keys to delete.}
}
\description{
Removes a tag from a given resource.

See \url{https://paws-r.github.io/docs/apigateway/untag_resource.html} for full documentation.
}
\keyword{internal}
