% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_update_account}
\alias{apigateway_update_account}
\title{Changes information about the current Account resource}
\usage{
apigateway_update_account(patchOperations = NULL)
}
\arguments{
\item{patchOperations}{For more information about supported patch operations, see \href{https://docs.aws.amazon.com/apigateway/latest/api/patch-operations.html}{Patch Operations}.}
}
\description{
Changes information about the current Account resource.

See \url{https://paws-r.github.io/docs/apigateway/update_account.html} for full documentation.
}
\keyword{internal}
