% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_update_api_key}
\alias{apigateway_update_api_key}
\title{Changes information about an ApiKey resource}
\usage{
apigateway_update_api_key(apiKey, patchOperations = NULL)
}
\arguments{
\item{apiKey}{[required] The identifier of the ApiKey resource to be updated.}

\item{patchOperations}{For more information about supported patch operations, see \href{https://docs.aws.amazon.com/apigateway/latest/api/patch-operations.html}{Patch Operations}.}
}
\description{
Changes information about an ApiKey resource.

See \url{https://www.paws-r-sdk.com/docs/apigateway_update_api_key/} for full documentation.
}
\keyword{internal}
