% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_update_authorizer}
\alias{apigateway_update_authorizer}
\title{Updates an existing Authorizer resource}
\usage{
apigateway_update_authorizer(restApiId, authorizerId, patchOperations = NULL)
}
\arguments{
\item{restApiId}{[required] The string identifier of the associated RestApi.}

\item{authorizerId}{[required] The identifier of the Authorizer resource.}

\item{patchOperations}{For more information about supported patch operations, see \href{https://docs.aws.amazon.com/apigateway/latest/api/patch-operations.html}{Patch Operations}.}
}
\description{
Updates an existing Authorizer resource.

See \url{https://www.paws-r-sdk.com/docs/apigateway_update_authorizer/} for full documentation.
}
\keyword{internal}
