% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_update_client_certificate}
\alias{apigateway_update_client_certificate}
\title{Changes information about an ClientCertificate resource}
\usage{
apigateway_update_client_certificate(
  clientCertificateId,
  patchOperations = NULL
)
}
\arguments{
\item{clientCertificateId}{[required] The identifier of the ClientCertificate resource to be updated.}

\item{patchOperations}{For more information about supported patch operations, see \href{https://docs.aws.amazon.com/apigateway/latest/api/patch-operations.html}{Patch Operations}.}
}
\description{
Changes information about an ClientCertificate resource.

See \url{https://www.paws-r-sdk.com/docs/apigateway_update_client_certificate/} for full documentation.
}
\keyword{internal}
