% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_update_deployment}
\alias{apigateway_update_deployment}
\title{Changes information about a Deployment resource}
\usage{
apigateway_update_deployment(restApiId, deploymentId, patchOperations = NULL)
}
\arguments{
\item{restApiId}{[required] The string identifier of the associated RestApi.}

\item{deploymentId}{[required] The replacement identifier for the Deployment resource to change
information about.}

\item{patchOperations}{For more information about supported patch operations, see \href{https://docs.aws.amazon.com/apigateway/latest/api/patch-operations.html}{Patch Operations}.}
}
\description{
Changes information about a Deployment resource.

See \url{https://www.paws-r-sdk.com/docs/apigateway_update_deployment/} for full documentation.
}
\keyword{internal}
