% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_update_documentation_part}
\alias{apigateway_update_documentation_part}
\title{Updates a documentation part}
\usage{
apigateway_update_documentation_part(
  restApiId,
  documentationPartId,
  patchOperations = NULL
)
}
\arguments{
\item{restApiId}{[required] The string identifier of the associated RestApi.}

\item{documentationPartId}{[required] The identifier of the to-be-updated documentation part.}

\item{patchOperations}{For more information about supported patch operations, see \href{https://docs.aws.amazon.com/apigateway/latest/api/patch-operations.html}{Patch Operations}.}
}
\description{
Updates a documentation part.

See \url{https://www.paws-r-sdk.com/docs/apigateway_update_documentation_part/} for full documentation.
}
\keyword{internal}
