% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_update_documentation_version}
\alias{apigateway_update_documentation_version}
\title{Updates a documentation version}
\usage{
apigateway_update_documentation_version(
  restApiId,
  documentationVersion,
  patchOperations = NULL
)
}
\arguments{
\item{restApiId}{[required] The string identifier of the associated RestApi..}

\item{documentationVersion}{[required] The version identifier of the to-be-updated documentation version.}

\item{patchOperations}{For more information about supported patch operations, see \href{https://docs.aws.amazon.com/apigateway/latest/api/patch-operations.html}{Patch Operations}.}
}
\description{
Updates a documentation version.

See \url{https://www.paws-r-sdk.com/docs/apigateway_update_documentation_version/} for full documentation.
}
\keyword{internal}
