% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_update_integration}
\alias{apigateway_update_integration}
\title{Represents an update integration}
\usage{
apigateway_update_integration(
  restApiId,
  resourceId,
  httpMethod,
  patchOperations = NULL
)
}
\arguments{
\item{restApiId}{[required] The string identifier of the associated RestApi.}

\item{resourceId}{[required] Represents an update integration request's resource identifier.}

\item{httpMethod}{[required] Represents an update integration request's HTTP method.}

\item{patchOperations}{For more information about supported patch operations, see \href{https://docs.aws.amazon.com/apigateway/latest/api/patch-operations.html}{Patch Operations}.}
}
\description{
Represents an update integration.

See \url{https://www.paws-r-sdk.com/docs/apigateway_update_integration/} for full documentation.
}
\keyword{internal}
