% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_update_model}
\alias{apigateway_update_model}
\title{Changes information about a model}
\usage{
apigateway_update_model(restApiId, modelName, patchOperations = NULL)
}
\arguments{
\item{restApiId}{[required] The string identifier of the associated RestApi.}

\item{modelName}{[required] The name of the model to update.}

\item{patchOperations}{For more information about supported patch operations, see \href{https://docs.aws.amazon.com/apigateway/latest/api/patch-operations.html}{Patch Operations}.}
}
\description{
Changes information about a model.

See \url{https://www.paws-r-sdk.com/docs/apigateway_update_model/} for full documentation.
}
\keyword{internal}
