% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_update_request_validator}
\alias{apigateway_update_request_validator}
\title{Updates a RequestValidator of a given RestApi}
\usage{
apigateway_update_request_validator(
  restApiId,
  requestValidatorId,
  patchOperations = NULL
)
}
\arguments{
\item{restApiId}{[required] The string identifier of the associated RestApi.}

\item{requestValidatorId}{[required] The identifier of RequestValidator to be updated.}

\item{patchOperations}{For more information about supported patch operations, see \href{https://docs.aws.amazon.com/apigateway/latest/api/patch-operations.html}{Patch Operations}.}
}
\description{
Updates a RequestValidator of a given RestApi.

See \url{https://paws-r.github.io/docs/apigateway/update_request_validator.html} for full documentation.
}
\keyword{internal}
