% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_update_rest_api}
\alias{apigateway_update_rest_api}
\title{Changes information about the specified API}
\usage{
apigateway_update_rest_api(restApiId, patchOperations = NULL)
}
\arguments{
\item{restApiId}{[required] The string identifier of the associated RestApi.}

\item{patchOperations}{For more information about supported patch operations, see \href{https://docs.aws.amazon.com/apigateway/latest/api/patch-operations.html}{Patch Operations}.}
}
\description{
Changes information about the specified API.

See \url{https://paws-r.github.io/docs/apigateway/update_rest_api.html} for full documentation.
}
\keyword{internal}
