% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_update_rest_api}
\alias{apigateway_update_rest_api}
\title{Changes information about the specified API}
\usage{
apigateway_update_rest_api(restApiId, patchOperations)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated RestApi.}

\item{patchOperations}{A list of update operations to be applied to the specified resource and
in the order specified in this list.}
}
\description{
Changes information about the specified API.
}
\section{Request syntax}{
\preformatted{svc$update_rest_api(
  restApiId = "string",
  patchOperations = list(
    list(
      op = "add"|"remove"|"replace"|"move"|"copy"|"test",
      path = "string",
      value = "string",
      from = "string"
    )
  )
)
}
}

\keyword{internal}
