% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_update_stage}
\alias{apigateway_update_stage}
\title{Changes information about a Stage resource}
\usage{
apigateway_update_stage(restApiId, stageName, patchOperations = NULL)
}
\arguments{
\item{restApiId}{[required] The string identifier of the associated RestApi.}

\item{stageName}{[required] The name of the Stage resource to change information about.}

\item{patchOperations}{For more information about supported patch operations, see \href{https://docs.aws.amazon.com/apigateway/latest/api/patch-operations.html}{Patch Operations}.}
}
\description{
Changes information about a Stage resource.

See \url{https://paws-r.github.io/docs/apigateway/update_stage.html} for full documentation.
}
\keyword{internal}
