% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_update_usage_plan}
\alias{apigateway_update_usage_plan}
\title{Updates a usage plan of a given plan Id}
\usage{
apigateway_update_usage_plan(usagePlanId, patchOperations)
}
\arguments{
\item{usagePlanId}{[required] [Required] The Id of the to-be-updated usage plan.}

\item{patchOperations}{A list of update operations to be applied to the specified resource and
in the order specified in this list.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  id = "string",
  name = "string",
  description = "string",
  apiStages = list(
    list(
      apiId = "string",
      stage = "string",
      throttle = list(
        list(
          burstLimit = 123,
          rateLimit = 123.0
        )
      )
    )
  ),
  throttle = list(
    burstLimit = 123,
    rateLimit = 123.0
  ),
  quota = list(
    limit = 123,
    offset = 123,
    period = "DAY"|"WEEK"|"MONTH"
  ),
  productCode = "string",
  tags = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Updates a usage plan of a given plan Id.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_usage_plan(
  usagePlanId = "string",
  patchOperations = list(
    list(
      op = "add"|"remove"|"replace"|"move"|"copy"|"test",
      path = "string",
      value = "string",
      from = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
