% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_update_vpc_link}
\alias{apigateway_update_vpc_link}
\title{Updates an existing VpcLink of a specified identifier}
\usage{
apigateway_update_vpc_link(vpcLinkId, patchOperations)
}
\arguments{
\item{vpcLinkId}{[required] [Required] The identifier of the VpcLink. It is used in an Integration
to reference this VpcLink.}

\item{patchOperations}{A list of update operations to be applied to the specified resource and
in the order specified in this list.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  id = "string",
  name = "string",
  description = "string",
  targetArns = list(
    "string"
  ),
  status = "AVAILABLE"|"PENDING"|"DELETING"|"FAILED",
  statusMessage = "string",
  tags = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Updates an existing VpcLink of a specified identifier.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_vpc_link(
  vpcLinkId = "string",
  patchOperations = list(
    list(
      op = "add"|"remove"|"replace"|"move"|"copy"|"test",
      path = "string",
      value = "string",
      from = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
