% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewaymanagementapi_operations.R
\name{apigatewaymanagementapi_delete_connection}
\alias{apigatewaymanagementapi_delete_connection}
\title{Delete the connection with the provided id}
\usage{
apigatewaymanagementapi_delete_connection(ConnectionId)
}
\arguments{
\item{ConnectionId}{[required]}
}
\description{
Delete the connection with the provided id.

See \url{https://www.paws-r-sdk.com/docs/apigatewaymanagementapi_delete_connection/} for full documentation.
}
\keyword{internal}
