% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewaymanagementapi_operations.R
\name{apigatewaymanagementapi_post_to_connection}
\alias{apigatewaymanagementapi_post_to_connection}
\title{Sends the provided data to the specified connection}
\usage{
apigatewaymanagementapi_post_to_connection(Data, ConnectionId)
}
\arguments{
\item{Data}{[required] The data to be sent to the client specified by its connection id.}

\item{ConnectionId}{[required] The identifier of the connection that a specific client is using.}
}
\value{
An empty list.
}
\description{
Sends the provided data to the specified connection.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$post_to_connection(
  Data = raw,
  ConnectionId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
