% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_create_api_mapping}
\alias{apigatewayv2_create_api_mapping}
\title{Creates an API mapping}
\usage{
apigatewayv2_create_api_mapping(ApiId, ApiMappingKey, DomainName, Stage)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{ApiMappingKey}{The API mapping key.}

\item{DomainName}{[required] The domain name.}

\item{Stage}{[required] The API stage.}
}
\value{
A list with the following syntax:\preformatted{list(
  ApiId = "string",
  ApiMappingId = "string",
  ApiMappingKey = "string",
  Stage = "string"
)
}
}
\description{
Creates an API mapping.
}
\section{Request syntax}{
\preformatted{svc$create_api_mapping(
  ApiId = "string",
  ApiMappingKey = "string",
  DomainName = "string",
  Stage = "string"
)
}
}

\keyword{internal}
