% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_create_deployment}
\alias{apigatewayv2_create_deployment}
\title{Creates a Deployment for an API}
\usage{
apigatewayv2_create_deployment(ApiId, Description, StageName)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{Description}{The description for the deployment resource.}

\item{StageName}{The name of the Stage resource for the Deployment resource to create.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  AutoDeployed = TRUE|FALSE,
  CreatedDate = as.POSIXct(
    "2015-01-01"
  ),
  DeploymentId = "string",
  DeploymentStatus = "PENDING"|"FAILED"|"DEPLOYED",
  DeploymentStatusMessage = "string",
  Description = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Creates a Deployment for an API.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_deployment(
  ApiId = "string",
  Description = "string",
  StageName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
