% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_create_deployment}
\alias{apigatewayv2_create_deployment}
\title{Creates a Deployment for an API}
\usage{
apigatewayv2_create_deployment(ApiId, Description = NULL, StageName = NULL)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{Description}{The description for the deployment resource.}

\item{StageName}{The name of the Stage resource for the Deployment resource to create.}
}
\description{
Creates a Deployment for an API.

See \url{https://www.paws-r-sdk.com/docs/apigatewayv2_create_deployment/} for full documentation.
}
\keyword{internal}
