% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_create_domain_name}
\alias{apigatewayv2_create_domain_name}
\title{Creates a domain name}
\usage{
apigatewayv2_create_domain_name(
  DomainName,
  DomainNameConfigurations = NULL,
  MutualTlsAuthentication = NULL,
  RoutingMode = NULL,
  Tags = NULL
)
}
\arguments{
\item{DomainName}{[required] The domain name.}

\item{DomainNameConfigurations}{The domain name configurations.}

\item{MutualTlsAuthentication}{The mutual TLS authentication configuration for a custom domain name.}

\item{RoutingMode}{The routing mode.}

\item{Tags}{The collection of tags associated with a domain name.}
}
\description{
Creates a domain name.

See \url{https://www.paws-r-sdk.com/docs/apigatewayv2_create_domain_name/} for full documentation.
}
\keyword{internal}
