% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_create_model}
\alias{apigatewayv2_create_model}
\title{Creates a Model for an API}
\usage{
apigatewayv2_create_model(
  ApiId,
  ContentType = NULL,
  Description = NULL,
  Name,
  Schema
)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{ContentType}{The content-type for the model, for example, "application/json".}

\item{Description}{The description of the model.}

\item{Name}{[required] The name of the model. Must be alphanumeric.}

\item{Schema}{[required] The schema for the model. For application/json models, this should be
JSON schema draft 4 model.}
}
\description{
Creates a Model for an API.

See \url{https://www.paws-r-sdk.com/docs/apigatewayv2_create_model/} for full documentation.
}
\keyword{internal}
