% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_create_route_response}
\alias{apigatewayv2_create_route_response}
\title{Creates a RouteResponse for a Route}
\usage{
apigatewayv2_create_route_response(ApiId, ModelSelectionExpression,
  ResponseModels, ResponseParameters, RouteId, RouteResponseKey)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{ModelSelectionExpression}{The model selection expression for the route response. Supported only
for WebSocket APIs.}

\item{ResponseModels}{The response models for the route response.}

\item{ResponseParameters}{The route response parameters.}

\item{RouteId}{[required] The route ID.}

\item{RouteResponseKey}{[required] The route response key.}
}
\value{
A list with the following syntax:\preformatted{list(
  ModelSelectionExpression = "string",
  ResponseModels = list(
    "string"
  ),
  ResponseParameters = list(
    list(
      Required = TRUE|FALSE
    )
  ),
  RouteResponseId = "string",
  RouteResponseKey = "string"
)
}
}
\description{
Creates a RouteResponse for a Route.
}
\section{Request syntax}{
\preformatted{svc$create_route_response(
  ApiId = "string",
  ModelSelectionExpression = "string",
  ResponseModels = list(
    "string"
  ),
  ResponseParameters = list(
    list(
      Required = TRUE|FALSE
    )
  ),
  RouteId = "string",
  RouteResponseKey = "string"
)
}
}

\keyword{internal}
