% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_create_stage}
\alias{apigatewayv2_create_stage}
\title{Creates a Stage for an API}
\usage{
apigatewayv2_create_stage(AccessLogSettings, ApiId, AutoDeploy,
  ClientCertificateId, DefaultRouteSettings, DeploymentId, Description,
  RouteSettings, StageName, StageVariables, Tags)
}
\arguments{
\item{AccessLogSettings}{Settings for logging access in this stage.}

\item{ApiId}{[required] The API identifier.}

\item{AutoDeploy}{Specifies whether updates to an API automatically trigger a new
deployment. The default value is false.}

\item{ClientCertificateId}{The identifier of a client certificate for a Stage. Supported only for
WebSocket APIs.}

\item{DefaultRouteSettings}{The default route settings for the stage.}

\item{DeploymentId}{The deployment identifier of the API stage.}

\item{Description}{The description for the API stage.}

\item{RouteSettings}{Route settings for the stage, by routeKey.}

\item{StageName}{[required] The name of the stage.}

\item{StageVariables}{A map that defines the stage variables for a Stage. Variable names can
have alphanumeric and underscore characters, and the values must match
[A-Za-z0-9-._~:/?#&=,]+.}

\item{Tags}{The collection of tags. Each tag element is associated with a given
resource.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  AccessLogSettings = list(
    DestinationArn = "string",
    Format = "string"
  ),
  ApiGatewayManaged = TRUE|FALSE,
  AutoDeploy = TRUE|FALSE,
  ClientCertificateId = "string",
  CreatedDate = as.POSIXct(
    "2015-01-01"
  ),
  DefaultRouteSettings = list(
    DataTraceEnabled = TRUE|FALSE,
    DetailedMetricsEnabled = TRUE|FALSE,
    LoggingLevel = "ERROR"|"INFO"|"OFF",
    ThrottlingBurstLimit = 123,
    ThrottlingRateLimit = 123.0
  ),
  DeploymentId = "string",
  Description = "string",
  LastDeploymentStatusMessage = "string",
  LastUpdatedDate = as.POSIXct(
    "2015-01-01"
  ),
  RouteSettings = list(
    list(
      DataTraceEnabled = TRUE|FALSE,
      DetailedMetricsEnabled = TRUE|FALSE,
      LoggingLevel = "ERROR"|"INFO"|"OFF",
      ThrottlingBurstLimit = 123,
      ThrottlingRateLimit = 123.0
    )
  ),
  StageName = "string",
  StageVariables = list(
    "string"
  ),
  Tags = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Creates a Stage for an API.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_stage(
  AccessLogSettings = list(
    DestinationArn = "string",
    Format = "string"
  ),
  ApiId = "string",
  AutoDeploy = TRUE|FALSE,
  ClientCertificateId = "string",
  DefaultRouteSettings = list(
    DataTraceEnabled = TRUE|FALSE,
    DetailedMetricsEnabled = TRUE|FALSE,
    LoggingLevel = "ERROR"|"INFO"|"OFF",
    ThrottlingBurstLimit = 123,
    ThrottlingRateLimit = 123.0
  ),
  DeploymentId = "string",
  Description = "string",
  RouteSettings = list(
    list(
      DataTraceEnabled = TRUE|FALSE,
      DetailedMetricsEnabled = TRUE|FALSE,
      LoggingLevel = "ERROR"|"INFO"|"OFF",
      ThrottlingBurstLimit = 123,
      ThrottlingRateLimit = 123.0
    )
  ),
  StageName = "string",
  StageVariables = list(
    "string"
  ),
  Tags = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
