% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_delete_api}
\alias{apigatewayv2_delete_api}
\title{Deletes an Api resource}
\usage{
apigatewayv2_delete_api(ApiId)
}
\arguments{
\item{ApiId}{[required] The API identifier.}
}
\description{
Deletes an Api resource.
}
\section{Request syntax}{
\preformatted{svc$delete_api(
  ApiId = "string"
)
}
}

\keyword{internal}
