% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_delete_deployment}
\alias{apigatewayv2_delete_deployment}
\title{Deletes a Deployment}
\usage{
apigatewayv2_delete_deployment(ApiId, DeploymentId)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{DeploymentId}{[required] The deployment ID.}
}
\description{
Deletes a Deployment.

See \url{https://www.paws-r-sdk.com/docs/apigatewayv2_delete_deployment/} for full documentation.
}
\keyword{internal}
