% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_delete_deployment}
\alias{apigatewayv2_delete_deployment}
\title{Deletes a Deployment}
\usage{
apigatewayv2_delete_deployment(ApiId, DeploymentId)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{DeploymentId}{[required] The deployment ID.}
}
\value{
An empty list.
}
\description{
Deletes a Deployment.
}
\section{Request syntax}{
\preformatted{svc$delete_deployment(
  ApiId = "string",
  DeploymentId = "string"
)
}
}

\keyword{internal}
