% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_delete_domain_name}
\alias{apigatewayv2_delete_domain_name}
\title{Deletes a domain name}
\usage{
apigatewayv2_delete_domain_name(DomainName)
}
\arguments{
\item{DomainName}{[required] The domain name.}
}
\description{
Deletes a domain name.

See \url{https://www.paws-r-sdk.com/docs/apigatewayv2_delete_domain_name/} for full documentation.
}
\keyword{internal}
