% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_delete_integration}
\alias{apigatewayv2_delete_integration}
\title{Deletes an Integration}
\usage{
apigatewayv2_delete_integration(ApiId, IntegrationId)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{IntegrationId}{[required] The integration ID.}
}
\description{
Deletes an Integration.

See \url{https://paws-r.github.io/docs/apigatewayv2/delete_integration.html} for full documentation.
}
\keyword{internal}
