% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_delete_integration}
\alias{apigatewayv2_delete_integration}
\title{Deletes an Integration}
\usage{
apigatewayv2_delete_integration(ApiId, IntegrationId)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{IntegrationId}{[required] The integration ID.}
}
\value{
An empty list.
}
\description{
Deletes an Integration.
}
\section{Request syntax}{
\preformatted{svc$delete_integration(
  ApiId = "string",
  IntegrationId = "string"
)
}
}

\keyword{internal}
