% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_delete_integration_response}
\alias{apigatewayv2_delete_integration_response}
\title{Deletes an IntegrationResponses}
\usage{
apigatewayv2_delete_integration_response(ApiId, IntegrationId,
  IntegrationResponseId)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{IntegrationId}{[required] The integration ID.}

\item{IntegrationResponseId}{[required] The integration response ID.}
}
\value{
An empty list.
}
\description{
Deletes an IntegrationResponses.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_integration_response(
  ApiId = "string",
  IntegrationId = "string",
  IntegrationResponseId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
