% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_delete_route}
\alias{apigatewayv2_delete_route}
\title{Deletes a Route}
\usage{
apigatewayv2_delete_route(ApiId, RouteId)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{RouteId}{[required] The route ID.}
}
\description{
Deletes a Route.

See \url{https://paws-r.github.io/docs/apigatewayv2/delete_route.html} for full documentation.
}
\keyword{internal}
