% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_delete_vpc_link}
\alias{apigatewayv2_delete_vpc_link}
\title{Deletes a VPC link}
\usage{
apigatewayv2_delete_vpc_link(VpcLinkId)
}
\arguments{
\item{VpcLinkId}{[required] The ID of the VPC link.}
}
\value{
An empty list.
}
\description{
Deletes a VPC link.
}
\section{Request syntax}{
\preformatted{svc$delete_vpc_link(
  VpcLinkId = "string"
)
}
}

\keyword{internal}
