% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_get_api}
\alias{apigatewayv2_get_api}
\title{Gets an Api resource}
\usage{
apigatewayv2_get_api(ApiId)
}
\arguments{
\item{ApiId}{[required] The API identifier.}
}
\description{
Gets an Api resource.

See \url{https://paws-r.github.io/docs/apigatewayv2/get_api.html} for full documentation.
}
\keyword{internal}
