% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_get_api_mapping}
\alias{apigatewayv2_get_api_mapping}
\title{Gets an API mapping}
\usage{
apigatewayv2_get_api_mapping(ApiMappingId, DomainName)
}
\arguments{
\item{ApiMappingId}{[required] The API mapping identifier.}

\item{DomainName}{[required] The domain name.}
}
\description{
Gets an API mapping.

See \url{https://paws-r.github.io/docs/apigatewayv2/get_api_mapping.html} for full documentation.
}
\keyword{internal}
