% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_get_api_mappings}
\alias{apigatewayv2_get_api_mappings}
\title{Gets API mappings}
\usage{
apigatewayv2_get_api_mappings(DomainName, MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{DomainName}{[required] The domain name.}

\item{MaxResults}{The maximum number of elements to be returned for this resource.}

\item{NextToken}{The next page of elements from this collection. Not valid for the last
element of the collection.}
}
\description{
Gets API mappings.

See \url{https://paws-r.github.io/docs/apigatewayv2/get_api_mappings.html} for full documentation.
}
\keyword{internal}
