% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_get_authorizer}
\alias{apigatewayv2_get_authorizer}
\title{Gets an Authorizer}
\usage{
apigatewayv2_get_authorizer(ApiId, AuthorizerId)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{AuthorizerId}{[required] The authorizer identifier.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  AuthorizerCredentialsArn = "string",
  AuthorizerId = "string",
  AuthorizerPayloadFormatVersion = "string",
  AuthorizerResultTtlInSeconds = 123,
  AuthorizerType = "REQUEST"|"JWT",
  AuthorizerUri = "string",
  EnableSimpleResponses = TRUE|FALSE,
  IdentitySource = list(
    "string"
  ),
  IdentityValidationExpression = "string",
  JwtConfiguration = list(
    Audience = list(
      "string"
    ),
    Issuer = "string"
  ),
  Name = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Gets an Authorizer.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_authorizer(
  ApiId = "string",
  AuthorizerId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
