% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_get_deployment}
\alias{apigatewayv2_get_deployment}
\title{Gets a Deployment}
\usage{
apigatewayv2_get_deployment(ApiId, DeploymentId)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{DeploymentId}{[required] The deployment ID.}
}
\description{
Gets a Deployment.

See \url{https://paws-r.github.io/docs/apigatewayv2/get_deployment.html} for full documentation.
}
\keyword{internal}
