% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_get_integration}
\alias{apigatewayv2_get_integration}
\title{Gets an Integration}
\usage{
apigatewayv2_get_integration(ApiId, IntegrationId)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{IntegrationId}{[required] The integration ID.}
}
\value{
A list with the following syntax:\preformatted{list(
  ApiGatewayManaged = TRUE|FALSE,
  ConnectionId = "string",
  ConnectionType = "INTERNET"|"VPC_LINK",
  ContentHandlingStrategy = "CONVERT_TO_BINARY"|"CONVERT_TO_TEXT",
  CredentialsArn = "string",
  Description = "string",
  IntegrationId = "string",
  IntegrationMethod = "string",
  IntegrationResponseSelectionExpression = "string",
  IntegrationSubtype = "string",
  IntegrationType = "AWS"|"HTTP"|"MOCK"|"HTTP_PROXY"|"AWS_PROXY",
  IntegrationUri = "string",
  PassthroughBehavior = "WHEN_NO_MATCH"|"NEVER"|"WHEN_NO_TEMPLATES",
  PayloadFormatVersion = "string",
  RequestParameters = list(
    "string"
  ),
  RequestTemplates = list(
    "string"
  ),
  ResponseParameters = list(
    list(
      "string"
    )
  ),
  TemplateSelectionExpression = "string",
  TimeoutInMillis = 123,
  TlsConfig = list(
    ServerNameToVerify = "string"
  )
)
}
}
\description{
Gets an Integration.
}
\section{Request syntax}{
\preformatted{svc$get_integration(
  ApiId = "string",
  IntegrationId = "string"
)
}
}

\keyword{internal}
