% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_get_integration_responses}
\alias{apigatewayv2_get_integration_responses}
\title{Gets the IntegrationResponses for an Integration}
\usage{
apigatewayv2_get_integration_responses(ApiId, IntegrationId, MaxResults,
  NextToken)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{IntegrationId}{[required] The integration ID.}

\item{MaxResults}{The maximum number of elements to be returned for this resource.}

\item{NextToken}{The next page of elements from this collection. Not valid for the last
element of the collection.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Items = list(
    list(
      ContentHandlingStrategy = "CONVERT_TO_BINARY"|"CONVERT_TO_TEXT",
      IntegrationResponseId = "string",
      IntegrationResponseKey = "string",
      ResponseParameters = list(
        "string"
      ),
      ResponseTemplates = list(
        "string"
      ),
      TemplateSelectionExpression = "string"
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Gets the IntegrationResponses for an Integration.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_integration_responses(
  ApiId = "string",
  IntegrationId = "string",
  MaxResults = "string",
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
