% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_get_model}
\alias{apigatewayv2_get_model}
\title{Gets a Model}
\usage{
apigatewayv2_get_model(ApiId, ModelId)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{ModelId}{[required] The model ID.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ContentType = "string",
  Description = "string",
  ModelId = "string",
  Name = "string",
  Schema = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Gets a Model.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_model(
  ApiId = "string",
  ModelId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
