% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_get_route}
\alias{apigatewayv2_get_route}
\title{Gets a Route}
\usage{
apigatewayv2_get_route(ApiId, RouteId)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{RouteId}{[required] The route ID.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ApiGatewayManaged = TRUE|FALSE,
  ApiKeyRequired = TRUE|FALSE,
  AuthorizationScopes = list(
    "string"
  ),
  AuthorizationType = "NONE"|"AWS_IAM"|"CUSTOM"|"JWT",
  AuthorizerId = "string",
  ModelSelectionExpression = "string",
  OperationName = "string",
  RequestModels = list(
    "string"
  ),
  RequestParameters = list(
    list(
      Required = TRUE|FALSE
    )
  ),
  RouteId = "string",
  RouteKey = "string",
  RouteResponseSelectionExpression = "string",
  Target = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Gets a Route.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_route(
  ApiId = "string",
  RouteId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
