% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_get_route_response}
\alias{apigatewayv2_get_route_response}
\title{Gets a RouteResponse}
\usage{
apigatewayv2_get_route_response(ApiId, RouteId, RouteResponseId)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{RouteId}{[required] The route ID.}

\item{RouteResponseId}{[required] The route response ID.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ModelSelectionExpression = "string",
  ResponseModels = list(
    "string"
  ),
  ResponseParameters = list(
    list(
      Required = TRUE|FALSE
    )
  ),
  RouteResponseId = "string",
  RouteResponseKey = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Gets a RouteResponse.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_route_response(
  ApiId = "string",
  RouteId = "string",
  RouteResponseId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
