% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_get_route_responses}
\alias{apigatewayv2_get_route_responses}
\title{Gets the RouteResponses for a Route}
\usage{
apigatewayv2_get_route_responses(ApiId, MaxResults, NextToken, RouteId)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{MaxResults}{The maximum number of elements to be returned for this resource.}

\item{NextToken}{The next page of elements from this collection. Not valid for the last
element of the collection.}

\item{RouteId}{[required] The route ID.}
}
\value{
A list with the following syntax:\preformatted{list(
  Items = list(
    list(
      ModelSelectionExpression = "string",
      ResponseModels = list(
        "string"
      ),
      ResponseParameters = list(
        list(
          Required = TRUE|FALSE
        )
      ),
      RouteResponseId = "string",
      RouteResponseKey = "string"
    )
  ),
  NextToken = "string"
)
}
}
\description{
Gets the RouteResponses for a Route.
}
\section{Request syntax}{
\preformatted{svc$get_route_responses(
  ApiId = "string",
  MaxResults = "string",
  NextToken = "string",
  RouteId = "string"
)
}
}

\keyword{internal}
