% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_get_routes}
\alias{apigatewayv2_get_routes}
\title{Gets the Routes for an API}
\usage{
apigatewayv2_get_routes(ApiId, MaxResults, NextToken)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{MaxResults}{The maximum number of elements to be returned for this resource.}

\item{NextToken}{The next page of elements from this collection. Not valid for the last
element of the collection.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Items = list(
    list(
      ApiGatewayManaged = TRUE|FALSE,
      ApiKeyRequired = TRUE|FALSE,
      AuthorizationScopes = list(
        "string"
      ),
      AuthorizationType = "NONE"|"AWS_IAM"|"CUSTOM"|"JWT",
      AuthorizerId = "string",
      ModelSelectionExpression = "string",
      OperationName = "string",
      RequestModels = list(
        "string"
      ),
      RequestParameters = list(
        list(
          Required = TRUE|FALSE
        )
      ),
      RouteId = "string",
      RouteKey = "string",
      RouteResponseSelectionExpression = "string",
      Target = "string"
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Gets the Routes for an API.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_routes(
  ApiId = "string",
  MaxResults = "string",
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
