% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_get_tags}
\alias{apigatewayv2_get_tags}
\title{Gets the Tags for an API}
\usage{
apigatewayv2_get_tags(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required]}
}
\description{
Gets the Tags for an API.
}
\section{Request syntax}{
\preformatted{svc$get_tags(
  ResourceArn = "string"
)
}
}

\keyword{internal}
