% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_get_tags}
\alias{apigatewayv2_get_tags}
\title{Gets a collection of Tag resources}
\usage{
apigatewayv2_get_tags(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The resource ARN for the tag.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Tags = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Gets a collection of Tag resources.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_tags(
  ResourceArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
