% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_get_vpc_link}
\alias{apigatewayv2_get_vpc_link}
\title{Gets a VPC link}
\usage{
apigatewayv2_get_vpc_link(VpcLinkId)
}
\arguments{
\item{VpcLinkId}{[required] The ID of the VPC link.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  CreatedDate = as.POSIXct(
    "2015-01-01"
  ),
  Name = "string",
  SecurityGroupIds = list(
    "string"
  ),
  SubnetIds = list(
    "string"
  ),
  Tags = list(
    "string"
  ),
  VpcLinkId = "string",
  VpcLinkStatus = "PENDING"|"AVAILABLE"|"DELETING"|"FAILED"|"INACTIVE",
  VpcLinkStatusMessage = "string",
  VpcLinkVersion = "V2"
)
}\if{html}{\out{</div>}}
}
\description{
Gets a VPC link.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_vpc_link(
  VpcLinkId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
