% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_get_vpc_links}
\alias{apigatewayv2_get_vpc_links}
\title{Gets a collection of VPC links}
\usage{
apigatewayv2_get_vpc_links(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{The maximum number of elements to be returned for this resource.}

\item{NextToken}{The next page of elements from this collection. Not valid for the last
element of the collection.}
}
\description{
Gets a collection of VPC links.

See \url{https://paws-r.github.io/docs/apigatewayv2/get_vpc_links.html} for full documentation.
}
\keyword{internal}
