% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_untag_resource}
\alias{apigatewayv2_untag_resource}
\title{Deletes a Tag}
\usage{
apigatewayv2_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The resource ARN for the tag.}

\item{TagKeys}{[required] The Tag keys to delete}
}
\value{
An empty list.
}
\description{
Deletes a Tag.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$untag_resource(
  ResourceArn = "string",
  TagKeys = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
