% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_untag_resource}
\alias{apigatewayv2_untag_resource}
\title{Untag an APIGW resource}
\usage{
apigatewayv2_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] AWS resource arn}

\item{TagKeys}{[required] The Tag keys to delete}
}
\description{
Untag an APIGW resource
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  ResourceArn = "string",
  TagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
