% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_update_api_mapping}
\alias{apigatewayv2_update_api_mapping}
\title{The API mapping}
\usage{
apigatewayv2_update_api_mapping(
  ApiId,
  ApiMappingId,
  ApiMappingKey = NULL,
  DomainName,
  Stage = NULL
)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{ApiMappingId}{[required] The API mapping identifier.}

\item{ApiMappingKey}{The API mapping key.}

\item{DomainName}{[required] The domain name.}

\item{Stage}{The API stage.}
}
\description{
The API mapping.

See \url{https://www.paws-r-sdk.com/docs/apigatewayv2_update_api_mapping/} for full documentation.
}
\keyword{internal}
