% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_update_api_mapping}
\alias{apigatewayv2_update_api_mapping}
\title{The API mapping}
\usage{
apigatewayv2_update_api_mapping(ApiId, ApiMappingId, ApiMappingKey,
  DomainName, Stage)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{ApiMappingId}{[required] The API mapping identifier.}

\item{ApiMappingKey}{The API mapping key.}

\item{DomainName}{[required] The domain name.}

\item{Stage}{The API stage.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ApiId = "string",
  ApiMappingId = "string",
  ApiMappingKey = "string",
  Stage = "string"
)
}\if{html}{\out{</div>}}
}
\description{
The API mapping.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_api_mapping(
  ApiId = "string",
  ApiMappingId = "string",
  ApiMappingKey = "string",
  DomainName = "string",
  Stage = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
