% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_update_deployment}
\alias{apigatewayv2_update_deployment}
\title{Updates a Deployment}
\usage{
apigatewayv2_update_deployment(ApiId, DeploymentId, Description = NULL)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{DeploymentId}{[required] The deployment ID.}

\item{Description}{The description for the deployment resource.}
}
\description{
Updates a Deployment.

See \url{https://www.paws-r-sdk.com/docs/apigatewayv2_update_deployment/} for full documentation.
}
\keyword{internal}
