% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_update_model}
\alias{apigatewayv2_update_model}
\title{Updates a Model}
\usage{
apigatewayv2_update_model(ApiId, ContentType, Description, ModelId,
  Name, Schema)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{ContentType}{The content-type for the model, for example, "application/json".}

\item{Description}{The description of the model.}

\item{ModelId}{[required] The model ID.}

\item{Name}{The name of the model.}

\item{Schema}{The schema for the model. For application/json models, this should be
JSON schema draft 4 model.}
}
\description{
Updates a Model.
}
\section{Request syntax}{
\preformatted{svc$update_model(
  ApiId = "string",
  ContentType = "string",
  Description = "string",
  ModelId = "string",
  Name = "string",
  Schema = "string"
)
}
}

\keyword{internal}
