% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appmesh_operations.R
\name{appmesh_delete_mesh}
\alias{appmesh_delete_mesh}
\title{Deletes an existing service mesh}
\usage{
appmesh_delete_mesh(meshName)
}
\arguments{
\item{meshName}{[required] The name of the service mesh to delete.}
}
\description{
Deletes an existing service mesh.

See \url{https://www.paws-r-sdk.com/docs/appmesh_delete_mesh/} for full documentation.
}
\keyword{internal}
