% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appmesh_operations.R
\name{appmesh_delete_virtual_service}
\alias{appmesh_delete_virtual_service}
\title{Deletes an existing virtual service}
\usage{
appmesh_delete_virtual_service(meshName, meshOwner = NULL, virtualServiceName)
}
\arguments{
\item{meshName}{[required] The name of the service mesh to delete the virtual service in.}

\item{meshOwner}{The Amazon Web Services IAM account ID of the service mesh owner. If the
account ID is not your own, then it's the ID of the account that shared
the mesh with your account. For more information about mesh sharing, see
\href{https://docs.aws.amazon.com/app-mesh/latest/userguide/sharing.html}{Working with shared meshes}.}

\item{virtualServiceName}{[required] The name of the virtual service to delete.}
}
\description{
Deletes an existing virtual service.

See \url{https://www.paws-r-sdk.com/docs/appmesh_delete_virtual_service/} for full documentation.
}
\keyword{internal}
