% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appmesh_operations.R
\name{appmesh_untag_resource}
\alias{appmesh_untag_resource}
\title{Deletes specified tags from a resource}
\usage{
appmesh_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource to delete tags from.}

\item{tagKeys}{[required] The keys of the tags to be removed.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Deletes specified tags from a resource.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  resourceArn = "string",
  tagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
