% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appmesh_operations.R
\name{appmesh_untag_resource}
\alias{appmesh_untag_resource}
\title{Deletes specified tags from a resource}
\usage{
appmesh_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource to delete tags from.}

\item{tagKeys}{[required] The keys of the tags to be removed.}
}
\description{
Deletes specified tags from a resource.

See \url{https://www.paws-r-sdk.com/docs/appmesh_untag_resource/} for full documentation.
}
\keyword{internal}
