% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arczonalshift_service.R
\name{arczonalshift}
\alias{arczonalshift}
\title{AWS ARC - Zonal Shift}
\usage{
arczonalshift(
  config = list(),
  credentials = list(),
  endpoint = NULL,
  region = NULL
)
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.
\itemize{
\item{\strong{credentials}:} {\itemize{
\item{\strong{creds}:} {\itemize{
\item{\strong{access_key_id}:} {AWS access key ID}
\item{\strong{secret_access_key}:} {AWS secret access key}
\item{\strong{session_token}:} {AWS temporary session token}
}}
\item{\strong{profile}:} {The name of a profile to use. If not given, then the default profile is used.}
\item{\strong{anonymous}:} {Set anonymous credentials.}
}}
\item{\strong{endpoint}:} {The complete URL to use for the constructed client.}
\item{\strong{region}:} {The AWS Region used in instantiating the client.}
\item{\strong{close_connection}:} {Immediately close all HTTP connections.}
\item{\strong{timeout}:} {The time in seconds till a timeout exception is thrown when attempting to make a connection. The default is 60 seconds.}
\item{\strong{s3_force_path_style}:} {Set this to \code{true} to force the request to use path-style addressing, i.e. \verb{http://s3.amazonaws.com/BUCKET/KEY}.}
\item{\strong{sts_regional_endpoint}:} {Set sts regional endpoint resolver to regional or legacy \url{https://docs.aws.amazon.com/sdkref/latest/guide/feature-sts-regionalized-endpoints.html}}
}}

\item{credentials}{Optional credentials shorthand for the config parameter
\itemize{
\item{\strong{creds}:} {\itemize{
\item{\strong{access_key_id}:} {AWS access key ID}
\item{\strong{secret_access_key}:} {AWS secret access key}
\item{\strong{session_token}:} {AWS temporary session token}
}}
\item{\strong{profile}:} {The name of a profile to use. If not given, then the default profile is used.}
\item{\strong{anonymous}:} {Set anonymous credentials.}
}}

\item{endpoint}{Optional shorthand for complete URL to use for the constructed client.}

\item{region}{Optional shorthand for AWS Region used in instantiating the client.}
}
\value{
A client for the service. You can call the service's operations using
syntax like \code{svc$operation(...)}, where \code{svc} is the name you've assigned
to the client. The available operations are listed in the
Operations section.
}
\description{
This is the API Reference Guide for the zonal shift feature of Amazon
Route 53 Application Recovery Controller. This guide is for developers
who need detailed information about zonal shift API actions, data types,
and errors.

Zonal shift is in preview release for Amazon Route 53 Application
Recovery Controller and is subject to change.

Zonal shift in Route 53 ARC enables you to move traffic for a load
balancer resource away from an Availability Zone. Starting a zonal shift
helps your application recover immediately, for example, from a
developer's bad code deployment or from an AWS infrastructure failure in
a single Availability Zone, reducing the impact and time lost from an
issue in one zone.

Supported AWS resources are automatically registered with Route 53 ARC.
Resources that are registered for zonal shifts in Route 53 ARC are
managed resources in Route 53 ARC. You can start a zonal shift for any
managed resource in your account in a Region. At this time, you can only
start a zonal shift for Network Load Balancers and Application Load
Balancers with cross-zone load balancing turned off.

Zonal shifts are temporary. You must specify an expiration when you
start a zonal shift, of up to three days initially. If you want to still
keep traffic away from an Availability Zone, you can update the zonal
shift and set a new expiration. You can also cancel a zonal shift,
before it expires, for example, if you're ready to restore traffic to
the Availability Zone.

For more information about using zonal shift, see the \href{https://docs.aws.amazon.com/r53recovery/latest/dg/what-is-route53-recovery.html}{Amazon Route 53 Application Recovery Controller Developer Guide}.
}
\section{Service syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc <- arczonalshift(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string",
      anonymous = "logical"
    ),
    endpoint = "string",
    region = "string",
    close_connection = "logical",
    timeout = "numeric",
    s3_force_path_style = "logical",
    sts_regional_endpoint = "string"
  ),
  credentials = list(
    creds = list(
      access_key_id = "string",
      secret_access_key = "string",
      session_token = "string"
    ),
    profile = "string",
    anonymous = "logical"
  ),
  endpoint = "string",
  region = "string"
)
}\if{html}{\out{</div>}}
}

\section{Operations}{

\tabular{ll}{
\link[=arczonalshift_cancel_zonal_shift]{cancel_zonal_shift} \tab Cancel a zonal shift in Amazon Route 53 Application Recovery Controller that you've started for a resource in your AWS account in an AWS Region\cr
\link[=arczonalshift_get_managed_resource]{get_managed_resource} \tab Get information about a resource that's been registered for zonal shifts with Amazon Route 53 Application Recovery Controller in this AWS Region\cr
\link[=arczonalshift_list_managed_resources]{list_managed_resources} \tab Lists all the resources in your AWS account in this AWS Region that are managed for zonal shifts in Amazon Route 53 Application Recovery Controller, and information about them\cr
\link[=arczonalshift_list_zonal_shifts]{list_zonal_shifts} \tab Lists all the active zonal shifts in Amazon Route 53 Application Recovery Controller in your AWS account in this AWS Region\cr
\link[=arczonalshift_start_zonal_shift]{start_zonal_shift} \tab You start a zonal shift to temporarily move load balancer traffic away from an Availability Zone in a AWS Region, to help your application recover immediately, for example, from a developer's bad code deployment or from an AWS infrastructure failure in a single Availability Zone\cr
\link[=arczonalshift_update_zonal_shift]{update_zonal_shift} \tab Update an active zonal shift in Amazon Route 53 Application Recovery Controller in your AWS account
}
}

\examples{
\dontrun{
svc <- arczonalshift()
svc$cancel_zonal_shift(
  Foo = 123
)
}

}
