% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arczonalshift_operations.R
\name{arczonalshift_cancel_zonal_shift}
\alias{arczonalshift_cancel_zonal_shift}
\title{Cancel a zonal shift in Amazon Application Recovery Controller}
\usage{
arczonalshift_cancel_zonal_shift(zonalShiftId)
}
\arguments{
\item{zonalShiftId}{[required] The internally-generated identifier of a zonal shift.}
}
\description{
Cancel a zonal shift in Amazon Application Recovery Controller. To cancel the zonal shift, specify the zonal shift ID.

See \url{https://www.paws-r-sdk.com/docs/arczonalshift_cancel_zonal_shift/} for full documentation.
}
\keyword{internal}
