% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arczonalshift_operations.R
\name{arczonalshift_create_practice_run_configuration}
\alias{arczonalshift_create_practice_run_configuration}
\title{A practice run configuration for zonal autoshift is required when you
enable zonal autoshift}
\usage{
arczonalshift_create_practice_run_configuration(
  resourceIdentifier,
  blockedWindows = NULL,
  blockedDates = NULL,
  blockingAlarms = NULL,
  allowedWindows = NULL,
  outcomeAlarms
)
}
\arguments{
\item{resourceIdentifier}{[required] The identifier of the resource that Amazon Web Services shifts traffic
for with a practice run zonal shift. The identifier is the Amazon
Resource Name (ARN) for the resource.

Amazon Application Recovery Controller currently supports enabling the
following resources for zonal shift and zonal autoshift:
\itemize{
\item \href{https://docs.aws.amazon.com/r53recovery/latest/dg/arc-zonal-shift.resource-types.ec2-auto-scaling-groups.html}{Amazon EC2 Auto Scaling groups}
\item \href{https://docs.aws.amazon.com/r53recovery/latest/dg/arc-zonal-shift.resource-types.eks.html}{Amazon Elastic Kubernetes Service}
\item \href{https://docs.aws.amazon.com/r53recovery/latest/dg/arc-zonal-shift.resource-types.app-load-balancers.html}{Application Load Balancer}
\item \href{https://docs.aws.amazon.com/r53recovery/latest/dg/arc-zonal-shift.resource-types.network-load-balancers.html}{Network Load Balancer}
}}

\item{blockedWindows}{Optionally, you can block ARC from starting practice runs for specific
windows of days and times.

The format for blocked windows is: DAY:HH:SS-DAY:HH:SS. Keep in mind,
when you specify dates, that dates and times for practice runs are in
UTC. Also, be aware of potential time adjustments that might be required
for daylight saving time differences. Separate multiple blocked windows
with spaces.

For example, say you run business report summaries three days a week.
For this scenario, you could set the following recurring days and times
as blocked windows, for example:
\verb{Mon:00:00-Mon:10:00 Wed-20:30-Wed:21:30 Fri-20:30-Fri:21:30}.

The \code{blockedWindows} have to start and end on the same day. Windows that
span multiple days aren't supported.}

\item{blockedDates}{Optionally, you can block ARC from starting practice runs for a resource
on specific calendar dates.

The format for blocked dates is: YYYY-MM-DD. Keep in mind, when you
specify dates, that dates and times for practice runs are in UTC.
Separate multiple blocked dates with spaces.

For example, if you have an application update scheduled to launch on
May 1, 2024, and you don't want practice runs to shift traffic away at
that time, you could set a blocked date for \code{2024-05-01}.}

\item{blockingAlarms}{\emph{Blocking alarms} for practice runs are optional alarms that you can
specify that block practice runs when one or more of the alarms is in an
\code{ALARM} state.}

\item{allowedWindows}{Optionally, you can allow ARC to start practice runs for specific
windows of days and times.

The format for allowed windows is: DAY:HH:SS-DAY:HH:SS. Keep in mind,
when you specify dates, that dates and times for practice runs are in
UTC. Also, be aware of potential time adjustments that might be required
for daylight saving time differences. Separate multiple allowed windows
with spaces.

For example, say you want to allow practice runs only on Wednesdays and
Fridays from noon to 5 p.m. For this scenario, you could set the
following recurring days and times as allowed windows, for example:
\verb{Wed-12:00-Wed:17:00 Fri-12:00-Fri:17:00}.

The \code{allowedWindows} have to start and end on the same day. Windows that
span multiple days aren't supported.}

\item{outcomeAlarms}{[required] \emph{Outcome alarms} for practice runs are alarms that you specify that end
a practice run when one or more of the alarms is in an \code{ALARM} state.

Configure one or more of these alarms to monitor the health of your
application when traffic is shifted away from an Availability Zone
during each practice run. You should configure these alarms to go into
an \code{ALARM} state if you want to stop a zonal shift, to let traffic for
the resource return to the original Availability Zone.}
}
\description{
A practice run configuration for zonal autoshift is required when you enable zonal autoshift. A practice run configuration includes specifications for blocked dates and blocked time windows, and for Amazon CloudWatch alarms that you create to use with practice runs. The alarms that you specify are an \emph{outcome alarm}, to monitor application health during practice runs and, optionally, a \emph{blocking alarm}, to block practice runs from starting.

See \url{https://www.paws-r-sdk.com/docs/arczonalshift_create_practice_run_configuration/} for full documentation.
}
\keyword{internal}
