% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arczonalshift_operations.R
\name{arczonalshift_get_autoshift_observer_notification_status}
\alias{arczonalshift_get_autoshift_observer_notification_status}
\title{Returns the status of autoshift observer notification}
\usage{
arczonalshift_get_autoshift_observer_notification_status()
}
\description{
Returns the status of autoshift observer notification. Autoshift observer notification enables you to be notified, through Amazon EventBridge, when there is an autoshift event for zonal autoshift.

See \url{https://www.paws-r-sdk.com/docs/arczonalshift_get_autoshift_observer_notification_status/} for full documentation.
}
\keyword{internal}
