% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backupgateway_operations.R
\name{backupgateway_update_hypervisor}
\alias{backupgateway_update_hypervisor}
\title{Updates a hypervisor metadata, including its host, username, and
password}
\usage{
backupgateway_update_hypervisor(
  Host = NULL,
  HypervisorArn,
  Name = NULL,
  Password = NULL,
  Username = NULL
)
}
\arguments{
\item{Host}{The updated host of the hypervisor. This can be either an IP address or
a fully-qualified domain name (FQDN).}

\item{HypervisorArn}{[required] The Amazon Resource Name (ARN) of the hypervisor to update.}

\item{Name}{The updated name for the hypervisor}

\item{Password}{The updated password for the hypervisor.}

\item{Username}{The updated username for the hypervisor.}
}
\description{
Updates a hypervisor metadata, including its host, username, and password. Specify which hypervisor to update using the Amazon Resource Name (ARN) of the hypervisor in your request.

See \url{https://paws-r.github.io/docs/backupgateway/update_hypervisor.html} for full documentation.
}
\keyword{internal}
