% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_associate_alias}
\alias{cloudfront_associate_alias}
\title{The AssociateAlias API operation only supports standard distributions}
\usage{
cloudfront_associate_alias(TargetDistributionId, Alias)
}
\arguments{
\item{TargetDistributionId}{[required] The ID of the standard distribution that you're associating the alias
with.}

\item{Alias}{[required] The alias (also known as a CNAME) to add to the target standard
distribution.}
}
\description{
The \code{\link[=cloudfront_associate_alias]{associate_alias}} API operation only supports standard distributions. To move domains between distribution tenants and/or standard distributions, we recommend that you use the \code{\link[=cloudfront_update_domain_association]{update_domain_association}} API operation instead.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_associate_alias/} for full documentation.
}
\keyword{internal}
