% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_create_cache_policy}
\alias{cloudfront_create_cache_policy}
\title{Creates a cache policy}
\usage{
cloudfront_create_cache_policy(CachePolicyConfig)
}
\arguments{
\item{CachePolicyConfig}{[required] A cache policy configuration.}
}
\description{
Creates a cache policy.

See \url{https://paws-r.github.io/docs/cloudfront/create_cache_policy.html} for full documentation.
}
\keyword{internal}
