% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_create_distribution}
\alias{cloudfront_create_distribution}
\title{Creates a new web distribution}
\usage{
cloudfront_create_distribution(DistributionConfig)
}
\arguments{
\item{DistributionConfig}{[required] The distribution's configuration information.}
}
\description{
Creates a new web distribution. You create a CloudFront distribution to tell CloudFront where you want content to be delivered from, and the details about how to track and manage content delivery. Send a \code{POST} request to the \verb{/CloudFront API version/distribution}/\verb{distribution ID} resource.

See \url{https://paws-r.github.io/docs/cloudfront/create_distribution.html} for full documentation.
}
\keyword{internal}
