% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_create_distribution_with_tags}
\alias{cloudfront_create_distribution_with_tags}
\title{Create a new distribution with tags}
\usage{
cloudfront_create_distribution_with_tags(DistributionConfigWithTags)
}
\arguments{
\item{DistributionConfigWithTags}{[required] The distribution's configuration information.}
}
\description{
Create a new distribution with tags. This API operation requires the following IAM permissions:

See \url{https://www.paws-r-sdk.com/docs/cloudfront_create_distribution_with_tags/} for full documentation.
}
\keyword{internal}
