% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_create_function}
\alias{cloudfront_create_function}
\title{Creates a CloudFront function}
\usage{
cloudfront_create_function(Name, FunctionConfig, FunctionCode)
}
\arguments{
\item{Name}{[required] A name to identify the function.}

\item{FunctionConfig}{[required] Configuration information about the function, including an optional
comment and the function’s runtime.}

\item{FunctionCode}{[required] The function code. For more information about writing a CloudFront
function, see \href{https://docs.aws.amazon.com/AmazonCloudFront/latest/DeveloperGuide/writing-function-code.html}{Writing function code for CloudFront Functions}
in the \emph{Amazon CloudFront Developer Guide}.}
}
\description{
Creates a CloudFront function.

See \url{https://paws-r.github.io/docs/cloudfront/create_function.html} for full documentation.
}
\keyword{internal}
