% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_create_invalidation}
\alias{cloudfront_create_invalidation}
\title{Create a new invalidation}
\usage{
cloudfront_create_invalidation(DistributionId, InvalidationBatch)
}
\arguments{
\item{DistributionId}{[required] The distribution's id.}

\item{InvalidationBatch}{[required] The batch information for the invalidation.}
}
\description{
Create a new invalidation. For more information, see \href{https://docs.aws.amazon.com/AmazonCloudFront/latest/DeveloperGuide/Invalidation.html}{Invalidating files} in the \emph{Amazon CloudFront Developer Guide}.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_create_invalidation/} for full documentation.
}
\keyword{internal}
