% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_create_key_group}
\alias{cloudfront_create_key_group}
\title{Creates a key group that you can use with CloudFront signed URLs and
signed cookies}
\usage{
cloudfront_create_key_group(KeyGroupConfig)
}
\arguments{
\item{KeyGroupConfig}{[required] A key group configuration.}
}
\description{
Creates a key group that you can use with \href{https://docs.aws.amazon.com/AmazonCloudFront/latest/DeveloperGuide/PrivateContent.html}{CloudFront signed URLs and signed cookies}.

See \url{https://paws-r.github.io/docs/cloudfront/create_key_group.html} for full documentation.
}
\keyword{internal}
