% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_create_key_value_store}
\alias{cloudfront_create_key_value_store}
\title{Specifies the Key Value Store resource to add to your account}
\usage{
cloudfront_create_key_value_store(Name, Comment = NULL, ImportSource = NULL)
}
\arguments{
\item{Name}{[required] The name of the Key Value Store. The maximum length of the name is 32
characters.}

\item{Comment}{The comment of the Key Value Store.}

\item{ImportSource}{The S3 bucket that provides the source for the import. The source must
be in a valid JSON format.}
}
\description{
Specifies the Key Value Store resource to add to your account. In your account, the Key Value Store names must be unique. You can also import Key Value Store data in JSON format from an S3 bucket by providing a valid \code{ImportSource} that you own.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_create_key_value_store/} for full documentation.
}
\keyword{internal}
