% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_create_public_key}
\alias{cloudfront_create_public_key}
\title{Uploads a public key to CloudFront that you can use with signed URLs and
signed cookies, or with field-level encryption}
\usage{
cloudfront_create_public_key(PublicKeyConfig)
}
\arguments{
\item{PublicKeyConfig}{[required] A CloudFront public key configuration.}
}
\description{
Uploads a public key to CloudFront that you can use with \href{https://docs.aws.amazon.com/AmazonCloudFront/latest/DeveloperGuide/PrivateContent.html}{signed URLs and signed cookies}, or with \href{https://docs.aws.amazon.com/AmazonCloudFront/latest/DeveloperGuide/field-level-encryption.html}{field-level encryption}.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_create_public_key/} for full documentation.
}
\keyword{internal}
