% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_create_realtime_log_config}
\alias{cloudfront_create_realtime_log_config}
\title{Creates a real-time log configuration}
\usage{
cloudfront_create_realtime_log_config(EndPoints, Fields, Name, SamplingRate)
}
\arguments{
\item{EndPoints}{[required] Contains information about the Amazon Kinesis data stream where you are
sending real-time log data.}

\item{Fields}{[required] A list of fields to include in each real-time log record.

For more information about fields, see \href{https://docs.aws.amazon.com/AmazonCloudFront/latest/DeveloperGuide/real-time-logs.html#understand-real-time-log-config-fields}{Real-time log configuration fields}
in the \emph{Amazon CloudFront Developer Guide}.}

\item{Name}{[required] A unique name to identify this real-time log configuration.}

\item{SamplingRate}{[required] The sampling rate for this real-time log configuration. You can specify
a whole number between 1 and 100 (inclusive) to determine the percentage
of viewer requests that are represented in the real-time log data.}
}
\description{
Creates a real-time log configuration.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_create_realtime_log_config/} for full documentation.
}
\keyword{internal}
