% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_create_streaming_distribution}
\alias{cloudfront_create_streaming_distribution}
\title{This API is deprecated}
\usage{
cloudfront_create_streaming_distribution(StreamingDistributionConfig)
}
\arguments{
\item{StreamingDistributionConfig}{[required] The streaming distribution's configuration information.}
}
\description{
This API is deprecated. Amazon CloudFront is deprecating real-time messaging protocol (RTMP) distributions on December 31, 2020. For more information, read the announcement on the Amazon CloudFront discussion forum.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_create_streaming_distribution/} for full documentation.
}
\keyword{internal}
