% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_create_trust_store}
\alias{cloudfront_create_trust_store}
\title{Creates a trust store}
\usage{
cloudfront_create_trust_store(Name, CaCertificatesBundleSource, Tags = NULL)
}
\arguments{
\item{Name}{[required] A name for the trust store.}

\item{CaCertificatesBundleSource}{[required] The CA certificates bundle source for the trust store.}

\item{Tags}{}
}
\description{
Creates a trust store.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_create_trust_store/} for full documentation.
}
\keyword{internal}
