% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_create_vpc_origin}
\alias{cloudfront_create_vpc_origin}
\title{Create an Amazon CloudFront VPC origin}
\usage{
cloudfront_create_vpc_origin(VpcOriginEndpointConfig, Tags = NULL)
}
\arguments{
\item{VpcOriginEndpointConfig}{[required] The VPC origin endpoint configuration.}

\item{Tags}{}
}
\description{
Create an Amazon CloudFront VPC origin.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_create_vpc_origin/} for full documentation.
}
\keyword{internal}
