% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_delete_cache_policy}
\alias{cloudfront_delete_cache_policy}
\title{Deletes a cache policy}
\usage{
cloudfront_delete_cache_policy(Id, IfMatch = NULL)
}
\arguments{
\item{Id}{[required] The unique identifier for the cache policy that you are deleting. To get
the identifier, you can use
\code{\link[=cloudfront_list_cache_policies]{list_cache_policies}}.}

\item{IfMatch}{The version of the cache policy that you are deleting. The version is
the cache policy's \code{ETag} value, which you can get using
\code{\link[=cloudfront_list_cache_policies]{list_cache_policies}},
\code{\link[=cloudfront_get_cache_policy]{get_cache_policy}}, or
\code{\link[=cloudfront_get_cache_policy_config]{get_cache_policy_config}}.}
}
\description{
Deletes a cache policy.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_delete_cache_policy/} for full documentation.
}
\keyword{internal}
