% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_delete_distribution}
\alias{cloudfront_delete_distribution}
\title{Delete a distribution}
\usage{
cloudfront_delete_distribution(Id, IfMatch)
}
\arguments{
\item{Id}{[required] The distribution ID.}

\item{IfMatch}{The value of the \code{ETag} header that you received when you disabled the
distribution. For example: \code{E2QWRUHAPOMQZL}.}
}
\value{
An empty list.
}
\description{
Delete a distribution.
}
\section{Request syntax}{
\preformatted{svc$delete_distribution(
  Id = "string",
  IfMatch = "string"
)
}
}

\keyword{internal}
