% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_delete_function}
\alias{cloudfront_delete_function}
\title{Deletes a CloudFront function}
\usage{
cloudfront_delete_function(Name, IfMatch)
}
\arguments{
\item{Name}{[required] The name of the function that you are deleting.}

\item{IfMatch}{[required] The current version (\code{ETag} value) of the function that you are
deleting, which you can get using
\code{\link[=cloudfront_describe_function]{describe_function}}.}
}
\description{
Deletes a CloudFront function.

See \url{https://paws-r.github.io/docs/cloudfront/delete_function.html} for full documentation.
}
\keyword{internal}
