% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_delete_monitoring_subscription}
\alias{cloudfront_delete_monitoring_subscription}
\title{Disables additional CloudWatch metrics for the specified CloudFront
distribution}
\usage{
cloudfront_delete_monitoring_subscription(DistributionId)
}
\arguments{
\item{DistributionId}{[required] The ID of the distribution that you are disabling metrics for.}
}
\description{
Disables additional CloudWatch metrics for the specified CloudFront
distribution.
}
\section{Request syntax}{
\preformatted{svc$delete_monitoring_subscription(
  DistributionId = "string"
)
}
}

\keyword{internal}
