% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_delete_origin_request_policy}
\alias{cloudfront_delete_origin_request_policy}
\title{Deletes an origin request policy}
\usage{
cloudfront_delete_origin_request_policy(Id, IfMatch = NULL)
}
\arguments{
\item{Id}{[required] The unique identifier for the origin request policy that you are
deleting. To get the identifier, you can use
\code{\link[=cloudfront_list_origin_request_policies]{list_origin_request_policies}}.}

\item{IfMatch}{The version of the origin request policy that you are deleting. The
version is the origin request policy's \code{ETag} value, which you can get
using
\code{\link[=cloudfront_list_origin_request_policies]{list_origin_request_policies}},
\code{\link[=cloudfront_get_origin_request_policy]{get_origin_request_policy}}, or
\code{\link[=cloudfront_get_origin_request_policy_config]{get_origin_request_policy_config}}.}
}
\description{
Deletes an origin request policy.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_delete_origin_request_policy/} for full documentation.
}
\keyword{internal}
