% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_delete_vpc_origin}
\alias{cloudfront_delete_vpc_origin}
\title{Delete an Amazon CloudFront VPC origin}
\usage{
cloudfront_delete_vpc_origin(Id, IfMatch)
}
\arguments{
\item{Id}{[required] The VPC origin ID.}

\item{IfMatch}{[required] The VPC origin to delete, if a match occurs.}
}
\description{
Delete an Amazon CloudFront VPC origin.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_delete_vpc_origin/} for full documentation.
}
\keyword{internal}
