% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_describe_key_value_store}
\alias{cloudfront_describe_key_value_store}
\title{Specifies the Key Value Store and its configuration}
\usage{
cloudfront_describe_key_value_store(Name)
}
\arguments{
\item{Name}{[required] The name of the Key Value Store.}
}
\description{
Specifies the Key Value Store and its configuration.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_describe_key_value_store/} for full documentation.
}
\keyword{internal}
