% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_get_cache_policy}
\alias{cloudfront_get_cache_policy}
\title{Gets a cache policy, including the following metadata:}
\usage{
cloudfront_get_cache_policy(Id)
}
\arguments{
\item{Id}{[required] The unique identifier for the cache policy. If the cache policy is
attached to a distribution's cache behavior, you can get the policy's
identifier using \code{\link[=cloudfront_list_distributions]{list_distributions}}
or \code{\link[=cloudfront_get_distribution]{get_distribution}}. If the cache
policy is not attached to a cache behavior, you can get the identifier
using \code{\link[=cloudfront_list_cache_policies]{list_cache_policies}}.}
}
\description{
Gets a cache policy, including the following metadata:

See \url{https://www.paws-r-sdk.com/docs/cloudfront_get_cache_policy/} for full documentation.
}
\keyword{internal}
