% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_get_cloud_front_origin_access_identity}
\alias{cloudfront_get_cloud_front_origin_access_identity}
\title{Get the information about an origin access identity}
\usage{
cloudfront_get_cloud_front_origin_access_identity(Id)
}
\arguments{
\item{Id}{[required] The identity's ID.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  CloudFrontOriginAccessIdentity = list(
    Id = "string",
    S3CanonicalUserId = "string",
    CloudFrontOriginAccessIdentityConfig = list(
      CallerReference = "string",
      Comment = "string"
    )
  ),
  ETag = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Get the information about an origin access identity.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_cloud_front_origin_access_identity(
  Id = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
