% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_get_cloud_front_origin_access_identity_config}
\alias{cloudfront_get_cloud_front_origin_access_identity_config}
\title{Get the configuration information about an origin access identity}
\usage{
cloudfront_get_cloud_front_origin_access_identity_config(Id)
}
\arguments{
\item{Id}{[required] The identity's ID.}
}
\description{
Get the configuration information about an origin access identity.

See \url{https://paws-r.github.io/docs/cloudfront/get_cloud_front_origin_access_identity_config.html} for full documentation.
}
\keyword{internal}
