% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_get_field_level_encryption}
\alias{cloudfront_get_field_level_encryption}
\title{Get the field-level encryption configuration information}
\usage{
cloudfront_get_field_level_encryption(Id)
}
\arguments{
\item{Id}{[required] Request the ID for the field-level encryption configuration information.}
}
\description{
Get the field-level encryption configuration information.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_get_field_level_encryption/} for full documentation.
}
\keyword{internal}
