% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_get_field_level_encryption_config}
\alias{cloudfront_get_field_level_encryption_config}
\title{Get the field-level encryption configuration information}
\usage{
cloudfront_get_field_level_encryption_config(Id)
}
\arguments{
\item{Id}{[required] Request the ID for the field-level encryption configuration information.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  FieldLevelEncryptionConfig = list(
    CallerReference = "string",
    Comment = "string",
    QueryArgProfileConfig = list(
      ForwardWhenQueryArgProfileIsUnknown = TRUE|FALSE,
      QueryArgProfiles = list(
        Quantity = 123,
        Items = list(
          list(
            QueryArg = "string",
            ProfileId = "string"
          )
        )
      )
    ),
    ContentTypeProfileConfig = list(
      ForwardWhenContentTypeIsUnknown = TRUE|FALSE,
      ContentTypeProfiles = list(
        Quantity = 123,
        Items = list(
          list(
            Format = "URLEncoded",
            ProfileId = "string",
            ContentType = "string"
          )
        )
      )
    )
  ),
  ETag = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Get the field-level encryption configuration information.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_field_level_encryption_config(
  Id = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
