% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_get_field_level_encryption_profile}
\alias{cloudfront_get_field_level_encryption_profile}
\title{Get the field-level encryption profile information}
\usage{
cloudfront_get_field_level_encryption_profile(Id)
}
\arguments{
\item{Id}{[required] Get the ID for the field-level encryption profile information.}
}
\description{
Get the field-level encryption profile information.
}
\section{Request syntax}{
\preformatted{svc$get_field_level_encryption_profile(
  Id = "string"
)
}
}

\keyword{internal}
