% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_get_field_level_encryption_profile_config}
\alias{cloudfront_get_field_level_encryption_profile_config}
\title{Get the field-level encryption profile configuration information}
\usage{
cloudfront_get_field_level_encryption_profile_config(Id)
}
\arguments{
\item{Id}{[required] Get the ID for the field-level encryption profile configuration
information.}
}
\description{
Get the field-level encryption profile configuration information.

See \url{https://paws-r.github.io/docs/cloudfront/get_field_level_encryption_profile_config.html} for full documentation.
}
\keyword{internal}
