% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_get_function}
\alias{cloudfront_get_function}
\title{Gets the code of a CloudFront function}
\usage{
cloudfront_get_function(Name, Stage = NULL)
}
\arguments{
\item{Name}{[required] The name of the function whose code you are getting.}

\item{Stage}{The function’s stage, either \code{DEVELOPMENT} or \code{LIVE}.}
}
\description{
Gets the code of a CloudFront function. To get configuration information and metadata about a function, use \code{\link[=cloudfront_describe_function]{describe_function}}.

See \url{https://paws-r.github.io/docs/cloudfront/get_function.html} for full documentation.
}
\keyword{internal}
