% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_get_key_group}
\alias{cloudfront_get_key_group}
\title{Gets a key group, including the date and time when the key group was
last modified}
\usage{
cloudfront_get_key_group(Id)
}
\arguments{
\item{Id}{[required] The identifier of the key group that you are getting. To get the
identifier, use \code{\link[=cloudfront_list_key_groups]{list_key_groups}}.}
}
\description{
Gets a key group, including the date and time when the key group was last modified.

See \url{https://paws-r.github.io/docs/cloudfront/get_key_group.html} for full documentation.
}
\keyword{internal}
