% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_get_key_group_config}
\alias{cloudfront_get_key_group_config}
\title{Gets a key group configuration}
\usage{
cloudfront_get_key_group_config(Id)
}
\arguments{
\item{Id}{[required] The identifier of the key group whose configuration you are getting. To
get the identifier, use \code{\link[=cloudfront_list_key_groups]{list_key_groups}}.}
}
\value{
A list with the following syntax:\preformatted{list(
  KeyGroupConfig = list(
    Name = "string",
    Items = list(
      "string"
    ),
    Comment = "string"
  ),
  ETag = "string"
)
}
}
\description{
Gets a key group configuration.

To get a key group configuration, you must provide the key group’s
identifier. If the key group is referenced in a distribution’s cache
behavior, you can get the key group’s identifier using
\code{\link[=cloudfront_list_distributions]{list_distributions}} or
\code{\link[=cloudfront_get_distribution]{get_distribution}}. If the key group is
not referenced in a cache behavior, you can get the identifier using
\code{\link[=cloudfront_list_key_groups]{list_key_groups}}.
}
\section{Request syntax}{
\preformatted{svc$get_key_group_config(
  Id = "string"
)
}
}

\keyword{internal}
