% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_get_origin_access_control_config}
\alias{cloudfront_get_origin_access_control_config}
\title{Gets a CloudFront origin access control configuration}
\usage{
cloudfront_get_origin_access_control_config(Id)
}
\arguments{
\item{Id}{[required] The unique identifier of the origin access control.}
}
\description{
Gets a CloudFront origin access control configuration.

See \url{https://paws-r.github.io/docs/cloudfront/get_origin_access_control_config.html} for full documentation.
}
\keyword{internal}
