% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_get_origin_request_policy_config}
\alias{cloudfront_get_origin_request_policy_config}
\title{Gets an origin request policy configuration}
\usage{
cloudfront_get_origin_request_policy_config(Id)
}
\arguments{
\item{Id}{[required] The unique identifier for the origin request policy. If the origin
request policy is attached to a distribution’s cache behavior, you can
get the policy’s identifier using
\code{\link[=cloudfront_list_distributions]{list_distributions}} or
\code{\link[=cloudfront_get_distribution]{get_distribution}}. If the origin request
policy is not attached to a cache behavior, you can get the identifier
using
\code{\link[=cloudfront_list_origin_request_policies]{list_origin_request_policies}}.}
}
\description{
Gets an origin request policy configuration.

See \url{https://paws-r.github.io/docs/cloudfront/get_origin_request_policy_config.html} for full documentation.
}
\keyword{internal}
