% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_get_public_key}
\alias{cloudfront_get_public_key}
\title{Gets a public key}
\usage{
cloudfront_get_public_key(Id)
}
\arguments{
\item{Id}{[required] The identifier of the public key you are getting.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  PublicKey = list(
    Id = "string",
    CreatedTime = as.POSIXct(
      "2015-01-01"
    ),
    PublicKeyConfig = list(
      CallerReference = "string",
      Name = "string",
      EncodedKey = "string",
      Comment = "string"
    )
  ),
  ETag = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Gets a public key.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_public_key(
  Id = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
