% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_get_public_key}
\alias{cloudfront_get_public_key}
\title{Gets a public key}
\usage{
cloudfront_get_public_key(Id)
}
\arguments{
\item{Id}{[required] The identifier of the public key you are getting.}
}
\description{
Gets a public key.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_get_public_key/} for full documentation.
}
\keyword{internal}
