% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_get_public_key_config}
\alias{cloudfront_get_public_key_config}
\title{Gets a public key configuration}
\usage{
cloudfront_get_public_key_config(Id)
}
\arguments{
\item{Id}{[required] The identifier of the public key whose configuration you are getting.}
}
\description{
Gets a public key configuration.

See \url{https://paws-r.github.io/docs/cloudfront/get_public_key_config.html} for full documentation.
}
\keyword{internal}
