% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_get_streaming_distribution}
\alias{cloudfront_get_streaming_distribution}
\title{Gets information about a specified RTMP distribution, including the
distribution configuration}
\usage{
cloudfront_get_streaming_distribution(Id)
}
\arguments{
\item{Id}{[required] The streaming distribution's ID.}
}
\description{
Gets information about a specified RTMP distribution, including the distribution configuration.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_get_streaming_distribution/} for full documentation.
}
\keyword{internal}
