% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_list_cache_policies}
\alias{cloudfront_list_cache_policies}
\title{Gets a list of cache policies}
\usage{
cloudfront_list_cache_policies(Type = NULL, Marker = NULL, MaxItems = NULL)
}
\arguments{
\item{Type}{A filter to return only the specified kinds of cache policies. Valid
values are:
\itemize{
\item \code{managed} – Returns only the managed policies created by Amazon Web
Services.
\item \code{custom} – Returns only the custom policies created in your Amazon Web
Services account.
}}

\item{Marker}{Use this field when paginating results to indicate where to begin in
your list of cache policies. The response includes cache policies in the
list that occur after the marker. To get the next page of the list, set
this field's value to the value of \code{NextMarker} from the current page's
response.}

\item{MaxItems}{The maximum number of cache policies that you want in the response.}
}
\description{
Gets a list of cache policies.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_list_cache_policies/} for full documentation.
}
\keyword{internal}
