% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_list_conflicting_aliases}
\alias{cloudfront_list_conflicting_aliases}
\title{The ListConflictingAliases API operation only supports standard
distributions}
\usage{
cloudfront_list_conflicting_aliases(
  DistributionId,
  Alias,
  Marker = NULL,
  MaxItems = NULL
)
}
\arguments{
\item{DistributionId}{[required] The ID of a standard distribution in your account that has an attached
TLS certificate that includes the provided alias.}

\item{Alias}{[required] The alias (also called a CNAME) to search for conflicting aliases.}

\item{Marker}{Use this field when paginating results to indicate where to begin in the
list of conflicting aliases. The response includes conflicting aliases
in the list that occur after the marker. To get the next page of the
list, set this field's value to the value of \code{NextMarker} from the
current page's response.}

\item{MaxItems}{The maximum number of conflicting aliases that you want in the response.}
}
\description{
The \code{\link[=cloudfront_list_conflicting_aliases]{list_conflicting_aliases}} API operation only supports standard distributions. To list domain conflicts for both standard distributions and distribution tenants, we recommend that you use the \code{\link[=cloudfront_list_domain_conflicts]{list_domain_conflicts}} API operation instead.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_list_conflicting_aliases/} for full documentation.
}
\keyword{internal}
