% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_list_distributions}
\alias{cloudfront_list_distributions}
\title{List CloudFront distributions}
\usage{
cloudfront_list_distributions(Marker, MaxItems)
}
\arguments{
\item{Marker}{Use this when paginating results to indicate where to begin in your list
of distributions. The results include distributions in the list that
occur after the marker. To get the next page of results, set the
\code{Marker} to the value of the \code{NextMarker} from the current page's
response (which is also the ID of the last distribution on that page).}

\item{MaxItems}{The maximum number of distributions you want in the response body.}
}
\value{
A list with the following syntax:\preformatted{list(
  DistributionList = list(
    Marker = "string",
    NextMarker = "string",
    MaxItems = 123,
    IsTruncated = TRUE|FALSE,
    Quantity = 123,
    Items = list(
      list(
        Id = "string",
        ARN = "string",
        Status = "string",
        LastModifiedTime = as.POSIXct(
          "2015-01-01"
        ),
        DomainName = "string",
        Aliases = list(
          Quantity = 123,
          Items = list(
            "string"
          )
        ),
        Origins = list(
          Quantity = 123,
          Items = list(
            list(
              Id = "string",
              DomainName = "string",
              OriginPath = "string",
              CustomHeaders = list(
                Quantity = 123,
                Items = list(
                  list(
                    HeaderName = "string",
                    HeaderValue = "string"
                  )
                )
              ),
              S3OriginConfig = list(
                OriginAccessIdentity = "string"
              ),
              CustomOriginConfig = list(
                HTTPPort = 123,
                HTTPSPort = 123,
                OriginProtocolPolicy = "http-only"|"match-viewer"|"https-only",
                OriginSslProtocols = list(
                  Quantity = 123,
                  Items = list(
                    "SSLv3"|"TLSv1"|"TLSv1.1"|"TLSv1.2"
                  )
                ),
                OriginReadTimeout = 123,
                OriginKeepaliveTimeout = 123
              ),
              ConnectionAttempts = 123,
              ConnectionTimeout = 123,
              OriginShield = list(
                Enabled = TRUE|FALSE,
                OriginShieldRegion = "string"
              )
            )
          )
        ),
        OriginGroups = list(
          Quantity = 123,
          Items = list(
            list(
              Id = "string",
              FailoverCriteria = list(
                StatusCodes = list(
                  Quantity = 123,
                  Items = list(
                    123
                  )
                )
              ),
              Members = list(
                Quantity = 123,
                Items = list(
                  list(
                    OriginId = "string"
                  )
                )
              )
            )
          )
        ),
        DefaultCacheBehavior = list(
          TargetOriginId = "string",
          TrustedSigners = list(
            Enabled = TRUE|FALSE,
            Quantity = 123,
            Items = list(
              "string"
            )
          ),
          TrustedKeyGroups = list(
            Enabled = TRUE|FALSE,
            Quantity = 123,
            Items = list(
              "string"
            )
          ),
          ViewerProtocolPolicy = "allow-all"|"https-only"|"redirect-to-https",
          AllowedMethods = list(
            Quantity = 123,
            Items = list(
              "GET"|"HEAD"|"POST"|"PUT"|"PATCH"|"OPTIONS"|"DELETE"
            ),
            CachedMethods = list(
              Quantity = 123,
              Items = list(
                "GET"|"HEAD"|"POST"|"PUT"|"PATCH"|"OPTIONS"|"DELETE"
              )
            )
          ),
          SmoothStreaming = TRUE|FALSE,
          Compress = TRUE|FALSE,
          LambdaFunctionAssociations = list(
            Quantity = 123,
            Items = list(
              list(
                LambdaFunctionARN = "string",
                EventType = "viewer-request"|"viewer-response"|"origin-request"|"origin-response",
                IncludeBody = TRUE|FALSE
              )
            )
          ),
          FunctionAssociations = list(
            Quantity = 123,
            Items = list(
              list(
                FunctionARN = "string",
                EventType = "viewer-request"|"viewer-response"|"origin-request"|"origin-response"
              )
            )
          ),
          FieldLevelEncryptionId = "string",
          RealtimeLogConfigArn = "string",
          CachePolicyId = "string",
          OriginRequestPolicyId = "string",
          ForwardedValues = list(
            QueryString = TRUE|FALSE,
            Cookies = list(
              Forward = "none"|"whitelist"|"all",
              WhitelistedNames = list(
                Quantity = 123,
                Items = list(
                  "string"
                )
              )
            ),
            Headers = list(
              Quantity = 123,
              Items = list(
                "string"
              )
            ),
            QueryStringCacheKeys = list(
              Quantity = 123,
              Items = list(
                "string"
              )
            )
          ),
          MinTTL = 123,
          DefaultTTL = 123,
          MaxTTL = 123
        ),
        CacheBehaviors = list(
          Quantity = 123,
          Items = list(
            list(
              PathPattern = "string",
              TargetOriginId = "string",
              TrustedSigners = list(
                Enabled = TRUE|FALSE,
                Quantity = 123,
                Items = list(
                  "string"
                )
              ),
              TrustedKeyGroups = list(
                Enabled = TRUE|FALSE,
                Quantity = 123,
                Items = list(
                  "string"
                )
              ),
              ViewerProtocolPolicy = "allow-all"|"https-only"|"redirect-to-https",
              AllowedMethods = list(
                Quantity = 123,
                Items = list(
                  "GET"|"HEAD"|"POST"|"PUT"|"PATCH"|"OPTIONS"|"DELETE"
                ),
                CachedMethods = list(
                  Quantity = 123,
                  Items = list(
                    "GET"|"HEAD"|"POST"|"PUT"|"PATCH"|"OPTIONS"|"DELETE"
                  )
                )
              ),
              SmoothStreaming = TRUE|FALSE,
              Compress = TRUE|FALSE,
              LambdaFunctionAssociations = list(
                Quantity = 123,
                Items = list(
                  list(
                    LambdaFunctionARN = "string",
                    EventType = "viewer-request"|"viewer-response"|"origin-request"|"origin-response",
                    IncludeBody = TRUE|FALSE
                  )
                )
              ),
              FunctionAssociations = list(
                Quantity = 123,
                Items = list(
                  list(
                    FunctionARN = "string",
                    EventType = "viewer-request"|"viewer-response"|"origin-request"|"origin-response"
                  )
                )
              ),
              FieldLevelEncryptionId = "string",
              RealtimeLogConfigArn = "string",
              CachePolicyId = "string",
              OriginRequestPolicyId = "string",
              ForwardedValues = list(
                QueryString = TRUE|FALSE,
                Cookies = list(
                  Forward = "none"|"whitelist"|"all",
                  WhitelistedNames = list(
                    Quantity = 123,
                    Items = list(
                      "string"
                    )
                  )
                ),
                Headers = list(
                  Quantity = 123,
                  Items = list(
                    "string"
                  )
                ),
                QueryStringCacheKeys = list(
                  Quantity = 123,
                  Items = list(
                    "string"
                  )
                )
              ),
              MinTTL = 123,
              DefaultTTL = 123,
              MaxTTL = 123
            )
          )
        ),
        CustomErrorResponses = list(
          Quantity = 123,
          Items = list(
            list(
              ErrorCode = 123,
              ResponsePagePath = "string",
              ResponseCode = "string",
              ErrorCachingMinTTL = 123
            )
          )
        ),
        Comment = "string",
        PriceClass = "PriceClass_100"|"PriceClass_200"|"PriceClass_All",
        Enabled = TRUE|FALSE,
        ViewerCertificate = list(
          CloudFrontDefaultCertificate = TRUE|FALSE,
          IAMCertificateId = "string",
          ACMCertificateArn = "string",
          SSLSupportMethod = "sni-only"|"vip"|"static-ip",
          MinimumProtocolVersion = "SSLv3"|"TLSv1"|"TLSv1_2016"|"TLSv1.1_2016"|"TLSv1.2_2018"|"TLSv1.2_2019",
          Certificate = "string",
          CertificateSource = "cloudfront"|"iam"|"acm"
        ),
        Restrictions = list(
          GeoRestriction = list(
            RestrictionType = "blacklist"|"whitelist"|"none",
            Quantity = 123,
            Items = list(
              "string"
            )
          )
        ),
        WebACLId = "string",
        HttpVersion = "http1.1"|"http2",
        IsIPV6Enabled = TRUE|FALSE,
        AliasICPRecordals = list(
          list(
            CNAME = "string",
            ICPRecordalStatus = "APPROVED"|"SUSPENDED"|"PENDING"
          )
        )
      )
    )
  )
)
}
}
\description{
List CloudFront distributions.
}
\section{Request syntax}{
\preformatted{svc$list_distributions(
  Marker = "string",
  MaxItems = "string"
)
}
}

\keyword{internal}
