% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_list_realtime_log_configs}
\alias{cloudfront_list_realtime_log_configs}
\title{Gets a list of real-time log configurations}
\usage{
cloudfront_list_realtime_log_configs(MaxItems = NULL, Marker = NULL)
}
\arguments{
\item{MaxItems}{The maximum number of real-time log configurations that you want in the
response.}

\item{Marker}{Use this field when paginating results to indicate where to begin in
your list of real-time log configurations. The response includes
real-time log configurations in the list that occur after the marker. To
get the next page of the list, set this field’s value to the value of
\code{NextMarker} from the current page’s response.}
}
\description{
Gets a list of real-time log configurations.

See \url{https://paws-r.github.io/docs/cloudfront/list_realtime_log_configs.html} for full documentation.
}
\keyword{internal}
