% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_list_tags_for_resource}
\alias{cloudfront_list_tags_for_resource}
\title{List tags for a CloudFront resource}
\usage{
cloudfront_list_tags_for_resource(Resource)
}
\arguments{
\item{Resource}{[required] An ARN of a CloudFront resource.}
}
\value{
A list with the following syntax:\preformatted{list(
  Tags = list(
    Items = list(
      list(
        Key = "string",
        Value = "string"
      )
    )
  )
)
}
}
\description{
List tags for a CloudFront resource.
}
\section{Request syntax}{
\preformatted{svc$list_tags_for_resource(
  Resource = "string"
)
}
}

\keyword{internal}
