% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_list_tags_for_resource}
\alias{cloudfront_list_tags_for_resource}
\title{List tags for a CloudFront resource}
\usage{
cloudfront_list_tags_for_resource(Resource)
}
\arguments{
\item{Resource}{[required] An ARN of a CloudFront resource.}
}
\description{
List tags for a CloudFront resource.

See \url{https://paws-r.github.io/docs/cloudfront/list_tags_for_resource.html} for full documentation.
}
\keyword{internal}
