% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_untag_resource}
\alias{cloudfront_untag_resource}
\title{Remove tags from a CloudFront resource}
\usage{
cloudfront_untag_resource(Resource, TagKeys)
}
\arguments{
\item{Resource}{[required] An ARN of a CloudFront resource.}

\item{TagKeys}{[required] A complex type that contains zero or more \code{Tag} key elements.}
}
\value{
An empty list.
}
\description{
Remove tags from a CloudFront resource.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  Resource = "string",
  TagKeys = list(
    Items = list(
      "string"
    )
  )
)
}
}

\keyword{internal}
