% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_update_cloud_front_origin_access_identity}
\alias{cloudfront_update_cloud_front_origin_access_identity}
\title{Update an origin access identity}
\usage{
cloudfront_update_cloud_front_origin_access_identity(
  CloudFrontOriginAccessIdentityConfig, Id, IfMatch)
}
\arguments{
\item{CloudFrontOriginAccessIdentityConfig}{[required] The identity's configuration information.}

\item{Id}{[required] The identity's id.}

\item{IfMatch}{The value of the \code{ETag} header that you received when retrieving the
identity's configuration. For example: \code{E2QWRUHAPOMQZL}.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  CloudFrontOriginAccessIdentity = list(
    Id = "string",
    S3CanonicalUserId = "string",
    CloudFrontOriginAccessIdentityConfig = list(
      CallerReference = "string",
      Comment = "string"
    )
  ),
  ETag = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Update an origin access identity.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_cloud_front_origin_access_identity(
  CloudFrontOriginAccessIdentityConfig = list(
    CallerReference = "string",
    Comment = "string"
  ),
  Id = "string",
  IfMatch = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
