% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_update_field_level_encryption_config}
\alias{cloudfront_update_field_level_encryption_config}
\title{Update a field-level encryption configuration}
\usage{
cloudfront_update_field_level_encryption_config(
  FieldLevelEncryptionConfig,
  Id,
  IfMatch = NULL
)
}
\arguments{
\item{FieldLevelEncryptionConfig}{[required] Request to update a field-level encryption configuration.}

\item{Id}{[required] The ID of the configuration you want to update.}

\item{IfMatch}{The value of the \code{ETag} header that you received when retrieving the
configuration identity to update. For example: \code{E2QWRUHAPOMQZL}.}
}
\description{
Update a field-level encryption configuration.

See \url{https://paws-r.github.io/docs/cloudfront/update_field_level_encryption_config.html} for full documentation.
}
\keyword{internal}
