% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_update_field_level_encryption_profile}
\alias{cloudfront_update_field_level_encryption_profile}
\title{Update a field-level encryption profile}
\usage{
cloudfront_update_field_level_encryption_profile(
  FieldLevelEncryptionProfileConfig,
  Id,
  IfMatch = NULL
)
}
\arguments{
\item{FieldLevelEncryptionProfileConfig}{[required] Request to update a field-level encryption profile.}

\item{Id}{[required] The ID of the field-level encryption profile request.}

\item{IfMatch}{The value of the \code{ETag} header that you received when retrieving the
profile identity to update. For example: \code{E2QWRUHAPOMQZL}.}
}
\description{
Update a field-level encryption profile.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_update_field_level_encryption_profile/} for full documentation.
}
\keyword{internal}
