% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_update_function}
\alias{cloudfront_update_function}
\title{Updates a CloudFront function}
\usage{
cloudfront_update_function(Name, IfMatch, FunctionConfig, FunctionCode)
}
\arguments{
\item{Name}{[required] The name of the function that you are updating.}

\item{IfMatch}{[required] The current version (\code{ETag} value) of the function that you are
updating, which you can get using
\code{\link[=cloudfront_describe_function]{describe_function}}.}

\item{FunctionConfig}{[required] Configuration information about the function.}

\item{FunctionCode}{[required] The function code. For more information about writing a CloudFront
function, see \href{https://docs.aws.amazon.com/AmazonCloudFront/latest/DeveloperGuide/writing-function-code.html}{Writing function code for CloudFront Functions}
in the \emph{Amazon CloudFront Developer Guide}.}
}
\value{
A list with the following syntax:\preformatted{list(
  FunctionSummary = list(
    Name = "string",
    Status = "string",
    FunctionConfig = list(
      Comment = "string",
      Runtime = "cloudfront-js-1.0"
    ),
    FunctionMetadata = list(
      FunctionARN = "string",
      Stage = "DEVELOPMENT"|"LIVE",
      CreatedTime = as.POSIXct(
        "2015-01-01"
      ),
      LastModifiedTime = as.POSIXct(
        "2015-01-01"
      )
    )
  ),
  ETag = "string"
)
}
}
\description{
Updates a CloudFront function.

You can update a function’s code or the comment that describes the
function. You cannot update a function’s name.

To update a function, you provide the function’s name and version
(\code{ETag} value) along with the updated function code. To get the name and
version, you can use \code{\link[=cloudfront_list_functions]{list_functions}} and
\code{\link[=cloudfront_describe_function]{describe_function}}.
}
\section{Request syntax}{
\preformatted{svc$update_function(
  Name = "string",
  IfMatch = "string",
  FunctionConfig = list(
    Comment = "string",
    Runtime = "cloudfront-js-1.0"
  ),
  FunctionCode = raw
)
}
}

\keyword{internal}
