% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_update_trust_store}
\alias{cloudfront_update_trust_store}
\title{Updates a trust store}
\usage{
cloudfront_update_trust_store(Id, CaCertificatesBundleSource, IfMatch)
}
\arguments{
\item{Id}{[required] The trust store ID.}

\item{CaCertificatesBundleSource}{[required] The CA certificates bundle source.}

\item{IfMatch}{[required] The current version (\code{ETag} value) of the trust store you are updating.}
}
\description{
Updates a trust store.

See \url{https://www.paws-r-sdk.com/docs/cloudfront_update_trust_store/} for full documentation.
}
\keyword{internal}
