% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_allocate_private_virtual_interface}
\alias{directconnect_allocate_private_virtual_interface}
\title{Provisions a private virtual interface to be owned by the specified
Amazon Web Services account}
\usage{
directconnect_allocate_private_virtual_interface(
  connectionId,
  ownerAccount,
  newPrivateVirtualInterfaceAllocation
)
}
\arguments{
\item{connectionId}{[required] The ID of the connection on which the private virtual interface is
provisioned.}

\item{ownerAccount}{[required] The ID of the Amazon Web Services account that owns the virtual private
interface.}

\item{newPrivateVirtualInterfaceAllocation}{[required] Information about the private virtual interface.}
}
\description{
Provisions a private virtual interface to be owned by the specified Amazon Web Services account.

See \url{https://www.paws-r-sdk.com/docs/directconnect_allocate_private_virtual_interface/} for full documentation.
}
\keyword{internal}
