% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_allocate_public_virtual_interface}
\alias{directconnect_allocate_public_virtual_interface}
\title{Provisions a public virtual interface to be owned by the specified
Amazon Web Services account}
\usage{
directconnect_allocate_public_virtual_interface(
  connectionId,
  ownerAccount,
  newPublicVirtualInterfaceAllocation
)
}
\arguments{
\item{connectionId}{[required] The ID of the connection on which the public virtual interface is
provisioned.}

\item{ownerAccount}{[required] The ID of the Amazon Web Services account that owns the public virtual
interface.}

\item{newPublicVirtualInterfaceAllocation}{[required] Information about the public virtual interface.}
}
\description{
Provisions a public virtual interface to be owned by the specified Amazon Web Services account.

See \url{https://paws-r.github.io/docs/directconnect/allocate_public_virtual_interface.html} for full documentation.
}
\keyword{internal}
