% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_associate_virtual_interface}
\alias{directconnect_associate_virtual_interface}
\title{Associates a virtual interface with a specified link aggregation group
(LAG) or connection}
\usage{
directconnect_associate_virtual_interface(virtualInterfaceId, connectionId)
}
\arguments{
\item{virtualInterfaceId}{[required] The ID of the virtual interface.}

\item{connectionId}{[required] The ID of the LAG or connection.}
}
\description{
Associates a virtual interface with a specified link aggregation group (LAG) or connection. Connectivity to Amazon Web Services is temporarily interrupted as the virtual interface is being migrated. If the target connection or LAG has an associated virtual interface with a conflicting VLAN number or a conflicting IP address, the operation fails.

See \url{https://www.paws-r-sdk.com/docs/directconnect_associate_virtual_interface/} for full documentation.
}
\keyword{internal}
