% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_confirm_connection}
\alias{directconnect_confirm_connection}
\title{Confirms the creation of the specified hosted connection on an
interconnect}
\usage{
directconnect_confirm_connection(connectionId)
}
\arguments{
\item{connectionId}{[required] The ID of the hosted connection.}
}
\description{
Confirms the creation of the specified hosted connection on an interconnect.

See \url{https://www.paws-r-sdk.com/docs/directconnect_confirm_connection/} for full documentation.
}
\keyword{internal}
