% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_confirm_public_virtual_interface}
\alias{directconnect_confirm_public_virtual_interface}
\title{Accepts ownership of a public virtual interface created by another AWS
account}
\usage{
directconnect_confirm_public_virtual_interface(virtualInterfaceId)
}
\arguments{
\item{virtualInterfaceId}{[required] The ID of the virtual interface.}
}
\description{
Accepts ownership of a public virtual interface created by another AWS
account.

After the virtual interface owner makes this call, the specified virtual
interface is created and made available to handle traffic.
}
\section{Request syntax}{
\preformatted{svc$confirm_public_virtual_interface(
  virtualInterfaceId = "string"
)
}
}

\keyword{internal}
