% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_confirm_transit_virtual_interface}
\alias{directconnect_confirm_transit_virtual_interface}
\title{Accepts ownership of a transit virtual interface created by another AWS
account}
\usage{
directconnect_confirm_transit_virtual_interface(virtualInterfaceId,
  directConnectGatewayId)
}
\arguments{
\item{virtualInterfaceId}{[required] The ID of the virtual interface.}

\item{directConnectGatewayId}{[required] The ID of the Direct Connect gateway.}
}
\value{
A list with the following syntax:\preformatted{list(
  virtualInterfaceState = "confirming"|"verifying"|"pending"|"available"|"down"|"deleting"|"deleted"|"rejected"|"unknown"
)
}
}
\description{
Accepts ownership of a transit virtual interface created by another AWS
account.

After the owner of the transit virtual interface makes this call, the
specified transit virtual interface is created and made available to
handle traffic.
}
\section{Request syntax}{
\preformatted{svc$confirm_transit_virtual_interface(
  virtualInterfaceId = "string",
  directConnectGatewayId = "string"
)
}
}

\keyword{internal}
