% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_create_bgp_peer}
\alias{directconnect_create_bgp_peer}
\title{Creates a BGP peer on the specified virtual interface}
\usage{
directconnect_create_bgp_peer(virtualInterfaceId = NULL, newBGPPeer = NULL)
}
\arguments{
\item{virtualInterfaceId}{The ID of the virtual interface.}

\item{newBGPPeer}{Information about the BGP peer.}
}
\description{
Creates a BGP peer on the specified virtual interface.

See \url{https://www.paws-r-sdk.com/docs/directconnect_create_bgp_peer/} for full documentation.
}
\keyword{internal}
