% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_create_direct_connect_gateway_association_proposal}
\alias{directconnect_create_direct_connect_gateway_association_proposal}
\title{Creates a proposal to associate the specified virtual private gateway or
transit gateway with the specified Direct Connect gateway}
\usage{
directconnect_create_direct_connect_gateway_association_proposal(
  directConnectGatewayId, directConnectGatewayOwnerAccount, gatewayId,
  addAllowedPrefixesToDirectConnectGateway,
  removeAllowedPrefixesToDirectConnectGateway)
}
\arguments{
\item{directConnectGatewayId}{[required] The ID of the Direct Connect gateway.}

\item{directConnectGatewayOwnerAccount}{[required] The ID of the AWS account that owns the Direct Connect gateway.}

\item{gatewayId}{[required] The ID of the virtual private gateway or transit gateway.}

\item{addAllowedPrefixesToDirectConnectGateway}{The Amazon VPC prefixes to advertise to the Direct Connect gateway.}

\item{removeAllowedPrefixesToDirectConnectGateway}{The Amazon VPC prefixes to no longer advertise to the Direct Connect
gateway.}
}
\value{
A list with the following syntax:\preformatted{list(
  directConnectGatewayAssociationProposal = list(
    proposalId = "string",
    directConnectGatewayId = "string",
    directConnectGatewayOwnerAccount = "string",
    proposalState = "requested"|"accepted"|"deleted",
    associatedGateway = list(
      id = "string",
      type = "virtualPrivateGateway"|"transitGateway",
      ownerAccount = "string",
      region = "string"
    ),
    existingAllowedPrefixesToDirectConnectGateway = list(
      list(
        cidr = "string"
      )
    ),
    requestedAllowedPrefixesToDirectConnectGateway = list(
      list(
        cidr = "string"
      )
    )
  )
)
}
}
\description{
Creates a proposal to associate the specified virtual private gateway or
transit gateway with the specified Direct Connect gateway.

You can associate a Direct Connect gateway and virtual private gateway
or transit gateway that is owned by any AWS account.
}
\section{Request syntax}{
\preformatted{svc$create_direct_connect_gateway_association_proposal(
  directConnectGatewayId = "string",
  directConnectGatewayOwnerAccount = "string",
  gatewayId = "string",
  addAllowedPrefixesToDirectConnectGateway = list(
    list(
      cidr = "string"
    )
  ),
  removeAllowedPrefixesToDirectConnectGateway = list(
    list(
      cidr = "string"
    )
  )
)
}
}

\keyword{internal}
