% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_create_private_virtual_interface}
\alias{directconnect_create_private_virtual_interface}
\title{Creates a private virtual interface}
\usage{
directconnect_create_private_virtual_interface(
  connectionId,
  newPrivateVirtualInterface
)
}
\arguments{
\item{connectionId}{[required] The ID of the connection.}

\item{newPrivateVirtualInterface}{[required] Information about the private virtual interface.}
}
\description{
Creates a private virtual interface. A virtual interface is the VLAN that transports Direct Connect traffic. A private virtual interface can be connected to either a Direct Connect gateway or a Virtual Private Gateway (VGW). Connecting the private virtual interface to a Direct Connect gateway enables the possibility for connecting to multiple VPCs, including VPCs in different Amazon Web Services Regions. Connecting the private virtual interface to a VGW only provides access to a single VPC within the same Region.

See \url{https://paws-r.github.io/docs/directconnect/create_private_virtual_interface.html} for full documentation.
}
\keyword{internal}
