% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_create_transit_virtual_interface}
\alias{directconnect_create_transit_virtual_interface}
\title{Creates a transit virtual interface}
\usage{
directconnect_create_transit_virtual_interface(connectionId,
  newTransitVirtualInterface)
}
\arguments{
\item{connectionId}{[required] The ID of the connection.}

\item{newTransitVirtualInterface}{[required] Information about the transit virtual interface.}
}
\description{
Creates a transit virtual interface. A transit virtual interface is a
VLAN that transports traffic from a Direct Connect gateway to one or
more transit gateways. A transit virtual interface enables the
connection of multiple VPCs attached to a transit gateway to a Direct
Connect gateway.
}
\section{Request syntax}{
\preformatted{svc$create_transit_virtual_interface(
  connectionId = "string",
  newTransitVirtualInterface = list(
    virtualInterfaceName = "string",
    vlan = 123,
    asn = 123,
    mtu = 123,
    authKey = "string",
    amazonAddress = "string",
    customerAddress = "string",
    addressFamily = "ipv4"|"ipv6",
    directConnectGatewayId = "string"
  )
)
}
}

\keyword{internal}
