% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_delete_connection}
\alias{directconnect_delete_connection}
\title{Deletes the specified connection}
\usage{
directconnect_delete_connection(connectionId)
}
\arguments{
\item{connectionId}{[required] The ID of the connection.}
}
\description{
Deletes the specified connection.

See \url{https://www.paws-r-sdk.com/docs/directconnect_delete_connection/} for full documentation.
}
\keyword{internal}
