% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_delete_direct_connect_gateway}
\alias{directconnect_delete_direct_connect_gateway}
\title{Deletes the specified Direct Connect gateway}
\usage{
directconnect_delete_direct_connect_gateway(directConnectGatewayId)
}
\arguments{
\item{directConnectGatewayId}{[required] The ID of the Direct Connect gateway.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  directConnectGateway = list(
    directConnectGatewayId = "string",
    directConnectGatewayName = "string",
    amazonSideAsn = 123,
    ownerAccount = "string",
    directConnectGatewayState = "pending"|"available"|"deleting"|"deleted",
    stateChangeError = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Deletes the specified Direct Connect gateway. You must first delete all
virtual interfaces that are attached to the Direct Connect gateway and
disassociate all virtual private gateways associated with the Direct
Connect gateway.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_direct_connect_gateway(
  directConnectGatewayId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
