% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_delete_interconnect}
\alias{directconnect_delete_interconnect}
\title{Deletes the specified interconnect}
\usage{
directconnect_delete_interconnect(interconnectId)
}
\arguments{
\item{interconnectId}{[required] The ID of the interconnect.}
}
\description{
Deletes the specified interconnect.

See \url{https://paws-r.github.io/docs/directconnect/delete_interconnect.html} for full documentation.
}
\keyword{internal}
