% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_delete_lag}
\alias{directconnect_delete_lag}
\title{Deletes the specified link aggregation group (LAG)}
\usage{
directconnect_delete_lag(lagId)
}
\arguments{
\item{lagId}{[required] The ID of the LAG.}
}
\description{
Deletes the specified link aggregation group (LAG). You cannot delete a
LAG if it has active virtual interfaces or hosted connections.
}
\section{Request syntax}{
\preformatted{svc$delete_lag(
  lagId = "string"
)
}
}

\keyword{internal}
