% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_delete_virtual_interface}
\alias{directconnect_delete_virtual_interface}
\title{Deletes a virtual interface}
\usage{
directconnect_delete_virtual_interface(virtualInterfaceId)
}
\arguments{
\item{virtualInterfaceId}{[required] The ID of the virtual interface.}
}
\description{
Deletes a virtual interface.

See \url{https://paws-r.github.io/docs/directconnect/delete_virtual_interface.html} for full documentation.
}
\keyword{internal}
