% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_describe_connections}
\alias{directconnect_describe_connections}
\title{Displays the specified connection or all connections in this Region}
\usage{
directconnect_describe_connections(
  connectionId = NULL,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{connectionId}{The ID of the connection.}

\item{maxResults}{The maximum number of results to return with a single call. To retrieve
the remaining results, make another call with the returned \code{nextToken}
value.

If \code{MaxResults} is given a value larger than 100, only 100 results are
returned.}

\item{nextToken}{The token for the next page of results.}
}
\description{
Displays the specified connection or all connections in this Region.

See \url{https://www.paws-r-sdk.com/docs/directconnect_describe_connections/} for full documentation.
}
\keyword{internal}
