% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_describe_connections_on_interconnect}
\alias{directconnect_describe_connections_on_interconnect}
\title{Deprecated}
\usage{
directconnect_describe_connections_on_interconnect(interconnectId)
}
\arguments{
\item{interconnectId}{[required] The ID of the interconnect.}
}
\description{
Deprecated. Use \code{\link[=directconnect_describe_hosted_connections]{describe_hosted_connections}} instead.

See \url{https://paws-r.github.io/docs/directconnect/describe_connections_on_interconnect.html} for full documentation.
}
\keyword{internal}
