% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_describe_customer_metadata}
\alias{directconnect_describe_customer_metadata}
\title{Get and view a list of customer agreements, along with their signed
status and whether the customer is an NNIPartner, NNIPartnerV2, or a
nonPartner}
\usage{
directconnect_describe_customer_metadata()
}
\description{
Get and view a list of customer agreements, along with their signed status and whether the customer is an NNIPartner, NNIPartnerV2, or a nonPartner.

See \url{https://www.paws-r-sdk.com/docs/directconnect_describe_customer_metadata/} for full documentation.
}
\keyword{internal}
