% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_describe_interconnect_loa}
\alias{directconnect_describe_interconnect_loa}
\title{Deprecated}
\usage{
directconnect_describe_interconnect_loa(
  interconnectId,
  providerName = NULL,
  loaContentType = NULL
)
}
\arguments{
\item{interconnectId}{[required] The ID of the interconnect.}

\item{providerName}{The name of the service provider who establishes connectivity on your
behalf. If you supply this parameter, the LOA-CFA lists the provider
name alongside your company name as the requester of the cross connect.}

\item{loaContentType}{The standard media type for the LOA-CFA document. The only supported
value is application/pdf.}
}
\description{
Deprecated. Use \code{\link[=directconnect_describe_loa]{describe_loa}} instead.

See \url{https://paws-r.github.io/docs/directconnect/describe_interconnect_loa.html} for full documentation.
}
\keyword{internal}
