% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_describe_interconnects}
\alias{directconnect_describe_interconnects}
\title{Lists the interconnects owned by the Amazon Web Services account or only
the specified interconnect}
\usage{
directconnect_describe_interconnects(interconnectId = NULL)
}
\arguments{
\item{interconnectId}{The ID of the interconnect.}
}
\description{
Lists the interconnects owned by the Amazon Web Services account or only the specified interconnect.

See \url{https://paws-r.github.io/docs/directconnect/describe_interconnects.html} for full documentation.
}
\keyword{internal}
