% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_describe_interconnects}
\alias{directconnect_describe_interconnects}
\title{Lists the interconnects owned by the AWS account or only the specified
interconnect}
\usage{
directconnect_describe_interconnects(interconnectId)
}
\arguments{
\item{interconnectId}{The ID of the interconnect.}
}
\description{
Lists the interconnects owned by the AWS account or only the specified
interconnect.
}
\section{Request syntax}{
\preformatted{svc$describe_interconnects(
  interconnectId = "string"
)
}
}

\keyword{internal}
