% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_describe_lags}
\alias{directconnect_describe_lags}
\title{Describes all your link aggregation groups (LAG) or the specified LAG}
\usage{
directconnect_describe_lags(lagId = NULL, maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{lagId}{The ID of the LAG.}

\item{maxResults}{The maximum number of results to return with a single call. To retrieve
the remaining results, make another call with the returned \code{nextToken}
value.

If \code{MaxResults} is given a value larger than 100, only 100 results are
returned.}

\item{nextToken}{The token for the next page of results.}
}
\description{
Describes all your link aggregation groups (LAG) or the specified LAG.

See \url{https://www.paws-r-sdk.com/docs/directconnect_describe_lags/} for full documentation.
}
\keyword{internal}
