% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_describe_locations}
\alias{directconnect_describe_locations}
\title{Lists the AWS Direct Connect locations in the current AWS Region}
\usage{
directconnect_describe_locations()
}
\value{
A list with the following syntax:\preformatted{list(
  locations = list(
    list(
      locationCode = "string",
      locationName = "string",
      region = "string",
      availablePortSpeeds = list(
        "string"
      ),
      availableProviders = list(
        "string"
      )
    )
  )
)
}
}
\description{
Lists the AWS Direct Connect locations in the current AWS Region. These
are the locations that can be selected when calling
\code{\link[=directconnect_create_connection]{create_connection}} or
\code{\link[=directconnect_create_interconnect]{create_interconnect}}.
}
\keyword{internal}
