% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_describe_tags}
\alias{directconnect_describe_tags}
\title{Describes the tags associated with the specified Direct Connect
resources}
\usage{
directconnect_describe_tags(resourceArns)
}
\arguments{
\item{resourceArns}{[required] The Amazon Resource Names (ARNs) of the resources.}
}
\description{
Describes the tags associated with the specified Direct Connect resources.

See \url{https://paws-r.github.io/docs/directconnect/describe_tags.html} for full documentation.
}
\keyword{internal}
