% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_describe_virtual_gateways}
\alias{directconnect_describe_virtual_gateways}
\title{Deprecated}
\usage{
directconnect_describe_virtual_gateways()
}
\description{
Deprecated. Use \code{DescribeVpnGateways} instead. See \href{https://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_DescribeVpnGateways.html}{DescribeVPNGateways} in the \emph{Amazon Elastic Compute Cloud API Reference}.

See \url{https://www.paws-r-sdk.com/docs/directconnect_describe_virtual_gateways/} for full documentation.
}
\keyword{internal}
