% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_start_bgp_failover_test}
\alias{directconnect_start_bgp_failover_test}
\title{Starts the virtual interface failover test that verifies your
configuration meets your resiliency requirements by placing the BGP
peering session in the DOWN state}
\usage{
directconnect_start_bgp_failover_test(
  virtualInterfaceId,
  bgpPeers = NULL,
  testDurationInMinutes = NULL
)
}
\arguments{
\item{virtualInterfaceId}{[required] The ID of the virtual interface you want to test.}

\item{bgpPeers}{The BGP peers to place in the DOWN state.}

\item{testDurationInMinutes}{The time in minutes that the virtual interface failover test will last.

Maximum value: 4,320 minutes (72 hours).

Default: 180 minutes (3 hours).}
}
\description{
Starts the virtual interface failover test that verifies your configuration meets your resiliency requirements by placing the BGP peering session in the DOWN state. You can then send traffic to verify that there are no outages.

See \url{https://paws-r.github.io/docs/directconnect/start_bgp_failover_test.html} for full documentation.
}
\keyword{internal}
