% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_stop_bgp_failover_test}
\alias{directconnect_stop_bgp_failover_test}
\title{Stops the virtual interface failover test}
\usage{
directconnect_stop_bgp_failover_test(virtualInterfaceId)
}
\arguments{
\item{virtualInterfaceId}{[required] The ID of the virtual interface you no longer want to test.}
}
\value{
A list with the following syntax:\preformatted{list(
  virtualInterfaceTest = list(
    testId = "string",
    virtualInterfaceId = "string",
    bgpPeers = list(
      "string"
    ),
    status = "string",
    ownerAccount = "string",
    testDurationInMinutes = 123,
    startTime = as.POSIXct(
      "2015-01-01"
    ),
    endTime = as.POSIXct(
      "2015-01-01"
    )
  )
)
}
}
\description{
Stops the virtual interface failover test.
}
\section{Request syntax}{
\preformatted{svc$stop_bgp_failover_test(
  virtualInterfaceId = "string"
)
}
}

\keyword{internal}
