% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_untag_resource}
\alias{directconnect_untag_resource}
\title{Removes one or more tags from the specified AWS Direct Connect resource}
\usage{
directconnect_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource.}

\item{tagKeys}{[required] The tag keys of the tags to remove.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Removes one or more tags from the specified AWS Direct Connect resource.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  resourceArn = "string",
  tagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
