% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_update_lag}
\alias{directconnect_update_lag}
\title{Updates the attributes of the specified link aggregation group (LAG)}
\usage{
directconnect_update_lag(lagId, lagName, minimumLinks, encryptionMode)
}
\arguments{
\item{lagId}{[required] The ID of the LAG.}

\item{lagName}{The name of the LAG.}

\item{minimumLinks}{The minimum number of physical connections that must be operational for
the LAG itself to be operational.}

\item{encryptionMode}{The LAG MAC Security (MACsec) encryption mode.

AWS applies the value to all connections which are part of the LAG.}
}
\value{
A list with the following syntax:\preformatted{list(
  connectionsBandwidth = "string",
  numberOfConnections = 123,
  lagId = "string",
  ownerAccount = "string",
  lagName = "string",
  lagState = "requested"|"pending"|"available"|"down"|"deleting"|"deleted"|"unknown",
  location = "string",
  region = "string",
  minimumLinks = 123,
  awsDevice = "string",
  awsDeviceV2 = "string",
  connections = list(
    list(
      ownerAccount = "string",
      connectionId = "string",
      connectionName = "string",
      connectionState = "ordering"|"requested"|"pending"|"available"|"down"|"deleting"|"deleted"|"rejected"|"unknown",
      region = "string",
      location = "string",
      bandwidth = "string",
      vlan = 123,
      partnerName = "string",
      loaIssueTime = as.POSIXct(
        "2015-01-01"
      ),
      lagId = "string",
      awsDevice = "string",
      jumboFrameCapable = TRUE|FALSE,
      awsDeviceV2 = "string",
      hasLogicalRedundancy = "unknown"|"yes"|"no",
      tags = list(
        list(
          key = "string",
          value = "string"
        )
      ),
      providerName = "string",
      macSecCapable = TRUE|FALSE,
      portEncryptionStatus = "string",
      encryptionMode = "string",
      macSecKeys = list(
        list(
          secretARN = "string",
          ckn = "string",
          state = "string",
          startOn = "string"
        )
      )
    )
  ),
  allowsHostedConnections = TRUE|FALSE,
  jumboFrameCapable = TRUE|FALSE,
  hasLogicalRedundancy = "unknown"|"yes"|"no",
  tags = list(
    list(
      key = "string",
      value = "string"
    )
  ),
  providerName = "string",
  macSecCapable = TRUE|FALSE,
  encryptionMode = "string",
  macSecKeys = list(
    list(
      secretARN = "string",
      ckn = "string",
      state = "string",
      startOn = "string"
    )
  )
)
}
}
\description{
Updates the attributes of the specified link aggregation group (LAG).

You can update the following LAG attributes:
\itemize{
\item The name of the LAG.
\item The value for the minimum number of connections that must be
operational for the LAG itself to be operational.
\item The LAG's MACsec encryption mode.

AWS assigns this value to each connection which is part of the LAG.
\item The tags
}

If you adjust the threshold value for the minimum number of operational
connections, ensure that the new value does not cause the LAG to fall
below the threshold and become non-operational.
}
\section{Request syntax}{
\preformatted{svc$update_lag(
  lagId = "string",
  lagName = "string",
  minimumLinks = 123,
  encryptionMode = "string"
)
}
}

\keyword{internal}
