% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_update_lag}
\alias{directconnect_update_lag}
\title{Updates the attributes of the specified link aggregation group (LAG)}
\usage{
directconnect_update_lag(
  lagId,
  lagName = NULL,
  minimumLinks = NULL,
  encryptionMode = NULL
)
}
\arguments{
\item{lagId}{[required] The ID of the LAG.}

\item{lagName}{The name of the LAG.}

\item{minimumLinks}{The minimum number of physical connections that must be operational for
the LAG itself to be operational.}

\item{encryptionMode}{The LAG MAC Security (MACsec) encryption mode.

Amazon Web Services applies the value to all connections which are part
of the LAG.}
}
\description{
Updates the attributes of the specified link aggregation group (LAG).

See \url{https://paws-r.github.io/docs/directconnect/update_lag.html} for full documentation.
}
\keyword{internal}
