% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_update_virtual_interface_attributes}
\alias{directconnect_update_virtual_interface_attributes}
\title{Updates the specified attributes of the specified virtual private
interface}
\usage{
directconnect_update_virtual_interface_attributes(
  virtualInterfaceId,
  mtu = NULL,
  enableSiteLink = NULL,
  virtualInterfaceName = NULL
)
}
\arguments{
\item{virtualInterfaceId}{[required] The ID of the virtual private interface.}

\item{mtu}{The maximum transmission unit (MTU), in bytes. The supported values are
1500 and 8500. The default value is 1500.}

\item{enableSiteLink}{Indicates whether to enable or disable SiteLink.}

\item{virtualInterfaceName}{The name of the virtual private interface.}
}
\description{
Updates the specified attributes of the specified virtual private interface.

See \url{https://www.paws-r-sdk.com/docs/directconnect_update_virtual_interface_attributes/} for full documentation.
}
\keyword{internal}
