% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elb_operations.R
\name{elb_add_tags}
\alias{elb_add_tags}
\title{Adds the specified tags to the specified load balancer}
\usage{
elb_add_tags(LoadBalancerNames, Tags)
}
\arguments{
\item{LoadBalancerNames}{[required] The name of the load balancer. You can specify one load balancer only.}

\item{Tags}{[required] The tags.}
}
\description{
Adds the specified tags to the specified load balancer. Each load balancer can have a maximum of 10 tags.

See \url{https://paws-r.github.io/docs/elb/add_tags.html} for full documentation.
}
\keyword{internal}
