% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elb_operations.R
\name{elb_describe_account_limits}
\alias{elb_describe_account_limits}
\title{Describes the current Elastic Load Balancing resource limits for your
AWS account}
\usage{
elb_describe_account_limits(Marker = NULL, PageSize = NULL)
}
\arguments{
\item{Marker}{The marker for the next set of results. (You received this marker from a
previous call.)}

\item{PageSize}{The maximum number of results to return with this call.}
}
\description{
Describes the current Elastic Load Balancing resource limits for your AWS account.

See \url{https://paws-r.github.io/docs/elb/describe_account_limits.html} for full documentation.
}
\keyword{internal}
