% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elb_operations.R
\name{elb_describe_load_balancer_attributes}
\alias{elb_describe_load_balancer_attributes}
\title{Describes the attributes for the specified load balancer}
\usage{
elb_describe_load_balancer_attributes(LoadBalancerName)
}
\arguments{
\item{LoadBalancerName}{[required] The name of the load balancer.}
}
\description{
Describes the attributes for the specified load balancer.

See \url{https://paws-r.github.io/docs/elb/describe_load_balancer_attributes.html} for full documentation.
}
\keyword{internal}
