% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elb_operations.R
\name{elb_describe_load_balancer_attributes}
\alias{elb_describe_load_balancer_attributes}
\title{Describes the attributes for the specified load balancer}
\usage{
elb_describe_load_balancer_attributes(LoadBalancerName)
}
\arguments{
\item{LoadBalancerName}{[required] The name of the load balancer.}
}
\description{
Describes the attributes for the specified load balancer.
}
\section{Request syntax}{
\preformatted{svc$describe_load_balancer_attributes(
  LoadBalancerName = "string"
)
}
}

\section{Examples}{
\preformatted{# This example describes the attributes of the specified load balancer.
svc$describe_load_balancer_attributes(
  LoadBalancerName = "my-load-balancer"
)
}
}

\keyword{internal}
