% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elb_operations.R
\name{elb_describe_load_balancer_policy_types}
\alias{elb_describe_load_balancer_policy_types}
\title{Describes the specified load balancer policy types or all load balancer
policy types}
\usage{
elb_describe_load_balancer_policy_types(PolicyTypeNames = NULL)
}
\arguments{
\item{PolicyTypeNames}{The names of the policy types. If no names are specified, describes all
policy types defined by Elastic Load Balancing.}
}
\description{
Describes the specified load balancer policy types or all load balancer policy types.

See \url{https://paws-r.github.io/docs/elb/describe_load_balancer_policy_types.html} for full documentation.
}
\keyword{internal}
