% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elb_operations.R
\name{elb_describe_load_balancers}
\alias{elb_describe_load_balancers}
\title{Describes the specified the load balancers}
\usage{
elb_describe_load_balancers(LoadBalancerNames, Marker, PageSize)
}
\arguments{
\item{LoadBalancerNames}{The names of the load balancers.}

\item{Marker}{The marker for the next set of results. (You received this marker from a
previous call.)}

\item{PageSize}{The maximum number of results to return with this call (a number from 1
to 400). The default is 400.}
}
\description{
Describes the specified the load balancers. If no load balancers are
specified, the call describes all of your load balancers.
}
\section{Request syntax}{
\preformatted{svc$describe_load_balancers(
  LoadBalancerNames = list(
    "string"
  ),
  Marker = "string",
  PageSize = 123
)
}
}

\examples{
# This example describes the specified load balancer.
\donttest{svc$describe_load_balancers(
  LoadBalancerNames = list(
    "my-load-balancer"
  )
)}

}
\keyword{internal}
