% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elb_operations.R
\name{elb_describe_tags}
\alias{elb_describe_tags}
\title{Describes the tags associated with the specified load balancers}
\usage{
elb_describe_tags(LoadBalancerNames)
}
\arguments{
\item{LoadBalancerNames}{[required] The names of the load balancers.}
}
\description{
Describes the tags associated with the specified load balancers.

See \url{https://paws-r.github.io/docs/elb/describe_tags.html} for full documentation.
}
\keyword{internal}
