% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elb_operations.R
\name{elb_describe_tags}
\alias{elb_describe_tags}
\title{Describes the tags associated with the specified load balancers}
\usage{
elb_describe_tags(LoadBalancerNames)
}
\arguments{
\item{LoadBalancerNames}{[required] The names of the load balancers.}
}
\description{
Describes the tags associated with the specified load balancers.
}
\section{Request syntax}{
\preformatted{svc$describe_tags(
  LoadBalancerNames = list(
    "string"
  )
)
}
}

\examples{
\dontrun{
# This example describes the tags for the specified load balancer.
svc$describe_tags(
  LoadBalancerNames = list(
    "my-load-balancer"
  )
)
}

}
\keyword{internal}
