% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elb_operations.R
\name{elb_detach_load_balancer_from_subnets}
\alias{elb_detach_load_balancer_from_subnets}
\title{Removes the specified subnets from the set of configured subnets for the
load balancer}
\usage{
elb_detach_load_balancer_from_subnets(LoadBalancerName, Subnets)
}
\arguments{
\item{LoadBalancerName}{[required] The name of the load balancer.}

\item{Subnets}{[required] The IDs of the subnets.}
}
\description{
Removes the specified subnets from the set of configured subnets for the load balancer.

See \url{https://paws-r.github.io/docs/elb/detach_load_balancer_from_subnets.html} for full documentation.
}
\keyword{internal}
