% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elb_operations.R
\name{elb_modify_load_balancer_attributes}
\alias{elb_modify_load_balancer_attributes}
\title{Modifies the attributes of the specified load balancer}
\usage{
elb_modify_load_balancer_attributes(LoadBalancerName, LoadBalancerAttributes)
}
\arguments{
\item{LoadBalancerName}{[required] The name of the load balancer.}

\item{LoadBalancerAttributes}{[required] The attributes for the load balancer.}
}
\description{
Modifies the attributes of the specified load balancer.

See \url{https://paws-r.github.io/docs/elb/modify_load_balancer_attributes.html} for full documentation.
}
\keyword{internal}
