% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elb_operations.R
\name{elb_set_load_balancer_listener_ssl_certificate}
\alias{elb_set_load_balancer_listener_ssl_certificate}
\title{Sets the certificate that terminates the specified listener's SSL
connections}
\usage{
elb_set_load_balancer_listener_ssl_certificate(
  LoadBalancerName,
  LoadBalancerPort,
  SSLCertificateId
)
}
\arguments{
\item{LoadBalancerName}{[required] The name of the load balancer.}

\item{LoadBalancerPort}{[required] The port that uses the specified SSL certificate.}

\item{SSLCertificateId}{[required] The Amazon Resource Name (ARN) of the SSL certificate.}
}
\description{
Sets the certificate that terminates the specified listener's SSL connections. The specified certificate replaces any prior certificate that was used on the same load balancer and port.

See \url{https://paws-r.github.io/docs/elb/set_load_balancer_listener_ssl_certificate.html} for full documentation.
}
\keyword{internal}
