% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbv2_operations.R
\name{elbv2_add_listener_certificates}
\alias{elbv2_add_listener_certificates}
\title{Adds the specified SSL server certificate to the certificate list for
the specified HTTPS or TLS listener}
\usage{
elbv2_add_listener_certificates(ListenerArn, Certificates)
}
\arguments{
\item{ListenerArn}{[required] The Amazon Resource Name (ARN) of the listener.}

\item{Certificates}{[required] The certificate to add. You can specify one certificate per call. Set
\code{CertificateArn} to the certificate ARN but do not set \code{IsDefault}.}
}
\description{
Adds the specified SSL server certificate to the certificate list for the specified HTTPS or TLS listener.

See \url{https://paws-r.github.io/docs/elbv2/add_listener_certificates.html} for full documentation.
}
\keyword{internal}
