% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbv2_operations.R
\name{elbv2_delete_listener}
\alias{elbv2_delete_listener}
\title{Deletes the specified listener}
\usage{
elbv2_delete_listener(ListenerArn)
}
\arguments{
\item{ListenerArn}{[required] The Amazon Resource Name (ARN) of the listener.}
}
\description{
Deletes the specified listener.

See \url{https://paws-r.github.io/docs/elbv2/delete_listener.html} for full documentation.
}
\keyword{internal}
