% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbv2_operations.R
\name{elbv2_describe_load_balancers}
\alias{elbv2_describe_load_balancers}
\title{Describes the specified load balancers or all of your load balancers}
\usage{
elbv2_describe_load_balancers(
  LoadBalancerArns = NULL,
  Names = NULL,
  Marker = NULL,
  PageSize = NULL
)
}
\arguments{
\item{LoadBalancerArns}{The Amazon Resource Names (ARN) of the load balancers. You can specify
up to 20 load balancers in a single call.}

\item{Names}{The names of the load balancers.}

\item{Marker}{The marker for the next set of results. (You received this marker from a
previous call.)}

\item{PageSize}{The maximum number of results to return with this call.}
}
\description{
Describes the specified load balancers or all of your load balancers.

See \url{https://www.paws-r-sdk.com/docs/elbv2_describe_load_balancers/} for full documentation.
}
\keyword{internal}
