% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbv2_operations.R
\name{elbv2_describe_rules}
\alias{elbv2_describe_rules}
\title{Describes the specified rules or the rules for the specified listener}
\usage{
elbv2_describe_rules(
  ListenerArn = NULL,
  RuleArns = NULL,
  Marker = NULL,
  PageSize = NULL
)
}
\arguments{
\item{ListenerArn}{The Amazon Resource Name (ARN) of the listener.}

\item{RuleArns}{The Amazon Resource Names (ARN) of the rules.}

\item{Marker}{The marker for the next set of results. (You received this marker from a
previous call.)}

\item{PageSize}{The maximum number of results to return with this call.}
}
\description{
Describes the specified rules or the rules for the specified listener. You must specify either a listener or rules.

See \url{https://www.paws-r-sdk.com/docs/elbv2_describe_rules/} for full documentation.
}
\keyword{internal}
