% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbv2_operations.R
\name{elbv2_describe_tags}
\alias{elbv2_describe_tags}
\title{Describes the tags for the specified Elastic Load Balancing resources}
\usage{
elbv2_describe_tags(ResourceArns)
}
\arguments{
\item{ResourceArns}{[required] The Amazon Resource Names (ARN) of the resources. You can specify up to
20 resources in a single call.}
}
\description{
Describes the tags for the specified Elastic Load Balancing resources. You can describe the tags for one or more Application Load Balancers, Network Load Balancers, Gateway Load Balancers, target groups, listeners, or rules.

See \url{https://www.paws-r-sdk.com/docs/elbv2_describe_tags/} for full documentation.
}
\keyword{internal}
