% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbv2_operations.R
\name{elbv2_describe_target_group_attributes}
\alias{elbv2_describe_target_group_attributes}
\title{Describes the attributes for the specified target group}
\usage{
elbv2_describe_target_group_attributes(TargetGroupArn)
}
\arguments{
\item{TargetGroupArn}{[required] The Amazon Resource Name (ARN) of the target group.}
}
\description{
Describes the attributes for the specified target group.

See \url{https://paws-r.github.io/docs/elbv2/describe_target_group_attributes.html} for full documentation.
}
\keyword{internal}
