% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbv2_operations.R
\name{elbv2_describe_target_groups}
\alias{elbv2_describe_target_groups}
\title{Describes the specified target groups or all of your target groups}
\usage{
elbv2_describe_target_groups(LoadBalancerArn, TargetGroupArns, Names,
  Marker, PageSize)
}
\arguments{
\item{LoadBalancerArn}{The Amazon Resource Name (ARN) of the load balancer.}

\item{TargetGroupArns}{The Amazon Resource Names (ARN) of the target groups.}

\item{Names}{The names of the target groups.}

\item{Marker}{The marker for the next set of results. (You received this marker from a
previous call.)}

\item{PageSize}{The maximum number of results to return with this call.}
}
\description{
Describes the specified target groups or all of your target groups. By
default, all target groups are described. Alternatively, you can specify
one of the following to filter the results: the ARN of the load
balancer, the names of one or more target groups, or the ARNs of one or
more target groups.
}
\details{
To describe the targets for a target group, use DescribeTargetHealth. To
describe the attributes of a target group, use
DescribeTargetGroupAttributes.
}
\section{Request syntax}{
\preformatted{svc$describe_target_groups(
  LoadBalancerArn = "string",
  TargetGroupArns = list(
    "string"
  ),
  Names = list(
    "string"
  ),
  Marker = "string",
  PageSize = 123
)
}
}

\examples{
\dontrun{
# This example describes the specified target group.
svc$describe_target_groups(
  TargetGroupArns = list(
    "arn:aws:elasticloadbalancing:us-west-2:123456789012:targetgroup/my-targets/73e2d6bc24d8..."
  )
)
}

}
\keyword{internal}
