% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbv2_operations.R
\name{elbv2_modify_load_balancer_attributes}
\alias{elbv2_modify_load_balancer_attributes}
\title{Modifies the specified attributes of the specified Application Load
Balancer, Network Load Balancer, or Gateway Load Balancer}
\usage{
elbv2_modify_load_balancer_attributes(LoadBalancerArn, Attributes)
}
\arguments{
\item{LoadBalancerArn}{[required] The Amazon Resource Name (ARN) of the load balancer.}

\item{Attributes}{[required] The load balancer attributes.}
}
\description{
Modifies the specified attributes of the specified Application Load Balancer, Network Load Balancer, or Gateway Load Balancer.

See \url{https://www.paws-r-sdk.com/docs/elbv2_modify_load_balancer_attributes/} for full documentation.
}
\keyword{internal}
