% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbv2_operations.R
\name{elbv2_modify_rule}
\alias{elbv2_modify_rule}
\title{Replaces the specified properties of the specified rule}
\usage{
elbv2_modify_rule(
  RuleArn,
  Conditions = NULL,
  Actions = NULL,
  Transforms = NULL,
  ResetTransforms = NULL
)
}
\arguments{
\item{RuleArn}{[required] The Amazon Resource Name (ARN) of the rule.}

\item{Conditions}{The conditions.}

\item{Actions}{The actions.}

\item{Transforms}{The transforms to apply to requests that match this rule. You can add
one host header rewrite transform and one URL rewrite transform. If you
specify \code{Transforms}, you can't specify \code{ResetTransforms}.}

\item{ResetTransforms}{Indicates whether to remove all transforms from the rule. If you specify
\code{ResetTransforms}, you can't specify \code{Transforms}.}
}
\description{
Replaces the specified properties of the specified rule. Any properties that you do not specify are unchanged.

See \url{https://www.paws-r-sdk.com/docs/elbv2_modify_rule/} for full documentation.
}
\keyword{internal}
