% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbv2_operations.R
\name{elbv2_modify_target_group_attributes}
\alias{elbv2_modify_target_group_attributes}
\title{Modifies the specified attributes of the specified target group}
\usage{
elbv2_modify_target_group_attributes(TargetGroupArn, Attributes)
}
\arguments{
\item{TargetGroupArn}{[required] The Amazon Resource Name (ARN) of the target group.}

\item{Attributes}{[required] The attributes.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Attributes = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Modifies the specified attributes of the specified target group.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$modify_target_group_attributes(
  TargetGroupArn = "string",
  Attributes = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}

\examples{
\dontrun{
# This example sets the deregistration delay timeout to the specified
# value for the specified target group.
svc$modify_target_group_attributes(
  Attributes = list(
    list(
      Key = "deregistration_delay.timeout_seconds",
      Value = "600"
    )
  ),
  TargetGroupArn = "arn:aws:elasticloadbalancing:us-west-2:123456789012:tar..."
)
}

}
\keyword{internal}
