% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbv2_operations.R
\name{elbv2_register_targets}
\alias{elbv2_register_targets}
\title{Registers the specified targets with the specified target group}
\usage{
elbv2_register_targets(TargetGroupArn, Targets)
}
\arguments{
\item{TargetGroupArn}{[required] The Amazon Resource Name (ARN) of the target group.}

\item{Targets}{[required] The targets.}
}
\description{
Registers the specified targets with the specified target group.

See \url{https://www.paws-r-sdk.com/docs/elbv2_register_targets/} for full documentation.
}
\keyword{internal}
