% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalaccelerator_operations.R
\name{globalaccelerator_advertise_byoip_cidr}
\alias{globalaccelerator_advertise_byoip_cidr}
\title{Advertises an IPv4 address range that is provisioned for use with your
AWS resources through bring your own IP addresses (BYOIP)}
\usage{
globalaccelerator_advertise_byoip_cidr(Cidr)
}
\arguments{
\item{Cidr}{[required] The address range, in CIDR notation. This must be the exact range that
you provisioned. You can't advertise only a portion of the provisioned
range.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ByoipCidr = list(
    Cidr = "string",
    State = "PENDING_PROVISIONING"|"READY"|"PENDING_ADVERTISING"|"ADVERTISING"|"PENDING_WITHDRAWING"|"PENDING_DEPROVISIONING"|"DEPROVISIONED"|"FAILED_PROVISION"|"FAILED_ADVERTISING"|"FAILED_WITHDRAW"|"FAILED_DEPROVISION",
    Events = list(
      list(
        Message = "string",
        Timestamp = as.POSIXct(
          "2015-01-01"
        )
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Advertises an IPv4 address range that is provisioned for use with your
AWS resources through bring your own IP addresses (BYOIP). It can take a
few minutes before traffic to the specified addresses starts routing to
AWS because of propagation delays.

To stop advertising the BYOIP address range, use
\href{https://docs.aws.amazon.com/global-accelerator/latest/api/}{WithdrawByoipCidr}.

For more information, see \href{https://docs.aws.amazon.com/global-accelerator/latest/dg/using-byoip.html}{Bring Your Own IP Addresses (BYOIP)}
in the \emph{AWS Global Accelerator Developer Guide}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$advertise_byoip_cidr(
  Cidr = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
