% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalaccelerator_operations.R
\name{globalaccelerator_delete_custom_routing_listener}
\alias{globalaccelerator_delete_custom_routing_listener}
\title{Delete a listener for a custom routing accelerator}
\usage{
globalaccelerator_delete_custom_routing_listener(ListenerArn)
}
\arguments{
\item{ListenerArn}{[required] The Amazon Resource Name (ARN) of the listener to delete.}
}
\value{
An empty list.
}
\description{
Delete a listener for a custom routing accelerator.
}
\section{Request syntax}{
\preformatted{svc$delete_custom_routing_listener(
  ListenerArn = "string"
)
}
}

\keyword{internal}
