% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalaccelerator_operations.R
\name{globalaccelerator_delete_endpoint_group}
\alias{globalaccelerator_delete_endpoint_group}
\title{Delete an endpoint group from a listener}
\usage{
globalaccelerator_delete_endpoint_group(EndpointGroupArn)
}
\arguments{
\item{EndpointGroupArn}{[required] The Amazon Resource Name (ARN) of the endpoint group to delete.}
}
\value{
An empty list.
}
\description{
Delete an endpoint group from a listener.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_endpoint_group(
  EndpointGroupArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
