% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalaccelerator_operations.R
\name{globalaccelerator_delete_listener}
\alias{globalaccelerator_delete_listener}
\title{Delete a listener from an accelerator}
\usage{
globalaccelerator_delete_listener(ListenerArn)
}
\arguments{
\item{ListenerArn}{[required] The Amazon Resource Name (ARN) of the listener.}
}
\description{
Delete a listener from an accelerator.

See \url{https://paws-r.github.io/docs/globalaccelerator/delete_listener.html} for full documentation.
}
\keyword{internal}
