% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalaccelerator_operations.R
\name{globalaccelerator_describe_accelerator_attributes}
\alias{globalaccelerator_describe_accelerator_attributes}
\title{Describe the attributes of an accelerator}
\usage{
globalaccelerator_describe_accelerator_attributes(AcceleratorArn)
}
\arguments{
\item{AcceleratorArn}{[required] The Amazon Resource Name (ARN) of the accelerator with the attributes
that you want to describe.}
}
\description{
Describe the attributes of an accelerator.

See \url{https://www.paws-r-sdk.com/docs/globalaccelerator_describe_accelerator_attributes/} for full documentation.
}
\keyword{internal}
