% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalaccelerator_operations.R
\name{globalaccelerator_describe_accelerator_attributes}
\alias{globalaccelerator_describe_accelerator_attributes}
\title{Describe the attributes of an accelerator}
\usage{
globalaccelerator_describe_accelerator_attributes(AcceleratorArn)
}
\arguments{
\item{AcceleratorArn}{[required] The Amazon Resource Name (ARN) of the accelerator with the attributes
that you want to describe.}
}
\value{
A list with the following syntax:\preformatted{list(
  AcceleratorAttributes = list(
    FlowLogsEnabled = TRUE|FALSE,
    FlowLogsS3Bucket = "string",
    FlowLogsS3Prefix = "string"
  )
)
}
}
\description{
Describe the attributes of an accelerator.
}
\section{Request syntax}{
\preformatted{svc$describe_accelerator_attributes(
  AcceleratorArn = "string"
)
}
}

\keyword{internal}
