% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalaccelerator_operations.R
\name{globalaccelerator_describe_custom_routing_listener}
\alias{globalaccelerator_describe_custom_routing_listener}
\title{The description of a listener for a custom routing accelerator}
\usage{
globalaccelerator_describe_custom_routing_listener(ListenerArn)
}
\arguments{
\item{ListenerArn}{[required] The Amazon Resource Name (ARN) of the listener to describe.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Listener = list(
    ListenerArn = "string",
    PortRanges = list(
      list(
        FromPort = 123,
        ToPort = 123
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
The description of a listener for a custom routing accelerator.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_custom_routing_listener(
  ListenerArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
