% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalaccelerator_operations.R
\name{globalaccelerator_list_accelerators}
\alias{globalaccelerator_list_accelerators}
\title{List the accelerators for an Amazon Web Services account}
\usage{
globalaccelerator_list_accelerators(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{The number of Global Accelerator objects that you want to return with
this call. The default value is 10.}

\item{NextToken}{The token for the next set of results. You receive this token from a
previous call.}
}
\description{
List the accelerators for an Amazon Web Services account.

See \url{https://www.paws-r-sdk.com/docs/globalaccelerator_list_accelerators/} for full documentation.
}
\keyword{internal}
