% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalaccelerator_operations.R
\name{globalaccelerator_list_accelerators}
\alias{globalaccelerator_list_accelerators}
\title{List the accelerators for an AWS account}
\usage{
globalaccelerator_list_accelerators(MaxResults, NextToken)
}
\arguments{
\item{MaxResults}{The number of Global Accelerator objects that you want to return with
this call. The default value is 10.}

\item{NextToken}{The token for the next set of results. You receive this token from a
previous call.}
}
\description{
List the accelerators for an AWS account. To see an AWS CLI example of
listing the accelerators for an AWS account, scroll down to \strong{Example}.
}
\section{Request syntax}{
\preformatted{svc$list_accelerators(
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
