% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalaccelerator_operations.R
\name{globalaccelerator_list_byoip_cidrs}
\alias{globalaccelerator_list_byoip_cidrs}
\title{Lists the IP address ranges that were specified in calls to
ProvisionByoipCidr, including the current state and a history of state
changes}
\usage{
globalaccelerator_list_byoip_cidrs(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{The maximum number of results to return with a single call. To retrieve
the remaining results, make another call with the returned \code{nextToken}
value.}

\item{NextToken}{The token for the next page of results.}
}
\description{
Lists the IP address ranges that were specified in calls to \code{\link[=globalaccelerator_provision_byoip_cidr]{provision_byoip_cidr}}, including the current state and a history of state changes.

See \url{https://paws-r.github.io/docs/globalaccelerator/list_byoip_cidrs.html} for full documentation.
}
\keyword{internal}
