% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalaccelerator_operations.R
\name{globalaccelerator_list_custom_routing_port_mappings}
\alias{globalaccelerator_list_custom_routing_port_mappings}
\title{Provides a complete mapping from the public accelerator IP address and
port to destination EC2 instance IP addresses and ports in the virtual
public cloud (VPC) subnet endpoint for a custom routing accelerator}
\usage{
globalaccelerator_list_custom_routing_port_mappings(AcceleratorArn,
  EndpointGroupArn, MaxResults, NextToken)
}
\arguments{
\item{AcceleratorArn}{[required] The Amazon Resource Name (ARN) of the accelerator to list the custom
routing port mappings for.}

\item{EndpointGroupArn}{The Amazon Resource Name (ARN) of the endpoint group to list the custom
routing port mappings for.}

\item{MaxResults}{The number of destination port mappings that you want to return with
this call. The default value is 10.}

\item{NextToken}{The token for the next set of results. You receive this token from a
previous call.}
}
\description{
Provides a complete mapping from the public accelerator IP address and
port to destination EC2 instance IP addresses and ports in the virtual
public cloud (VPC) subnet endpoint for a custom routing accelerator. For
each subnet endpoint that you add, Global Accelerator creates a new
static port mapping for the accelerator. The port mappings don't change
after Global Accelerator generates them, so you can retrieve and cache
the full mapping on your servers.
}
\details{
If you remove a subnet from your accelerator, Global Accelerator removes
(reclaims) the port mappings. If you add a subnet to your accelerator,
Global Accelerator creates new port mappings (the existing ones don't
change). If you add or remove EC2 instances in your subnet, the port
mappings don't change, because the mappings are created when you add the
subnet to Global Accelerator.

The mappings also include a flag for each destination denoting which
destination IP addresses and ports are allowed or denied traffic.
}
\section{Request syntax}{
\preformatted{svc$list_custom_routing_port_mappings(
  AcceleratorArn = "string",
  EndpointGroupArn = "string",
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
