% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalaccelerator_operations.R
\name{globalaccelerator_list_endpoint_groups}
\alias{globalaccelerator_list_endpoint_groups}
\title{List the endpoint groups that are associated with a listener}
\usage{
globalaccelerator_list_endpoint_groups(ListenerArn, MaxResults,
  NextToken)
}
\arguments{
\item{ListenerArn}{[required] The Amazon Resource Name (ARN) of the listener.}

\item{MaxResults}{The number of endpoint group objects that you want to return with this
call. The default value is 10.}

\item{NextToken}{The token for the next set of results. You receive this token from a
previous call.}
}
\description{
List the endpoint groups that are associated with a listener.
}
\section{Request syntax}{
\preformatted{svc$list_endpoint_groups(
  ListenerArn = "string",
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
