% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalaccelerator_operations.R
\name{globalaccelerator_list_tags_for_resource}
\alias{globalaccelerator_list_tags_for_resource}
\title{List all tags for an accelerator}
\usage{
globalaccelerator_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the accelerator to list tags for. An
ARN uniquely identifies an accelerator.}
}
\description{
List all tags for an accelerator.

See \url{https://paws-r.github.io/docs/globalaccelerator/list_tags_for_resource.html} for full documentation.
}
\keyword{internal}
