% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalaccelerator_operations.R
\name{globalaccelerator_remove_custom_routing_endpoints}
\alias{globalaccelerator_remove_custom_routing_endpoints}
\title{Remove endpoints from a custom routing accelerator}
\usage{
globalaccelerator_remove_custom_routing_endpoints(
  EndpointIds,
  EndpointGroupArn
)
}
\arguments{
\item{EndpointIds}{[required] The IDs for the endpoints. For custom routing accelerators, endpoint IDs
are the virtual private cloud (VPC) subnet IDs.}

\item{EndpointGroupArn}{[required] The Amazon Resource Name (ARN) of the endpoint group to remove endpoints
from.}
}
\description{
Remove endpoints from a custom routing accelerator.

See \url{https://www.paws-r-sdk.com/docs/globalaccelerator_remove_custom_routing_endpoints/} for full documentation.
}
\keyword{internal}
