% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalaccelerator_operations.R
\name{globalaccelerator_update_accelerator}
\alias{globalaccelerator_update_accelerator}
\title{Update an accelerator to make changes, such as the following:}
\usage{
globalaccelerator_update_accelerator(
  AcceleratorArn,
  Name = NULL,
  IpAddressType = NULL,
  Enabled = NULL
)
}
\arguments{
\item{AcceleratorArn}{[required] The Amazon Resource Name (ARN) of the accelerator to update.}

\item{Name}{The name of the accelerator. The name can have a maximum of 64
characters, must contain only alphanumeric characters, periods (.), or
hyphens (-), and must not begin or end with a hyphen or period.}

\item{IpAddressType}{The IP address type that an accelerator supports. For a standard
accelerator, the value can be IPV4 or DUAL_STACK.}

\item{Enabled}{Indicates whether an accelerator is enabled. The value is true or false.
The default value is true.

If the value is set to true, the accelerator cannot be deleted. If set
to false, the accelerator can be deleted.}
}
\description{
Update an accelerator to make changes, such as the following:

See \url{https://www.paws-r-sdk.com/docs/globalaccelerator_update_accelerator/} for full documentation.
}
\keyword{internal}
