% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalaccelerator_operations.R
\name{globalaccelerator_update_custom_routing_accelerator_attributes}
\alias{globalaccelerator_update_custom_routing_accelerator_attributes}
\title{Update the attributes for a custom routing accelerator}
\usage{
globalaccelerator_update_custom_routing_accelerator_attributes(
  AcceleratorArn, FlowLogsEnabled, FlowLogsS3Bucket, FlowLogsS3Prefix)
}
\arguments{
\item{AcceleratorArn}{[required] The Amazon Resource Name (ARN) of the custom routing accelerator to
update attributes for.}

\item{FlowLogsEnabled}{Update whether flow logs are enabled. The default value is false. If the
value is true, \code{FlowLogsS3Bucket} and \code{FlowLogsS3Prefix} must be
specified.

For more information, see \href{https://docs.aws.amazon.com/global-accelerator/latest/dg/monitoring-global-accelerator.flow-logs.html}{Flow Logs}
in the \emph{AWS Global Accelerator Developer Guide}.}

\item{FlowLogsS3Bucket}{The name of the Amazon S3 bucket for the flow logs. Attribute is
required if \code{FlowLogsEnabled} is \code{true}. The bucket must exist and have
a bucket policy that grants AWS Global Accelerator permission to write
to the bucket.}

\item{FlowLogsS3Prefix}{Update the prefix for the location in the Amazon S3 bucket for the flow
logs. Attribute is required if \code{FlowLogsEnabled} is \code{true}.

If you don’t specify a prefix, the flow logs are stored in the root of
the bucket. If you specify slash (/) for the S3 bucket prefix, the log
file bucket folder structure will include a double slash (//), like the
following:

DOC-EXAMPLE-BUCKET//AWSLogs/aws_account_id}
}
\value{
A list with the following syntax:\preformatted{list(
  AcceleratorAttributes = list(
    FlowLogsEnabled = TRUE|FALSE,
    FlowLogsS3Bucket = "string",
    FlowLogsS3Prefix = "string"
  )
)
}
}
\description{
Update the attributes for a custom routing accelerator.
}
\section{Request syntax}{
\preformatted{svc$update_custom_routing_accelerator_attributes(
  AcceleratorArn = "string",
  FlowLogsEnabled = TRUE|FALSE,
  FlowLogsS3Bucket = "string",
  FlowLogsS3Prefix = "string"
)
}
}

\keyword{internal}
