% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalaccelerator_operations.R
\name{globalaccelerator_update_endpoint_group}
\alias{globalaccelerator_update_endpoint_group}
\title{Update an endpoint group}
\usage{
globalaccelerator_update_endpoint_group(EndpointGroupArn,
  EndpointConfigurations, TrafficDialPercentage, HealthCheckPort,
  HealthCheckProtocol, HealthCheckPath, HealthCheckIntervalSeconds,
  ThresholdCount, PortOverrides)
}
\arguments{
\item{EndpointGroupArn}{[required] The Amazon Resource Name (ARN) of the endpoint group.}

\item{EndpointConfigurations}{The list of endpoint objects. A resource must be valid and active when
you add it as an endpoint.}

\item{TrafficDialPercentage}{The percentage of traffic to send to an AWS Region. Additional traffic
is distributed to other endpoint groups for this listener.

Use this action to increase (dial up) or decrease (dial down) traffic to
a specific Region. The percentage is applied to the traffic that would
otherwise have been routed to the Region based on optimal routing.

The default value is 100.}

\item{HealthCheckPort}{The port that AWS Global Accelerator uses to check the health of
endpoints that are part of this endpoint group. The default port is the
listener port that this endpoint group is associated with. If the
listener port is a list of ports, Global Accelerator uses the first port
in the list.}

\item{HealthCheckProtocol}{The protocol that AWS Global Accelerator uses to check the health of
endpoints that are part of this endpoint group. The default value is
TCP.}

\item{HealthCheckPath}{If the protocol is HTTP/S, then this specifies the path that is the
destination for health check targets. The default value is slash (/).}

\item{HealthCheckIntervalSeconds}{The time—10 seconds or 30 seconds—between each health check for an
endpoint. The default value is 30.}

\item{ThresholdCount}{The number of consecutive health checks required to set the state of a
healthy endpoint to unhealthy, or to set an unhealthy endpoint to
healthy. The default value is 3.}

\item{PortOverrides}{Override specific listener ports used to route traffic to endpoints that
are part of this endpoint group. For example, you can create a port
override in which the listener receives user traffic on ports 80 and
443, but your accelerator routes that traffic to ports 1080 and 1443,
respectively, on the endpoints.

For more information, see \href{https://docs.aws.amazon.com/global-accelerator/latest/dg/about-endpoint-groups-port-override.html}{Port overrides}
in the \emph{AWS Global Accelerator Developer Guide}.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  EndpointGroup = list(
    EndpointGroupArn = "string",
    EndpointGroupRegion = "string",
    EndpointDescriptions = list(
      list(
        EndpointId = "string",
        Weight = 123,
        HealthState = "INITIAL"|"HEALTHY"|"UNHEALTHY",
        HealthReason = "string",
        ClientIPPreservationEnabled = TRUE|FALSE
      )
    ),
    TrafficDialPercentage = 123.0,
    HealthCheckPort = 123,
    HealthCheckProtocol = "TCP"|"HTTP"|"HTTPS",
    HealthCheckPath = "string",
    HealthCheckIntervalSeconds = 123,
    ThresholdCount = 123,
    PortOverrides = list(
      list(
        ListenerPort = 123,
        EndpointPort = 123
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Update an endpoint group. A resource must be valid and active when you
add it as an endpoint.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_endpoint_group(
  EndpointGroupArn = "string",
  EndpointConfigurations = list(
    list(
      EndpointId = "string",
      Weight = 123,
      ClientIPPreservationEnabled = TRUE|FALSE
    )
  ),
  TrafficDialPercentage = 123.0,
  HealthCheckPort = 123,
  HealthCheckProtocol = "TCP"|"HTTP"|"HTTPS",
  HealthCheckPath = "string",
  HealthCheckIntervalSeconds = 123,
  ThresholdCount = 123,
  PortOverrides = list(
    list(
      ListenerPort = 123,
      EndpointPort = 123
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
