% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalaccelerator_operations.R
\name{globalaccelerator_withdraw_byoip_cidr}
\alias{globalaccelerator_withdraw_byoip_cidr}
\title{Stops advertising an address range that is provisioned as an address
pool}
\usage{
globalaccelerator_withdraw_byoip_cidr(Cidr)
}
\arguments{
\item{Cidr}{[required] The address range, in CIDR notation.

For more information, see \href{https://docs.aws.amazon.com/global-accelerator/latest/dg/using-byoip.html}{Bring your own IP addresses (BYOIP)}
in the Global Accelerator Developer Guide.}
}
\description{
Stops advertising an address range that is provisioned as an address pool. You can perform this operation at most once every 10 seconds, even if you specify different address ranges each time.

See \url{https://www.paws-r-sdk.com/docs/globalaccelerator_withdraw_byoip_cidr/} for full documentation.
}
\keyword{internal}
