% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkfirewall_operations.R
\name{networkfirewall_update_logging_configuration}
\alias{networkfirewall_update_logging_configuration}
\title{Sets the logging configuration for the specified firewall}
\usage{
networkfirewall_update_logging_configuration(
  FirewallArn = NULL,
  FirewallName = NULL,
  LoggingConfiguration = NULL,
  EnableMonitoringDashboard = NULL
)
}
\arguments{
\item{FirewallArn}{The Amazon Resource Name (ARN) of the firewall.

You must specify the ARN or the name, and you can specify both.}

\item{FirewallName}{The descriptive name of the firewall. You can't change the name of a
firewall after you create it.

You must specify the ARN or the name, and you can specify both.}

\item{LoggingConfiguration}{Defines how Network Firewall performs logging for a firewall. If you
omit this setting, Network Firewall disables logging for the firewall.}

\item{EnableMonitoringDashboard}{A boolean that lets you enable or disable the detailed firewall
monitoring dashboard on the firewall.

The monitoring dashboard provides comprehensive visibility into your
firewall's flow logs and alert logs. After you enable detailed
monitoring, you can access these dashboards directly from the
\strong{Monitoring} page of the Network Firewall console.

Specify \code{TRUE} to enable the the detailed monitoring dashboard on the
firewall. Specify \code{FALSE} to disable the the detailed monitoring
dashboard on the firewall.}
}
\description{
Sets the logging configuration for the specified firewall.

See \url{https://www.paws-r-sdk.com/docs/networkfirewall_update_logging_configuration/} for full documentation.
}
\keyword{internal}
