% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_create_site_to_site_vpn_attachment}
\alias{networkmanager_create_site_to_site_vpn_attachment}
\title{Creates an Amazon Web Services site-to-site VPN attachment on an edge
location of a core network}
\usage{
networkmanager_create_site_to_site_vpn_attachment(
  CoreNetworkId,
  VpnConnectionArn,
  RoutingPolicyLabel = NULL,
  Tags = NULL,
  ClientToken = NULL
)
}
\arguments{
\item{CoreNetworkId}{[required] The ID of a core network where you're creating a site-to-site VPN
attachment.}

\item{VpnConnectionArn}{[required] The ARN identifying the VPN attachment.}

\item{RoutingPolicyLabel}{The routing policy label to apply to the Site-to-Site VPN attachment for
traffic routing decisions.}

\item{Tags}{The tags associated with the request.}

\item{ClientToken}{The client token associated with the request.}
}
\description{
Creates an Amazon Web Services site-to-site VPN attachment on an edge location of a core network.

See \url{https://www.paws-r-sdk.com/docs/networkmanager_create_site_to_site_vpn_attachment/} for full documentation.
}
\keyword{internal}
