% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_create_transit_gateway_route_table_attachment}
\alias{networkmanager_create_transit_gateway_route_table_attachment}
\title{Creates a transit gateway route table attachment}
\usage{
networkmanager_create_transit_gateway_route_table_attachment(
  PeeringId,
  TransitGatewayRouteTableArn,
  RoutingPolicyLabel = NULL,
  Tags = NULL,
  ClientToken = NULL
)
}
\arguments{
\item{PeeringId}{[required] The ID of the peer for the}

\item{TransitGatewayRouteTableArn}{[required] The ARN of the transit gateway route table for the attachment request.
For example,
\code{"TransitGatewayRouteTableArn": "arn:aws:ec2:us-west-2:123456789012:transit-gateway-route-table/tgw-rtb-9876543210123456"}.}

\item{RoutingPolicyLabel}{The routing policy label to apply to the Transit Gateway route table
attachment for traffic routing decisions.}

\item{Tags}{The list of key-value tags associated with the request.}

\item{ClientToken}{The client token associated with the request.}
}
\description{
Creates a transit gateway route table attachment.

See \url{https://www.paws-r-sdk.com/docs/networkmanager_create_transit_gateway_route_table_attachment/} for full documentation.
}
\keyword{internal}
