% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_execute_core_network_change_set}
\alias{networkmanager_execute_core_network_change_set}
\title{Executes a change set on your core network}
\usage{
networkmanager_execute_core_network_change_set(CoreNetworkId, PolicyVersionId)
}
\arguments{
\item{CoreNetworkId}{[required] The ID of a core network.}

\item{PolicyVersionId}{[required] The ID of the policy version.}
}
\description{
Executes a change set on your core network. Deploys changes globally based on the policy submitted..

See \url{https://paws-r.github.io/docs/networkmanager/execute_core_network_change_set.html} for full documentation.
}
\keyword{internal}
