% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_get_connect_attachment}
\alias{networkmanager_get_connect_attachment}
\title{Returns information about a core network Connect attachment}
\usage{
networkmanager_get_connect_attachment(AttachmentId)
}
\arguments{
\item{AttachmentId}{[required] The ID of the attachment.}
}
\description{
Returns information about a core network Connect attachment.

See \url{https://paws-r.github.io/docs/networkmanager/get_connect_attachment.html} for full documentation.
}
\keyword{internal}
