% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_get_connect_peer}
\alias{networkmanager_get_connect_peer}
\title{Returns information about a core network Connect peer}
\usage{
networkmanager_get_connect_peer(ConnectPeerId)
}
\arguments{
\item{ConnectPeerId}{[required] The ID of the Connect peer.}
}
\description{
Returns information about a core network Connect peer.

See \url{https://www.paws-r-sdk.com/docs/networkmanager_get_connect_peer/} for full documentation.
}
\keyword{internal}
