% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmanager_operations.R
\name{networkmanager_get_route_analysis}
\alias{networkmanager_get_route_analysis}
\title{Gets information about the specified route analysis}
\usage{
networkmanager_get_route_analysis(GlobalNetworkId, RouteAnalysisId)
}
\arguments{
\item{GlobalNetworkId}{[required] The ID of the global network.}

\item{RouteAnalysisId}{[required] The ID of the route analysis.}
}
\description{
Gets information about the specified route analysis.

See \url{https://www.paws-r-sdk.com/docs/networkmanager_get_route_analysis/} for full documentation.
}
\keyword{internal}
