% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53_operations.R
\name{route53_delete_reusable_delegation_set}
\alias{route53_delete_reusable_delegation_set}
\title{Deletes a reusable delegation set}
\usage{
route53_delete_reusable_delegation_set(Id)
}
\arguments{
\item{Id}{[required] The ID of the reusable delegation set that you want to delete.}
}
\value{
An empty list.
}
\description{
Deletes a reusable delegation set.

You can delete a reusable delegation set only if it isn't associated
with any hosted zones.

To verify that the reusable delegation set is not associated with any
hosted zones, submit a
\code{\link[=route53_get_reusable_delegation_set]{get_reusable_delegation_set}}
request and specify the ID of the reusable delegation set that you want
to delete.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_reusable_delegation_set(
  Id = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
