% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53_operations.R
\name{route53_enable_hosted_zone_dnssec}
\alias{route53_enable_hosted_zone_dnssec}
\title{Enables DNSSEC signing in a specific hosted zone}
\usage{
route53_enable_hosted_zone_dnssec(HostedZoneId)
}
\arguments{
\item{HostedZoneId}{[required] A unique string used to identify a hosted zone.}
}
\description{
Enables DNSSEC signing in a specific hosted zone.

See \url{https://paws-r.github.io/docs/route53/enable_hosted_zone_dnssec.html} for full documentation.
}
\keyword{internal}
