% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53_operations.R
\name{route53_get_hosted_zone_count}
\alias{route53_get_hosted_zone_count}
\title{Retrieves the number of hosted zones that are associated with the
current AWS account}
\usage{
route53_get_hosted_zone_count()
}
\value{
A list with the following syntax:\preformatted{list(
  HostedZoneCount = 123
)
}
}
\description{
Retrieves the number of hosted zones that are associated with the
current AWS account.
}
\section{Request syntax}{
\preformatted{svc$get_hosted_zone_count()
}
}

\keyword{internal}
