% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53_operations.R
\name{route53_get_reusable_delegation_set_limit}
\alias{route53_get_reusable_delegation_set_limit}
\title{Gets the maximum number of hosted zones that you can associate with the
specified reusable delegation set}
\usage{
route53_get_reusable_delegation_set_limit(Type, DelegationSetId)
}
\arguments{
\item{Type}{[required] Specify \code{MAX_ZONES_BY_REUSABLE_DELEGATION_SET} to get the maximum number
of hosted zones that you can associate with the specified reusable
delegation set.}

\item{DelegationSetId}{[required] The ID of the delegation set that you want to get the limit for.}
}
\description{
Gets the maximum number of hosted zones that you can associate with the specified reusable delegation set.

See \url{https://www.paws-r-sdk.com/docs/route53_get_reusable_delegation_set_limit/} for full documentation.
}
\keyword{internal}
