% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53_operations.R
\name{route53_update_hosted_zone_comment}
\alias{route53_update_hosted_zone_comment}
\title{Updates the comment for a specified hosted zone}
\usage{
route53_update_hosted_zone_comment(Id, Comment = NULL)
}
\arguments{
\item{Id}{[required] The ID for the hosted zone that you want to update the comment for.}

\item{Comment}{The new comment for the hosted zone. If you don't specify a value for
\code{Comment}, Amazon Route 53 deletes the existing value of the \code{Comment}
element, if any.}
}
\description{
Updates the comment for a specified hosted zone.

See \url{https://paws-r.github.io/docs/route53/update_hosted_zone_comment.html} for full documentation.
}
\keyword{internal}
