% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53domains_operations.R
\name{route53domains_cancel_domain_transfer_to_another_aws_account}
\alias{route53domains_cancel_domain_transfer_to_another_aws_account}
\title{Cancels the transfer of a domain from the current Amazon Web Services
account to another Amazon Web Services account}
\usage{
route53domains_cancel_domain_transfer_to_another_aws_account(DomainName)
}
\arguments{
\item{DomainName}{[required] The name of the domain for which you want to cancel the transfer to
another Amazon Web Services account.}
}
\description{
Cancels the transfer of a domain from the current Amazon Web Services account to another Amazon Web Services account. You initiate a transfer betweenAmazon Web Services accounts using \code{\link[=route53domains_transfer_domain_to_another_aws_account]{transfer_domain_to_another_aws_account}}.

See \url{https://paws-r.github.io/docs/route53domains/cancel_domain_transfer_to_another_aws_account.html} for full documentation.
}
\keyword{internal}
