% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53domains_operations.R
\name{route53domains_enable_domain_auto_renew}
\alias{route53domains_enable_domain_auto_renew}
\title{This operation configures Amazon Route 53 to automatically renew the
specified domain before the domain registration expires}
\usage{
route53domains_enable_domain_auto_renew(DomainName)
}
\arguments{
\item{DomainName}{[required] The name of the domain that you want to enable automatic renewal for.}
}
\description{
This operation configures Amazon Route 53 to automatically renew the specified domain before the domain registration expires. The cost of renewing your domain registration is billed to your Amazon Web Services account.

See \url{https://paws-r.github.io/docs/route53domains/enable_domain_auto_renew.html} for full documentation.
}
\keyword{internal}
