% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53domains_operations.R
\name{route53domains_get_domain_detail}
\alias{route53domains_get_domain_detail}
\title{This operation returns detailed information about a specified domain
that is associated with the current Amazon Web Services account}
\usage{
route53domains_get_domain_detail(DomainName)
}
\arguments{
\item{DomainName}{[required] The name of the domain that you want to get detailed information about.}
}
\description{
This operation returns detailed information about a specified domain that is associated with the current Amazon Web Services account. Contact information for the domain is also returned as part of the output.

See \url{https://www.paws-r-sdk.com/docs/route53domains_get_domain_detail/} for full documentation.
}
\keyword{internal}
