% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53domains_operations.R
\name{route53domains_get_operation_detail}
\alias{route53domains_get_operation_detail}
\title{This operation returns the current status of an operation that is not
completed}
\usage{
route53domains_get_operation_detail(OperationId)
}
\arguments{
\item{OperationId}{[required] The identifier for the operation for which you want to get the status.
Route 53 returned the identifier in the response to the original
request.}
}
\description{
This operation returns the current status of an operation that is not completed.

See \url{https://www.paws-r-sdk.com/docs/route53domains_get_operation_detail/} for full documentation.
}
\keyword{internal}
