% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53domains_operations.R
\name{route53domains_list_operations}
\alias{route53domains_list_operations}
\title{Returns information about all of the operations that return an operation
ID and that have ever been performed on domains that were registered by
the current account}
\usage{
route53domains_list_operations(
  SubmittedSince = NULL,
  Marker = NULL,
  MaxItems = NULL,
  Status = NULL,
  Type = NULL,
  SortBy = NULL,
  SortOrder = NULL
)
}
\arguments{
\item{SubmittedSince}{An optional parameter that lets you get information about all the
operations that you submitted after a specified date and time. Specify
the date and time in Unix time format and Coordinated Universal time
(UTC).}

\item{Marker}{For an initial request for a list of operations, omit this element. If
the number of operations that are not yet complete is greater than the
value that you specified for \code{MaxItems}, you can use \code{Marker} to return
additional operations. Get the value of \code{NextPageMarker} from the
previous response, and submit another request that includes the value of
\code{NextPageMarker} in the \code{Marker} element.}

\item{MaxItems}{Number of domains to be returned.

Default: 20}

\item{Status}{The status of the operations.}

\item{Type}{An arrays of the domains operation types.}

\item{SortBy}{The sort type for returned values.}

\item{SortOrder}{The sort order for returned values, either ascending or descending.}
}
\description{
Returns information about all of the operations that return an operation ID and that have ever been performed on domains that were registered by the current account.

See \url{https://www.paws-r-sdk.com/docs/route53domains_list_operations/} for full documentation.
}
\keyword{internal}
