% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53domains_operations.R
\name{route53domains_list_tags_for_domain}
\alias{route53domains_list_tags_for_domain}
\title{This operation returns all of the tags that are associated with the
specified domain}
\usage{
route53domains_list_tags_for_domain(DomainName)
}
\arguments{
\item{DomainName}{[required] The domain for which you want to get a list of tags.}
}
\description{
This operation returns all of the tags that are associated with the specified domain.

See \url{https://paws-r.github.io/docs/route53domains/list_tags_for_domain.html} for full documentation.
}
\keyword{internal}
