% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53domains_operations.R
\name{route53domains_update_tags_for_domain}
\alias{route53domains_update_tags_for_domain}
\title{This operation adds or updates tags for a specified domain}
\usage{
route53domains_update_tags_for_domain(DomainName, TagsToUpdate = NULL)
}
\arguments{
\item{DomainName}{[required] The domain for which you want to add or update tags.}

\item{TagsToUpdate}{A list of the tag keys and values that you want to add or update. If you
specify a key that already exists, the corresponding value will be
replaced.}
}
\description{
This operation adds or updates tags for a specified domain.

See \url{https://www.paws-r-sdk.com/docs/route53domains_update_tags_for_domain/} for full documentation.
}
\keyword{internal}
