% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53recoverycontrolconfig_operations.R
\name{route53recoverycontrolconfig_delete_cluster}
\alias{route53recoverycontrolconfig_delete_cluster}
\title{Delete a cluster}
\usage{
route53recoverycontrolconfig_delete_cluster(ClusterArn)
}
\arguments{
\item{ClusterArn}{[required] The Amazon Resource Name (ARN) of the cluster that you're deleting.}
}
\description{
Delete a cluster.

See \url{https://paws-r.github.io/docs/route53recoverycontrolconfig/delete_cluster.html} for full documentation.
}
\keyword{internal}
