% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53recoverycontrolconfig_operations.R
\name{route53recoverycontrolconfig_describe_routing_control}
\alias{route53recoverycontrolconfig_describe_routing_control}
\title{Displays details about a routing control}
\usage{
route53recoverycontrolconfig_describe_routing_control(RoutingControlArn)
}
\arguments{
\item{RoutingControlArn}{[required] The Amazon Resource Name (ARN) of the routing control.}
}
\description{
Displays details about a routing control. A routing control has one of two states: ON and OFF. You can map the routing control state to the state of an Amazon Route 53 health check, which can be used to control routing.

See \url{https://www.paws-r-sdk.com/docs/route53recoverycontrolconfig_describe_routing_control/} for full documentation.
}
\keyword{internal}
