% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53recoverycontrolconfig_operations.R
\name{route53recoverycontrolconfig_list_tags_for_resource}
\alias{route53recoverycontrolconfig_list_tags_for_resource}
\title{Lists the tags for a resource}
\usage{
route53recoverycontrolconfig_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) for the resource that's tagged.}
}
\description{
Lists the tags for a resource.

See \url{https://paws-r.github.io/docs/route53recoverycontrolconfig/list_tags_for_resource.html} for full documentation.
}
\keyword{internal}
