% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53recoveryreadiness_operations.R
\name{route53recoveryreadiness_delete_cross_account_authorization}
\alias{route53recoveryreadiness_delete_cross_account_authorization}
\title{Deletes cross account readiness authorization}
\usage{
route53recoveryreadiness_delete_cross_account_authorization(
  CrossAccountAuthorization
)
}
\arguments{
\item{CrossAccountAuthorization}{[required] The cross-account authorization.}
}
\description{
Deletes cross account readiness authorization.

See \url{https://paws-r.github.io/docs/route53recoveryreadiness/delete_cross_account_authorization.html} for full documentation.
}
\keyword{internal}
