% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53recoveryreadiness_operations.R
\name{route53recoveryreadiness_delete_resource_set}
\alias{route53recoveryreadiness_delete_resource_set}
\title{Deletes a resource set}
\usage{
route53recoveryreadiness_delete_resource_set(ResourceSetName)
}
\arguments{
\item{ResourceSetName}{[required] Name of a resource set.}
}
\description{
Deletes a resource set.

See \url{https://paws-r.github.io/docs/route53recoveryreadiness/delete_resource_set.html} for full documentation.
}
\keyword{internal}
