% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53recoveryreadiness_operations.R
\name{route53recoveryreadiness_tag_resource}
\alias{route53recoveryreadiness_tag_resource}
\title{Adds a tag to a resource}
\usage{
route53recoveryreadiness_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) for a resource.}

\item{Tags}{[required]}
}
\description{
Adds a tag to a resource.

See \url{https://paws-r.github.io/docs/route53recoveryreadiness/tag_resource.html} for full documentation.
}
\keyword{internal}
