% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53recoveryreadiness_operations.R
\name{route53recoveryreadiness_update_recovery_group}
\alias{route53recoveryreadiness_update_recovery_group}
\title{Updates a recovery group}
\usage{
route53recoveryreadiness_update_recovery_group(Cells, RecoveryGroupName)
}
\arguments{
\item{Cells}{[required] A list of cell Amazon Resource Names (ARNs). This list completely
replaces the previous list.}

\item{RecoveryGroupName}{[required] The name of a recovery group.}
}
\description{
Updates a recovery group.

See \url{https://paws-r.github.io/docs/route53recoveryreadiness/update_recovery_group.html} for full documentation.
}
\keyword{internal}
