% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53resolver_operations.R
\name{route53resolver_create_firewall_domain_list}
\alias{route53resolver_create_firewall_domain_list}
\title{Creates an empty firewall domain list for use in DNS Firewall rules}
\usage{
route53resolver_create_firewall_domain_list(
  CreatorRequestId,
  Name,
  Tags = NULL
)
}
\arguments{
\item{CreatorRequestId}{[required] A unique string that identifies the request and that allows you to retry
failed requests without the risk of running the operation twice.
\code{CreatorRequestId} can be any unique string, for example, a date/time
stamp.}

\item{Name}{[required] A name that lets you identify the domain list to manage and use it.}

\item{Tags}{A list of the tag keys and values that you want to associate with the
domain list.}
}
\description{
Creates an empty firewall domain list for use in DNS Firewall rules. You can populate the domains for the new list with a file, using \code{\link[=route53resolver_import_firewall_domains]{import_firewall_domains}}, or with domain strings, using \code{\link[=route53resolver_update_firewall_domains]{update_firewall_domains}}.

See \url{https://paws-r.github.io/docs/route53resolver/create_firewall_domain_list.html} for full documentation.
}
\keyword{internal}
