% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53resolver_operations.R
\name{route53resolver_create_firewall_rule_group}
\alias{route53resolver_create_firewall_rule_group}
\title{Creates an empty DNS Firewall rule group for filtering DNS network
traffic in a VPC}
\usage{
route53resolver_create_firewall_rule_group(CreatorRequestId, Name, Tags)
}
\arguments{
\item{CreatorRequestId}{[required] A unique string defined by you to identify the request. This allows you
to retry failed requests without the risk of executing the operation
twice. This can be any unique string, for example, a timestamp.}

\item{Name}{[required] A name that lets you identify the rule group, to manage and use it.}

\item{Tags}{A list of the tag keys and values that you want to associate with the
rule group.}
}
\value{
A list with the following syntax:\preformatted{list(
  FirewallRuleGroup = list(
    Id = "string",
    Arn = "string",
    Name = "string",
    RuleCount = 123,
    Status = "COMPLETE"|"DELETING"|"UPDATING",
    StatusMessage = "string",
    OwnerId = "string",
    CreatorRequestId = "string",
    ShareStatus = "NOT_SHARED"|"SHARED_WITH_ME"|"SHARED_BY_ME",
    CreationTime = "string",
    ModificationTime = "string"
  )
)
}
}
\description{
Creates an empty DNS Firewall rule group for filtering DNS network
traffic in a VPC. You can add rules to the new rule group by calling
\code{\link[=route53resolver_create_firewall_rule]{create_firewall_rule}}.
}
\section{Request syntax}{
\preformatted{svc$create_firewall_rule_group(
  CreatorRequestId = "string",
  Name = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
