% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53resolver_operations.R
\name{route53resolver_create_resolver_endpoint}
\alias{route53resolver_create_resolver_endpoint}
\title{Creates a Resolver endpoint}
\usage{
route53resolver_create_resolver_endpoint(
  CreatorRequestId,
  Name = NULL,
  SecurityGroupIds,
  Direction,
  IpAddresses,
  OutpostArn = NULL,
  PreferredInstanceType = NULL,
  Tags = NULL,
  ResolverEndpointType = NULL,
  Protocols = NULL,
  RniEnhancedMetricsEnabled = NULL,
  TargetNameServerMetricsEnabled = NULL
)
}
\arguments{
\item{CreatorRequestId}{[required] A unique string that identifies the request and that allows failed
requests to be retried without the risk of running the operation twice.
\code{CreatorRequestId} can be any unique string, for example, a date/time
stamp.}

\item{Name}{A friendly name that lets you easily find a configuration in the
Resolver dashboard in the Route 53 console.}

\item{SecurityGroupIds}{[required] The ID of one or more security groups that you want to use to control
access to this VPC. The security group that you specify must include one
or more inbound rules (for inbound Resolver endpoints) or outbound rules
(for outbound Resolver endpoints). Inbound and outbound rules must allow
TCP and UDP access. For inbound access, open port 53. For outbound
access, open the port that you're using for DNS queries on your network.

Some security group rules will cause your connection to be tracked. For
outbound resolver endpoint, it can potentially impact the maximum
queries per second from outbound endpoint to your target name server.
For inbound resolver endpoint, it can bring down the overall maximum
queries per second per IP address to as low as 1500. To avoid connection
tracking caused by security group, see \href{https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/security-group-connection-tracking.html#untracked-connectionsl}{Untracked connections}.}

\item{Direction}{[required] Specify the applicable value:
\itemize{
\item \code{INBOUND}: Resolver forwards DNS queries to the DNS service for a
VPC from your network.
\item \code{OUTBOUND}: Resolver forwards DNS queries from the DNS service for a
VPC to your network.
\item \code{INBOUND_DELEGATION}: Resolver delegates queries to Route 53 private
hosted zones from your network.
}}

\item{IpAddresses}{[required] The subnets and IP addresses in your VPC that DNS queries originate from
(for outbound endpoints) or that you forward DNS queries to (for inbound
endpoints). The subnet ID uniquely identifies a VPC.

Even though the minimum is 1, Route 53 requires that you create at least
two.}

\item{OutpostArn}{The Amazon Resource Name (ARN) of the Outpost. If you specify this, you
must also specify a value for the \code{PreferredInstanceType}.}

\item{PreferredInstanceType}{The instance type. If you specify this, you must also specify a value
for the \code{OutpostArn}.}

\item{Tags}{A list of the tag keys and values that you want to associate with the
endpoint.}

\item{ResolverEndpointType}{For the endpoint type you can choose either IPv4, IPv6, or dual-stack. A
dual-stack endpoint means that it will resolve via both IPv4 and IPv6.
This endpoint type is applied to all IP addresses.}

\item{Protocols}{The protocols you want to use for the endpoint. DoH-FIPS is applicable
for default inbound endpoints only.

For a default inbound endpoint you can apply the protocols as follows:
\itemize{
\item Do53 and DoH in combination.
\item Do53 and DoH-FIPS in combination.
\item Do53 alone.
\item DoH alone.
\item DoH-FIPS alone.
\item None, which is treated as Do53.
}

For a delegation inbound endpoint you can use Do53 only.

For an outbound endpoint you can apply the protocols as follows:
\itemize{
\item Do53 and DoH in combination.
\item Do53 alone.
\item DoH alone.
\item None, which is treated as Do53.
}}

\item{RniEnhancedMetricsEnabled}{Specifies whether RNI enhanced metrics are enabled for the Resolver
endpoints. When set to true, one-minute granular metrics are published
in CloudWatch for each RNI associated with this endpoint. When set to
false, metrics are not published. Default is false.

Standard CloudWatch pricing and charges are applied for using the Route
53 Resolver endpoint RNI enhanced metrics. For more information, see
\href{https://docs.aws.amazon.com/Route53/latest/DeveloperGuide/monitoring-resolver-with-cloudwatch.html}{Detailed metrics}.}

\item{TargetNameServerMetricsEnabled}{Specifies whether target name server metrics are enabled for the
outbound Resolver endpoints. When set to true, one-minute granular
metrics are published in CloudWatch for each target name server
associated with this endpoint. When set to false, metrics are not
published. Default is false. This is not supported for inbound Resolver
endpoints.

Standard CloudWatch pricing and charges are applied for using the Route
53 Resolver endpoint target name server metrics. For more information,
see \href{https://docs.aws.amazon.com/Route53/latest/DeveloperGuide/monitoring-resolver-with-cloudwatch.html}{Detailed metrics}.}
}
\description{
Creates a Resolver endpoint. There are two types of Resolver endpoints, inbound and outbound:

See \url{https://www.paws-r-sdk.com/docs/route53resolver_create_resolver_endpoint/} for full documentation.
}
\keyword{internal}
