% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53resolver_operations.R
\name{route53resolver_delete_firewall_domain_list}
\alias{route53resolver_delete_firewall_domain_list}
\title{Deletes the specified domain list}
\usage{
route53resolver_delete_firewall_domain_list(FirewallDomainListId)
}
\arguments{
\item{FirewallDomainListId}{[required] The ID of the domain list that you want to delete.}
}
\value{
A list with the following syntax:\preformatted{list(
  FirewallDomainList = list(
    Id = "string",
    Arn = "string",
    Name = "string",
    DomainCount = 123,
    Status = "COMPLETE"|"COMPLETE_IMPORT_FAILED"|"IMPORTING"|"DELETING"|"UPDATING",
    StatusMessage = "string",
    ManagedOwnerName = "string",
    CreatorRequestId = "string",
    CreationTime = "string",
    ModificationTime = "string"
  )
)
}
}
\description{
Deletes the specified domain list.
}
\section{Request syntax}{
\preformatted{svc$delete_firewall_domain_list(
  FirewallDomainListId = "string"
)
}
}

\keyword{internal}
