% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53resolver_operations.R
\name{route53resolver_delete_firewall_domain_list}
\alias{route53resolver_delete_firewall_domain_list}
\title{Deletes the specified domain list}
\usage{
route53resolver_delete_firewall_domain_list(FirewallDomainListId)
}
\arguments{
\item{FirewallDomainListId}{[required] The ID of the domain list that you want to delete.}
}
\description{
Deletes the specified domain list.

See \url{https://www.paws-r-sdk.com/docs/route53resolver_delete_firewall_domain_list/} for full documentation.
}
\keyword{internal}
