% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53resolver_operations.R
\name{route53resolver_delete_firewall_rule}
\alias{route53resolver_delete_firewall_rule}
\title{Deletes the specified firewall rule}
\usage{
route53resolver_delete_firewall_rule(FirewallRuleGroupId,
  FirewallDomainListId)
}
\arguments{
\item{FirewallRuleGroupId}{[required] The unique identifier of the firewall rule group that you want to delete
the rule from.}

\item{FirewallDomainListId}{[required] The ID of the domain list that's used in the rule.}
}
\value{
A list with the following syntax:\preformatted{list(
  FirewallRule = list(
    FirewallRuleGroupId = "string",
    FirewallDomainListId = "string",
    Name = "string",
    Priority = 123,
    Action = "ALLOW"|"BLOCK"|"ALERT",
    BlockResponse = "NODATA"|"NXDOMAIN"|"OVERRIDE",
    BlockOverrideDomain = "string",
    BlockOverrideDnsType = "CNAME",
    BlockOverrideTtl = 123,
    CreatorRequestId = "string",
    CreationTime = "string",
    ModificationTime = "string"
  )
)
}
}
\description{
Deletes the specified firewall rule.
}
\section{Request syntax}{
\preformatted{svc$delete_firewall_rule(
  FirewallRuleGroupId = "string",
  FirewallDomainListId = "string"
)
}
}

\keyword{internal}
