% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53resolver_operations.R
\name{route53resolver_delete_firewall_rule}
\alias{route53resolver_delete_firewall_rule}
\title{Deletes the specified firewall rule}
\usage{
route53resolver_delete_firewall_rule(FirewallRuleGroupId, FirewallDomainListId)
}
\arguments{
\item{FirewallRuleGroupId}{[required] The unique identifier of the firewall rule group that you want to delete
the rule from.}

\item{FirewallDomainListId}{[required] The ID of the domain list that's used in the rule.}
}
\description{
Deletes the specified firewall rule.

See \url{https://www.paws-r-sdk.com/docs/route53resolver_delete_firewall_rule/} for full documentation.
}
\keyword{internal}
