% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53resolver_operations.R
\name{route53resolver_disassociate_firewall_rule_group}
\alias{route53resolver_disassociate_firewall_rule_group}
\title{Disassociates a FirewallRuleGroup from a VPC, to remove DNS filtering
from the VPC}
\usage{
route53resolver_disassociate_firewall_rule_group(
  FirewallRuleGroupAssociationId)
}
\arguments{
\item{FirewallRuleGroupAssociationId}{[required] The identifier of the FirewallRuleGroupAssociation.}
}
\value{
A list with the following syntax:\preformatted{list(
  FirewallRuleGroupAssociation = list(
    Id = "string",
    Arn = "string",
    FirewallRuleGroupId = "string",
    VpcId = "string",
    Name = "string",
    Priority = 123,
    MutationProtection = "ENABLED"|"DISABLED",
    ManagedOwnerName = "string",
    Status = "COMPLETE"|"DELETING"|"UPDATING",
    StatusMessage = "string",
    CreatorRequestId = "string",
    CreationTime = "string",
    ModificationTime = "string"
  )
)
}
}
\description{
Disassociates a FirewallRuleGroup from a VPC, to remove DNS filtering
from the VPC.
}
\section{Request syntax}{
\preformatted{svc$disassociate_firewall_rule_group(
  FirewallRuleGroupAssociationId = "string"
)
}
}

\keyword{internal}
