% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53resolver_operations.R
\name{route53resolver_get_firewall_config}
\alias{route53resolver_get_firewall_config}
\title{Retrieves the configuration of the firewall behavior provided by DNS
Firewall for a single VPC from Amazon Virtual Private Cloud (Amazon VPC)}
\usage{
route53resolver_get_firewall_config(ResourceId)
}
\arguments{
\item{ResourceId}{[required] The ID of the VPC from Amazon VPC that the configuration is for.}
}
\description{
Retrieves the configuration of the firewall behavior provided by DNS Firewall for a single VPC from Amazon Virtual Private Cloud (Amazon VPC).

See \url{https://www.paws-r-sdk.com/docs/route53resolver_get_firewall_config/} for full documentation.
}
\keyword{internal}
