% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53resolver_operations.R
\name{route53resolver_get_firewall_domain_list}
\alias{route53resolver_get_firewall_domain_list}
\title{Retrieves the specified firewall domain list}
\usage{
route53resolver_get_firewall_domain_list(FirewallDomainListId)
}
\arguments{
\item{FirewallDomainListId}{[required] The ID of the domain list.}
}
\description{
Retrieves the specified firewall domain list.

See \url{https://paws-r.github.io/docs/route53resolver/get_firewall_domain_list.html} for full documentation.
}
\keyword{internal}
