% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53resolver_operations.R
\name{route53resolver_get_firewall_rule_group}
\alias{route53resolver_get_firewall_rule_group}
\title{Retrieves the specified firewall rule group}
\usage{
route53resolver_get_firewall_rule_group(FirewallRuleGroupId)
}
\arguments{
\item{FirewallRuleGroupId}{[required] The unique identifier of the firewall rule group.}
}
\description{
Retrieves the specified firewall rule group.

See \url{https://paws-r.github.io/docs/route53resolver/get_firewall_rule_group.html} for full documentation.
}
\keyword{internal}
