% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53resolver_operations.R
\name{route53resolver_get_firewall_rule_group_association}
\alias{route53resolver_get_firewall_rule_group_association}
\title{Retrieves a firewall rule group association, which enables DNS filtering
for a VPC with one rule group}
\usage{
route53resolver_get_firewall_rule_group_association(
  FirewallRuleGroupAssociationId
)
}
\arguments{
\item{FirewallRuleGroupAssociationId}{[required] The identifier of the FirewallRuleGroupAssociation.}
}
\description{
Retrieves a firewall rule group association, which enables DNS filtering for a VPC with one rule group. A VPC can have more than one firewall rule group association, and a rule group can be associated with more than one VPC.

See \url{https://www.paws-r-sdk.com/docs/route53resolver_get_firewall_rule_group_association/} for full documentation.
}
\keyword{internal}
