% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53resolver_operations.R
\name{route53resolver_get_firewall_rule_group_policy}
\alias{route53resolver_get_firewall_rule_group_policy}
\title{Returns the Identity and Access Management (Amazon Web Services IAM)
policy for sharing the specified rule group}
\usage{
route53resolver_get_firewall_rule_group_policy(Arn)
}
\arguments{
\item{Arn}{[required] The ARN (Amazon Resource Name) for the rule group.}
}
\description{
Returns the Identity and Access Management (Amazon Web Services IAM) policy for sharing the specified rule group. You can use the policy to share the rule group using Resource Access Manager (RAM).

See \url{https://www.paws-r-sdk.com/docs/route53resolver_get_firewall_rule_group_policy/} for full documentation.
}
\keyword{internal}
