% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53resolver_operations.R
\name{route53resolver_get_resolver_dnssec_config}
\alias{route53resolver_get_resolver_dnssec_config}
\title{Gets DNSSEC validation information for a specified resource}
\usage{
route53resolver_get_resolver_dnssec_config(ResourceId)
}
\arguments{
\item{ResourceId}{[required] The ID of the virtual private cloud (VPC) for the DNSSEC validation
status.}
}
\value{
A list with the following syntax:\preformatted{list(
  ResolverDNSSECConfig = list(
    Id = "string",
    OwnerId = "string",
    ResourceId = "string",
    ValidationStatus = "ENABLING"|"ENABLED"|"DISABLING"|"DISABLED"
  )
)
}
}
\description{
Gets DNSSEC validation information for a specified resource.
}
\section{Request syntax}{
\preformatted{svc$get_resolver_dnssec_config(
  ResourceId = "string"
)
}
}

\keyword{internal}
