% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53resolver_operations.R
\name{route53resolver_get_resolver_query_log_config_association}
\alias{route53resolver_get_resolver_query_log_config_association}
\title{Gets information about a specified association between a Resolver query
logging configuration and an Amazon VPC}
\usage{
route53resolver_get_resolver_query_log_config_association(
  ResolverQueryLogConfigAssociationId
)
}
\arguments{
\item{ResolverQueryLogConfigAssociationId}{[required] The ID of the Resolver query logging configuration association that you
want to get information about.}
}
\description{
Gets information about a specified association between a Resolver query logging configuration and an Amazon VPC. When you associate a VPC with a query logging configuration, Resolver logs DNS queries that originate in that VPC.

See \url{https://paws-r.github.io/docs/route53resolver/get_resolver_query_log_config_association.html} for full documentation.
}
\keyword{internal}
