% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53resolver_operations.R
\name{route53resolver_list_firewall_rule_types}
\alias{route53resolver_list_firewall_rule_types}
\title{Retrieves the available rule types that can be used in DNS Firewall
rules}
\usage{
route53resolver_list_firewall_rule_types(
  RuleType = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{RuleType}{The rule type to filter by. If specified, only rule types matching this value are returned.}

\item{MaxResults}{The maximum number of objects that you want Resolver to return for this request. If more objects are available, in the response, Resolver provides a \code{NextToken} value that you can use in a subsequent call to get the next batch of objects.}

\item{NextToken}{For the first call to this list request, omit this value. When you request a list of objects, Resolver returns at most the number of objects specified in \code{MaxResults}. If more objects are available for retrieval, Resolver provides a \code{NextToken} value in the response. To retrieve the next batch of objects, use the token that was returned for the prior request in your next request.}
}
\description{
Retrieves the available rule types that can be used in DNS Firewall rules.

See \url{https://www.paws-r-sdk.com/docs/route53resolver_list_firewall_rule_types/} for full documentation.
}
\keyword{internal}
