% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53resolver_operations.R
\name{route53resolver_list_resolver_rule_associations}
\alias{route53resolver_list_resolver_rule_associations}
\title{Lists the associations that were created between Resolver rules and VPCs
using the current AWS account}
\usage{
route53resolver_list_resolver_rule_associations(MaxResults, NextToken,
  Filters)
}
\arguments{
\item{MaxResults}{The maximum number of rule associations that you want to return in the
response to a
\code{\link[=route53resolver_list_resolver_rule_associations]{list_resolver_rule_associations}}
request. If you don't specify a value for \code{MaxResults}, Resolver returns
up to 100 rule associations.}

\item{NextToken}{For the first \code{ListResolverRuleAssociation} request, omit this value.

If you have more than \code{MaxResults} rule associations, you can submit
another \code{ListResolverRuleAssociation} request to get the next group of
rule associations. In the next request, specify the value of \code{NextToken}
from the previous response.}

\item{Filters}{An optional specification to return a subset of Resolver rules, such as
Resolver rules that are associated with the same VPC ID.

If you submit a second or subsequent
\code{\link[=route53resolver_list_resolver_rule_associations]{list_resolver_rule_associations}}
request and specify the \code{NextToken} parameter, you must use the same
values for \code{Filters}, if any, as in the previous request.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  NextToken = "string",
  MaxResults = 123,
  ResolverRuleAssociations = list(
    list(
      Id = "string",
      ResolverRuleId = "string",
      Name = "string",
      VPCId = "string",
      Status = "CREATING"|"COMPLETE"|"DELETING"|"FAILED"|"OVERRIDDEN",
      StatusMessage = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Lists the associations that were created between Resolver rules and VPCs
using the current AWS account.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_resolver_rule_associations(
  MaxResults = 123,
  NextToken = "string",
  Filters = list(
    list(
      Name = "string",
      Values = list(
        "string"
      )
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
