% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53resolver_operations.R
\name{route53resolver_put_firewall_rule_group_policy}
\alias{route53resolver_put_firewall_rule_group_policy}
\title{Attaches an Identity and Access Management (Amazon Web Services IAM)
policy for sharing the rule group}
\usage{
route53resolver_put_firewall_rule_group_policy(Arn, FirewallRuleGroupPolicy)
}
\arguments{
\item{Arn}{[required] The ARN (Amazon Resource Name) for the rule group that you want to
share.}

\item{FirewallRuleGroupPolicy}{[required] The Identity and Access Management (Amazon Web Services IAM) policy to
attach to the rule group.}
}
\description{
Attaches an Identity and Access Management (Amazon Web Services IAM) policy for sharing the rule group. You can use the policy to share the rule group using Resource Access Manager (RAM).

See \url{https://paws-r.github.io/docs/route53resolver/put_firewall_rule_group_policy.html} for full documentation.
}
\keyword{internal}
