% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicediscovery_operations.R
\name{servicediscovery_delete_namespace}
\alias{servicediscovery_delete_namespace}
\title{Deletes a namespace from the current account}
\usage{
servicediscovery_delete_namespace(Id)
}
\arguments{
\item{Id}{[required] The ID of the namespace that you want to delete.}
}
\description{
Deletes a namespace from the current account. If the namespace still contains one or more services, the request fails.

See \url{https://paws-r.github.io/docs/servicediscovery/delete_namespace.html} for full documentation.
}
\keyword{internal}
