% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicediscovery_operations.R
\name{servicediscovery_delete_service}
\alias{servicediscovery_delete_service}
\title{Deletes a specified service}
\usage{
servicediscovery_delete_service(Id)
}
\arguments{
\item{Id}{[required] The ID of the service that you want to delete.}
}
\description{
Deletes a specified service. If the service still contains one or more registered instances, the request fails.

See \url{https://paws-r.github.io/docs/servicediscovery/delete_service.html} for full documentation.
}
\keyword{internal}
