% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicediscovery_operations.R
\name{servicediscovery_delete_service}
\alias{servicediscovery_delete_service}
\title{Deletes a specified service and all associated service attributes}
\usage{
servicediscovery_delete_service(Id)
}
\arguments{
\item{Id}{[required] The ID or Amazon Resource Name (ARN) of the service that you want to
delete. If the namespace associated with the service is shared with your
Amazon Web Services account, specify the service ARN. For more
information about shared namespaces, see \href{https://docs.aws.amazon.com/cloud-map/latest/dg/sharing-namespaces.html}{Cross-account Cloud Map namespace sharing}.}
}
\description{
Deletes a specified service and all associated service attributes. If the service still contains one or more registered instances, the request fails.

See \url{https://www.paws-r-sdk.com/docs/servicediscovery_delete_service/} for full documentation.
}
\keyword{internal}
