% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicediscovery_operations.R
\name{servicediscovery_get_instance}
\alias{servicediscovery_get_instance}
\title{Gets information about a specified instance}
\usage{
servicediscovery_get_instance(ServiceId, InstanceId)
}
\arguments{
\item{ServiceId}{[required] The ID or Amazon Resource Name (ARN) of the service that the instance is
associated with. For services created in a shared namespace, specify the
service ARN. For more information about shared namespaces, see
\href{https://docs.aws.amazon.com/cloud-map/latest/dg/sharing-namespaces.html}{Cross-account Cloud Map namespace sharing}
in the \emph{Cloud Map Developer Guide}.}

\item{InstanceId}{[required] The ID of the instance that you want to get information about.}
}
\description{
Gets information about a specified instance.

See \url{https://www.paws-r-sdk.com/docs/servicediscovery_get_instance/} for full documentation.
}
\keyword{internal}
