% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicediscovery_operations.R
\name{servicediscovery_get_operation}
\alias{servicediscovery_get_operation}
\title{Gets information about any operation that returns an operation ID in the
response, such as a CreateService request}
\usage{
servicediscovery_get_operation(OperationId)
}
\arguments{
\item{OperationId}{[required] The ID of the operation that you want to get more information about.}
}
\description{
Gets information about any operation that returns an operation ID in the response, such as a \code{\link[=servicediscovery_create_service]{create_service}} request.

See \url{https://www.paws-r-sdk.com/docs/servicediscovery_get_operation/} for full documentation.
}
\keyword{internal}
