% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicediscovery_operations.R
\name{servicediscovery_get_service}
\alias{servicediscovery_get_service}
\title{Gets the settings for a specified service}
\usage{
servicediscovery_get_service(Id)
}
\arguments{
\item{Id}{[required] The ID or Amazon Resource Name (ARN) of the service that you want to get
settings for. For services created by consumers in a shared namespace,
specify the service ARN. For more information about shared namespaces,
see \href{https://docs.aws.amazon.com/cloud-map/latest/dg/sharing-namespaces.html}{Cross-account Cloud Map namespace sharing}
in the \emph{Cloud Map Developer Guide}.}
}
\description{
Gets the settings for a specified service.

See \url{https://www.paws-r-sdk.com/docs/servicediscovery_get_service/} for full documentation.
}
\keyword{internal}
