% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicediscovery_operations.R
\name{servicediscovery_get_service_attributes}
\alias{servicediscovery_get_service_attributes}
\title{Returns the attributes associated with a specified service}
\usage{
servicediscovery_get_service_attributes(ServiceId)
}
\arguments{
\item{ServiceId}{[required] The ID or Amazon Resource Name (ARN) of the service that you want to get
attributes for. For services created in a namespace shared with your
Amazon Web Services account, specify the service ARN. For more
information about shared namespaces, see \href{https://docs.aws.amazon.com/cloud-map/latest/dg/sharing-namespaces.html}{Cross-account Cloud Map namespace sharing}
in the \emph{Cloud Map Developer Guide}.}
}
\description{
Returns the attributes associated with a specified service.

See \url{https://www.paws-r-sdk.com/docs/servicediscovery_get_service_attributes/} for full documentation.
}
\keyword{internal}
