% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicediscovery_operations.R
\name{servicediscovery_list_tags_for_resource}
\alias{servicediscovery_list_tags_for_resource}
\title{Lists tags for the specified resource}
\usage{
servicediscovery_list_tags_for_resource(ResourceARN)
}
\arguments{
\item{ResourceARN}{[required] The Amazon Resource Name (ARN) of the resource that you want to retrieve
tags for.}
}
\description{
Lists tags for the specified resource.

See \url{https://paws-r.github.io/docs/servicediscovery/list_tags_for_resource.html} for full documentation.
}
\keyword{internal}
