% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicediscovery_operations.R
\name{servicediscovery_update_instance_custom_health_status}
\alias{servicediscovery_update_instance_custom_health_status}
\title{Submits a request to change the health status of a custom health check
to healthy or unhealthy}
\usage{
servicediscovery_update_instance_custom_health_status(
  ServiceId,
  InstanceId,
  Status
)
}
\arguments{
\item{ServiceId}{[required] The ID or Amazon Resource Name (ARN) of the service that includes the
configuration for the custom health check that you want to change the
status for. For services created in a shared namespace, specify the
service ARN. For more information about shared namespaces, see
\href{https://docs.aws.amazon.com/cloud-map/latest/dg/sharing-namespaces.html}{Cross-account Cloud Map namespace sharing}
in the \emph{Cloud Map Developer Guide}.}

\item{InstanceId}{[required] The ID of the instance that you want to change the health status for.}

\item{Status}{[required] The new status of the instance, \code{HEALTHY} or \code{UNHEALTHY}.}
}
\description{
Submits a request to change the health status of a custom health check to healthy or unhealthy.

See \url{https://www.paws-r-sdk.com/docs/servicediscovery_update_instance_custom_health_status/} for full documentation.
}
\keyword{internal}
