% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telconetworkbuilder_operations.R
\name{telconetworkbuilder_update_sol_network_instance}
\alias{telconetworkbuilder_update_sol_network_instance}
\title{Update a network instance}
\usage{
telconetworkbuilder_update_sol_network_instance(
  modifyVnfInfoData = NULL,
  nsInstanceId,
  tags = NULL,
  updateType
)
}
\arguments{
\item{modifyVnfInfoData}{Identifies the network function information parameters and/or the
configurable properties of the network function to be modified.}

\item{nsInstanceId}{[required] ID of the network instance.}

\item{tags}{A tag is a label that you assign to an Amazon Web Services resource.
Each tag consists of a key and an optional value. When you use this API,
the tags are transferred to the network operation that is created. Use
tags to search and filter your resources or track your Amazon Web
Services costs.}

\item{updateType}{[required] The type of update.}
}
\description{
Update a network instance.

See \url{https://www.paws-r-sdk.com/docs/telconetworkbuilder_update_sol_network_instance/} for full documentation.
}
\keyword{internal}
