% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpclattice_operations.R
\name{vpclattice_batch_update_rule}
\alias{vpclattice_batch_update_rule}
\title{Updates the listener rules in a batch}
\usage{
vpclattice_batch_update_rule(serviceIdentifier, listenerIdentifier, rules)
}
\arguments{
\item{serviceIdentifier}{[required] The ID or ARN of the service.}

\item{listenerIdentifier}{[required] The ID or ARN of the listener.}

\item{rules}{[required] The rules for the specified listener.}
}
\description{
Updates the listener rules in a batch. You can use this operation to change the priority of listener rules. This can be useful when bulk updating or swapping rule priority.

See \url{https://www.paws-r-sdk.com/docs/vpclattice_batch_update_rule/} for full documentation.
}
\keyword{internal}
