% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpclattice_operations.R
\name{vpclattice_create_service}
\alias{vpclattice_create_service}
\title{Creates a service}
\usage{
vpclattice_create_service(
  clientToken = NULL,
  name,
  tags = NULL,
  customDomainName = NULL,
  certificateArn = NULL,
  authType = NULL
)
}
\arguments{
\item{clientToken}{A unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. If you retry a request that completed
successfully using the same client token and parameters, the retry
succeeds without performing any actions. If the parameters aren't
identical, the retry fails.}

\item{name}{[required] The name of the service. The name must be unique within the account. The
valid characters are a-z, 0-9, and hyphens (-). You can't use a hyphen
as the first or last character, or immediately after another hyphen.}

\item{tags}{The tags for the service.}

\item{customDomainName}{The custom domain name of the service.}

\item{certificateArn}{The Amazon Resource Name (ARN) of the certificate.}

\item{authType}{The type of IAM policy.
\itemize{
\item \code{NONE}: The resource does not use an IAM policy. This is the
default.
\item \code{AWS_IAM}: The resource uses an IAM policy. When this type is used,
auth is enabled and an auth policy is required.
}}
}
\description{
Creates a service. A service is any software application that can run on instances containers, or serverless functions within an account or virtual private cloud (VPC).

See \url{https://www.paws-r-sdk.com/docs/vpclattice_create_service/} for full documentation.
}
\keyword{internal}
