% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpclattice_operations.R
\name{vpclattice_get_listener}
\alias{vpclattice_get_listener}
\title{Retrieves information about the specified listener for the specified
service}
\usage{
vpclattice_get_listener(serviceIdentifier, listenerIdentifier)
}
\arguments{
\item{serviceIdentifier}{[required] The ID or ARN of the service.}

\item{listenerIdentifier}{[required] The ID or ARN of the listener.}
}
\description{
Retrieves information about the specified listener for the specified service.

See \url{https://www.paws-r-sdk.com/docs/vpclattice_get_listener/} for full documentation.
}
\keyword{internal}
