% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpclattice_operations.R
\name{vpclattice_list_listeners}
\alias{vpclattice_list_listeners}
\title{Lists the listeners for the specified service}
\usage{
vpclattice_list_listeners(
  serviceIdentifier,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{serviceIdentifier}{[required] The ID or ARN of the service.}

\item{maxResults}{The maximum number of results to return.}

\item{nextToken}{A pagination token for the next page of results.}
}
\description{
Lists the listeners for the specified service.

See \url{https://www.paws-r-sdk.com/docs/vpclattice_list_listeners/} for full documentation.
}
\keyword{internal}
