% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpclattice_operations.R
\name{vpclattice_list_rules}
\alias{vpclattice_list_rules}
\title{Lists the rules for the specified listener}
\usage{
vpclattice_list_rules(
  serviceIdentifier,
  listenerIdentifier,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{serviceIdentifier}{[required] The ID or ARN of the service.}

\item{listenerIdentifier}{[required] The ID or ARN of the listener.}

\item{maxResults}{The maximum number of results to return.}

\item{nextToken}{A pagination token for the next page of results.}
}
\description{
Lists the rules for the specified listener.

See \url{https://www.paws-r-sdk.com/docs/vpclattice_list_rules/} for full documentation.
}
\keyword{internal}
