% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpclattice_operations.R
\name{vpclattice_list_service_network_service_associations}
\alias{vpclattice_list_service_network_service_associations}
\title{Lists the associations between a service network and a service}
\usage{
vpclattice_list_service_network_service_associations(
  serviceNetworkIdentifier = NULL,
  serviceIdentifier = NULL,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{serviceNetworkIdentifier}{The ID or ARN of the service network.}

\item{serviceIdentifier}{The ID or ARN of the service.}

\item{maxResults}{The maximum number of results to return.}

\item{nextToken}{A pagination token for the next page of results.}
}
\description{
Lists the associations between a service network and a service. You can filter the list either by service or service network. You must provide either the service network identifier or the service identifier.

See \url{https://www.paws-r-sdk.com/docs/vpclattice_list_service_network_service_associations/} for full documentation.
}
\keyword{internal}
