% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpclattice_operations.R
\name{vpclattice_list_service_networks}
\alias{vpclattice_list_service_networks}
\title{Lists the service networks owned by or shared with this account}
\usage{
vpclattice_list_service_networks(maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{maxResults}{The maximum number of results to return.}

\item{nextToken}{A pagination token for the next page of results.}
}
\description{
Lists the service networks owned by or shared with this account. The account ID in the ARN shows which account owns the service network.

See \url{https://www.paws-r-sdk.com/docs/vpclattice_list_service_networks/} for full documentation.
}
\keyword{internal}
