% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpclattice_operations.R
\name{vpclattice_list_targets}
\alias{vpclattice_list_targets}
\title{Lists the targets for the target group}
\usage{
vpclattice_list_targets(
  targetGroupIdentifier,
  maxResults = NULL,
  nextToken = NULL,
  targets = NULL
)
}
\arguments{
\item{targetGroupIdentifier}{[required] The ID or ARN of the target group.}

\item{maxResults}{The maximum number of results to return.}

\item{nextToken}{A pagination token for the next page of results.}

\item{targets}{The targets.}
}
\description{
Lists the targets for the target group. By default, all targets are included. You can use this API to check the health status of targets. You can also filter the results by target.

See \url{https://www.paws-r-sdk.com/docs/vpclattice_list_targets/} for full documentation.
}
\keyword{internal}
