% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpclattice_operations.R
\name{vpclattice_put_resource_policy}
\alias{vpclattice_put_resource_policy}
\title{Attaches a resource-based permission policy to a service or service
network}
\usage{
vpclattice_put_resource_policy(resourceArn, policy)
}
\arguments{
\item{resourceArn}{[required] The ID or ARN of the service network or service for which the policy is
created.}

\item{policy}{[required] An IAM policy. The policy string in JSON must not contain newlines or
blank lines.}
}
\description{
Attaches a resource-based permission policy to a service or service network. The policy must contain the same actions and condition statements as the Amazon Web Services Resource Access Manager permission for sharing services and service networks.

See \url{https://www.paws-r-sdk.com/docs/vpclattice_put_resource_policy/} for full documentation.
}
\keyword{internal}
