% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpclattice_operations.R
\name{vpclattice_update_service}
\alias{vpclattice_update_service}
\title{Updates the specified service}
\usage{
vpclattice_update_service(
  serviceIdentifier,
  certificateArn = NULL,
  authType = NULL
)
}
\arguments{
\item{serviceIdentifier}{[required] The ID or ARN of the service.}

\item{certificateArn}{The Amazon Resource Name (ARN) of the certificate.}

\item{authType}{The type of IAM policy.
\itemize{
\item \code{NONE}: The resource does not use an IAM policy. This is the
default.
\item \code{AWS_IAM}: The resource uses an IAM policy. When this type is used,
auth is enabled and an auth policy is required.
}}
}
\description{
Updates the specified service.

See \url{https://www.paws-r-sdk.com/docs/vpclattice_update_service/} for full documentation.
}
\keyword{internal}
