% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpclattice_operations.R
\name{vpclattice_update_service_network_vpc_association}
\alias{vpclattice_update_service_network_vpc_association}
\title{Updates the service network and VPC association}
\usage{
vpclattice_update_service_network_vpc_association(
  securityGroupIds,
  serviceNetworkVpcAssociationIdentifier
)
}
\arguments{
\item{securityGroupIds}{[required] The IDs of the security groups. Once you add a security group, it cannot
be removed.}

\item{serviceNetworkVpcAssociationIdentifier}{[required] The ID or Amazon Resource Name (ARN) of the association.}
}
\description{
Updates the service network and VPC association. Once you add a security group, it cannot be removed.

See \url{https://www.paws-r-sdk.com/docs/vpclattice_update_service_network_vpc_association/} for full documentation.
}
\keyword{internal}
