% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpclattice_operations.R
\name{vpclattice_update_target_group}
\alias{vpclattice_update_target_group}
\title{Updates the specified target group}
\usage{
vpclattice_update_target_group(targetGroupIdentifier, healthCheck)
}
\arguments{
\item{targetGroupIdentifier}{[required] The ID or ARN of the target group.}

\item{healthCheck}{[required] The health check configuration.}
}
\description{
Updates the specified target group.

See \url{https://www.paws-r-sdk.com/docs/vpclattice_update_target_group/} for full documentation.
}
\keyword{internal}
